Crate libc

Raw FFI bindings to platform system libraries.

Usage Guidelines

libc exposes non-Rust interfaces in Rust, which makes for some caveats to its use that are not present in most Rust libraries. Observing the following guidelines are recommended to help avoid soundness and stability pitfalls.

Never construct a libc struct with MaybeUninit::uninit(), initialize it, then call assume_init. Many structures have padding fields or may gain fields in the future, and it is far too easy to end up calling assume_init on partially initialized data.

Instead, use MaybeUninit::zeroed() or the Default implementations that are slowly being added. Alternatively, access fields only via raw pointer without ever using assume_init.

Avoid relying on the exact value of constants, the exact length of arrays, or the exact types of type aliases, as they may change across libc versions. That is, if libc contains code like:

const IFNAMSIZ: usize = 16;

pub struct ifreq {
    pub ifr_name: [c_char; IFNAMSIZ],
    // ...
}

extern "C" {
    pub fn time(time: *mut time_t) -> time_t;
}

Then avoid writing code like:

// Bad assumption that the length will always be 16.
fn takes_ifr_name(ifr_name: [c_char; 16]) { /* ... */ }

fn process_ifr(ifr: ifreq) {
    takes_ifr_name(ifr.ifr_name);
}

// Bad assumption that `time_t` will always be an `i64`. Use `-> time_t` instead, or
// explicitly cast to an `i64`.`
fn get_time() -> i64 {
    unsafe { time(ptr::null_mut()) }
}

For takes_ifr_name, use [c_char; IFNAMSIZ] or just &[c_char] instead. For get_time, return a time_t or explicitly cast to an i64.

Along the same lines, if you write code along the lines of assert_eq!(libc::ELAST, 97), expect that there may be a release where this starts to fail.

Do not name __c_anonymous_* types anywhere, which exist to represent anonymous fields in C. For example, FreeBSD defines:

struct filestat {
    int fs_type;
    // ...
    struct { struct filestat stqe_next; } next;
};

Which is represented in libc as:

struct filestat {
    fs_type: c_int,
    // ...
    next: __c_anonymous_filestat,
}

struct __c_anonymous_filestat { stqe_next: *mut filestat }

Accessing some_filestat.next.stqe_next is completely fine, but __c_anonymous_filestat should not be used anywhere (e.g. in a function signature). This is done to permit libc to switch to anonymous fields if the feature is ever added to Rust.

Avoid accessing fields with names such as __reserved, _pad, or _spare. Usually the platform libraries use these to allow adding new fields without changing the size of a struct, but this means their types change frequently.

Be aware of deprecation warnings. These are used as a way to migrate necessary API changes.

Cargo Features

std: by default libc assumes that the standard library contains link directives necessary to use the APIs in this crate. If std is disabled, libc will emit the directives instead.

This feature is slated for removal in libc 1.0. The intention is that no-std users of libc should use their own #[link] attributes, rustc-link-lib build script directives, or -l arguments for only the system libraries they need to link, rather than libc possibly linking more than is needed or available. If you are using libc without the std feature, consider starting to add link directives now for a smoother 1.0 transition.

extra_traits: all types in libc implement Clone, Copy, and Debug. The extra_traits feature adds Eq, Hash, and PartialEq.

This feature is expected to be removed in libc 1.0. Libraries should instead hash or check equality of only needed fields.

The features const-extern-fn, align, and use_std are all deprecated and do nothing.

Stability Expectations

Due to libc’s position in the ecosystem, it can effectively never publish semver-breaking releases. However, the API that libc binds changes all the time; sometimes in ways that are harmless, sometimes in ways that are technically API-breaking for all users but unlikely to be noticed (e.g. removing deprecated API), and sometimes in ways that are nonbreaking in C but translate to breaking changes in Rust (e.g. changing the type of an integer). libc tries to strike a balance but all of this means that unfortunately, libc must occasionally ship changes within a semver-compatible release that are technically semver-breaking.

The following are examples of changes that fall into this category:

Fields are added to a struct that is otherwise exhaustive.

Fields with names such as padding or reserved change type or are removed.

The length of an array type changes.

A struct field (with available padding) is changed from int to long.

In general, libc aims to follow platform API changes, even when this means changes that are user-visible in Rust. There are a few guidelines used here:

Adding struct fields is not considered breaking, nor is changing fields named reserved, padding, or similar. This is because users are expected to use field-by-field initialization.

Changing type aliases, values of constants, or array lengths is not considered breaking.

If the platform libc has accepted breakage on the C side (typically in the form of removing old API), the libc crate will follow suit.

Where possible, #[deprecated(...)] will be used to warn about changes before applying them. Alternative mitigations may be considered.

Potentially breaking changes will be well-identified in release notes.

Beyond this, public API is not expected to change on Tier 1 targets. Tier 2 targets have relaxed API stability requirements, and API stability is not enforced on tier 3 targets.

While this section seems scary, keep in mind that it is meant to cover worst-case scenarios. In practice, breakage is rare and following the above-discussed Usage Guidelines means that most libc users will never encounter a problem.

Structs

AdhocPoolStat

Atrac3BufferInfo

AudioInputParams

DebugProfilerRegs

GeBreakParam

GeCallbackData

GeContext

GeListArgs

GeStack

IntrHandlerOptionParam

Key

Mp3Handle

OpenPSID

RegHandle

SceCtrlData

SceCtrlLatch

SceIoDirent

SceIoStat

SceKernelAlarmInfo

SceKernelCallbackInfo

SceKernelEventFlagInfo

SceKernelEventFlagOptParam

SceKernelFplInfo

SceKernelFplOptParam

SceKernelLMOption

SceKernelLoadExecParam

SceKernelMbxInfo

SceKernelMbxOptParam

SceKernelModuleInfo

SceKernelMppInfo

SceKernelSMOption

SceKernelSemaInfo

SceKernelSemaOptParam

SceKernelSysClock

SceKernelSystemStatus

SceKernelThreadEventHandlerInfo

SceKernelThreadInfo

SceKernelThreadOptParam

SceKernelThreadRunStatus

SceKernelUtilsMd5Context

SceKernelUtilsMt19937Context

SceKernelUtilsSha1Context

SceKernelVTimerInfo

SceKernelVTimerOptParam

SceKernelVplInfo

SceKernelVplOptParam

SceMp3InitArg

SceMpeg

SceMpegAu

SceMpegAvcMode

SceMpegLLI

SceMpegRingbuffer

SceMpegStream

SceMpegYCrCbBuffer

SceNetAdhocPdpStat

SceNetAdhocPtpStat

SceNetAdhocctlAdhocId

SceNetAdhocctlGameModeInfo

SceNetAdhocctlParams

SceNetAdhocctlPeerInfo

SceNetAdhocctlScanInfo

SceNetMallocStat

ScePspDateTime

ScePspFMatrix2

ScePspFMatrix3

ScePspFMatrix4

ScePspFMatrix4Unaligned

ScePspFRect

ScePspFVector2

ScePspFVector3

ScePspFVector4

ScePspFVector4Unaligned

ScePspIMatrix2

ScePspIMatrix3

ScePspIMatrix4

ScePspIMatrix4Unaligned

ScePspIRect

ScePspIVector2

ScePspIVector3

ScePspIVector4

ScePspL64Rect

ScePspL64Vector2

ScePspL64Vector3

ScePspL64Vector4

ScePspSRect

ScePspSVector2

ScePspSVector3

ScePspSVector4

SceUid

SceUtilityOskData

SceUtilityOskParams

SceUtilitySavedataParam

UmdInfo

UsbCamSetupStillExParam

UsbCamSetupStillParam

UsbCamSetupVideoExParam

UsbCamSetupVideoParam

UtilityDialogCommon

UtilityGameSharingParams

UtilityHtmlViewerParam

UtilityMsgDialogParams

UtilityNetconfAdhoc

UtilityNetconfData

UtilitySavedataFileData

UtilitySavedataListSaveNewData

UtilitySavedataSFOParam

in_addr

sockaddr

timeval

timezone

Enums

AdhocMatchingMode

AlphaFunc

ApctlEvent

ApctlInfo

ApctlInfoSecurityType

ApctlState

AudioFormat

AudioInputFrequency

AudioOutputFrequency

AvModule

BlendDst

BlendOp

BlendSrc

ClutPixelFormat

ColorFunc

CtrlMode

DepthFunc

DisplayMode

DisplayPixelFormat

DisplaySetBufSync

FrontFaceDirection

GeCommand

GeListState

GeMatrixType

GuCallbackId

GuContextType

GuPrimitive

GuQueueMode

GuState

GuSyncBehavior

GuSyncMode

GuTexWrapMode

HttpAuthType

HttpMethod

Interrupt

IoAssignPerms

IoWhence

KeyType

LightMode

LightType

LogicalOperation

MatrixMode

MipmapLevel

Module

NetModule

NetParam

PatchPrimitive

PowerTick

RtcCheckValidError

SceKernelIdListType

ScePspnetAdhocPtpState

SceSysMemBlockTypes

SceSysMemPartitionId

SceUtilityOskInputLanguage

SceUtilityOskInputType

SceUtilityOskResult

SceUtilityOskState

ShadingModel

SignalBehavior

SplineMode

StencilFunc

StencilOperation

SubInterrupt

SystemParamAdhocChannel

SystemParamDateFormat

SystemParamDaylightSavings

SystemParamId

SystemParamLanguage

SystemParamTimeFormat

SystemParamWlanPowerSaveState

TextureColorComponent

TextureEffect

TextureFilter

TextureLevelMode

TextureMapMode

TexturePixelFormat

TextureProjectionMapMode

UmdType

UsbCamDelay

UsbCamEffectMode

UsbCamEvLevel

UsbCamFrameRate

UsbCamResolution

UsbCamResolutionEx

UsbCamWb

UsbModule

UtilityDialogButtonAccept

UtilityGameSharingDataType

UtilityGameSharingMode

UtilityHtmlViewerConnectMode

UtilityHtmlViewerCookieMode

UtilityHtmlViewerDisconnectMode

UtilityHtmlViewerDisplayMode

UtilityHtmlViewerInterfaceMode

UtilityHtmlViewerTextSize

UtilityMsgDialogMode

UtilityMsgDialogPressed

UtilityNetconfAction

UtilitySavedataFocus

UtilitySavedataMode

c_void

Equivalent to C’s void type when used as a [pointer].

Constants

ACTIVATED

AUDIO_CHANNEL_MAX

AUDIO_NEXT_CHANNEL

AUDIO_SAMPLE_MAX

AUDIO_SAMPLE_MIN

AUDIO_VOLUME_MAX

BACK

CONNECTED

ESTABLISHED

EVENT_WAIT_AND

EVENT_WAIT_CLEAR

EVENT_WAIT_MULTIPLE

EVENT_WAIT_OR

FIO_SO_IFDIR

FIO_SO_IFLNK

FIO_SO_IFREG

FIO_SO_IROTH

FIO_SO_IWOTH

FIO_SO_IXOTH

FIO_S_IFDIR

FIO_S_IFLNK

FIO_S_IFREG

FIO_S_IRGRP

FIO_S_IROTH

FIO_S_IRUSR

FIO_S_ISGID

FIO_S_ISUID

FIO_S_ISVTX

FIO_S_IWGRP

FIO_S_IWOTH

FIO_S_IWUSR

FIO_S_IXGRP

FIO_S_IXOTH

FIO_S_IXUSR

FORWARD

GU_AMBIENT

GU_COLOR_4444

GU_COLOR_5551

GU_COLOR_5650

GU_COLOR_8888

GU_COLOR_BUFFER_BIT

GU_DEPTH_BUFFER_BIT

GU_DIFFUSE

GU_FAST_CLEAR_BIT

GU_INDEX_8BIT

GU_INDEX_16BIT

GU_NORMAL_8BIT

GU_NORMAL_16BIT

GU_NORMAL_32BITF

GU_PI

GU_SPECULAR

GU_STENCIL_BUFFER_BIT

GU_TEXTURE_8BIT

GU_TEXTURE_16BIT

GU_TEXTURE_32BITF

GU_TRANSFORM_2D

GU_TRANSFORM_3D

GU_UNKNOWN_LIGHT_COMPONENT

GU_VERTEX_8BIT

GU_VERTEX_16BIT

GU_VERTEX_32BITF

GU_VERTICES1

GU_VERTICES2

GU_VERTICES3

GU_VERTICES4

GU_VERTICES5

GU_VERTICES6

GU_VERTICES7

GU_VERTICES8

GU_WEIGHTS1

GU_WEIGHTS2

GU_WEIGHTS3

GU_WEIGHTS4

GU_WEIGHTS5

GU_WEIGHTS6

GU_WEIGHTS7

GU_WEIGHTS8

GU_WEIGHT_8BIT

GU_WEIGHT_16BIT

GU_WEIGHT_32BITF

HOLD

INT_MAX

INT_MIN

PLAY_PAUSE

POWER_INFO_AC_POWER

POWER_INFO_BATTERY_EXIST

POWER_INFO_BATTERY_LOW

POWER_INFO_BATTERY_POWER

POWER_INFO_HOLD_SWITCH

POWER_INFO_POWER_SWITCH

POWER_INFO_RESUME_COMPLETE

POWER_INFO_RESUMING

POWER_INFO_STANDBY

POWER_INFO_SUSPENDING

PSP_CTRL_CIRCLE

PSP_CTRL_CROSS

PSP_CTRL_DISC

PSP_CTRL_DOWN

PSP_CTRL_HOLD

PSP_CTRL_HOME

PSP_CTRL_LEFT

PSP_CTRL_LTRIGGER

PSP_CTRL_MS

PSP_CTRL_NOTE

PSP_CTRL_REMOTE

PSP_CTRL_RIGHT

PSP_CTRL_RTRIGGER

PSP_CTRL_SCREEN

PSP_CTRL_SELECT

PSP_CTRL_SQUARE

PSP_CTRL_START

PSP_CTRL_TRIANGLE

PSP_CTRL_UP

PSP_CTRL_VOLDOWN

PSP_CTRL_VOLUP

PSP_CTRL_WLAN_UP

PSP_O_APPEND

PSP_O_CREAT

PSP_O_DIR

PSP_O_EXCL

PSP_O_NBLOCK

PSP_O_NO_WAIT

PSP_O_RD_ONLY

PSP_O_RD_WR

PSP_O_TRUNC

PSP_O_WR_ONLY

REG_KEYNAME_SIZE

SYSTEM_REGISTRY

THREAD_ATTR_CLEAR_STACK

THREAD_ATTR_NO_FILLSTACK

THREAD_ATTR_SCRATCH_SRAM

THREAD_ATTR_USBWLAN

THREAD_ATTR_USER

THREAD_ATTR_VFPU

THREAD_ATTR_VSH

UMD_CHANGED

UMD_INITED

UMD_INITING

UMD_NOT_PRESENT

UMD_PRESENT

UMD_READY

USB_BUS_DRIVER_NAME

USB_CAM_DRIVER_NAME

USB_CAM_FLIP

USB_CAM_MIC_DRIVER_NAME

USB_CAM_MIRROR

USB_CAM_PID

USB_STOR_DRIVER_NAME

UTILITY_HTMLVIEWER_DISABLE_CURSOR

UTILITY_HTMLVIEWER_DISABLE_DOWNLOAD_COMPLETE_DIALOG

UTILITY_HTMLVIEWER_DISABLE_DOWNLOAD_DESTINATION_DIALOG

UTILITY_HTMLVIEWER_DISABLE_DOWNLOAD_START_DIALOG

UTILITY_HTMLVIEWER_DISABLE_EXIT_DIALOG

UTILITY_HTMLVIEWER_DISABLE_LRTRIGGER

UTILITY_HTMLVIEWER_DISABLE_STARTUP_LIMITS

UTILITY_HTMLVIEWER_DISABLE_TAB_DISPLAY

UTILITY_HTMLVIEWER_ENABLE_ANALOG_HOLD

UTILITY_HTMLVIEWER_ENABLE_FLASH

UTILITY_HTMLVIEWER_LOCK_DOWNLOAD_DESTINATION_DIALOG

UTILITY_HTMLVIEWER_OPEN_SCE_START_PAGE

UTILITY_MSGDIALOG_DEFAULT_NO

UTILITY_MSGDIALOG_ERROR

UTILITY_MSGDIALOG_TEXT

UTILITY_MSGDIALOG_YES_NO_BUTTONS

VOL_DOWN

VOL_UP

Functions

QueryIntrHandlerInfo⚠

sceAtracAddStreamData⚠

sceAtracDecodeData⚠

sceAtracGetAtracID⚠

sceAtracGetBitrate⚠

sceAtracGetBufferInfoForReseting⚠

sceAtracGetChannel⚠

sceAtracGetInternalErrorInfo⚠

sceAtracGetLoopStatus⚠

sceAtracGetMaxSample⚠

sceAtracGetNextDecodePosition⚠

sceAtracGetNextSample⚠

sceAtracGetRemainFrame⚠

sceAtracGetSecondBufferInfo⚠

sceAtracGetSoundSample⚠

sceAtracGetStreamDataInfo⚠

sceAtracReleaseAtracID⚠

sceAtracResetPlayPosition⚠

sceAtracSetData⚠

sceAtracSetDataAndGetID⚠

sceAtracSetHalfwayBuffer⚠

sceAtracSetHalfwayBufferAndGetID⚠

sceAtracSetLoopNum⚠

sceAtracSetSecondBuffer⚠

sceAudioChRelease⚠

sceAudioChReserve⚠

sceAudioChangeChannelConfig⚠

sceAudioChangeChannelVolume⚠

sceAudioGetChannelRestLen⚠

sceAudioGetChannelRestLength⚠

sceAudioGetInputLength⚠

sceAudioInput⚠

sceAudioInputBlocking⚠

sceAudioInputInit⚠

sceAudioInputInitEx⚠

sceAudioOutput⚠

sceAudioOutput2ChangeLength⚠

sceAudioOutput2GetRestSample⚠

sceAudioOutput2OutputBlocking⚠

sceAudioOutput2Release⚠

sceAudioOutput2Reserve⚠

sceAudioOutputBlocking⚠

sceAudioOutputPanned⚠

sceAudioOutputPannedBlocking⚠

sceAudioPollInputEnd⚠

sceAudioSRCChRelease⚠

sceAudioSRCChReserve⚠

sceAudioSRCOutputBlocking⚠

sceAudioSetChannelDataLen⚠

sceAudioWaitInputEnd⚠

sceCtrlGetIdleCancelThreshold⚠

sceCtrlGetSamplingCycle⚠

sceCtrlGetSamplingMode⚠

sceCtrlPeekBufferNegative⚠

sceCtrlPeekBufferPositive⚠

sceCtrlPeekLatch⚠

sceCtrlReadBufferNegative⚠

sceCtrlReadBufferPositive⚠

sceCtrlReadLatch⚠

sceCtrlSetIdleCancelThreshold⚠

sceCtrlSetSamplingCycle⚠

sceCtrlSetSamplingMode⚠

sceDisplayGetAccumulatedHcount⚠

sceDisplayGetCurrentHcount⚠

sceDisplayGetFrameBuf⚠

sceDisplayGetFramePerSec⚠

sceDisplayGetMode⚠

sceDisplayGetVcount⚠

sceDisplayIsForeground⚠

sceDisplayIsVblank⚠

sceDisplaySetFrameBuf⚠

sceDisplaySetMode⚠

sceDisplayWaitVblank⚠

sceDisplayWaitVblankCB⚠

sceDisplayWaitVblankStart⚠

sceDisplayWaitVblankStartCB⚠

sceGeBreak⚠

sceGeContinue⚠

sceGeDrawSync⚠

sceGeEdramGetAddr⚠

sceGeEdramGetSize⚠

sceGeEdramSetAddrTranslation⚠

sceGeGetCmd⚠

sceGeGetMtx⚠

sceGeGetStack⚠

sceGeListDeQueue⚠

sceGeListEnQueue⚠

sceGeListEnQueueHead⚠

sceGeListSync⚠

sceGeListUpdateStallAddr⚠

sceGeRestoreContext⚠

sceGeSaveContext⚠

sceGeSetCallback⚠

sceGeUnsetCallback⚠

sceGuAlphaFunc⚠

sceGuAmbient⚠

sceGuAmbientColor⚠

sceGuBeginObject⚠

sceGuBlendFunc⚠

sceGuBoneMatrix⚠

sceGuBreak⚠

sceGuCallList⚠

sceGuCallMode⚠

sceGuCheckList⚠

sceGuClear⚠

sceGuClearColor⚠

sceGuClearDepth⚠

sceGuClearStencil⚠

sceGuClutLoad⚠

sceGuClutMode⚠

sceGuColor⚠

sceGuColorFunc⚠

sceGuColorMaterial⚠

sceGuContinue⚠

sceGuCopyImage⚠

sceGuDepthBuffer⚠

sceGuDepthFunc⚠

sceGuDepthMask⚠

sceGuDepthOffset⚠

sceGuDepthRange⚠

sceGuDisable⚠

sceGuDispBuffer⚠

sceGuDisplay⚠

sceGuDrawArray⚠

sceGuDrawArrayN⚠

sceGuDrawBezier⚠

sceGuDrawBuffer⚠

sceGuDrawBufferList⚠

sceGuDrawSpline⚠

sceGuEnable⚠

sceGuEndObject⚠

sceGuFinish⚠

sceGuFinishId⚠

sceGuFog⚠

sceGuFrontFace⚠

sceGuGetAllStatus⚠

sceGuGetMemory⚠

sceGuGetStatus⚠

sceGuInit⚠

sceGuLight⚠

sceGuLightAtt⚠

sceGuLightColor⚠

sceGuLightMode⚠

sceGuLightSpot⚠

sceGuLogicalOp⚠

sceGuMaterial⚠

sceGuModelColor⚠

sceGuMorphWeight⚠

sceGuOffset⚠

sceGuPatchDivide⚠

sceGuPatchFrontFace⚠

sceGuPatchPrim⚠

sceGuPixelMask⚠

sceGuScissor⚠

sceGuSendCommandf⚠

sceGuSendCommandi⚠

sceGuSendList⚠

sceGuSetAllStatus⚠

sceGuSetCallback⚠

sceGuSetDither⚠

sceGuSetMatrix⚠

sceGuSetStatus⚠

sceGuShadeModel⚠

sceGuSignal⚠

sceGuSpecular⚠

sceGuStart⚠

sceGuStencilFunc⚠

sceGuStencilOp⚠

sceGuSwapBuffers⚠

sceGuSync⚠

sceGuTerm⚠

sceGuTexEnvColor⚠

sceGuTexFilter⚠

sceGuTexFlush⚠

sceGuTexFunc⚠

sceGuTexImage⚠

sceGuTexLevelMode⚠

sceGuTexMapMode⚠

sceGuTexMode⚠

sceGuTexOffset⚠

sceGuTexProjMapMode⚠

sceGuTexScale⚠

sceGuTexSlope⚠

sceGuTexSync⚠

sceGuTexWrap⚠

sceGuViewport⚠

sceGumDrawArray⚠

sceGumDrawArrayN⚠

sceGumDrawBezier⚠

sceGumDrawSpline⚠

sceGumFastInverse⚠

sceGumFullInverse⚠

sceGumLoadIdentity⚠

sceGumLoadMatrix⚠

sceGumLookAt⚠

sceGumMatrixMode⚠

sceGumMultMatrix⚠

sceGumOrtho⚠

sceGumPerspective⚠

sceGumPopMatrix⚠

sceGumPushMatrix⚠

sceGumRotateX⚠

sceGumRotateXYZ⚠

sceGumRotateY⚠

sceGumRotateZ⚠

sceGumRotateZYX⚠

sceGumScale⚠

sceGumStoreMatrix⚠

sceGumTranslate⚠

sceGumUpdateMatrix⚠

sceHprmIsHeadphoneExist⚠

sceHprmIsMicrophoneExist⚠

sceHprmIsRemoteExist⚠

sceHprmPeekCurrentKey⚠

sceHprmPeekLatch⚠

sceHprmReadLatch⚠

sceHttpAbortRequest⚠

sceHttpAddExtraHeader⚠

sceHttpCreateConnection⚠

sceHttpCreateConnectionWithURL⚠

sceHttpCreateRequest⚠

sceHttpCreateRequestWithURL⚠

sceHttpCreateTemplate⚠

sceHttpDeleteConnection⚠

sceHttpDeleteHeader⚠

sceHttpDeleteRequest⚠

sceHttpDeleteTemplate⚠

sceHttpDisableAuth⚠

sceHttpDisableCache⚠

sceHttpDisableCookie⚠

sceHttpDisableKeepAlive⚠

sceHttpDisableRedirect⚠

sceHttpEnableAuth⚠

sceHttpEnableCache⚠

sceHttpEnableCookie⚠

sceHttpEnableKeepAlive⚠

sceHttpEnableRedirect⚠

sceHttpEnd⚠

sceHttpEndCache⚠

sceHttpGetAllHeader⚠

sceHttpGetContentLength⚠

sceHttpGetNetworkErrno⚠

sceHttpGetProxy⚠

sceHttpGetStatusCode⚠

sceHttpInit⚠

sceHttpInitCache⚠

sceHttpLoadSystemCookie⚠

sceHttpReadData⚠

sceHttpSaveSystemCookie⚠

sceHttpSendRequest⚠

sceHttpSetAuthInfoCB⚠

sceHttpSetConnectTimeOut⚠

sceHttpSetMallocFunction⚠

sceHttpSetProxy⚠

sceHttpSetRecvTimeOut⚠

sceHttpSetResHeaderMaxSize⚠

sceHttpSetResolveRetry⚠

sceHttpSetResolveTimeOut⚠

sceHttpSetSendTimeOut⚠

sceHttpsEnd⚠

sceHttpsInit⚠

sceHttpsLoadDefaultCert⚠

sceIoAssign⚠

sceIoCancel⚠

sceIoChangeAsyncPriority⚠

sceIoChdir⚠

sceIoChstat⚠

sceIoClose⚠

sceIoCloseAsync⚠

sceIoDclose⚠

sceIoDevctl⚠

sceIoDopen⚠

sceIoDread⚠

sceIoGetAsyncStat⚠

sceIoGetDevType⚠

sceIoGetstat⚠

sceIoIoctl⚠

sceIoIoctlAsync⚠

sceIoLseek⚠

sceIoLseek32⚠

sceIoLseek32Async⚠

sceIoLseekAsync⚠

sceIoMkdir⚠

sceIoOpen⚠

sceIoOpenAsync⚠

sceIoPollAsync⚠

sceIoRead⚠

sceIoReadAsync⚠

sceIoRemove⚠

sceIoRename⚠

sceIoRmdir⚠

sceIoSetAsyncCallback⚠

sceIoSync⚠

sceIoUnassign⚠

sceIoWaitAsync⚠

sceIoWaitAsyncCB⚠

sceIoWrite⚠

sceIoWriteAsync⚠

sceJpegCreateMJpeg⚠

sceJpegDecodeMJpeg⚠

sceJpegDeleteMJpeg⚠

sceJpegFinishMJpeg⚠

sceJpegInitMJpeg⚠

sceKernelAllocPartitionMemory⚠

sceKernelAllocateFpl⚠

sceKernelAllocateFplCB⚠

sceKernelAllocateVpl⚠

sceKernelAllocateVplCB⚠

sceKernelCancelAlarm⚠

sceKernelCancelCallback⚠

sceKernelCancelFpl⚠

sceKernelCancelMsgPipe⚠

sceKernelCancelReceiveMbx⚠

sceKernelCancelVTimerHandler⚠

sceKernelCancelVpl⚠

sceKernelCancelWakeupThread⚠

sceKernelChangeCurrentThreadAttr⚠

sceKernelChangeThreadPriority⚠

sceKernelCheckCallback⚠

sceKernelCheckThreadStack⚠

sceKernelClearEventFlag⚠

sceKernelCpuResumeIntr⚠

sceKernelCpuResumeIntrWithSync⚠

sceKernelCpuSuspendIntr⚠

sceKernelCreateCallback⚠

sceKernelCreateEventFlag⚠

sceKernelCreateFpl⚠

sceKernelCreateMbx⚠

sceKernelCreateMsgPipe⚠

sceKernelCreateSema⚠

sceKernelCreateThread⚠

sceKernelCreateVTimer⚠

sceKernelCreateVpl⚠

sceKernelDcacheInvalidateRange⚠

sceKernelDcacheWritebackAll⚠

sceKernelDcacheWritebackInvalidateAll⚠

sceKernelDcacheWritebackInvalidateRange⚠

sceKernelDcacheWritebackRange⚠

sceKernelDelaySysClockThread⚠

sceKernelDelaySysClockThreadCB⚠

sceKernelDelayThread⚠

sceKernelDelayThreadCB⚠

sceKernelDeleteCallback⚠

sceKernelDeleteEventFlag⚠

sceKernelDeleteFpl⚠

sceKernelDeleteMbx⚠

sceKernelDeleteMsgPipe⚠

sceKernelDeleteSema⚠

sceKernelDeleteThread⚠

sceKernelDeleteVTimer⚠

sceKernelDeleteVpl⚠

sceKernelDevkitVersion⚠

sceKernelDisableSubIntr⚠

sceKernelEnableSubIntr⚠

sceKernelExitDeleteThread⚠

sceKernelExitGame⚠

sceKernelExitThread⚠

sceKernelFreeFpl⚠

sceKernelFreePartitionMemory⚠

sceKernelFreeVpl⚠

sceKernelGetBlockHeadAddr⚠

sceKernelGetCallbackCount⚠

sceKernelGetCompiledSdkVersion⚠

sceKernelGetModuleIdList⚠

sceKernelGetSystemTime⚠

sceKernelGetSystemTimeLow⚠

sceKernelGetSystemTimeWide⚠

sceKernelGetThreadCurrentPriority⚠

sceKernelGetThreadExitStatus⚠

sceKernelGetThreadId⚠

sceKernelGetThreadStackFreeSize⚠

sceKernelGetThreadmanIdList⚠

sceKernelGetThreadmanIdType⚠

sceKernelGetVTimerBase⚠

sceKernelGetVTimerBaseWide⚠

sceKernelGetVTimerTime⚠

sceKernelGetVTimerTimeWide⚠

sceKernelIcacheInvalidateAll⚠

sceKernelIcacheInvalidateRange⚠

sceKernelIsCpuIntrEnable⚠

sceKernelIsCpuIntrSuspended⚠

sceKernelLibcClock⚠

sceKernelLibcGettimeofday⚠

sceKernelLibcTime⚠

sceKernelLoadExec⚠

sceKernelLoadModule⚠

sceKernelLoadModuleBufferUsbWlan⚠

sceKernelLoadModuleByID⚠

sceKernelLoadModuleMs⚠

sceKernelMaxFreeMemSize⚠

sceKernelNotifyCallback⚠

sceKernelPollEventFlag⚠

sceKernelPollMbx⚠

sceKernelPollSema⚠

sceKernelQueryModuleInfo⚠

sceKernelReceiveMbx⚠

sceKernelReceiveMbxCB⚠

sceKernelReceiveMsgPipe⚠

sceKernelReceiveMsgPipeCB⚠

sceKernelReferAlarmStatus⚠

sceKernelReferCallbackStatus⚠

sceKernelReferEventFlagStatus⚠

sceKernelReferFplStatus⚠

sceKernelReferGlobalProfiler⚠

sceKernelReferMbxStatus⚠

sceKernelReferMsgPipeStatus⚠

sceKernelReferSemaStatus⚠

sceKernelReferSystemStatus⚠

sceKernelReferThreadEventHandlerStatus⚠

sceKernelReferThreadProfiler⚠

sceKernelReferThreadRunStatus⚠

sceKernelReferThreadStatus⚠

sceKernelReferVTimerStatus⚠

sceKernelReferVplStatus⚠

sceKernelRegisterExitCallback⚠

sceKernelRegisterSubIntrHandler⚠

sceKernelRegisterThreadEventHandler⚠

sceKernelReleaseSubIntrHandler⚠

sceKernelReleaseThreadEventHandler⚠

sceKernelReleaseWaitThread⚠

sceKernelResumeDispatchThread⚠

sceKernelResumeThread⚠

sceKernelRotateThreadReadyQueue⚠

sceKernelSelfStopUnloadModule⚠

sceKernelSendMbx⚠

sceKernelSendMsgPipe⚠

sceKernelSendMsgPipeCB⚠

sceKernelSetAlarm⚠

sceKernelSetCompiledSdkVersion⚠

sceKernelSetEventFlag⚠

sceKernelSetSysClockAlarm⚠

sceKernelSetVTimerHandler⚠

sceKernelSetVTimerHandlerWide⚠

sceKernelSetVTimerTime⚠

sceKernelSetVTimerTimeWide⚠

sceKernelSignalSema⚠

sceKernelSleepThread⚠

sceKernelSleepThreadCB⚠

sceKernelStartModule⚠

sceKernelStartThread⚠

sceKernelStartVTimer⚠

sceKernelStderr⚠

sceKernelStdin⚠

sceKernelStdout⚠

sceKernelStopModule⚠

sceKernelStopUnloadSelfModule⚠

sceKernelStopVTimer⚠

sceKernelSuspendDispatchThread⚠

sceKernelSuspendThread⚠

sceKernelSysClock2USec⚠

sceKernelSysClock2USecWide⚠

sceKernelTerminateDeleteThread⚠

sceKernelTerminateThread⚠

sceKernelTotalFreeMemSize⚠

sceKernelTryAllocateFpl⚠

sceKernelTryAllocateVpl⚠

sceKernelTryReceiveMsgPipe⚠

sceKernelTrySendMsgPipe⚠

sceKernelUSec2SysClock⚠

sceKernelUSec2SysClockWide⚠

sceKernelUnloadModule⚠

sceKernelUtilsMd5BlockInit⚠

sceKernelUtilsMd5BlockResult⚠

sceKernelUtilsMd5BlockUpdate⚠

sceKernelUtilsMd5Digest⚠

sceKernelUtilsMt19937Init⚠

sceKernelUtilsMt19937UInt⚠

sceKernelUtilsSha1BlockInit⚠

sceKernelUtilsSha1BlockResult⚠

sceKernelUtilsSha1BlockUpdate⚠

sceKernelUtilsSha1Digest⚠

sceKernelVolatileMemLock⚠

sceKernelVolatileMemTryLock⚠

sceKernelVolatileMemUnlock⚠

sceKernelWaitEventFlag⚠

sceKernelWaitEventFlagCB⚠

sceKernelWaitSema⚠

sceKernelWaitSemaCB⚠

sceKernelWaitThreadEnd⚠

sceKernelWaitThreadEndCB⚠

sceKernelWakeupThread⚠

sceMp3CheckStreamDataNeeded⚠

sceMp3Decode⚠

sceMp3GetBitRate⚠

sceMp3GetInfoToAddStreamData⚠

sceMp3GetLoopNum⚠

sceMp3GetMaxOutputSample⚠

sceMp3GetMp3ChannelNum⚠

sceMp3GetSamplingRate⚠

sceMp3GetSumDecodedSample⚠

sceMp3Init⚠

sceMp3InitResource⚠

sceMp3NotifyAddStreamData⚠

sceMp3ReleaseMp3Handle⚠

sceMp3ReserveMp3Handle⚠

sceMp3ResetPlayPosition⚠

sceMp3SetLoopNum⚠

sceMp3TermResource⚠

sceMpegAtracDecode⚠

sceMpegAvcDecode⚠

sceMpegAvcDecodeMode⚠

sceMpegAvcDecodeStop⚠

sceMpegBaseCscInit⚠

sceMpegBaseCscVme⚠

sceMpegBaseYCrCbCopyVme⚠

sceMpegCreate⚠

sceMpegDelete⚠

sceMpegFinish⚠

sceMpegFlushAllStream⚠

sceMpegFreeAvcEsBuf⚠

sceMpegGetAtracAu⚠

sceMpegGetAvcAu⚠

sceMpegInit⚠

sceMpegInitAu⚠

sceMpegMallocAvcEsBuf⚠

sceMpegQueryAtracEsSize⚠

sceMpegQueryMemSize⚠

sceMpegQueryStreamOffset⚠

sceMpegQueryStreamSize⚠

sceMpegRegistStream⚠

sceMpegRingbufferAvailableSize⚠

sceMpegRingbufferConstruct⚠

sceMpegRingbufferDestruct⚠

sceMpegRingbufferPut⚠

sceMpegRingbufferQueryMemSize⚠

sceMpegUnRegistStream⚠

sceMpegbase_BEA18F91⚠

sceNetAdhocGameModeCreateMaster⚠

sceNetAdhocGameModeCreateReplica⚠

sceNetAdhocGameModeDeleteMaster⚠

sceNetAdhocGameModeDeleteReplica⚠

sceNetAdhocGameModeUpdateMaster⚠

sceNetAdhocGameModeUpdateReplica⚠

sceNetAdhocGetPdpStat⚠

sceNetAdhocGetPtpStat⚠

sceNetAdhocInit⚠

sceNetAdhocMatchingAbortSendData⚠

sceNetAdhocMatchingCancelTarget⚠

sceNetAdhocMatchingCancelTargetWithOpt⚠

sceNetAdhocMatchingCreate⚠

sceNetAdhocMatchingDelete⚠

sceNetAdhocMatchingGetHelloOpt⚠

sceNetAdhocMatchingGetMembers⚠

sceNetAdhocMatchingGetPoolMaxAlloc⚠

sceNetAdhocMatchingGetPoolStat⚠

sceNetAdhocMatchingInit⚠

sceNetAdhocMatchingSelectTarget⚠

sceNetAdhocMatchingSendData⚠

sceNetAdhocMatchingSetHelloOpt⚠

sceNetAdhocMatchingStart⚠

sceNetAdhocMatchingStop⚠

sceNetAdhocMatchingTerm⚠

sceNetAdhocPdpCreate⚠

sceNetAdhocPdpDelete⚠

sceNetAdhocPdpRecv⚠

sceNetAdhocPdpSend⚠

sceNetAdhocPtpAccept⚠

sceNetAdhocPtpClose⚠

sceNetAdhocPtpConnect⚠

sceNetAdhocPtpFlush⚠

sceNetAdhocPtpListen⚠

sceNetAdhocPtpOpen⚠

sceNetAdhocPtpRecv⚠

sceNetAdhocPtpSend⚠

sceNetAdhocTerm⚠

sceNetAdhocctlAddHandler⚠

sceNetAdhocctlConnect⚠

sceNetAdhocctlCreate⚠

sceNetAdhocctlCreateEnterGameMode⚠

sceNetAdhocctlDelHandler⚠

sceNetAdhocctlDisconnect⚠

sceNetAdhocctlExitGameMode⚠

sceNetAdhocctlGetAddrByName⚠

sceNetAdhocctlGetAdhocId⚠

sceNetAdhocctlGetGameModeInfo⚠

sceNetAdhocctlGetNameByAddr⚠

sceNetAdhocctlGetParameter⚠

sceNetAdhocctlGetPeerInfo⚠

sceNetAdhocctlGetPeerList⚠

sceNetAdhocctlGetScanInfo⚠

sceNetAdhocctlGetState⚠

sceNetAdhocctlInit⚠

sceNetAdhocctlJoin⚠

sceNetAdhocctlJoinEnterGameMode⚠

sceNetAdhocctlScan⚠

sceNetAdhocctlTerm⚠

sceNetApctlAddHandler⚠

sceNetApctlConnect⚠

sceNetApctlDelHandler⚠

sceNetApctlDisconnect⚠

sceNetApctlGetInfo⚠

sceNetApctlGetState⚠

sceNetApctlInit⚠

sceNetApctlTerm⚠

sceNetEtherNtostr⚠

sceNetEtherStrton⚠

sceNetFreeThreadinfo⚠

sceNetGetLocalEtherAddr⚠

sceNetGetMallocStat⚠

sceNetInetAccept⚠

sceNetInetBind⚠

sceNetInetClose⚠

sceNetInetConnect⚠

sceNetInetGetErrno⚠

sceNetInetGetsockopt⚠

sceNetInetInit⚠

sceNetInetListen⚠

sceNetInetRecv⚠

sceNetInetRecvfrom⚠

sceNetInetSend⚠

sceNetInetSendto⚠

sceNetInetSetsockopt⚠

sceNetInetShutdown⚠

sceNetInetSocket⚠

sceNetInetTerm⚠

sceNetInit⚠

sceNetResolverCreate⚠

sceNetResolverDelete⚠

sceNetResolverInit⚠

sceNetResolverStartAtoN⚠

sceNetResolverStartNtoA⚠

sceNetResolverStop⚠

sceNetResolverTerm⚠

sceNetTerm⚠

sceNetThreadAbort⚠

sceOpenPSIDGetOpenPSID⚠

scePowerGetBatteryChargingStatus⚠

scePowerGetBatteryElec⚠

scePowerGetBatteryLifePercent⚠

scePowerGetBatteryLifeTime⚠

scePowerGetBatteryTemp⚠

scePowerGetBatteryVolt⚠

scePowerGetBusClockFrequency⚠

scePowerGetBusClockFrequencyFloat⚠

scePowerGetBusClockFrequencyInt⚠

scePowerGetCpuClockFrequency⚠

scePowerGetCpuClockFrequencyFloat⚠

scePowerGetCpuClockFrequencyInt⚠

scePowerGetIdleTimer⚠

scePowerIdleTimerDisable⚠

scePowerIdleTimerEnable⚠

scePowerIsBatteryCharging⚠

scePowerIsBatteryExist⚠

scePowerIsLowBattery⚠

scePowerIsPowerOnline⚠

scePowerLock⚠

scePowerRegisterCallback⚠

scePowerRequestStandby⚠

scePowerRequestSuspend⚠

scePowerSetBusClockFrequency⚠

scePowerSetClockFrequency⚠

scePowerSetCpuClockFrequency⚠

scePowerTick⚠

scePowerUnlock⚠

scePowerUnregisterCallback⚠

sceRegCloseCategory⚠

sceRegCloseRegistry⚠

sceRegCreateKey⚠

sceRegFlushCategory⚠

sceRegFlushRegistry⚠

sceRegGetKeyInfo⚠

sceRegGetKeyInfoByName⚠

sceRegGetKeyValue⚠

sceRegGetKeyValueByName⚠

sceRegGetKeys⚠

sceRegGetKeysNum⚠

sceRegOpenCategory⚠

sceRegOpenRegistry⚠

sceRegRemoveCategory⚠

sceRegRemoveRegistry⚠

sceRegSetKeyValue⚠

sceRtcCheckValid⚠

sceRtcCompareTick⚠

sceRtcConvertLocalTimeToUTC⚠

sceRtcConvertUtcToLocalTime⚠

sceRtcFormatRFC2822⚠

sceRtcFormatRFC3339⚠

sceRtcFormatRFC2822LocalTime⚠

sceRtcFormatRFC3339LocalTime⚠

sceRtcGetCurrentClock⚠

sceRtcGetCurrentClockLocalTime⚠

sceRtcGetCurrentTick⚠

sceRtcGetDayOfWeek⚠

sceRtcGetDaysInMonth⚠

sceRtcGetDosTime⚠

sceRtcGetTick⚠

sceRtcGetTickResolution⚠

sceRtcGetTime64_t⚠

sceRtcGetTime_t⚠

sceRtcGetWin32FileTime⚠

sceRtcIsLeapYear⚠

sceRtcParseDateTime⚠

sceRtcParseRFC3339⚠

sceRtcSetDosTime⚠

sceRtcSetTick⚠

sceRtcSetTime64_t⚠

sceRtcSetTime_t⚠

sceRtcSetWin32FileTime⚠

sceRtcTickAddDays⚠

sceRtcTickAddHours⚠

sceRtcTickAddMicroseconds⚠

sceRtcTickAddMinutes⚠

sceRtcTickAddMonths⚠

sceRtcTickAddSeconds⚠

sceRtcTickAddTicks⚠

sceRtcTickAddWeeks⚠

sceRtcTickAddYears⚠

sceSslEnd⚠

sceSslGetUsedMemoryCurrent⚠

sceSslGetUsedMemoryMax⚠

sceSslInit⚠

sceUmdActivate⚠

sceUmdCancelWaitDriveStat⚠

sceUmdCheckMedium⚠

sceUmdDeactivate⚠

sceUmdGetDiscInfo⚠

sceUmdGetDriveStat⚠

sceUmdGetErrorStat⚠

sceUmdRegisterUMDCallBack⚠

sceUmdReplacePermit⚠

sceUmdReplaceProhibit⚠

sceUmdUnRegisterUMDCallBack⚠

sceUmdWaitDriveStat⚠

sceUmdWaitDriveStatCB⚠

sceUmdWaitDriveStatWithTimer⚠

sceUsbActivate⚠

sceUsbCamAutoImageReverseSW⚠

sceUsbCamGetAutoImageReverseState⚠

sceUsbCamGetBrightness⚠

sceUsbCamGetContrast⚠

sceUsbCamGetEvLevel⚠

sceUsbCamGetImageEffectMode⚠

sceUsbCamGetLensDirection⚠

sceUsbCamGetReadVideoFrameSize⚠

sceUsbCamGetReverseMode⚠

sceUsbCamGetSaturation⚠

sceUsbCamGetSharpness⚠

sceUsbCamGetZoom⚠

sceUsbCamPollReadVideoFrameEnd⚠

sceUsbCamReadVideoFrame⚠

sceUsbCamReadVideoFrameBlocking⚠

sceUsbCamSetBrightness⚠

sceUsbCamSetContrast⚠

sceUsbCamSetEvLevel⚠

sceUsbCamSetImageEffectMode⚠

sceUsbCamSetReverseMode⚠

sceUsbCamSetSaturation⚠

sceUsbCamSetSharpness⚠

sceUsbCamSetZoom⚠

sceUsbCamSetupStill⚠

sceUsbCamSetupStillEx⚠

sceUsbCamSetupVideo⚠

sceUsbCamSetupVideoEx⚠

sceUsbCamStartVideo⚠

sceUsbCamStillCancelInput⚠

sceUsbCamStillGetInputLength⚠

sceUsbCamStillInput⚠

sceUsbCamStillInputBlocking⚠

sceUsbCamStillPollInputEnd⚠

sceUsbCamStillWaitInputEnd⚠

sceUsbCamStopVideo⚠

sceUsbCamWaitReadVideoFrameEnd⚠

sceUsbDeactivate⚠

sceUsbGetDrvState⚠

sceUsbGetState⚠

sceUsbStart⚠

sceUsbStop⚠

sceUsbstorBootRegisterNotify⚠

sceUsbstorBootSetCapacity⚠

sceUsbstorBootUnregisterNotify⚠

sceUtilityCheckNetParam⚠

sceUtilityCopyNetParam⚠

sceUtilityCreateNetParam⚠

sceUtilityDeleteNetParam⚠

sceUtilityGameSharingGetStatus⚠

sceUtilityGameSharingInitStart⚠

sceUtilityGameSharingShutdownStart⚠

sceUtilityGameSharingUpdate⚠

sceUtilityGetNetParam⚠

sceUtilityGetSystemParamInt⚠

sceUtilityGetSystemParamString⚠

sceUtilityHtmlViewerGetStatus⚠

sceUtilityHtmlViewerInitStart⚠

sceUtilityHtmlViewerShutdownStart⚠

sceUtilityHtmlViewerUpdate⚠

sceUtilityLoadAvModule⚠

sceUtilityLoadModule⚠

sceUtilityLoadNetModule⚠

sceUtilityLoadUsbModule⚠

sceUtilityMsgDialogAbort⚠

sceUtilityMsgDialogGetStatus⚠

sceUtilityMsgDialogInitStart⚠

sceUtilityMsgDialogShutdownStart⚠

sceUtilityMsgDialogUpdate⚠

sceUtilityNetconfGetStatus⚠

sceUtilityNetconfInitStart⚠

sceUtilityNetconfShutdownStart⚠

sceUtilityNetconfUpdate⚠

sceUtilityOskGetStatus⚠

sceUtilityOskInitStart⚠

sceUtilityOskShutdownStart⚠

sceUtilityOskUpdate⚠

sceUtilitySavedataGetStatus⚠

sceUtilitySavedataInitStart⚠

sceUtilitySavedataShutdownStart⚠

sceUtilitySavedataUpdate⚠

sceUtilitySetNetParam⚠

sceUtilitySetSystemParamInt⚠

sceUtilitySetSystemParamString⚠

sceUtilityUnloadAvModule⚠

sceUtilityUnloadModule⚠

sceUtilityUnloadNetModule⚠

sceUtilityUnloadUsbModule⚠

sceWlanDevAttach⚠

sceWlanDevDetach⚠

sceWlanDevIsPowerOn⚠

sceWlanGetEtherAddr⚠

sceWlanGetSwitchState⚠

Type Aliases

AdhocMatchingCallback

GuCallback

GuSwapBuffersCallback

HttpFreeFunction

HttpMallocFunction

HttpPasswordCB

HttpReallocFunction

IoPermissions

PowerCallback

SceKernelAlarmHandler

SceKernelCallbackFunction

SceKernelThreadEntry

SceKernelThreadEventHandler

SceKernelVTimerHandler

SceKernelVTimerHandlerWide

SceMpegRingbufferCb

SceNetAdhocctlHandler

SceNetApctlHandler

UmdCallback

c_char

Equivalent to C’s char type.

c_double

Equivalent to C’s double type.

c_float

Equivalent to C’s float type.

c_int

Equivalent to C’s signed int (int) type.

c_long

Equivalent to C’s signed long (long) type.

c_longlong

Equivalent to C’s signed long long (long long) type.

c_schar

Equivalent to C’s signed char type.

c_short

Equivalent to C’s signed short (short) type.

c_uchar

Equivalent to C’s unsigned char type.

c_uint

Equivalent to C’s unsigned int type.

c_ulong

Equivalent to C’s unsigned long type.

c_ulonglong

Equivalent to C’s unsigned long long type.

c_ushort

Equivalent to C’s unsigned short type.

int8_tDeprecated

int16_tDeprecated

int32_tDeprecated

int64_tDeprecated

intmax_t

intptr_t

ptrdiff_t

size_t

socklen_t

ssize_t

uint8_tDeprecated

uint16_tDeprecated

uint32_tDeprecated

uint64_tDeprecated

uintmax_t

uintptr_t

Unions

SceNetApctlInfo

ScePspMatrix2

ScePspMatrix3

ScePspMatrix4

ScePspVector2

ScePspVector3

ScePspVector4

UtilityNetData