libc::new::glibc::sysdeps::unix::linux::bits::types

Module siginfo_t

Source header: sysdeps/unix/sysv/linux/bits/types/siginfo_t.h

Note this module corresponds with multiple header files upstream. The above header file contains the generic definition, while other architecture-specific definitions live upstream at sysdeps/unix/sysv/linux/<arch>/bits/siginfo-arch.h. Currently, the following set of header files are contained in this one Rust module:

sysdeps/unix/sysv/linux/bits/types/siginfo_t.h

sysdeps/unix/sysv/linux/x86/bits/siginfo-arch.h

sysdeps/unix/sysv/linux/mips/bits/siginfo-arch.h

sysdeps/unix/sysv/linux/sparc/bits/siginfo-arch.h

https://github.com/sailfishos-mirror/glibc/blob/master/sysdeps/unix/sysv/linux/bits/types/siginfo_t.h

Structs

__SI_CLOCK_T 🔒

__c_anonymous__bounds__addr_bnd 🔒

__c_anonymous__si_fields__kill 🔒

__c_anonymous__si_fields__rt 🔒

__c_anonymous__si_fields__sigchld 🔒

__c_anonymous__si_fields__sigfault 🔒

__c_anonymous__si_fields__sigpoll 🔒

__c_anonymous__si_fields__sigsys 🔒

__c_anonymous__si_fields__timer 🔒

siginfo_t

Constants

__SI_MAX_SIZE 🔒

__SI_PAD_SIZE 🔒

Type Aliases

__SI_BAND_TYPE 🔒

Unions

__c_anonymous__sigfault__bounds 🔒

__c_anonymous_siginfo_t__si_fields 🔒