libc

Module unix

Definitions found commonly among almost all Unix derivatives

More functions and definitions can be found in the more specific modules according to the platform in question.

Re-exports

pub use self::linux_like::*;

Modules

linux_like 🔒

Structs

DIR

This is an extern type (“opaque” or “incomplete” type in C).

FILE

This is an extern type (“opaque” or “incomplete” type in C).

group

hostent

in6_addr

iovec

ipv6_mreq

itimerval

linger

pollfd

protoent

rlimit

rusage

servent

sigval

timeval

tms

utimbuf

winsize

Constants

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

FD_CLOEXEC

FNM_CASEFOLD

FNM_NOESCAPE

FNM_NOMATCH

FNM_PATHNAME

FNM_PERIOD

GRPQUOTA

IFNAMSIZ

IF_NAMESIZE

IN6ADDR_ANY_INIT

IN6ADDR_LOOPBACK_INIT

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INT_MAX

INT_MIN

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_TCP

IPPROTO_UDP

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

PRIO_MAX

PRIO_MIN

SIGIOT

SIG_DFL

SIG_ERR

SIG_IGN

S_ISGID

S_ISUID

S_ISVTX

USRQUOTA

Statics

in6addr_any⚠

in6addr_loopback⚠

Functions

_exit⚠

abort⚠

accept⚠

access⚠

adjtime⚠

alarm⚠

aligned_alloc⚠

atexit⚠

atof⚠

atoi⚠

atol⚠

atoll⚠

bsearch⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clearerr⚠

close⚠

closedir⚠

closelog⚠

confstr⚠

connect⚠

creat⚠

difftime⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

endservent⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

exit⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgets⚠

fileno⚠

flock⚠

fmemopen⚠

fnmatch⚠

fopen⚠

fork⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fstat⚠

fstatat⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftruncate⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgroups⚠

gethostname⚠

getline⚠

getlogin⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getprotobyname⚠

getprotobynumber⚠

getpwnam⚠

getpwuid⚠

getrusage⚠

getservbyname⚠

getservbyport⚠

getservent⚠

getsid⚠

getsockname⚠

getsockopt⚠

getuid⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

hstrerror⚠

htonl

htons

if_indextoname⚠

if_nametoindex⚠

isalnum⚠

isalpha⚠

isatty⚠

isblank⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kill⚠

killpg⚠

lchown⚠

link⚠

linkat⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lockf⚠

lseek⚠

lstat⚠

malloc⚠

memccpy⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memset⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mknod⚠

mkstemp⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

nice⚠

ntohl

ntohs

open⚠

open_memstream⚠

open_wmemstream⚠

openat⚠

opendir⚠

openlog⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

poll⚠

posix_memalign⚠

posix_openpt⚠

pread⚠

printf⚠

pselect⚠

pthread_attr_destroy⚠

pthread_attr_getstacksize⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_init⚠

pthread_detach⚠

pthread_equal⚠

pthread_exit⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_init⚠

pthread_self⚠

pthread_setspecific⚠

ptsname⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pwrite⚠

qsort⚠

raise⚠

read⚠

readdir⚠

readdir_r⚠

The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos:

readlink⚠

readlinkat⚠

realloc⚠

realpath⚠

recv⚠

remove⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

scanf⚠

sched_yield⚠

select⚠

sem_post⚠

sem_trywait⚠

sem_wait⚠

send⚠

sendto⚠

setbuf⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setregid⚠

setreuid⚠

setservent⚠

setsid⚠

setsockopt⚠

setuid⚠

setvbuf⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sigqueue⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

sscanf⚠

stat⚠

statvfs⚠

stpcpy⚠

stpncpy⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsignal⚠

strspn⚠

strstr⚠

strtod⚠

strtof⚠

strtok⚠

strtok_r⚠

strtol⚠

strtoll⚠

strtoul⚠

strtoull⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sysconf⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

usleep⚠

utime⚠

utimes⚠

wait⚠

waitpid⚠

wcslen⚠

wcstombs⚠

wmemchr⚠

write⚠

Type Aliases

cc_t

gid_t

in_addr_t

in_port_t

intmax_t

intptr_t

locale_t

pid_t

ptrdiff_t

sighandler_t

size_t

ssize_t

uid_t

uintmax_t

uintptr_t