Crate libc

libc - Raw FFI bindings to platforms’ system libraries

Structs

Dl_info

addrinfo

cmsghdr

cpu_set_t

dirent

fd_set

group

hostent

in6_addr

in_addr

iovec

ip_mreq

ipv6_mreq

itimerval

lconv

linger

msghdr

passwd

pollfd

protoent

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

rlimit

rusage

sem_t

servent

sigaction

sigset_t

sigval

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_storage

sockaddr_un

stack_t

stat

statvfs

termios

timespec

timeval

tm

tms

utimbuf

utsname

winsize

Enums

DIR

FILE

c_void

Equivalent to C’s void type when used as a [pointer].

fpos_t

Constants

AF_INET

AF_INET6

AF_UNIX

AF_UNSPEC

AI_ADDRCONFIG

AI_CANONNAME

AI_NUMERICHOST

AI_NUMERICSERV

AI_PASSIVE

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

CLOCK_BOOTTIME

CLOCK_MONOTONIC

CLOCK_REALTIME

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EAFNOSUPPORT

EAGAIN

EAI_FAMILY

EAI_MEMORY

EAI_NONAME

EAI_SOCKTYPE

EALREADY

EBADF

EBADMSG

EBUSY

ECANCELED

ECHILD

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDESTADDRREQ

EDOM

EDQUOT

EEXIST

EFAULT

EFBIG

EFTYPE

EHOSTDOWN

EHOSTUNREACH

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

ELOOP

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOBUFS

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOLCK

ENOLINK

ENOMEM

ENOMSG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENXIO

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERANGE

EROFS

ESPIPE

ESRCH

ESTALE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EWOULDBLOCK

EXDEV

EXIT_FAILURE

EXIT_SUCCESS

FD_CLOEXEC

FD_SETSIZE

FIOCLEX

FIONBIO

FIONCLEX

FNM_CASEFOLD

FNM_NOESCAPE

FNM_NOMATCH

FNM_PATHNAME

FNM_PERIOD

F_CNVT

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLK

F_GETOWN

F_RGETLK

F_RSETLK

F_RSETLKW

F_SETFD

F_SETFL

F_SETLK

F_SETLKW

F_SETOWN

GRPQUOTA

HOST_NOT_FOUND

IFF_ALLMULTI

IFF_ALTPHYS

IFF_BROADCAST

IFF_DEBUG

IFF_LINK0

IFF_LINK1

IFF_LINK2

IFF_LOOPBACK

IFF_MULTICAST

IFF_NOARP

IFF_NOTRAILERS

IFF_OACTIVE

IFF_POINTOPOINT

IFF_PROMISC

IFF_RUNNING

IFF_SIMPLEX

IFF_UP

IFNAMSIZ

IF_NAMESIZE

IN6ADDR_ANY_INIT

IN6ADDR_LOOPBACK_INIT

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INET_ADDRSTRLEN

INT_MAX

INT_MIN

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_TCP

IPPROTO_UDP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IPV6_JOIN_GROUP

IPV6_LEAVE_GROUP

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_LOOP

IPV6_UNICAST_HOPS

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TOS

IP_TTL

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOR

MSG_MORE

MSG_NOSIGNAL

MSG_OOB

MSG_PEEK

MSG_TRUNC

MSG_WAITALL

NCCS

NI_DGRAM

NI_MAXHOST

NI_MAXSERV

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSERV

NO_ADDRESS

NO_DATA

NO_RECOVERY

NSIG

O_ACCMODE

O_APPEND

O_CLOEXEC

O_CREAT

O_EXCL

O_NONBLOCK

O_RDONLY

O_RDWR

O_SYNC

O_TRUNC

O_WRONLY

PF_INET

PF_INET6

PF_UNSPEC

POLLERR

POLLHUP

POLLIN

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

POLLWRBAND

POLLWRNORM

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PTHREAD_COND_INITIALIZER

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

RTLD_LAZY

SEEK_CUR

SEEK_END

SEEK_SET

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIOT

SIGQUIT

SIGSEGV

SIGTERM

SIG_DFL

SIG_ERR

SIG_IGN

SOCK_CLOEXEC

SOCK_DGRAM

SOCK_STREAM

SOL_SOCKET

SOL_TCP

SO_ACCEPTCONN

SO_ACCEPTFILTER

SO_BINTIME

SO_BROADCAST

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LABEL

SO_LINGER

SO_LISTENINCQLEN

SO_LISTENQLEN

SO_LISTENQLIMIT

SO_NOSIGPIPE

SO_NO_DDP

SO_NO_OFFLOAD

SO_OOBINLINE

SO_PEERLABEL

SO_PROTOCOL

SO_PROTOTYPE

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_REUSEPORT_LB

SO_SETFIB

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TYPE

SO_USELOOPBACK

SO_USER_COOKIE

SO_VENDOR

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

S_BLKSIZE

S_ENFMT

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_MAXSEG

TCP_NODELAY

TCP_NOOPT

TCP_NOPUSH

TRY_AGAIN

USRQUOTA

__PTHREAD_MUTEX_HAVE_PREV

__PTHREAD_RWLOCK_INT_FLAGS_SHARED

__SIZEOF_PTHREAD_ATTR_T

__SIZEOF_PTHREAD_BARRIERATTR_T

__SIZEOF_PTHREAD_BARRIER_T

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

Statics

in6addr_any⚠

in6addr_loopback⚠

Functions

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

_exit⚠

abort⚠

abs⚠

accept⚠

access⚠

adjtime⚠

alarm⚠

aligned_alloc⚠

atexit⚠

atof⚠

atoi⚠

atol⚠

atoll⚠

bind⚠

bsearch⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clearerr⚠

clock_getres⚠

clock_gettime⚠

clock_settime⚠

close⚠

closedir⚠

closelog⚠

closesocket⚠

confstr⚠

connect⚠

creat⚠

daemon⚠

difftime⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

endservent⚠

eventfd⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

exit⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fexecve⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgets⚠

fileno⚠

flock⚠

fmemopen⚠

fnmatch⚠

fopen⚠

fork⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fstat⚠

fstatat⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftruncate⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getdtablesize⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgroups⚠

gethostname⚠

getline⚠

getlogin⚠

getnameinfo⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getprotobyname⚠

getprotobynumber⚠

getpwnam⚠

getpwnam_r⚠

getpwuid⚠

getpwuid_r⚠

getrandom⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getservbyport⚠

getservent⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

hstrerror⚠

htonl

htons

if_indextoname⚠

if_nametoindex⚠

ioctl⚠

isalnum⚠

isalpha⚠

isatty⚠

isblank⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kill⚠

killpg⚠

labs⚠

lchown⚠

link⚠

linkat⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lockf⚠

lseek⚠

lstat⚠

malloc⚠

memalign⚠

memccpy⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memset⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mknod⚠

mkstemp⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

nice⚠

ntohl

ntohs

open⚠

open_memstream⚠

open_wmemstream⚠

openat⚠

opendir⚠

openlog⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

poll⚠

popen⚠

posix_memalign⚠

posix_openpt⚠

pread⚠

printf⚠

pselect⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_getstacksize⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_init⚠

pthread_create⚠

pthread_detach⚠

pthread_equal⚠

pthread_exit⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_init⚠

pthread_self⚠

pthread_setspecific⚠

pthread_sigmask⚠

ptsname⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pwrite⚠

qsort⚠

raise⚠

rand⚠

read⚠

readdir⚠

readdir_r⚠

The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

realloc⚠

realpath⚠

recv⚠

recvfrom⚠

recvmsg⚠

remove⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

scanf⚠

sched_yield⚠

select⚠

sem_close⚠

sem_destroy⚠

sem_init⚠

sem_open⚠

sem_post⚠

sem_trywait⚠

sem_unlink⚠

sem_wait⚠

send⚠

sendmsg⚠

sendto⚠

setbuf⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setregid⚠

setreuid⚠

setrlimit⚠

setservent⚠

setsid⚠

setsockopt⚠

setuid⚠

setvbuf⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sigwait⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

srand⚠

sscanf⚠

stat⚠

statvfs⚠

stpcpy⚠

stpncpy⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsignal⚠

strspn⚠

strstr⚠

strtod⚠

strtof⚠

strtok⚠

strtok_r⚠

strtol⚠

strtoll⚠

strtoul⚠

strtoull⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sysconf⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

usleep⚠

utime⚠

utimes⚠

wait⚠

waitpid⚠

wcslen⚠

wcstombs⚠

wmemchr⚠

write⚠

Type Aliases

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

dev_t

fsblkcnt_t

fsfilcnt_t

gid_t

id_t

in_addr_t

in_port_t

ino_t

int8_tDeprecated

int16_tDeprecated

int32_tDeprecated

int64_tDeprecated

intmax_t

intptr_t

key_t

locale_t

loff_t

mode_t

nfds_t

nlink_t

off_t

pid_t

pthread_key_t

pthread_t

ptrdiff_t

rlim_t

sa_family_t

sighandler_t

size_t

socklen_t

speed_t

ssize_t

suseconds_t

tcflag_t

time_t

uid_t

uint8_tDeprecated

uint16_tDeprecated

uint32_tDeprecated

uint64_tDeprecated

uintmax_t

uintptr_t

useconds_t

wchar_t