Crate libc

libc - Raw FFI bindings to platforms' system libraries

Documentation for other platforms.

Structs

Dl_info

Elf32_Ehdr

Elf32_Sym

Elf32_Phdr

Elf32_Shdr

Elf32_Chdr

Elf64_Ehdr

Elf64_Sym

Elf64_Phdr

Elf64_Shdr

Elf64_Chdr

__exit_status

__timeval

_libc_fpstate

_libc_fpxreg

_libc_xmmreg

addrinfo

af_alg_iv

aiocb

arpd_request

arphdr

arpreq

arpreq_old

cmsghdr

cpu_set_t

dirent

dirent64

dl_phdr_info

dqblk

epoll_event

fd_set

ff_condition_effect

ff_constant_effect

ff_effect

ff_envelope

ff_periodic_effect

ff_ramp_effect

ff_replay

ff_rumble_effect

ff_trigger

flock

fsid_t

genlmsghdr

glob64_t

glob_t

group

hostent

if_nameindex

ifaddrs

in6_rtmsg

in6_pktinfo

in6_addr

in_addr

in_pktinfo

inotify_event

input_absinfo

input_event

input_id

input_keymap_entry

input_mask

iovec

ip_mreq

ipc_perm

ipv6_mreq

itimerspec

itimerval

lconv

linger

mallinfo

mcontext_t

mmsghdr

mntent

mq_attr

msghdr

msginfo

msqid_ds

nl_mmap_hdr

nl_mmap_req

nl_pktinfo

nlattr

nlmsgerr

nlmsghdr

packet_mreq

passwd

pollfd

posix_spawn_file_actions_t

posix_spawnattr_t

protoent

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

rlimit

rlimit64

rtentry

rusage

sched_param

sem_t

sembuf

servent

shmid_ds

sigaction

sigevent

siginfo_t

signalfd_siginfo

sigset_t

sigval

sockaddr

sockaddr_alg

sockaddr_in

sockaddr_in6

sockaddr_ll

sockaddr_nl

sockaddr_storage

sockaddr_un

spwd

stack_t

stat

stat64

statfs

statfs64

statvfs

statvfs64

statx

statx_timestamp

sysinfo

termios

termios2

timespec

timeval

tm

tms

ucontext_t

ucred

user

user_fpregs_struct

user_regs_struct

utimbuf

utmpx

utsname

winsize

Enums

DIR

FILE

c_void
Equivalent to C's void type when used as a pointer.

fpos64_t

fpos_t

timezone

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

ACCOUNTING

ADFS_SUPER_MAGIC

AFFS_SUPER_MAGIC

AF_ALG

AF_APPLETALK

AF_ASH

AF_ATMPVC

AF_ATMSVC

AF_AX25

AF_BLUETOOTH

AF_BRIDGE

AF_CAIF

AF_CAN

AF_DECnet

AF_ECONET

AF_IB

AF_IEEE802154

AF_INET

AF_INET6

AF_IPX

AF_IRDA

AF_ISDN

AF_IUCV

AF_KEY

AF_LLC

AF_LOCAL

AF_MPLS

AF_NETBEUI

AF_NETLINK

AF_NETROM

AF_NFC

AF_PACKET

AF_PHONET

AF_PPPOX

AF_RDS

AF_ROSE

AF_ROUTE

AF_RXRPC

AF_SECURITY

AF_SNA

AF_TIPC

AF_UNIX

AF_UNSPEC

AF_VSOCK

AF_WANPIPE

AF_X25

AF_XDP

AIO_ALLDONE

AIO_CANCELED

AIO_NOTCANCELED

AI_ADDRCONFIG

AI_ALL

AI_CANONNAME

AI_NUMERICHOST

AI_NUMERICSERV

AI_PASSIVE

AI_V4MAPPED

ALG_OP_DECRYPT

ALG_OP_ENCRYPT

ALG_SET_AEAD_ASSOCLEN

ALG_SET_AEAD_AUTHSIZE

ALG_SET_IV

ALG_SET_KEY

ALG_SET_OP

ALT_DIGITS

AM_STR

ARPD_FLUSH

ARPD_LOOKUP

ARPD_UPDATE

ARPHRD_ADAPT

ARPHRD_APPLETLK

ARPHRD_ARCNET

ARPHRD_ASH

ARPHRD_ATM

ARPHRD_AX25

ARPHRD_BIF

ARPHRD_CHAOS

ARPHRD_CISCO

ARPHRD_CSLIP

ARPHRD_CSLIP6

ARPHRD_DDCMP

ARPHRD_DLCI

ARPHRD_ECONET

ARPHRD_EETHER

ARPHRD_ETHER

ARPHRD_EUI64

ARPHRD_FCAL

ARPHRD_FCFABRIC

ARPHRD_FCPL

ARPHRD_FCPP

ARPHRD_FDDI

ARPHRD_FRAD

ARPHRD_HDLC

ARPHRD_HIPPI

ARPHRD_HWX25

ARPHRD_IEEE802

ARPHRD_IEEE802_TR

ARPHRD_IEEE1394

ARPHRD_IEEE80211

ARPHRD_IEEE80211_PRISM

ARPHRD_IEEE80211_RADIOTAP

ARPHRD_IEEE802154

ARPHRD_INFINIBAND

ARPHRD_IPDDP

ARPHRD_IPGRE

ARPHRD_IRDA

ARPHRD_LAPB

ARPHRD_LOCALTLK

ARPHRD_LOOPBACK

ARPHRD_METRICOM

ARPHRD_NETROM

ARPHRD_NONE

ARPHRD_PIMREG

ARPHRD_PPP

ARPHRD_PRONET

ARPHRD_RAWHDLC

ARPHRD_ROSE

ARPHRD_RSRVD

ARPHRD_SIT

ARPHRD_SKIP

ARPHRD_SLIP

ARPHRD_SLIP6

ARPHRD_TUNNEL

ARPHRD_TUNNEL6

ARPHRD_VOID

ARPHRD_X25

ARPOP_InREPLY

ARPOP_InREQUEST

ARPOP_NAK

ARPOP_REPLY

ARPOP_REQUEST

ARPOP_RREPLY

ARPOP_RREQUEST

ATF_COM

ATF_DONTPUB

ATF_MAGIC

ATF_NETMASK

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

AT_EMPTY_PATH

AT_FDCWD

AT_NO_AUTOMOUNT

AT_REMOVEDIR

AT_STATX_DONT_SYNC

AT_STATX_FORCE_SYNC

AT_STATX_SYNC_AS_STAT

AT_STATX_SYNC_TYPE

AT_SYMLINK_FOLLOW

AT_SYMLINK_NOFOLLOW

B0

B50

B75

B110

B134

B150

B200

B300

B600

B1200

B1800

B2400

B4800

B9600

B19200

B38400

B57600

B115200

B230400

B460800

B500000

B576000

B921600

B1000000

B1152000

B1500000

B2000000

B2500000

B3000000

B3500000

B4000000

BOOT_TIME

BOTHER

BRKINT

BS0

BS1

BSDLY

BUFSIZ

CBAUD

CBAUDEX

CIBAUD

CLOCAL

CLOCK_BOOTTIME

CLOCK_BOOTTIME_ALARM

CLOCK_MONOTONIC

CLOCK_MONOTONIC_COARSE

CLOCK_MONOTONIC_RAW

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_REALTIME_ALARM

CLOCK_REALTIME_COARSE

CLOCK_THREAD_CPUTIME_ID

CLONE_CHILD_CLEARTID

CLONE_CHILD_SETTID

CLONE_DETACHED

CLONE_FILES

CLONE_FS

CLONE_IO

CLONE_NEWCGROUP

CLONE_NEWIPC

CLONE_NEWNET

CLONE_NEWNS

CLONE_NEWPID

CLONE_NEWUSER

CLONE_NEWUTS

CLONE_PARENT

CLONE_PARENT_SETTID

CLONE_PTRACE

CLONE_SETTLS

CLONE_SIGHAND

CLONE_SYSVSEM

CLONE_THREAD

CLONE_UNTRACED

CLONE_VFORK

CLONE_VM

CMSPAR

CODA_SUPER_MAGIC

CODESET

CPU_SETSIZE

CR0

CR1

CR2

CR3

CRAMFS_MAGIC

CRDLY

CREAD

CRNCYSTR

CRTSCTS

CS

CS5

CS6

CS7

CS8

CSIZE

CSTOPB

CTRL_ATTR_FAMILY_ID

CTRL_ATTR_FAMILY_NAME

CTRL_ATTR_HDRSIZE

CTRL_ATTR_MAXATTR

CTRL_ATTR_MCAST_GROUPS

CTRL_ATTR_MCAST_GRP_ID

CTRL_ATTR_MCAST_GRP_NAME

CTRL_ATTR_MCAST_GRP_UNSPEC

CTRL_ATTR_OPS

CTRL_ATTR_OP_FLAGS

CTRL_ATTR_OP_ID

CTRL_ATTR_OP_UNSPEC

CTRL_ATTR_UNSPEC

CTRL_ATTR_VERSION

CTRL_CMD_DELFAMILY

CTRL_CMD_DELMCAST_GRP

CTRL_CMD_DELOPS

CTRL_CMD_GETFAMILY

CTRL_CMD_GETMCAST_GRP

CTRL_CMD_GETOPS

CTRL_CMD_NEWFAMILY

CTRL_CMD_NEWMCAST_GRP

CTRL_CMD_NEWOPS

CTRL_CMD_UNSPEC

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

DCCP_SERVICE_LIST_MAX_LEN
maximum number of services provided on the same listening port

DCCP_SOCKOPT_AVAILABLE_CCIDS

DCCP_SOCKOPT_CCID

DCCP_SOCKOPT_CCID_RX_INFO

DCCP_SOCKOPT_CCID_TX_INFO

DCCP_SOCKOPT_CHANGE_L

DCCP_SOCKOPT_CHANGE_R

DCCP_SOCKOPT_GET_CUR_MPS

DCCP_SOCKOPT_PACKET_SIZE

DCCP_SOCKOPT_QPOLICY_ID

DCCP_SOCKOPT_QPOLICY_TXQLEN

DCCP_SOCKOPT_RECV_CSCOV

DCCP_SOCKOPT_RX_CCID

DCCP_SOCKOPT_SEND_CSCOV

DCCP_SOCKOPT_SERVER_TIMEWAIT

DCCP_SOCKOPT_SERVICE

DCCP_SOCKOPT_TX_CCID

DEAD_PROCESS

DS

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

D_FMT

D_T_FMT

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADV

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NODATA

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDOTDOT

EDQUOT

EEXIST

EFAULT

EFBIG

EFD_CLOEXEC

EFD_NONBLOCK

EFD_SEMAPHORE

EFLAGS

EFS_SUPER_MAGIC

EHOSTDOWN

EHOSTUNREACH

EHWPOISON

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EISNAM

EKEYEXPIRED

EKEYREJECTED

EKEYREVOKED

EL2NSYNC

EL2HLT

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOOP

EMEDIUMTYPE

EMFILE

EMLINK

EMPTY

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENAVAIL

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOATTR	Deprecated

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOKEY

ENOLCK

ENOLINK

ENOMEDIUM

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTNAM

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENOTUNIQ

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPOLLERR

EPOLLET

EPOLLEXCLUSIVE

EPOLLHUP

EPOLLIN

EPOLLMSG

EPOLLONESHOT

EPOLLOUT

EPOLLPRI

EPOLLRDBAND

EPOLLRDHUP

EPOLLRDNORM

EPOLLWAKEUP

EPOLLWRBAND

EPOLLWRNORM

EPOLL_CLOEXEC

EPOLL_CTL_ADD

EPOLL_CTL_DEL

EPOLL_CTL_MOD

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EREMCHG

EREMOTE

EREMOTEIO

ERESTART

ERFKILL

EROFS

ES

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETH_ALEN

ETH_DATA_LEN

ETH_FCS_LEN

ETH_FRAME_LEN

ETH_HLEN

ETH_P_1588

ETH_P_8021Q

ETH_P_8021AD

ETH_P_8021AH

ETH_P_80221

ETH_P_802_2

ETH_P_802_3_MIN

ETH_P_802_3

ETH_P_802_EX1

ETH_P_AARP

ETH_P_AF_IUCV

ETH_P_ALL

ETH_P_AOE

ETH_P_ARCNET

ETH_P_ARP

ETH_P_ATALK

ETH_P_ATMFATE

ETH_P_ATMMPOA

ETH_P_AX25

ETH_P_BATMAN

ETH_P_BPQ

ETH_P_CAIF

ETH_P_CANFD

ETH_P_CONTROL

ETH_P_CUST

ETH_P_DDCMP

ETH_P_DEC

ETH_P_DIAG

ETH_P_DNA_DL

ETH_P_DNA_RC

ETH_P_DNA_RT

ETH_P_DSA

ETH_P_ECONET

ETH_P_EDSA

ETH_P_FCOE

ETH_P_FIP

ETH_P_HDLC

ETH_P_IEEE802154

ETH_P_IEEEPUP

ETH_P_IEEEPUPAT

ETH_P_IP

ETH_P_IPV6

ETH_P_IPX

ETH_P_IRDA

ETH_P_LAT

ETH_P_LINK_CTL

ETH_P_LOCALTALK

ETH_P_LOOP

ETH_P_LOOPBACK

ETH_P_MACSEC

ETH_P_MOBITEX

ETH_P_MPLS_MC

ETH_P_MPLS_UC

ETH_P_MVRP

ETH_P_PAE

ETH_P_PAUSE

ETH_P_PHONET

ETH_P_PPPTALK

ETH_P_PPP_DISC

ETH_P_PPP_MP

ETH_P_PPP_SES

ETH_P_PRP

ETH_P_PUP

ETH_P_PUPAT

ETH_P_QINQ1

ETH_P_QINQ2

ETH_P_QINQ3

ETH_P_RARP

ETH_P_SCA

ETH_P_SLOW

ETH_P_SNAP

ETH_P_TDLS

ETH_P_TEB

ETH_P_TIPC

ETH_P_TRAILER

ETH_P_TR_802_2

ETH_P_WAN_PPP

ETH_P_WCCP

ETH_P_X25

ETH_ZLEN

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUCLEAN

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

EXT2_SUPER_MAGIC

EXT3_SUPER_MAGIC

EXT4_SUPER_MAGIC

EXTA

EXTB

EXTPROC

FALLOC_FL_COLLAPSE_RANGE

FALLOC_FL_INSERT_RANGE

FALLOC_FL_KEEP_SIZE

FALLOC_FL_PUNCH_HOLE

FALLOC_FL_UNSHARE_RANGE

FALLOC_FL_ZERO_RANGE

FD_CLOEXEC

FD_SETSIZE

FF0

FF1

FFDLY

FILENAME_MAX

FIOCLEX

FIONBIO

FIONREAD

FLUSHO

FOPEN_MAX

FS

FS_BASE

FUTEX_CLOCK_REALTIME

FUTEX_CMD_MASK

FUTEX_CMP_REQUEUE

FUTEX_CMP_REQUEUE_PI

FUTEX_FD

FUTEX_LOCK_PI

FUTEX_PRIVATE_FLAG

FUTEX_REQUEUE

FUTEX_TRYLOCK_PI

FUTEX_UNLOCK_PI

FUTEX_WAIT

FUTEX_WAIT_BITSET

FUTEX_WAIT_REQUEUE_PI

FUTEX_WAKE

FUTEX_WAKE_BITSET

FUTEX_WAKE_OP

F_ADD_SEALS

F_CANCELLK

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLEASE

F_GETLK

F_GETOWN

F_GETPIPE_SZ

F_GET_SEALS

F_LOCK

F_NOTIFY

F_OK

F_RDLCK

F_SEAL_FUTURE_WRITE

F_SEAL_GROW

F_SEAL_SEAL

F_SEAL_SHRINK

F_SEAL_WRITE

F_SETFD

F_SETFL

F_SETLEASE

F_SETLK

F_SETLKW

F_SETOWN

F_SETPIPE_SZ

F_TEST

F_TLOCK

F_ULOCK

F_UNLCK

F_WRLCK

GENL_ADMIN_PERM

GENL_CMD_CAP_DO

GENL_CMD_CAP_DUMP

GENL_CMD_CAP_HASPOL

GENL_ID_CTRL

GENL_ID_PMCRAID

GENL_ID_VFS_DQUOT

GENL_MAX_ID

GENL_MIN_ID

GENL_NAMSIZ

GENL_UNS_ADMIN_PERM

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRND_NONBLOCK

GRND_RANDOM

GRPQUOTA

GS

GS_BASE

HPFS_SUPER_MAGIC

HUGETLBFS_MAGIC

HUPCL

ICANON

ICRNL

IEXTEN

IFA_ADDRESS

IFA_ANYCAST

IFA_BROADCAST

IFA_CACHEINFO

IFA_FLAGS

IFA_F_DADFAILED

IFA_F_DEPRECATED

IFA_F_HOMEADDRESS

IFA_F_MANAGETEMPADDR

IFA_F_MCAUTOJOIN

IFA_F_NODAD

IFA_F_NOPREFIXROUTE

IFA_F_OPTIMISTIC

IFA_F_PERMANENT

IFA_F_SECONDARY

IFA_F_STABLE_PRIVACY

IFA_F_TEMPORARY

IFA_F_TENTATIVE

IFA_LABEL

IFA_LOCAL

IFA_MULTICAST

IFA_UNSPEC

IFF_ALLMULTI

IFF_ATTACH_QUEUE

IFF_AUTOMEDIA

IFF_BROADCAST

IFF_DEBUG

IFF_DETACH_QUEUE

IFF_DORMANT

IFF_DYNAMIC

IFF_ECHO

IFF_LOOPBACK

IFF_LOWER_UP

IFF_MASTER

IFF_MULTICAST

IFF_MULTI_QUEUE

IFF_NOARP

IFF_NOFILTER

IFF_NOTRAILERS

IFF_NO_PI

IFF_ONE_QUEUE

IFF_PERSIST

IFF_POINTOPOINT

IFF_PORTSEL

IFF_PROMISC

IFF_RUNNING

IFF_SLAVE

IFF_TAP

IFF_TUN

IFF_TUN_EXCL

IFF_UP

IFF_VNET_HDR

IFLA_ADDRESS

IFLA_BROADCAST

IFLA_IFNAME

IFLA_LINK

IFLA_MTU

IFLA_QDISC

IFLA_STATS

IFLA_UNSPEC

IFNAMSIZ

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

IMAXBEL

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INIT_PROCESS

INLCR

INPCK

INT_MAX

INT_MIN

IN_ACCESS

IN_ALL_EVENTS

IN_ATTRIB

IN_CLOEXEC

IN_CLOSE

IN_CLOSE_NOWRITE

IN_CLOSE_WRITE

IN_CREATE

IN_DELETE

IN_DELETE_SELF

IN_DONT_FOLLOW

IN_IGNORED

IN_ISDIR

IN_MODIFY

IN_MOVE

IN_MOVED_FROM

IN_MOVED_TO

IN_MOVE_SELF

IN_NONBLOCK

IN_ONESHOT

IN_ONLYDIR

IN_OPEN

IN_Q_OVERFLOW

IN_UNMOUNT

IPC_CREAT

IPC_EXCL

IPC_INFO

IPC_NOWAIT

IPC_PRIVATE

IPC_RMID

IPC_SET

IPC_STAT

IPDEFTTL

IPOPT_CLASS_MASK

IPOPT_CONTROL

IPOPT_COPY

IPOPT_END

IPOPT_EOL

IPOPT_LSRR

IPOPT_MEASUREMENT

IPOPT_MINOFF

IPOPT_NOOP

IPOPT_NOP

IPOPT_NUMBER_MASK

IPOPT_OFFSET

IPOPT_OLEN

IPOPT_OPTVAL

IPOPT_RA

IPOPT_RESERVED1

IPOPT_RESERVED2

IPOPT_RR

IPOPT_SEC

IPOPT_SID

IPOPT_SSRR

IPOPT_TIMESTAMP

IPOPT_TS

IPOPT_TS_PRESPEC

IPOPT_TS_TSANDADDR

IPOPT_TS_TSONLY

IPPROTO_AH
IP6 Auth Header

IPPROTO_BEETPH

IPPROTO_COMP
IP Payload Comp. Protocol

IPPROTO_DCCP
DCCP

IPPROTO_DSTOPTS
IP6 destination option

IPPROTO_EGP
exterior gateway protocol

IPPROTO_ENCAP
encapsulation header

IPPROTO_ESP
IP6 Encap Sec. Payload

IPPROTO_FRAGMENT
IP6 fragmentation header

IPPROTO_GRE
General Routing Encap.

IPPROTO_HOPOPTS
Hop-by-hop option header

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IDP
xns idp

IPPROTO_IGMP
group mgmt protocol

IPPROTO_IP

IPPROTO_IPIP
for compatibility

IPPROTO_IPV6

IPPROTO_MAX

IPPROTO_MH

IPPROTO_MPLS

IPPROTO_MTP

IPPROTO_NONE
IP6 no next header

IPPROTO_PIM
Protocol indep. multicast

IPPROTO_PUP
pup

IPPROTO_RAW
raw IP packet

IPPROTO_ROUTING
IP6 routing header

IPPROTO_RSVP
resource reservation

IPPROTO_SCTP
SCTP

IPPROTO_TCP

IPPROTO_TP
tp-4 w/ class negotiation

IPPROTO_UDP

IPPROTO_UDPLITE

IPTOS_ECN_CE

IPTOS_ECN_ECT0

IPTOS_ECN_ECT1

IPTOS_ECN_MASK

IPTOS_ECN_NOT_ECT

IPTOS_LOWDELAY

IPTOS_MINCOST

IPTOS_PREC_CRITIC_ECP

IPTOS_PREC_FLASH

IPTOS_PREC_FLASHOVERRIDE

IPTOS_PREC_IMMEDIATE

IPTOS_PREC_INTERNETCONTROL

IPTOS_PREC_MASK

IPTOS_PREC_NETCONTROL

IPTOS_PREC_PRIORITY

IPTOS_PREC_ROUTINE

IPTOS_RELIABILITY

IPTOS_THROUGHPUT

IPTOS_TOS_MASK

IPV6_UNICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_LOOP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IPV6_RECVPKTINFO

IPV6_PKTINFO

IPV6_RECVTCLASS

IPV6_TCLASS

IPV6_ORIGDSTADDR

IPV6_RECVORIGDSTADDR

IPV6_FLOWINFO

IPV6_FLOWLABEL_MGR

IPV6_FLOWINFO_SEND

IPV6_FLOWINFO_FLOWLABEL

IPV6_FLOWINFO_PRIORITY

IPV6_V6ONLY

IPVERSION

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_ORIGDSTADDR

IP_PKTINFO

IP_RECVORIGDSTADDR

IP_RECVTOS

IP_TOS

IP_TRANSPARENT

IP_TTL

ISIG

ISOFS_SUPER_MAGIC

ISTRIP

ITIMER_PROF

ITIMER_REAL

ITIMER_VIRTUAL

IUTF8

IXANY

IXOFF

IXON

JFFS2_SUPER_MAGIC

LC_ADDRESS

LC_ADDRESS_MASK

LC_ALL

LC_ALL_MASK

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_IDENTIFICATION

LC_IDENTIFICATION_MASK

LC_MEASUREMENT

LC_MEASUREMENT_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NAME

LC_NAME_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_PAPER

LC_PAPER_MASK

LC_TELEPHONE

LC_TELEPHONE_MASK

LC_TIME

LC_TIME_MASK

LINUX_REBOOT_CMD_CAD_OFF

LINUX_REBOOT_CMD_CAD_ON

LINUX_REBOOT_CMD_HALT

LINUX_REBOOT_CMD_KEXEC

LINUX_REBOOT_CMD_POWER_OFF

LINUX_REBOOT_CMD_RESTART

LINUX_REBOOT_CMD_RESTART2

LINUX_REBOOT_CMD_SW_SUSPEND

LINUX_REBOOT_MAGIC1

LINUX_REBOOT_MAGIC2

LINUX_REBOOT_MAGIC2A

LINUX_REBOOT_MAGIC2B

LINUX_REBOOT_MAGIC2C

LIO_NOP

LIO_NOWAIT

LIO_READ

LIO_WAIT

LIO_WRITE

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

LOGIN_PROCESS

LOG_ALERT

LOG_AUTH

LOG_AUTHPRIV

LOG_CONS

LOG_CRIT

LOG_CRON

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_FTP

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NFACILITIES

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PERROR

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

L_tmpnam

MADV_DODUMP

MADV_DOFORK

MADV_DONTDUMP

MADV_DONTFORK

MADV_DONTNEED

MADV_FREE

MADV_HUGEPAGE

MADV_HWPOISON

MADV_MERGEABLE

MADV_NOHUGEPAGE

MADV_NORMAL

MADV_RANDOM

MADV_REMOVE

MADV_SEQUENTIAL

MADV_SOFT_OFFLINE

MADV_UNMERGEABLE

MADV_WILLNEED

MAP_32BIT

MAP_ANON

MAP_ANONYMOUS

MAP_DENYWRITE

MAP_EXECUTABLE

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_FIXED_NOREPLACE

MAP_GROWSDOWN

MAP_HUGETLB

MAP_LOCKED

MAP_NONBLOCK

MAP_NORESERVE

MAP_POPULATE

MAP_PRIVATE

MAP_SHARED

MAP_SHARED_VALIDATE

MAP_STACK

MAP_TYPE

MAXTTL

MAX_ADDR_LEN

MAX_IPOPTLEN

MAX_LINKS

MCL_CURRENT

MCL_FUTURE

MFD_ALLOW_SEALING

MFD_CLOEXEC

MFD_HUGETLB

MINIX2_SUPER_MAGIC

MINIX2_SUPER_MAGIC2

MINIX_SUPER_MAGIC

MINIX_SUPER_MAGIC2

MINSIGSTKSZ

MNT_DETACH

MNT_EXPIRE

MNT_FORCE

MODULE_INIT_IGNORE_MODVERSIONS

MODULE_INIT_IGNORE_VERMAGIC

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

MREMAP_FIXED

MREMAP_MAYMOVE

MSDOS_SUPER_MAGIC

MSG_CMSG_CLOEXEC

MSG_CONFIRM

MSG_COPY

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOR

MSG_ERRQUEUE

MSG_EXCEPT

MSG_FASTOPEN

MSG_FIN

MSG_INFO

MSG_MORE

MSG_NOERROR

MSG_NOSIGNAL

MSG_OOB

MSG_PEEK

MSG_RST

MSG_STAT

MSG_SYN

MSG_TRUNC

MSG_TRYHARD

MSG_WAITALL

MSG_WAITFORONE

MS_ACTIVE

MS_ASYNC

MS_BIND

MS_DIRSYNC

MS_INVALIDATE

MS_I_VERSION

MS_KERNMOUNT

MS_MANDLOCK

MS_MGC_MSK

MS_MGC_VAL

MS_MOVE

MS_NOATIME

MS_NODEV

MS_NODIRATIME

MS_NOEXEC

MS_NOSUID

MS_NOUSER

MS_POSIXACL

MS_PRIVATE

MS_RDONLY

MS_REC

MS_RELATIME

MS_REMOUNT

MS_RMT_MASK

MS_SHARED

MS_SILENT

MS_SLAVE

MS_STRICTATIME

MS_SYNC

MS_SYNCHRONOUS

MS_UNBINDABLE

M_ARENA_MAX

M_ARENA_TEST

M_CHECK_ACTION

M_GRAIN

M_KEEP

M_MMAP_MAX

M_MMAP_THRESHOLD

M_MXFAST

M_NLBLKS

M_PERTURB

M_TOP_PAD

M_TRIM_THRESHOLD

NCCS

NCP_SUPER_MAGIC

NDA_CACHEINFO

NDA_DST

NDA_IFINDEX

NDA_LINK_NETNSID

NDA_LLADDR

NDA_MASTER

NDA_PORT

NDA_PROBES

NDA_SRC_VNI

NDA_UNSPEC

NDA_VLAN

NDA_VNI

NETLINK_ADD_MEMBERSHIP

NETLINK_AUDIT

NETLINK_BROADCAST_ERROR

NETLINK_CAP_ACK

NETLINK_CONNECTOR

NETLINK_CRYPTO

NETLINK_DNRTMSG

NETLINK_DROP_MEMBERSHIP

NETLINK_ECRYPTFS

NETLINK_FIB_LOOKUP

NETLINK_FIREWALL

NETLINK_GENERIC

NETLINK_INET_DIAG

NETLINK_IP6_FW

NETLINK_ISCSI

NETLINK_KOBJECT_UEVENT

NETLINK_LISTEN_ALL_NSID

NETLINK_LIST_MEMBERSHIPS

NETLINK_NETFILTER

NETLINK_NFLOG

NETLINK_NO_ENOBUFS

NETLINK_PKTINFO

NETLINK_RDMA

NETLINK_ROUTE

NETLINK_RX_RING

NETLINK_SCSITRANSPORT

NETLINK_SELINUX

NETLINK_SOCK_DIAG

NETLINK_TX_RING

NETLINK_UNUSED

NETLINK_USERSOCK

NETLINK_XFRM

NEW_TIME

NFPROTO_ARP

NFPROTO_BRIDGE

NFPROTO_DECNET

NFPROTO_INET

NFPROTO_IPV4

NFPROTO_IPV6

NFPROTO_NETDEV

NFPROTO_NUMPROTO

NFPROTO_UNSPEC

NFS_SUPER_MAGIC

NFT_BREAK

NFT_BYTEORDER_HTON

NFT_BYTEORDER_NTOH

NFT_CHAIN_MAXNAMELEN

NFT_CMP_EQ

NFT_CMP_GT

NFT_CMP_GTE

NFT_CMP_LT

NFT_CMP_LTE

NFT_CMP_NEQ

NFT_CONTINUE

NFT_CT_BYTES

NFT_CT_DIRECTION

NFT_CT_DST

NFT_CT_EXPIRATION

NFT_CT_HELPER

NFT_CT_L3PROTOCOL

NFT_CT_LABELS

NFT_CT_MARK

NFT_CT_PKTS

NFT_CT_PROTOCOL

NFT_CT_PROTO_DST

NFT_CT_PROTO_SRC

NFT_CT_SECMARK

NFT_CT_SRC

NFT_CT_STATE

NFT_CT_STATUS

NFT_DATA_RESERVED_MASK

NFT_DATA_VALUE

NFT_DATA_VALUE_MAXLEN

NFT_DATA_VERDICT

NFT_DYNSET_F_INV

NFT_DYNSET_OP_ADD

NFT_DYNSET_OP_UPDATE

NFT_GOTO

NFT_JUMP

NFT_LIMIT_F_INV

NFT_LIMIT_PKTS

NFT_LIMIT_PKT_BYTES

NFT_LOOKUP_F_INV

NFT_META_BRI_IIFNAME

NFT_META_BRI_OIFNAME

NFT_META_CGROUP

NFT_META_CPU

NFT_META_IIF

NFT_META_IIFGROUP

NFT_META_IIFNAME

NFT_META_IIFTYPE

NFT_META_L4PROTO

NFT_META_LEN

NFT_META_MARK

NFT_META_NFPROTO

NFT_META_NFTRACE

NFT_META_OIF

NFT_META_OIFGROUP

NFT_META_OIFNAME

NFT_META_OIFTYPE

NFT_META_PKTTYPE

NFT_META_PRANDOM

NFT_META_PRIORITY

NFT_META_PROTOCOL

NFT_META_RTCLASSID

NFT_META_SECMARK

NFT_META_SKGID

NFT_META_SKUID

NFT_MSG_DELCHAIN

NFT_MSG_DELOBJ

NFT_MSG_DELRULE

NFT_MSG_DELSET

NFT_MSG_DELSETELEM

NFT_MSG_DELTABLE

NFT_MSG_GETCHAIN

NFT_MSG_GETGEN

NFT_MSG_GETOBJ

NFT_MSG_GETOBJ_RESET

NFT_MSG_GETRULE

NFT_MSG_GETSET

NFT_MSG_GETSETELEM

NFT_MSG_GETTABLE

NFT_MSG_MAX

NFT_MSG_NEWCHAIN

NFT_MSG_NEWGEN

NFT_MSG_NEWOBJ

NFT_MSG_NEWRULE

NFT_MSG_NEWSET

NFT_MSG_NEWSETELEM

NFT_MSG_NEWTABLE

NFT_MSG_TRACE

NFT_NAT_DNAT

NFT_NAT_SNAT

NFT_NG_INCREMENTAL

NFT_NG_RANDOM

NFT_OBJ_MAXNAMELEN

NFT_PAYLOAD_CSUM_INET

NFT_PAYLOAD_CSUM_NONE

NFT_PAYLOAD_LL_HEADER

NFT_PAYLOAD_NETWORK_HEADER

NFT_PAYLOAD_TRANSPORT_HEADER

NFT_QUEUE_FLAG_BYPASS

NFT_QUEUE_FLAG_CPU_FANOUT

NFT_QUEUE_FLAG_MASK

NFT_QUOTA_F_INV

NFT_RANGE_EQ

NFT_RANGE_NEQ

NFT_REG32_SIZE

NFT_REG32_00

NFT_REG32_01

NFT_REG32_02

NFT_REG32_03

NFT_REG32_04

NFT_REG32_05

NFT_REG32_06

NFT_REG32_07

NFT_REG32_08

NFT_REG32_09

NFT_REG32_10

NFT_REG32_11

NFT_REG32_12

NFT_REG32_13

NFT_REG32_14

NFT_REG32_15

NFT_REG_1

NFT_REG_2

NFT_REG_3

NFT_REG_4

NFT_REG_SIZE

NFT_REG_VERDICT

NFT_REJECT_ICMPX_ADMIN_PROHIBITED

NFT_REJECT_ICMPX_HOST_UNREACH

NFT_REJECT_ICMPX_NO_ROUTE

NFT_REJECT_ICMPX_PORT_UNREACH

NFT_REJECT_ICMPX_UNREACH

NFT_REJECT_ICMP_UNREACH

NFT_REJECT_TCP_RST

NFT_RETURN

NFT_SET_ANONYMOUS

NFT_SET_CONSTANT

NFT_SET_ELEM_INTERVAL_END

NFT_SET_EVAL

NFT_SET_INTERVAL

NFT_SET_MAP

NFT_SET_MAXNAMELEN

NFT_SET_POL_MEMORY

NFT_SET_POL_PERFORMANCE

NFT_SET_TIMEOUT

NFT_TABLE_MAXNAMELEN

NFT_TRACETYPE_POLICY

NFT_TRACETYPE_RETURN

NFT_TRACETYPE_RULE

NFT_TRACETYPE_UNSPEC

NFT_USERDATA_MAXLEN

NF_ACCEPT

NF_DROP

NF_INET_FORWARD

NF_INET_LOCAL_IN

NF_INET_LOCAL_OUT

NF_INET_NUMHOOKS

NF_INET_POST_ROUTING

NF_INET_PRE_ROUTING

NF_IP6_PRE_ROUTING

NF_IP6_LOCAL_IN

NF_IP6_FORWARD

NF_IP6_LOCAL_OUT

NF_IP6_POST_ROUTING

NF_IP6_NUMHOOKS

NF_IP6_PRI_FIRST

NF_IP6_PRI_CONNTRACK_DEFRAG

NF_IP6_PRI_RAW

NF_IP6_PRI_SELINUX_FIRST

NF_IP6_PRI_CONNTRACK

NF_IP6_PRI_MANGLE

NF_IP6_PRI_NAT_DST

NF_IP6_PRI_FILTER

NF_IP6_PRI_SECURITY

NF_IP6_PRI_NAT_SRC

NF_IP6_PRI_SELINUX_LAST

NF_IP6_PRI_CONNTRACK_HELPER

NF_IP6_PRI_LAST

NF_IP_FORWARD

NF_IP_LOCAL_IN

NF_IP_LOCAL_OUT

NF_IP_NUMHOOKS

NF_IP_POST_ROUTING

NF_IP_PRE_ROUTING

NF_IP_PRI_CONNTRACK

NF_IP_PRI_CONNTRACK_CONFIRM

NF_IP_PRI_CONNTRACK_DEFRAG

NF_IP_PRI_CONNTRACK_HELPER

NF_IP_PRI_FILTER

NF_IP_PRI_FIRST

NF_IP_PRI_LAST

NF_IP_PRI_MANGLE

NF_IP_PRI_NAT_DST

NF_IP_PRI_NAT_SRC

NF_IP_PRI_RAW

NF_IP_PRI_SECURITY

NF_IP_PRI_SELINUX_FIRST

NF_IP_PRI_SELINUX_LAST

NF_MAX_VERDICT

NF_NETDEV_INGRESS

NF_NETDEV_NUMHOOKS

NF_QUEUE

NF_REPEAT

NF_STOLEN

NF_STOP

NF_VERDICT_BITS

NF_VERDICT_FLAG_QUEUE_BYPASS

NF_VERDICT_MASK

NF_VERDICT_QBITS

NF_VERDICT_QMASK

NI_DGRAM

NI_MAXHOST

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSERV

NL0

NL1

NLA_ALIGNTO

NLA_F_NESTED

NLA_F_NET_BYTEORDER

NLA_TYPE_MASK

NLDLY

NLMSG_DONE

NLMSG_ERROR

NLMSG_MIN_TYPE

NLMSG_NOOP

NLMSG_OVERRUN

NLM_F_ACK

NLM_F_APPEND

NLM_F_ATOMIC

NLM_F_CREATE

NLM_F_DUMP

NLM_F_DUMP_FILTERED

NLM_F_DUMP_INTR

NLM_F_ECHO

NLM_F_EXCL

NLM_F_MATCH

NLM_F_MULTI

NLM_F_REPLACE

NLM_F_REQUEST

NLM_F_ROOT

NOEXPR

NOFLSH

NOSTR

NTF_EXT_LEARNED

NTF_MASTER

NTF_OFFLOADED

NTF_PROXY

NTF_ROUTER

NTF_SELF

NTF_USE

NUD_DELAY

NUD_FAILED

NUD_INCOMPLETE

NUD_NOARP

NUD_NONE

NUD_PERMANENT

NUD_PROBE

NUD_REACHABLE

NUD_STALE

OCRNL

OFDEL

OFILL

OLCUC

OLD_TIME

ONLCR

ONLRET

ONOCR

OPENPROM_SUPER_MAGIC

OPOST

ORIG_RAX

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECT

O_DIRECTORY

O_DSYNC

O_EXCL

O_FSYNC

O_LARGEFILE

O_NDELAY

O_NOATIME

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_PATH

O_RDONLY

O_RDWR

O_RSYNC

O_SYNC

O_TMPFILE

O_TRUNC

O_WRONLY

PACKET_ADD_MEMBERSHIP

PACKET_DROP_MEMBERSHIP

PACKET_MR_ALLMULTI

PACKET_MR_MULTICAST

PACKET_MR_PROMISC

PACKET_MR_UNICAST

PARENB

PARMRK

PARODD

PATH_MAX

PENDIN

PF_ALG

PF_APPLETALK

PF_ASH

PF_ATMPVC

PF_ATMSVC

PF_AX25

PF_BLUETOOTH

PF_BRIDGE

PF_CAIF

PF_CAN

PF_DECnet

PF_ECONET

PF_IB

PF_IEEE802154

PF_INET

PF_INET6

PF_IPX

PF_IRDA

PF_ISDN

PF_IUCV

PF_KEY

PF_LLC

PF_LOCAL

PF_MPLS

PF_NETBEUI

PF_NETLINK

PF_NETROM

PF_NFC

PF_PACKET

PF_PHONET

PF_PPPOX

PF_RDS

PF_ROSE

PF_ROUTE

PF_RXRPC

PF_SECURITY

PF_SNA

PF_TIPC

PF_UNIX

PF_UNSPEC

PF_VSOCK

PF_WANPIPE

PF_X25

PF_XDP

PIPE_BUF

PM_STR

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

POLLWRBAND

POLLWRNORM

POSIX_FADV_DONTNEED

POSIX_FADV_NOREUSE

POSIX_FADV_NORMAL

POSIX_FADV_RANDOM

POSIX_FADV_SEQUENTIAL

POSIX_FADV_WILLNEED

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

POSIX_SPAWN_RESETIDS

POSIX_SPAWN_SETPGROUP

POSIX_SPAWN_SETSCHEDPARAM

POSIX_SPAWN_SETSCHEDULER

POSIX_SPAWN_SETSIGDEF

POSIX_SPAWN_SETSIGMASK

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PROC_SUPER_MAGIC

PROT_EXEC

PROT_GROWSDOWN

PROT_GROWSUP

PROT_NONE

PROT_READ

PROT_WRITE

PR_CAPBSET_DROP

PR_CAPBSET_READ

PR_CAP_AMBIENT

PR_CAP_AMBIENT_CLEAR_ALL

PR_CAP_AMBIENT_IS_SET

PR_CAP_AMBIENT_LOWER

PR_CAP_AMBIENT_RAISE

PR_ENDIAN_BIG

PR_ENDIAN_LITTLE

PR_ENDIAN_PPC_LITTLE

PR_FPEMU_NOPRINT

PR_FPEMU_SIGFPE

PR_FP_EXC_ASYNC

PR_FP_EXC_DISABLED

PR_FP_EXC_DIV

PR_FP_EXC_INV

PR_FP_EXC_NONRECOV

PR_FP_EXC_OVF

PR_FP_EXC_PRECISE

PR_FP_EXC_RES

PR_FP_EXC_SW_ENABLE

PR_FP_EXC_UND

PR_FP_MODE_FR

PR_FP_MODE_FRE

PR_GET_CHILD_SUBREAPER

PR_GET_DUMPABLE

PR_GET_ENDIAN

PR_GET_FPEMU

PR_GET_FPEXC

PR_GET_FP_MODE

PR_GET_KEEPCAPS

PR_GET_NAME

PR_GET_NO_NEW_PRIVS

PR_GET_PDEATHSIG

PR_GET_SECCOMP

PR_GET_SECUREBITS

PR_GET_THP_DISABLE

PR_GET_TID_ADDRESS

PR_GET_TIMERSLACK

PR_GET_TIMING

PR_GET_TSC

PR_GET_UNALIGN

PR_MCE_KILL

PR_MCE_KILL_CLEAR

PR_MCE_KILL_DEFAULT

PR_MCE_KILL_EARLY

PR_MCE_KILL_GET

PR_MCE_KILL_LATE

PR_MCE_KILL_SET

PR_MPX_DISABLE_MANAGEMENT

PR_MPX_ENABLE_MANAGEMENT

PR_SET_CHILD_SUBREAPER

PR_SET_DUMPABLE

PR_SET_ENDIAN

PR_SET_FPEMU

PR_SET_FPEXC

PR_SET_FP_MODE

PR_SET_KEEPCAPS

PR_SET_MM

PR_SET_MM_ARG_END

PR_SET_MM_ARG_START

PR_SET_MM_AUXV

PR_SET_MM_BRK

PR_SET_MM_END_CODE

PR_SET_MM_END_DATA

PR_SET_MM_ENV_END

PR_SET_MM_ENV_START

PR_SET_MM_EXE_FILE

PR_SET_MM_MAP

PR_SET_MM_MAP_SIZE

PR_SET_MM_START_BRK

PR_SET_MM_START_CODE

PR_SET_MM_START_DATA

PR_SET_MM_START_STACK

PR_SET_NAME

PR_SET_NO_NEW_PRIVS

PR_SET_PDEATHSIG

PR_SET_PTRACER

PR_SET_SECCOMP

PR_SET_SECUREBITS

PR_SET_THP_DISABLE

PR_SET_TIMERSLACK

PR_SET_TIMING

PR_SET_TSC

PR_SET_UNALIGN

PR_TASK_PERF_EVENTS_DISABLE

PR_TASK_PERF_EVENTS_ENABLE

PR_TIMING_STATISTICAL

PR_TIMING_TIMESTAMP

PR_TSC_ENABLE

PR_TSC_SIGSEGV

PR_UNALIGN_NOPRINT

PR_UNALIGN_SIGBUS

PTHREAD_ADAPTIVE_MUTEX_INITIALIZER_NP

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_ERRORCHECK_MUTEX_INITIALIZER_NP

PTHREAD_MUTEX_ADAPTIVE_NP

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RECURSIVE_MUTEX_INITIALIZER_NP

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

PTRACE_ATTACH

PTRACE_CONT

PTRACE_DETACH

PTRACE_EVENT_CLONE

PTRACE_EVENT_EXEC

PTRACE_EVENT_EXIT

PTRACE_EVENT_FORK

PTRACE_EVENT_SECCOMP

PTRACE_EVENT_VFORK

PTRACE_EVENT_VFORK_DONE

PTRACE_GETEVENTMSG

PTRACE_GETFPREGS

PTRACE_GETFPXREGS

PTRACE_GETREGS

PTRACE_GETREGSET

PTRACE_GETSIGINFO

PTRACE_INTERRUPT

PTRACE_KILL

PTRACE_LISTEN

PTRACE_O_EXITKILL

PTRACE_O_MASK

PTRACE_O_SUSPEND_SECCOMP

PTRACE_O_TRACECLONE

PTRACE_O_TRACEEXEC

PTRACE_O_TRACEEXIT

PTRACE_O_TRACEFORK

PTRACE_O_TRACESECCOMP

PTRACE_O_TRACESYSGOOD

PTRACE_O_TRACEVFORK

PTRACE_O_TRACEVFORKDONE

PTRACE_PEEKDATA

PTRACE_PEEKSIGINFO

PTRACE_PEEKSIGINFO_SHARED

PTRACE_PEEKTEXT

PTRACE_PEEKUSER

PTRACE_POKEDATA

PTRACE_POKETEXT

PTRACE_POKEUSER

PTRACE_SEIZE

PTRACE_SETFPREGS

PTRACE_SETFPXREGS

PTRACE_SETOPTIONS

PTRACE_SETREGS

PTRACE_SETREGSET

PTRACE_SETSIGINFO

PTRACE_SINGLESTEP

PTRACE_SYSCALL

PTRACE_TRACEME

PT_DYNAMIC

PT_GNU_EH_FRAME

PT_GNU_RELRO

PT_GNU_STACK

PT_INTERP

PT_LOAD

PT_LOOS

PT_NOTE

PT_NULL

PT_NUM

PT_PHDR

PT_SHLIB

PT_TLS

P_ALL

P_PGID

P_PID

QFMT_VFS_OLD

QFMT_VFS_V0

QFMT_VFS_V1

QIF_ALL

QIF_BLIMITS

QIF_BTIME

QIF_ILIMITS

QIF_INODES

QIF_ITIME

QIF_LIMITS

QIF_SPACE

QIF_TIMES

QIF_USAGE

QNX4_SUPER_MAGIC

Q_GETFMT

Q_GETINFO

Q_GETQUOTA

Q_QUOTAOFF

Q_QUOTAON

Q_SETINFO

Q_SETQUOTA

Q_SYNC

R8

R9

R10

R11

R12

R13

R14

R15

RADIXCHAR

RAND_MAX

RAX

RBP

RBX

RB_AUTOBOOT

RB_DISABLE_CAD

RB_ENABLE_CAD

RB_HALT_SYSTEM

RB_KEXEC

RB_POWER_OFF

RB_SW_SUSPEND

RCX

RDI

RDX

REG_CR2

REG_CSGSFS

REG_EFL

REG_ERR

REG_OLDMASK

REG_R8

REG_R9

REG_R10

REG_R11

REG_R12

REG_R13

REG_R14

REG_R15

REG_RAX

REG_RBP

REG_RBX

REG_RCX

REG_RDI

REG_RDX

REG_RIP

REG_RSI

REG_RSP

REG_TRAPNO

REISERFS_SUPER_MAGIC

RENAME_EXCHANGE

RENAME_NOREPLACE

RENAME_WHITEOUT

RIP

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_LOCKS

RLIMIT_MEMLOCK

RLIMIT_MSGQUEUE

RLIMIT_NICE

RLIMIT_NLIMITS

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_RTPRIO

RLIMIT_RTTIME

RLIMIT_SIGPENDING

RLIMIT_STACK

RLIM_INFINITY

RLIM_SAVED_CUR

RLIM_SAVED_MAX

RSI

RSP

RTA_CACHEINFO

RTA_DST

RTA_ENCAP

RTA_ENCAP_TYPE

RTA_EXPIRES

RTA_FLOW

RTA_GATEWAY

RTA_IIF

RTA_MARK

RTA_METRICS

RTA_MFC_STATS

RTA_MP_ALGO

RTA_MULTIPATH

RTA_NEWDST

RTA_OIF

RTA_PAD

RTA_PREF

RTA_PREFSRC

RTA_PRIORITY

RTA_PROTOINFO

RTA_SESSION

RTA_SRC

RTA_TABLE

RTA_TTL_PROPAGATE

RTA_UID

RTA_UNSPEC

RTA_VIA

RTCF_DIRECTSRC

RTCF_DOREDIRECT

RTCF_LOG

RTCF_MASQ

RTCF_NAT

RTCF_VALVE

RTF_ADDRCLASSMASK

RTF_ADDRCONF

RTF_ALLONLINK

RTF_BROADCAST

RTF_CACHE

RTF_DEFAULT

RTF_DYNAMIC

RTF_FLOW

RTF_GATEWAY

RTF_HOST

RTF_INTERFACE

RTF_IRTT

RTF_LINKRT

RTF_LOCAL

RTF_MODIFIED

RTF_MSS

RTF_MTU

RTF_MULTICAST

RTF_NAT

RTF_NOFORWARD

RTF_NONEXTHOP

RTF_NOPMTUDISC

RTF_POLICY

RTF_REINSTATE

RTF_REJECT

RTF_STATIC

RTF_THROW

RTF_UP

RTF_WINDOW

RTF_XRESOLVE

RTLD_DEEPBIND

RTLD_DEFAULT

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NEXT

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

RTMSG_AR_FAILED

RTMSG_CONTROL

RTMSG_DELDEVICE

RTMSG_DELROUTE

RTMSG_DELRULE

RTMSG_NEWDEVICE

RTMSG_NEWROUTE

RTMSG_NEWRULE

RTMSG_OVERRUN

RTM_F_CLONED

RTM_F_EQUALIZE

RTM_F_FIB_MATCH

RTM_F_LOOKUP_TABLE

RTM_F_NOTIFY

RTM_F_PREFIX

RTN_ANYCAST

RTN_BLACKHOLE

RTN_BROADCAST

RTN_LOCAL

RTN_MULTICAST

RTN_NAT

RTN_PROHIBIT

RTN_THROW

RTN_UNICAST

RTN_UNREACHABLE

RTN_UNSPEC

RTN_XRESOLVE

RTPROT_BOOT

RTPROT_KERNEL

RTPROT_REDIRECT

RTPROT_STATIC

RTPROT_UNSPEC

RT_CLASS_DEFAULT

RT_CLASS_LOCAL

RT_CLASS_MAIN

RT_CLASS_MAX

RT_CLASS_UNSPEC

RT_SCOPE_HOST

RT_SCOPE_LINK

RT_SCOPE_NOWHERE

RT_SCOPE_SITE

RT_SCOPE_UNIVERSE

RT_TABLE_COMPAT

RT_TABLE_DEFAULT

RT_TABLE_LOCAL

RT_TABLE_MAIN

RT_TABLE_UNSPEC

RUN_LVL

RUSAGE_CHILDREN

RUSAGE_SELF

RUSAGE_THREAD

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SCHED_BATCH

SCHED_FIFO

SCHED_IDLE

SCHED_OTHER

SCHED_RR

SCM_CREDENTIALS

SCM_RIGHTS

SCM_TIMESTAMP

SCM_TIMESTAMPING

SCM_TIMESTAMPNS

SCM_WIFI_STATUS

SECCOMP_MODE_DISABLED

SECCOMP_MODE_FILTER

SECCOMP_MODE_STRICT

SEEK_CUR

SEEK_DATA

SEEK_END

SEEK_HOLE

SEEK_SET

SEM_FAILED

SFD_CLOEXEC

SFD_NONBLOCK

SHM_EXEC

SHM_HUGETLB

SHM_LOCK

SHM_NORESERVE

SHM_R

SHM_RDONLY

SHM_REMAP

SHM_RND

SHM_UNLOCK

SHM_W

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGEV_THREAD_ID

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPOLL

SIGPROF

SIGPWR

SIGQUIT

SIGSEGV

SIGSTKFLT

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGUNUSED	Deprecated

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SIOCADDMULTI

SIOCADDRT

SIOCDARP

SIOCDELMULTI

SIOCDELRT

SIOCDRARP

SIOCGARP

SIOCGIFADDR

SIOCGIFBRDADDR

SIOCGIFCONF

SIOCGIFDSTADDR

SIOCGIFENCAP

SIOCGIFFLAGS

SIOCGIFHWADDR

SIOCGIFMAP

SIOCGIFMEM

SIOCGIFMETRIC

SIOCGIFMTU

SIOCGIFNAME

SIOCGIFNETMASK

SIOCGIFSLAVE

SIOCGRARP

SIOCSARP

SIOCSIFADDR

SIOCSIFBRDADDR

SIOCSIFDSTADDR

SIOCSIFENCAP

SIOCSIFFLAGS

SIOCSIFHWADDR

SIOCSIFLINK

SIOCSIFMAP

SIOCSIFMEM

SIOCSIFMETRIC

SIOCSIFMTU

SIOCSIFNETMASK

SIOCSIFSLAVE

SIOCSRARP

SI_LOAD_SHIFT

SMB_SUPER_MAGIC

SOCK_CLOEXEC

SOCK_DCCP

SOCK_DGRAM

SOCK_NONBLOCK

SOCK_PACKET

SOCK_RAW

SOCK_RDM

SOCK_SEQPACKET

SOCK_STREAM

SOF_TIMESTAMPING_RAW_HARDWARE

SOF_TIMESTAMPING_RX_HARDWARE

SOF_TIMESTAMPING_RX_SOFTWARE

SOF_TIMESTAMPING_SOFTWARE

SOF_TIMESTAMPING_SYS_HARDWARE

SOF_TIMESTAMPING_TX_HARDWARE

SOF_TIMESTAMPING_TX_SOFTWARE

SOL_AAL

SOL_ALG

SOL_ATM

SOL_BLUETOOTH

SOL_CAIF

SOL_DCCP

SOL_DECNET

SOL_ICMPV6

SOL_IP

SOL_IPV6

SOL_IRDA

SOL_IUCV

SOL_LLC

SOL_NETBEUI

SOL_NETLINK

SOL_NFC

SOL_PACKET

SOL_PNPIPE

SOL_PPPOL2TP

SOL_RAW

SOL_RDS

SOL_RXRPC

SOL_SOCKET

SOL_TCP

SOL_TIPC

SOL_UDP

SOL_X25

SOL_XDP

SOMAXCONN

SO_ACCEPTCONN

SO_ATTACH_BPF

SO_ATTACH_FILTER

SO_BINDTODEVICE

SO_BPF_EXTENSIONS

SO_BROADCAST

SO_BSDCOMPAT

SO_BUSY_POLL

SO_DEBUG

SO_DETACH_BPF

SO_DETACH_FILTER

SO_DOMAIN

SO_DONTROUTE

SO_ERROR

SO_GET_FILTER

SO_INCOMING_CPU

SO_KEEPALIVE

SO_LINGER

SO_LOCK_FILTER

SO_MARK

SO_MAX_PACING_RATE

SO_NOFCS

SO_NO_CHECK

SO_OOBINLINE

SO_ORIGINAL_DST

SO_PASSCRED

SO_PASSSEC

SO_PEEK_OFF

SO_PEERCRED

SO_PEERNAME

SO_PEERSEC

SO_PRIORITY

SO_PROTOCOL

SO_RCVBUF

SO_RCVBUFFORCE

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_RXQ_OVFL

SO_SECURITY_AUTHENTICATION

SO_SECURITY_ENCRYPTION_NETWORK

SO_SECURITY_ENCRYPTION_TRANSPORT

SO_SELECT_ERR_QUEUE

SO_SNDBUF

SO_SNDBUFFORCE

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TIMESTAMPING

SO_TIMESTAMPNS

SO_TYPE

SO_WIFI_STATUS

SPLICE_F_GIFT

SPLICE_F_MORE

SPLICE_F_MOVE

SPLICE_F_NONBLOCK

SS

SS_DISABLE

SS_ONSTACK

STATX_ALL

STATX_ATIME

STATX_ATTR_APPEND

STATX_ATTR_AUTOMOUNT

STATX_ATTR_COMPRESSED

STATX_ATTR_ENCRYPTED

STATX_ATTR_IMMUTABLE

STATX_ATTR_NODUMP

STATX_BASIC_STATS

STATX_BLOCKS

STATX_BTIME

STATX_CTIME

STATX_GID

STATX_INO

STATX_MODE

STATX_MTIME

STATX_NLINK

STATX_SIZE

STATX_TYPE

STATX_UID

STATX__RESERVED

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

ST_APPEND

ST_IMMUTABLE

ST_MANDLOCK

ST_NOATIME

ST_NODEV

ST_NODIRATIME

ST_NOEXEC

ST_NOSUID

ST_RDONLY

ST_RELATIME

ST_SYNCHRONOUS

ST_WRITE

SYNC_FILE_RANGE_WAIT_AFTER

SYNC_FILE_RANGE_WAIT_BEFORE

SYNC_FILE_RANGE_WRITE

SYS__sysctl

SYS_accept

SYS_accept4

SYS_access

SYS_acct

SYS_add_key

SYS_adjtimex

SYS_afs_syscall

SYS_alarm

SYS_arch_prctl

SYS_bind

SYS_bpf

SYS_brk

SYS_capget

SYS_capset

SYS_chdir

SYS_chmod

SYS_chown

SYS_chroot

SYS_clock_adjtime

SYS_clock_getres

SYS_clock_gettime

SYS_clock_nanosleep

SYS_clock_settime

SYS_clone

SYS_close

SYS_connect

SYS_copy_file_range

SYS_creat

SYS_create_module

SYS_delete_module

SYS_dup

SYS_dup2

SYS_dup3

SYS_epoll_create

SYS_epoll_create1

SYS_epoll_ctl

SYS_epoll_ctl_old

SYS_epoll_pwait

SYS_epoll_wait

SYS_epoll_wait_old

SYS_eventfd

SYS_eventfd2

SYS_execve

SYS_execveat

SYS_exit

SYS_exit_group

SYS_faccessat

SYS_fadvise64

SYS_fallocate

SYS_fanotify_init

SYS_fanotify_mark

SYS_fchdir

SYS_fchmod

SYS_fchmodat

SYS_fchown

SYS_fchownat

SYS_fcntl

SYS_fdatasync

SYS_fgetxattr

SYS_finit_module

SYS_flistxattr

SYS_flock

SYS_fork

SYS_fremovexattr

SYS_fsetxattr

SYS_fstat

SYS_fstatfs

SYS_fsync

SYS_ftruncate

SYS_futex

SYS_futimesat

SYS_get_kernel_syms

SYS_get_mempolicy

SYS_get_robust_list

SYS_get_thread_area

SYS_getcpu

SYS_getcwd

SYS_getdents

SYS_getdents64

SYS_getegid

SYS_geteuid

SYS_getgid

SYS_getgroups

SYS_getitimer

SYS_getpeername

SYS_getpgid

SYS_getpgrp

SYS_getpid

SYS_getpmsg

SYS_getppid

SYS_getpriority

SYS_getrandom

SYS_getresgid

SYS_getresuid

SYS_getrlimit

SYS_getrusage

SYS_getsid

SYS_getsockname

SYS_getsockopt

SYS_gettid

SYS_gettimeofday

SYS_getuid

SYS_getxattr

SYS_init_module

SYS_inotify_add_watch

SYS_inotify_init

SYS_inotify_init1

SYS_inotify_rm_watch

SYS_io_cancel

SYS_io_destroy

SYS_io_getevents

SYS_io_setup

SYS_io_submit

SYS_ioctl

SYS_ioperm

SYS_iopl

SYS_ioprio_get

SYS_ioprio_set

SYS_kcmp

SYS_kexec_file_load

SYS_kexec_load

SYS_keyctl

SYS_kill

SYS_lchown

SYS_lgetxattr

SYS_link

SYS_linkat

SYS_listen

SYS_listxattr

SYS_llistxattr

SYS_lookup_dcookie

SYS_lremovexattr

SYS_lseek

SYS_lsetxattr

SYS_lstat

SYS_madvise

SYS_mbind

SYS_membarrier

SYS_memfd_create

SYS_migrate_pages

SYS_mincore

SYS_mkdir

SYS_mkdirat

SYS_mknod

SYS_mknodat

SYS_mlock

SYS_mlock2

SYS_mlockall

SYS_mmap

SYS_modify_ldt

SYS_mount

SYS_move_pages

SYS_mprotect

SYS_mq_getsetattr

SYS_mq_notify

SYS_mq_open

SYS_mq_timedreceive

SYS_mq_timedsend

SYS_mq_unlink

SYS_mremap

SYS_msgctl

SYS_msgget

SYS_msgrcv

SYS_msgsnd

SYS_msync

SYS_munlock

SYS_munlockall

SYS_munmap

SYS_name_to_handle_at

SYS_nanosleep

SYS_newfstatat

SYS_nfsservctl

SYS_open

SYS_open_by_handle_at

SYS_openat

SYS_pause

SYS_perf_event_open

SYS_personality

SYS_pipe

SYS_pipe2

SYS_pivot_root

SYS_pkey_alloc

SYS_pkey_free

SYS_pkey_mprotect

SYS_poll

SYS_ppoll

SYS_prctl

SYS_pread64

SYS_preadv

SYS_preadv2

SYS_prlimit64

SYS_process_vm_readv

SYS_process_vm_writev

SYS_pselect6

SYS_ptrace

SYS_putpmsg

SYS_pwrite64

SYS_pwritev

SYS_pwritev2

SYS_query_module

SYS_quotactl

SYS_read

SYS_readahead

SYS_readlink

SYS_readlinkat

SYS_readv

SYS_reboot

SYS_recvfrom

SYS_recvmmsg

SYS_recvmsg

SYS_remap_file_pages

SYS_removexattr

SYS_rename

SYS_renameat

SYS_renameat2

SYS_request_key

SYS_restart_syscall

SYS_rmdir

SYS_rt_sigaction

SYS_rt_sigpending

SYS_rt_sigprocmask

SYS_rt_sigqueueinfo

SYS_rt_sigreturn

SYS_rt_sigsuspend

SYS_rt_sigtimedwait

SYS_rt_tgsigqueueinfo

SYS_sched_get_priority_max

SYS_sched_get_priority_min

SYS_sched_getaffinity

SYS_sched_getattr

SYS_sched_getparam

SYS_sched_getscheduler

SYS_sched_rr_get_interval

SYS_sched_setaffinity

SYS_sched_setattr

SYS_sched_setparam

SYS_sched_setscheduler

SYS_sched_yield

SYS_seccomp

SYS_security

SYS_select

SYS_semctl

SYS_semget

SYS_semop

SYS_semtimedop

SYS_sendfile

SYS_sendmmsg

SYS_sendmsg

SYS_sendto

SYS_set_mempolicy

SYS_set_robust_list

SYS_set_thread_area

SYS_set_tid_address

SYS_setdomainname

SYS_setfsgid

SYS_setfsuid

SYS_setgid

SYS_setgroups

SYS_sethostname

SYS_setitimer

SYS_setns

SYS_setpgid

SYS_setpriority

SYS_setregid

SYS_setresgid

SYS_setresuid

SYS_setreuid

SYS_setrlimit

SYS_setsid

SYS_setsockopt

SYS_settimeofday

SYS_setuid

SYS_setxattr

SYS_shmat

SYS_shmctl

SYS_shmdt

SYS_shmget

SYS_shutdown

SYS_sigaltstack

SYS_signalfd

SYS_signalfd4

SYS_socket

SYS_socketpair

SYS_splice

SYS_stat

SYS_statfs

SYS_statx

SYS_swapoff

SYS_swapon

SYS_symlink

SYS_symlinkat

SYS_sync

SYS_sync_file_range

SYS_syncfs

SYS_sysfs

SYS_sysinfo

SYS_syslog

SYS_tee

SYS_tgkill

SYS_time

SYS_timer_create

SYS_timer_delete

SYS_timer_getoverrun

SYS_timer_gettime

SYS_timer_settime

SYS_timerfd_create

SYS_timerfd_gettime

SYS_timerfd_settime

SYS_times

SYS_tkill

SYS_truncate

SYS_tuxcall

SYS_umask

SYS_umount2

SYS_uname

SYS_unlink

SYS_unlinkat

SYS_unshare

SYS_uselib

SYS_userfaultfd

SYS_ustat

SYS_utime

SYS_utimensat

SYS_utimes

SYS_vfork

SYS_vhangup

SYS_vmsplice

SYS_vserver

SYS_wait4

SYS_waitid

SYS_write

SYS_writev

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TAB0

TAB1

TAB2

TAB3

TABDLY

TCA_CHAIN

TCA_DUMP_INVISIBLE

TCA_FCNT

TCA_HW_OFFLOAD

TCA_KIND

TCA_OPTIONS

TCA_PAD

TCA_RATE

TCA_STAB

TCA_STATS

TCA_STATS2

TCA_UNSPEC

TCA_XSTATS

TCFLSH

TCGETA

TCGETS

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_CONGESTION

TCP_COOKIE_TRANSACTIONS

TCP_CORK

TCP_DEFER_ACCEPT

TCP_FASTOPEN

TCP_INFO

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_LINGER2

TCP_MAXSEG

TCP_MD5SIG

TCP_NODELAY

TCP_QUEUE_SEQ

TCP_QUICKACK

TCP_REPAIR

TCP_REPAIR_OPTIONS

TCP_REPAIR_QUEUE

TCP_SYNCNT

TCP_THIN_DUPACK

TCP_THIN_LINEAR_TIMEOUTS

TCP_TIMESTAMP

TCP_USER_TIMEOUT

TCP_WINDOW_CLAMP

TCSADRAIN

TCSAFLUSH

TCSANOW

TCSBRK

TCSETA

TCSETAF

TCSETAW

TCSETS

TCSETSF

TCSETSW

TCXONC

TFD_CLOEXEC

TFD_NONBLOCK

TFD_TIMER_ABSTIME

THOUSEP

TIMER_ABSTIME

TIOCCONS

TIOCEXCL

TIOCGPGRP

TIOCGRS485

TIOCGSERIAL

TIOCGSOFTCAR

TIOCGWINSZ

TIOCINQ

TIOCLINUX

TIOCMBIC

TIOCMBIS

TIOCMGET

TIOCMSET

TIOCM_CAR

TIOCM_CD

TIOCM_CTS

TIOCM_DSR

TIOCM_DTR

TIOCM_LE

TIOCM_RI

TIOCM_RNG

TIOCM_RTS

TIOCM_SR

TIOCM_ST

TIOCNXCL

TIOCOUTQ

TIOCSCTTY

TIOCSPGRP

TIOCSRS485

TIOCSSOFTCAR

TIOCSTI

TIOCSWINSZ

TMPFS_MAGIC

TMP_MAX

TOSTOP

TUN_READQ_SIZE

TUN_TAP_DEV

TUN_TUN_DEV

TUN_TYPE_MASK

T_FMT

T_FMT_AMPM

USBDEVICE_SUPER_MAGIC

USER_PROCESS

USRQUOTA

UTIME_NOW

UTIME_OMIT

VDISCARD

VEOF

VEOL

VEOL2

VERASE

VINTR

VKILL

VLNEXT

VMIN

VQUIT

VREPRINT

VSTART

VSTOP

VSUSP

VSWTC

VT0

VT1

VTDLY

VTIME

VWERASE

WCONTINUED

WEXITED

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

W_OK

XATTR_CREATE

XATTR_REPLACE

XTABS

X_OK

YESEXPR

YESSTR

_IOFBF

_IOLBF

_IONBF

_PC_2_SYMLINKS

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SOCK_MAXBUF

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_POSIX_VDISABLE

_SC_2_VERSION

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_SW_DEV

_SC_2_LOCALEDEF

_SC_2_CHAR_TERM

_SC_2_UPE

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_PBS_CHECKPOINT

_SC_2_C_VERSION

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BASE

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHARCLASS_NAME_MAX

_SC_CHAR_BIT

_SC_CHAR_MAX

_SC_CHAR_MIN

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_C_LANG_SUPPORT

_SC_C_LANG_SUPPORT_R

_SC_DELAYTIMER_MAX

_SC_DEVICE_IO

_SC_DEVICE_SPECIFIC

_SC_DEVICE_SPECIFIC_R

_SC_EQUIV_CLASS_MAX

_SC_EXPR_NEST_MAX

_SC_FD_MGMT

_SC_FIFO

_SC_FILE_ATTRIBUTES

_SC_FILE_LOCKING

_SC_FILE_SYSTEM

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_INT_MAX

_SC_INT_MIN

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LEVEL1_ICACHE_SIZE

_SC_LEVEL1_ICACHE_ASSOC

_SC_LEVEL1_ICACHE_LINESIZE

_SC_LEVEL1_DCACHE_SIZE

_SC_LEVEL1_DCACHE_ASSOC

_SC_LEVEL1_DCACHE_LINESIZE

_SC_LEVEL2_CACHE_SIZE

_SC_LEVEL2_CACHE_ASSOC

_SC_LEVEL2_CACHE_LINESIZE

_SC_LEVEL3_CACHE_SIZE

_SC_LEVEL3_CACHE_ASSOC

_SC_LEVEL3_CACHE_LINESIZE

_SC_LEVEL4_CACHE_SIZE

_SC_LEVEL4_CACHE_ASSOC

_SC_LEVEL4_CACHE_LINESIZE

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_LONG_BIT

_SC_MAPPED_FILES

_SC_MB_LEN_MAX

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_MULTI_PROCESS

_SC_NETWORKING

_SC_NGROUPS_MAX

_SC_NL_ARGMAX

_SC_NL_LANGMAX

_SC_NL_MSGMAX

_SC_NL_NMAX

_SC_NL_SETMAX

_SC_NL_TEXTMAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_NZERO

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PII

_SC_PII_INTERNET

_SC_PII_INTERNET_DGRAM

_SC_PII_INTERNET_STREAM

_SC_PII_OSI

_SC_PII_OSI_CLTS

_SC_PII_OSI_COTS

_SC_PII_OSI_M

_SC_PII_SOCKET

_SC_PII_XTI

_SC_PIPE

_SC_POLL

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_REGEX_VERSION

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SCHAR_MAX

_SC_SCHAR_MIN

_SC_SELECT

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SHRT_MAX

_SC_SHRT_MIN

_SC_SIGNALS

_SC_SIGQUEUE_MAX

_SC_SINGLE_PROCESS

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SSIZE_MAX

_SC_SS_REPL_MAX

_SC_STREAMS

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_SYSTEM_DATABASE

_SC_SYSTEM_DATABASE_R

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_ROBUST_PRIO_INHERIT

_SC_THREAD_ROBUST_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_T_IOV_MAX

_SC_UCHAR_MAX

_SC_UINT_MAX

_SC_UIO_MAXIOV

_SC_ULONG_MAX

_SC_USER_GROUPS

_SC_USER_GROUPS_R

_SC_USHRT_MAX

_SC_V6_LPBIG_OFFBIG

_SC_V7_LPBIG_OFFBIG

_SC_V6_ILP32_OFFBIG

_SC_V6_ILP32_OFF32

_SC_V6_LP64_OFF64

_SC_V7_ILP32_OFFBIG

_SC_V7_ILP32_OFF32

_SC_V7_LP64_OFF64

_SC_VERSION

_SC_WORD_BIT

_SC_XBS5_LPBIG_OFFBIG

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_ILP32_OFF32

_SC_XBS5_LP64_OFF64

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_SC_XOPEN_XPG2

_SC_XOPEN_XPG3

_SC_XOPEN_XPG4

__NFT_REG_MAX

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

__UT_HOSTSIZE

__UT_LINESIZE

__UT_NAMESIZE

__WALL

__WCLONE

__WNOTHREAD

Functions

CMSG_DATA⚠

CMSG_FIRSTHDR⚠

CMSG_LEN⚠

CMSG_NXTHDR⚠

CMSG_SPACE⚠

CPU_CLR⚠

CPU_EQUAL⚠

CPU_ISSET⚠

CPU_SET⚠

CPU_ZERO⚠

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

IPOPT_CLASS⚠

IPOPT_COPIED⚠

IPOPT_NUMBER⚠

IPTOS_ECN⚠

IPTOS_PREC⚠

IPTOS_TOS⚠

NLA_ALIGN⚠

QCMD⚠

RT_ADDRCLASS⚠

RT_LOCALADDR⚠

RT_TOS⚠

WCOREDUMP⚠

WEXITSTATUS⚠

WIFCONTINUED⚠

WIFEXITED⚠

WIFSIGNALED⚠

WIFSTOPPED⚠

WSTOPSIG⚠

WTERMSIG⚠

__errno_location⚠

_exit⚠

abort⚠

abs⚠

accept⚠

accept4⚠

access⚠

acct⚠

addmntent⚠

aio_cancel⚠

aio_error⚠

aio_fsync⚠

aio_read⚠

aio_return⚠

aio_suspend⚠

aio_write⚠

alarm⚠

atexit⚠

atof⚠

atoi⚠

backtrace⚠

bind⚠

brk⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clearenv⚠

clock_getres⚠

clock_gettime⚠

clock_nanosleep⚠

clock_settime⚠

clone⚠

close⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

creat64⚠

daemon⚠

difftime⚠

dirfd⚠

dl_iterate_phdr⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

dup3⚠

duplocale⚠

endgrent⚠

endmntent⚠

endpwent⚠

endspent⚠

endutxent⚠

epoll_create⚠

epoll_create1⚠

epoll_ctl⚠

epoll_pwait⚠

epoll_wait⚠

eventfd⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

execvpe⚠

exit⚠

faccessat⚠

fallocate⚠

fallocate64⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdatasync⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fexecve⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgetpos64⚠

fgets⚠

fgetxattr⚠

fileno⚠

flistxattr⚠

flock⚠

fopen⚠

fopen64⚠

fork⚠

forkpty⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

fread_unlocked⚠

free⚠

freeaddrinfo⚠

freeifaddrs⚠

freelocale⚠

fremovexattr⚠

freopen⚠

freopen64⚠

fscanf⚠

fseek⚠

fseeko⚠

fseeko64⚠

fsetpos⚠

fsetpos64⚠

fsetxattr⚠

fstat⚠

fstat64⚠

fstatat⚠

fstatat64⚠

fstatfs⚠

fstatfs64⚠

fstatvfs⚠

fstatvfs64⚠

fsync⚠

ftell⚠

ftello⚠

ftello64⚠

ftok⚠

ftruncate⚠

ftruncate64⚠

futimens⚠

futimes⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getauxval⚠

getchar⚠

getchar_unlocked⚠

getcontext⚠

getcwd⚠

getdomainname⚠

getdtablesize⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgrent⚠

getgrent_r⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgrouplist⚠

getgroups⚠

gethostname⚠

getifaddrs⚠

getline⚠

getloadavg⚠

getlogin⚠

getmntent⚠

getnameinfo⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getpriority⚠

getprotobyname⚠

getprotobynumber⚠

getpt⚠

getpwent⚠

getpwent_r⚠

getpwnam⚠

getpwnam_r⚠

getpwuid⚠

getpwuid_r⚠

getresgid⚠

getresuid⚠

getrlimit⚠

getrlimit64⚠

getrusage⚠

getservbyname⚠

getsid⚠

getsockname⚠

getsockopt⚠

getspent⚠

getspnam⚠

gettimeofday⚠

getuid⚠

getutxent⚠

getutxid⚠

getutxline⚠

getxattr⚠

glob⚠

glob64⚠

globfree⚠

globfree64⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

hasmntopt⚠

if_freenameindex⚠

if_indextoname⚠

if_nameindex⚠

if_nametoindex⚠

initgroups⚠

inotify_add_watch⚠

inotify_init⚠

inotify_init1⚠

inotify_rm_watch⚠

ioctl⚠

ioperm⚠

iopl⚠

isalnum⚠

isalpha⚠

isatty⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kill⚠

killpg⚠

labs⚠

lchown⚠

lgetxattr⚠

link⚠

linkat⚠

lio_listio⚠

listen⚠

listxattr⚠

llistxattr⚠

localeconv⚠

localtime⚠

localtime_r⚠

login_tty⚠

lremovexattr⚠

lseek⚠

lseek64⚠

lsetxattr⚠

lstat⚠

lstat64⚠

lutimes⚠

madvise⚠

major⚠

makecontext⚠

makedev⚠

mallinfo⚠

malloc⚠

malloc_usable_size⚠

mallopt⚠

memalign⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memrchr⚠

memset⚠

mincore⚠

minor⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mkfifoat⚠

mknod⚠

mknodat⚠

mkostemp⚠

mkostemps⚠

mkstemp⚠

mkstemps⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

mmap64⚠

mount⚠

mprotect⚠

mq_close⚠

mq_getattr⚠

mq_open⚠

mq_receive⚠

mq_send⚠

mq_setattr⚠

mq_unlink⚠

mremap⚠

msgctl⚠

msgget⚠

msgrcv⚠

msgsnd⚠

msync⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

newlocale⚠

nice⚠

nl_langinfo⚠

nl_langinfo_l⚠

open⚠

open64⚠

openat⚠

openat64⚠

opendir⚠

openlog⚠

openpty⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

personality⚠

pipe⚠

pipe2⚠

poll⚠

popen⚠

posix_fadvise⚠

posix_fallocate⚠

posix_fallocate64⚠

posix_madvise⚠

posix_memalign⚠

posix_openpt⚠

posix_spawn⚠

posix_spawn_file_actions_addclose⚠

posix_spawn_file_actions_adddup2⚠

posix_spawn_file_actions_addopen⚠

posix_spawn_file_actions_destroy⚠

posix_spawn_file_actions_init⚠

posix_spawnattr_destroy⚠

posix_spawnattr_getflags⚠

posix_spawnattr_getpgroup⚠

posix_spawnattr_getschedparam⚠

posix_spawnattr_getschedpolicy⚠

posix_spawnattr_getsigdefault⚠

posix_spawnattr_getsigmask⚠

posix_spawnattr_init⚠

posix_spawnattr_setflags⚠

posix_spawnattr_setpgroup⚠

posix_spawnattr_setschedparam⚠

posix_spawnattr_setschedpolicy⚠

posix_spawnattr_setsigdefault⚠

posix_spawnattr_setsigmask⚠

posix_spawnp⚠

ppoll⚠

prctl⚠

pread⚠

pread64⚠

preadv⚠

preadv64⚠

printf⚠

prlimit⚠

prlimit64⚠

process_vm_readv⚠

process_vm_writev⚠

pselect⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_getaffinity_np⚠

pthread_attr_getguardsize⚠

pthread_attr_getstack⚠

pthread_attr_init⚠

pthread_attr_setaffinity_np⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cancel⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getclock⚠

pthread_condattr_getpshared⚠

pthread_condattr_init⚠

pthread_condattr_setclock⚠

pthread_condattr_setpshared⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_getaffinity_np⚠

pthread_getattr_np⚠

pthread_getschedparam⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_timedlock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_getpshared⚠

pthread_mutexattr_init⚠

pthread_mutexattr_setpshared⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_getkind_np⚠

pthread_rwlockattr_getpshared⚠

pthread_rwlockattr_init⚠

pthread_rwlockattr_setkind_np⚠

pthread_rwlockattr_setpshared⚠

pthread_self⚠

pthread_setaffinity_np⚠

pthread_setschedparam⚠

pthread_setschedprio⚠

pthread_setspecific⚠

pthread_sigmask⚠

ptrace⚠

ptsname⚠

ptsname_r⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pututxline⚠

pwrite⚠

pwrite64⚠

pwritev⚠

pwritev64⚠

quotactl⚠

raise⚠

rand⚠

read⚠

readahead⚠

readdir⚠

readdir64⚠

readdir64_r⚠

readdir_r⚠
The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

readv⚠

realloc⚠

realpath⚠

reboot⚠

recv⚠

recvfrom⚠

recvmmsg⚠

recvmsg⚠

remap_file_pages⚠

remove⚠

removexattr⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

sbrk⚠

scanf⚠

sched_get_priority_max⚠

sched_get_priority_min⚠

sched_getaffinity⚠

sched_getcpu⚠

sched_getparam⚠

sched_getscheduler⚠

sched_rr_get_interval⚠

sched_setaffinity⚠

sched_setparam⚠

sched_setscheduler⚠

sched_yield⚠

seekdir⚠

select⚠

sem_close⚠

sem_destroy⚠

sem_getvalue⚠

sem_init⚠

sem_open⚠

sem_post⚠

sem_timedwait⚠

sem_trywait⚠

sem_unlink⚠

sem_wait⚠

semctl⚠

semget⚠

semop⚠

send⚠

sendfile⚠

sendmmsg⚠

sendmsg⚠

sendto⚠

setbuf⚠

setcontext⚠

setdomainname⚠

setegid⚠

setenv⚠

seteuid⚠

setfsgid⚠

setfsuid⚠

setgid⚠

setgrent⚠

setgroups⚠

sethostname⚠

setlocale⚠

setlogmask⚠

setmntent⚠

setns⚠

setpgid⚠

setpriority⚠

setpwent⚠

setregid⚠

setresgid⚠

setresuid⚠

setreuid⚠

setrlimit⚠

setrlimit64⚠

setsid⚠

setsockopt⚠

setspent⚠

settimeofday⚠

setuid⚠

setutxent⚠

setvbuf⚠

setxattr⚠

shm_open⚠

shm_unlink⚠

shmat⚠

shmctl⚠

shmdt⚠

shmget⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

signalfd⚠

sigpending⚠

sigprocmask⚠

sigsuspend⚠

sigtimedwait⚠

sigwait⚠

sigwaitinfo⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

splice⚠

sprintf⚠

srand⚠

sscanf⚠

stat⚠

stat64⚠

statfs⚠

statfs64⚠

statvfs⚠

statvfs64⚠

statx⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strspn⚠

strstr⚠

strtod⚠

strtok⚠

strtol⚠

strtoul⚠

strxfrm⚠

swapcontext⚠

swapoff⚠

swapon⚠

symlink⚠

symlinkat⚠

sync⚠

sync_file_range⚠

syscall⚠

sysconf⚠

sysctl⚠

sysinfo⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

tee⚠

telldir⚠

time⚠

timegm⚠

timerfd_create⚠

timerfd_gettime⚠

timerfd_settime⚠

times⚠

tmpfile⚠

tmpfile64⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

truncate64⚠

ttyname⚠

umask⚠

umount⚠

umount2⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

unshare⚠

uselocale⚠

usleep⚠

utime⚠

utimensat⚠

utimes⚠

utmpxname⚠

vfork⚠

vhangup⚠

vmsplice⚠

wait⚠

wait4⚠

waitid⚠

waitpid⚠

wcslen⚠

wcstombs⚠

write⚠

writev⚠

Type Definitions

Elf32_Half

Elf32_Word

Elf32_Off

Elf32_Addr

Elf32_Section

Elf64_Half

Elf64_Word

Elf64_Off

Elf64_Addr

Elf64_Xword

Elf64_Sxword

Elf64_Section

__fsword_t

__priority_which_t

__rlimit_resource_t

__s16

__s32

__u8

__u16

__u32

__u64

blkcnt64_t

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

dev_t

fsblkcnt_t

fsfilcnt_t

gid_t

greg_t

id_t

idtype_t

in_addr_t

in_port_t

ino64_t

ino_t

int8_t	Deprecated

int16_t	Deprecated

int32_t	Deprecated

int64_t	Deprecated

intmax_t

intptr_t

key_t

locale_t

loff_t

mode_t

mqd_t

msglen_t

msgqnum_t

nfds_t

nl_item

nlink_t

off64_t

off_t

pid_t

pthread_key_t

pthread_t

ptrdiff_t

rlim64_t

rlim_t

sa_family_t

shmatt_t

sighandler_t

size_t

socklen_t

speed_t

ssize_t

suseconds_t

tcflag_t

time_t

uid_t

uint8_t	Deprecated

uint16_t	Deprecated

uint32_t	Deprecated

uint64_t	Deprecated

uintmax_t

uintptr_t

useconds_t

wchar_t