Crate libc

libc - Raw FFI bindings to platforms’ system libraries

Documentation for other platforms.

Structs

Dl_info

Elf64_Phdr

__c_anonymous_eax_0

__c_anonymous_eax_1

__c_anonymous_eax_2

__c_anonymous_eax_3

__c_anonymous_regs

addrinfo

area_info

attr_info

cmsghdr

cpu_info

dirent

entry

fd_set

flock

fpu_state

fs_info

glob_t

group

hostent

ifaddrs

image_info

in6_addr

in_addr

index_info

iovec

ip_mreq

ipc_perm

ipv6_mreq

itimerval

lconv

linger

mcontext_t

msghdr

msqid_ds

object_wait_info

passwd

pollfd

port_info

port_message_info

protoent

pthread_cond_t

pthread_mutex_t

pthread_rwlock_t

pthread_spinlock_t

regex_t

regmatch_t

rlimit

rusage

savefpu

sem_info

sem_t

sembuf

servent

sigaction

sigevent

siginfo_t

sigval

sockaddr

sockaddr_dl

sockaddr_in

sockaddr_in6

sockaddr_storage

sockaddr_un

spwd

stack_t

stat

statvfs

system_info

team_info

team_usage_info

termios

thread_info

timespec

timeval

tm

tms

ucontext_t

ucred

utimbuf

utmpx

utsname

winsize

xstate_hdr

Enums

DIR

FILE

be_task_flags

c_void

Equivalent to C’s void type when used as a [pointer].

directory_which

fpos_t

image_type

path_base_directory

schduler_mode

thread_state

timezone

Constants

AF_APPLETALK

AF_BLUETOOTH

AF_DLI

AF_INET

AF_INET6

AF_IPX

AF_LINK

AF_LOCAL

AF_NOTIFY

AF_ROUTE

AF_UNIX

AF_UNSPEC

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

AT_EACCESS

AT_FDCWD

AT_REMOVEDIR

AT_SYMLINK_FOLLOW

AT_SYMLINK_NOFOLLOW

B0

B50

B75

B110

B134

B150

B200

B300

B600

B1200

B1800

B2400

B4800

B9600

B19200

B31250

B38400

B57600

B115200

B230400

BOOT_TIME

BRKINT

BS0

BS1

BSDLY

BUFSIZ

BUS_ADRALN

BUS_ADRERR

BUS_OBJERR

B_32_BIT_CONTIGUOUS

B_32_BIT_FULL_LOCK

B_ABSOLUTE_REAL_TIME_TIMEOUT

B_ABSOLUTE_TIMEOUT

B_AFFINE_TRANSFORM_TYPE

B_ALIGNMENT_TYPE

B_ALREADY_RUNNING

B_AMBIGUOUS_APP_LAUNCH

B_ANY_ADDRESS

B_ANY_KERNEL_ADDRESS

B_ANY_NODE

B_ANY_TYPE

B_APP_ERROR_BASE

B_ASCII_TYPE

B_ATOMREF_TYPE

B_ATOM_TYPE

B_ATTR_NAME_LENGTH

B_BAD_ADDRESS

B_BAD_DATA

B_BAD_HANDLER

B_BAD_IMAGE_ID

B_BAD_INDEX

B_BAD_MIME_SNIFFER_RULE

B_BAD_PORT_ID

B_BAD_REPLY

B_BAD_SCRIPT_SYNTAX

B_BAD_SEM_ID

B_BAD_SUBSCRIBER

B_BAD_TEAM_ID

B_BAD_THREAD_ID

B_BAD_THREAD_STATE

B_BAD_TYPE

B_BAD_VALUE

B_BASE_ADDRESS

B_BOOL_TYPE

B_BUFFER_NOT_AVAILABLE

B_BUFFER_OVERFLOW

B_BUSTED_PIPE

B_BUSY

B_CANCELED

B_CAN_INTERRUPT

B_CHAR_TYPE

B_CHECK_PERMISSION

B_CLONEABLE_AREA

B_CLONE_ADDRESS

B_COLOR_8_BIT_TYPE

B_CONTIGUOUS

B_CREATE_FILE

B_CROSS_DEVICE_LINK

B_CURRENT_TEAM

B_DEBUGGER_ALREADY_INSTALLED

B_DEVICE_ERROR_BASE

B_DEVICE_FULL

B_DEVICE_NOT_FOUND

B_DEV_BAD_DRIVE_NUM

B_DEV_BAD_PID

B_DEV_CONFIGURATION_ERROR

B_DEV_CRC_ERROR

B_DEV_DATA_OVERRUN

B_DEV_DATA_UNDERRUN

B_DEV_DISABLED_BY_USER

B_DEV_DOOR_OPEN

B_DEV_FIFO_OVERRUN

B_DEV_FIFO_UNDERRUN

B_DEV_FORMAT_ERROR

B_DEV_ID_ERROR

B_DEV_INVALID_IOCTL

B_DEV_INVALID_PIPE

B_DEV_MEDIA_CHANGED

B_DEV_MEDIA_CHANGE_REQUESTED

B_DEV_MULTIPLE_ERRORS

B_DEV_NAME_LENGTH

B_DEV_NOT_READY

B_DEV_NO_MEDIA

B_DEV_NO_MEMORY

B_DEV_PENDING

B_DEV_READ_ERROR

B_DEV_RECALIBRATE_ERROR

B_DEV_RESOURCE_CONFLICT

B_DEV_SEEK_ERROR

B_DEV_STALLED

B_DEV_TIMEOUT

B_DEV_TOO_LATE

B_DEV_UNEXPECTED_PID

B_DEV_UNREADABLE

B_DEV_WRITE_ERROR

B_DIRECTORY_NODE

B_DIRECTORY_NOT_EMPTY

B_DISPLAY_PRIORITY

B_DONT_DO_THAT

B_DOUBLE_TYPE

B_DO_NOT_RESCHEDULE

B_DUPLICATE_REPLY

B_ENTRY_NOT_FOUND

B_ERASE_FILE

B_ERROR

B_ERRORS_END

B_EVENT_ACQUIRE_SEMAPHORE

B_EVENT_DISCONNECTED

B_EVENT_ERROR

B_EVENT_HIGH_PRIORITY_READ

B_EVENT_HIGH_PRIORITY_WRITE

B_EVENT_INVALID

B_EVENT_PRIORITY_READ

B_EVENT_PRIORITY_WRITE

B_EVENT_READ

B_EVENT_WRITE

B_EXACT_ADDRESS

B_EXECUTE_AREA

B_FAIL_IF_EXISTS

B_FILE_ERROR

B_FILE_EXISTS

B_FILE_NAME_LENGTH

B_FILE_NODE

B_FILE_TOO_LARGE

B_FIRST_REAL_TIME_PRIORITY

B_FLOAT_TYPE

B_FLUSH_DCACHE

B_FLUSH_ICACHE

B_FORCE_UNMOUNT

B_FS_HAS_ALIASES

B_FS_HAS_ATTR

B_FS_HAS_MIME

B_FS_HAS_QUERY

B_FS_HAS_SELF_HEALING_LINKS

B_FS_IS_PERSISTENT

B_FS_IS_READONLY

B_FS_IS_REMOVABLE

B_FS_IS_SHARED

B_FS_SUPPORTS_MONITOR_CHILDREN

B_FS_SUPPORTS_NODE_MONITORING

B_FULL_LOCK

B_GENERAL_ERROR_BASE

B_GRAYSCALE_8_BIT_TYPE

B_IDLE_PRIORITY

B_ILLEGAL_DATA

B_INFINITE_TIMEOUT

B_INT8_TYPE

B_INT16_TYPE

B_INT32_TYPE

B_INT64_TYPE

B_INTERFACE_ERROR_BASE

B_INTERRUPTED

B_INVALIDATE_DCACHE

B_INVALIDATE_ICACHE

B_IO_ERROR

B_IS_A_DIRECTORY

B_KILL_CAN_INTERRUPT

B_LARGE_ICON_TYPE

B_LAST_BUFFER_ERROR

B_LAUNCH_FAILED

B_LAUNCH_FAILED_APP_IN_TRASH

B_LAUNCH_FAILED_APP_NOT_FOUND

B_LAUNCH_FAILED_EXECUTABLE

B_LAUNCH_FAILED_FILES_APP_NOT_FOUND

B_LAUNCH_FAILED_NO_PREFERRED_APP

B_LAUNCH_FAILED_NO_RESOLVE_LINK

B_LAZY_LOCK

B_LEGACY_EXECUTABLE

B_LINK_LIMIT

B_LIVE_QUERY

B_LOMEM

B_LOWEST_ACTIVE_PRIORITY

B_LOW_PRIORITY

B_MAIL_ACCESS_ERROR

B_MAIL_ERROR_BASE

B_MAIL_INVALID_MAIL

B_MAIL_NO_DAEMON

B_MAIL_NO_RECIPIENT

B_MAIL_UNKNOWN_FIELD

B_MAIL_UNKNOWN_HOST

B_MAIL_UNKNOWN_USER

B_MAIL_WRONG_PASSWORD

B_MAX_SYMLINKS

B_MEDIA_ADDON_DISABLED

B_MEDIA_ADDON_FAILED

B_MEDIA_ADDON_RESTRICTED

B_MEDIA_ALREADY_CONNECTED

B_MEDIA_APP_ALREADY_REGISTERED

B_MEDIA_APP_NOT_REGISTERED

B_MEDIA_BAD_BUFFER

B_MEDIA_BAD_CLIP_FORMAT

B_MEDIA_BAD_DESTINATION

B_MEDIA_BAD_FORMAT

B_MEDIA_BAD_NODE

B_MEDIA_BAD_SOURCE

B_MEDIA_BUFFERS_NOT_RECLAIMED

B_MEDIA_BUFFER_ALREADY_EXISTS

B_MEDIA_CANNOT_CHANGE_RUN_MODE

B_MEDIA_CANNOT_RECLAIM_BUFFERS

B_MEDIA_CANNOT_SEEK

B_MEDIA_CHANGE_IN_PROGRESS

B_MEDIA_DUPLICATE_FORMAT

B_MEDIA_ERROR_BASE

B_MEDIA_NODE_ALREADY_EXISTS

B_MEDIA_NODE_BUSY

B_MEDIA_NOT_CONNECTED

B_MEDIA_NO_HANDLER

B_MEDIA_PARAMETER_GROUP_TYPE

B_MEDIA_PARAMETER_TYPE

B_MEDIA_PARAMETER_WEB_TYPE

B_MEDIA_REALTIME_DISABLED

B_MEDIA_REALTIME_UNAVAILABLE

B_MEDIA_STALE_CHANGE_COUNT

B_MEDIA_SYSTEM_FAILURE

B_MEDIA_TIME_SOURCE_BUSY

B_MEDIA_TIME_SOURCE_STOPPED

B_MEDIA_TOO_MANY_BUFFERS

B_MEDIA_TOO_MANY_NODES

B_MESSAGE_TO_SELF

B_MESSAGE_TYPE

B_MESSENGER_TYPE

B_MIDI_ERROR_BASE

B_MIME_STRING_TYPE

B_MIME_TYPE

B_MIME_TYPE_LENGTH

B_MINI_ICON_TYPE

B_MISMATCHED_VALUES

B_MISSING_LIBRARY

B_MISSING_SYMBOL

B_MONOCHROME_1_BIT_TYPE

B_MOUNT_READ_ONLY

B_MOUNT_VIRTUAL_DEVICE

B_NAME_IN_USE

B_NAME_NOT_FOUND

B_NAME_TOO_LONG

B_NETWORK_ADDRESS_TYPE

B_NORMAL_PRIORITY

B_NOT_ALLOWED

B_NOT_AN_EXECUTABLE

B_NOT_A_DIRECTORY

B_NOT_A_MESSAGE

B_NOT_INITIALIZED

B_NOT_SUPPORTED

B_NO_ERROR

B_NO_INIT

B_NO_LOCK

B_NO_MEMORY

B_NO_MORE_FDS

B_NO_MORE_PORTS

B_NO_MORE_SEMS

B_NO_MORE_TEAMS

B_NO_MORE_THREADS

B_NO_PRINT_SERVER

B_NO_TRANSLATOR

B_OBJECT_TYPE

B_OBJECT_TYPE_FD

B_OBJECT_TYPE_PORT

B_OBJECT_TYPE_SEMAPHORE

B_OBJECT_TYPE_THREAD

B_OFF_T_TYPE

B_OK

B_ONE_SHOT_ABSOLUTE_ALARM

B_ONE_SHOT_RELATIVE_ALARM

B_OPEN_AT_END

B_OS_ERROR_BASE

B_OS_NAME_LENGTH

B_PAGE_SIZE

B_PARTIAL_READ

B_PARTIAL_WRITE

B_PARTITION_TOO_SMALL

B_PATH_NAME_LENGTH

B_PATTERN_TYPE

B_PERIODIC_ALARM

B_PERMISSION_DENIED

B_POINTER_TYPE

B_POINT_TYPE

B_POSIX_ERROR_BASE

B_PRINT_ERROR_BASE

B_PROPERTY_INFO_TYPE

B_QUERY_NON_INDEXED

B_RANDOMIZED_ANY_ADDRESS

B_RANDOMIZED_BASE_ADDRESS

B_RAW_TYPE

B_READ_AREA

B_READ_ONLY

B_READ_ONLY_DEVICE

B_READ_WRITE

B_REAL_TIME_DISPLAY_PRIORITY

B_REAL_TIME_PRIORITY

B_RECT_TYPE

B_REF_TYPE

B_RELATIVE_TIMEOUT

B_RELEASE_ALL

B_RELEASE_IF_WAITING_ONLY

B_RESOURCE_NOT_FOUND

B_RESOURCE_UNAVAILABLE

B_RESULT_NOT_REPRESENTABLE

B_RGB_32_BIT_TYPE

B_RGB_COLOR_TYPE

B_SERVER_NOT_FOUND

B_SHUTDOWN_CANCELLED

B_SHUTTING_DOWN

B_SIZE_TYPE

B_SIZE_T_TYPE

B_SSIZE_T_TYPE

B_STACK_AREA

B_STORAGE_ERROR_BASE

B_STREAM_NOT_FOUND

B_STRING_LIST_TYPE

B_STRING_TYPE

B_SUBSCRIBER_NOT_ENTERED

B_SYMBOL_TYPE_ANY

B_SYMBOL_TYPE_DATA

B_SYMBOL_TYPE_TEXT

B_SYMLINK_NODE

B_SYSTEM_TEAM

B_SYSTEM_TIMEBASE

B_TEAM_USAGE_CHILDREN

B_TEAM_USAGE_SELF

B_TIMED_OUT

B_TIMEOUT_REAL_TIME_BASE

B_TIME_TYPE

B_TOO_MANY_ARGS

B_TRANSLATION_BASE_ERROR

B_TRANSLATION_ERROR_BASE

B_UINT8_TYPE

B_UINT16_TYPE

B_UINT32_TYPE

B_UINT64_TYPE

B_UNKNOWN_EXECUTABLE

B_UNKNOWN_MIME_TYPE

B_UNSUPPORTED

B_URGENT_DISPLAY_PRIORITY

B_URGENT_PRIORITY

B_VECTOR_ICON_TYPE

B_WOULD_BLOCK

B_WRITE_AREA

B_WRITE_ONLY

B_XATTR_TYPE

CLD_CONTINUED

CLD_DUMPED

CLD_EXITED

CLD_KILLED

CLD_STOPPED

CLD_TRAPPED

CLOCAL

CLOCK_MONOTONIC

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_THREAD_CPUTIME_ID

CR0

CR1

CR2

CR3

CRDLY

CREAD

CRTSCTS

CS5

CS6

CS7

CS8

CSIZE

CSTOPB

CTSFLOW

DEAD_PROCESS

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NODATA

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EBADF

EBADMSG

EBUSY

ECANCELED

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDESTADDRREQ

EDOM

EDQUOT

EEXIST

EFAULT

EFBIG

EFPOS

EHOSTDOWN

EHOSTUNREACH

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

ELOOP

EMFILE

EMLINK

EMPTY

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOATTR

ENOBUFS

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOLCK

ENOLINK

ENOMEM

ENOMSG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTSOCK

ENOTSUP

ENOTTY

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EPERM

EPFNOSUPPORT

EPIPE

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERANGE

EROFS

ESHUTDOWN

ESIGPARM

ESPIPE

ESRCH

ESTALE

ETIME

ETIMEDOUT

ETXTBSY

EWOULDBLOCK

EXDEV

EXIT_FAILURE

EXIT_SUCCESS

FD_CLOEXEC

FD_SETSIZE

FF0

FF1

FFDLY

FILENAME_MAX

FIOCLEX

FIONBIO

FIONREAD

FIOSEEKDATA

FIOSEEKHOLE

FLUSHO

FOPEN_MAX

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLK

F_OK

F_RDLCK

F_SETFD

F_SETFL

F_SETLK

F_SETLKW

F_UNLCK

F_WRLCK

GRPQUOTA

HUPCL

ICANON

ICRNL

IEXTEN

IFF_ALLMULTI

IFF_AUTOUP

IFF_AUTO_CONFIGURED

IFF_BROADCAST

IFF_CONFIGURING

IFF_LINK

IFF_LOOPBACK

IFF_MULTICAST

IFF_NOARP

IFF_POINTOPOINT

IFF_PROMISC

IFF_SIMPLEX

IFF_UP

IFNAMSIZ

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INIT_PROCESS

INLCR

INPCK

INT_MAX

INT_MIN

IOV_MAX

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_RAW

IPPROTO_TCP

IPPROTO_UDP

IPV6_DSTOPTS

IPV6_HOPLIMIT

IPV6_HOPOPTS

IPV6_JOIN_GROUP

IPV6_LEAVE_GROUP

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_LOOP

IPV6_PKTINFO

IPV6_RECVHOPLIMIT

IPV6_RECVPKTINFO

IPV6_RTHDR

IPV6_UNICAST_HOPS

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_ADD_SOURCE_MEMBERSHIP

IP_BLOCK_SOURCE

IP_DROP_MEMBERSHIP

IP_DROP_SOURCE_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_OPTIONS

IP_RECVDSTADDR

IP_RECVOPTS

IP_RECVRETOPTS

IP_RETOPTS

IP_TOS

IP_TTL

IP_UNBLOCK_SOURCE

ISIG

ISTRIP

IUCLC

IXANY

IXOFF

IXON

LC_ALL

LC_COLLATE

LC_CTYPE

LC_MESSAGES

LC_MONETARY

LC_NUMERIC

LC_TIME

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

LOGIN_PROCESS

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PERROR

LOG_PID

LOG_PRIMASK

LOG_SERIAL

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

L_tmpnam

MADV_DONTNEED

MADV_FREE

MADV_NORMAL

MADV_RANDOM

MADV_SEQUENTIAL

MADV_WILLNEED

MAP_ANON

MAP_ANONYMOUS

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_NORESERVE

MAP_PRIVATE

MAP_SHARED

MINSIGSTKSZ

MSG_BCAST

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOF

MSG_EOR

MSG_MCAST

MSG_NOSIGNAL

MSG_OOB

MSG_PEEK

MSG_TRUNC

MSG_WAITALL

MS_ASYNC

MS_INVALIDATE

MS_SYNC

NCCS

NEW_TIME

NI_MAXHOST

NL0

NL1

NLDLY

NOFLSH

OCRNL

OFDEL

OFILL

OLCUC

OLD_TIME

ONLCR

ONLRET

ONOCR

OPOST

O_ACCMODE

O_APPEND

O_CLOEXEC

O_CREAT

O_DIRECTORY

O_DSYNC

O_EXCL

O_NOCACHE

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_NOTRAVERSE

O_RDONLY

O_RDWR

O_RSYNC

O_SYNC

O_TRUNC

O_WRONLY

PARENB

PARMRK

PARODD

PATH_MAX

PENDIN

PF_BLUETOOTH

PF_INET

PF_INET6

PF_LINK

PF_LOCAL

PF_ROUTE

PF_UNIX

PF_UNSPEC

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

POLLWRBAND

POLLWRNORM

POSIX_SPAWN_RESETIDS

POSIX_SPAWN_SETPGROUP

POSIX_SPAWN_SETSID

POSIX_SPAWN_SETSIGDEF

POSIX_SPAWN_SETSIGMASK

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PROT_EXEC

PROT_NONE

PROT_READ

PROT_WRITE

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

P_ALL

P_PGID

P_PID

RAND_MAX

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_NOFILE

RLIMIT_NOVMON

RLIMIT_STACK

RLIM_INFINITY

RLIM_NLIMITS

RTLD_DEFAULT

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NOW

RTSFLOW

RUSAGE_CHILDREN

RUSAGE_SELF

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_NOMASK

SA_ONESHOT

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SA_STACK

SCM_RIGHTS

SEEK_CUR

SEEK_END

SEEK_SET

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIOT

SIGKILL

SIGKILLTHR

SIGPIPE

SIGPOLL

SIGPROF

SIGQUIT

SIGSEGV

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SOCK_DGRAM

SOCK_RAW

SOCK_SEQPACKET

SOCK_STREAM

SOL_SOCKET

SOMAXCONN

SO_ACCEPTCONN

SO_BINDTODEVICE

SO_BROADCAST

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_NONBLOCK

SO_OOBINLINE

SO_PEERCRED

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TYPE

SO_USELOOPBACK

SS_DISABLE

SS_ONSTACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TAB0

TAB1

TAB2

TAB3

TABDLY

TCFLSH

TCGB_CTS

TCGB_DCD

TCGB_DSR

TCGB_RI

TCGETA

TCGETBITS

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_MAXSEG

TCP_NODELAY

TCP_NOOPT

TCP_NOPUSH

TCQUERYCONNECTED

TCSADRAIN

TCSAFLUSH

TCSANOW

TCSBRK

TCSETA

TCSETAF

TCSETAW

TCSETDTR

TCSETRTS

TCVTIME

TCWAITEVENT

TCXONC

TIOCCBRK

TIOCGPGRP

TIOCGWINSZ

TIOCMBIC

TIOCMBIS

TIOCMGET

TIOCMSET

TIOCM_CAR

TIOCM_CD

TIOCM_CTS

TIOCM_DSR

TIOCM_DTR

TIOCM_RI

TIOCM_RTS

TIOCSBRK

TIOCSCTTY

TIOCSPGRP

TIOCSWINSZ

TMP_MAX

TOSTOP

USER_PROCESS

USRQUOTA

UTIME_NOW

UTIME_OMIT

VEOF

VEOL

VEOL2

VERASE

VINTR

VKILL

VMIN

VQUIT

VSTART

VSTOP

VSUSP

VSWTCH

VT0

VT1

VTDLY

VTIME

WCONTINUED

WEXITED

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

W_OK

XCASE

XLOBLK

X_OK

_IOFBF

_IOLBF

_IONBF

_PC_2_SYMLINKS

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SOCK_MAXBUF

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_PC_XATTR_ENABLED

_PC_XATTR_EXISTS

_SC_ARG_MAX

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_JOB_CONTROL

_SC_MAPPED_FILES

_SC_MEMORY_PROTECTION

_SC_MONOTONIC_CLOCK

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PIPE

_SC_POLL

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SELECT

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_STACK_MIN

_SC_TIMERS

_SC_TIMER_MAX

_SC_TTY_NAME_MAX

_SC_TZNAME_MAX

_SC_UIO_MAXIOV

_SC_VERSION

Functions

CMSG_DATA⚠

CMSG_FIRSTHDR⚠

CMSG_LEN⚠

CMSG_NXTHDR⚠

CMSG_SPACE⚠

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

WCOREDUMP

WEXITSTATUS

WIFCONTINUED

WIFEXITED

WIFSIGNALED

WIFSTOPPED

WSTOPSIG

WTERMSIG

_errnop⚠

_exit⚠

_get_area_info⚠

_get_cpu_info_etc⚠

_get_image_info⚠

_get_next_area_info⚠

_get_next_image_info⚠

_get_next_port_info⚠

_get_next_sem_info⚠

_get_next_team_info⚠

_get_next_thread_info⚠

_get_port_info⚠

_get_port_message_info_etc⚠

_get_sem_info⚠

_get_team_info⚠

_get_team_usage_info⚠

_get_thread_info⚠

abort⚠

abs⚠

accept⚠

access⚠

acquire_sem⚠

acquire_sem_etc⚠

alarm⚠

area_for⚠

atexit⚠

atof⚠

atoi⚠

bind⚠

brk⚠

bsearch⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clear_caches⚠

clearenv⚠

clearerr⚠

clock_getcpuclockid⚠

clock_getres⚠

clock_gettime⚠

clock_settime⚠

clone_area⚠

close⚠

close_port⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

create_area⚠

create_port⚠

create_sem⚠

ctermid⚠

daemon⚠

debugger⚠

delete_area⚠

delete_port⚠

delete_sem⚠

dev_for_path⚠

difftime⚠

dirfd⚠

disable_debugger⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

drand48⚠

dup⚠

dup2⚠

endgrent⚠

endpwent⚠

endservent⚠

endspent⚠

endusershell⚠

endutxent⚠

erand48⚠

estimate_max_scheduling_latency⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

execvpe⚠

exit⚠

exit_thread⚠

explicit_bzero⚠

faccessat⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgets⚠

fgetspent⚠

fgetspent_r⚠

fileno⚠

find_area⚠

find_directory⚠

find_path⚠

find_path_etc⚠

find_path_for_path⚠

find_path_for_path_etc⚠

find_paths⚠

find_paths_etc⚠

find_port⚠

find_thread⚠

flock⚠

fmemopen⚠

fopen⚠

fork⚠

forkpty⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freeifaddrs⚠

freopen⚠

fs_close_attr⚠

fs_close_attr_dir⚠

fs_close_index_dir⚠

fs_close_query⚠

fs_create_index⚠

fs_fopen_attr⚠

fs_fopen_attr_dir⚠

fs_lopen_attr_dir⚠

fs_mount_volume⚠

fs_open_attr⚠

fs_open_attr_dir⚠

fs_open_index_dir⚠

fs_open_live_query⚠

fs_open_query⚠

fs_read_attr⚠

fs_read_attr_dir⚠

fs_read_index_dir⚠

fs_read_query⚠

fs_remove_attr⚠

fs_remove_index⚠

fs_rewind_attr_dir⚠

fs_rewind_index_dir⚠

fs_stat_attr⚠

fs_stat_dev⚠

fs_stat_index⚠

fs_unmount_volume⚠

fs_write_attr⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fstat⚠

fstatat⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftok⚠

ftruncate⚠

futimens⚠

fwrite⚠

gai_strerror⚠

get_area_info⚠

get_cpu_info⚠

get_cpuid⚠

get_image_info⚠

get_image_symbol⚠

get_next_area_info⚠

get_next_image_info⚠

get_next_port_info⚠

get_next_sem_info⚠

get_next_team_info⚠

get_next_thread_info⚠

get_nth_image_symbol⚠

get_path_for_dirent⚠

get_port_info⚠

get_port_message_info_etc⚠

get_pthread_thread_id⚠

get_scheduler_mode⚠

get_sem_count⚠

get_sem_info⚠

get_system_info⚠

get_team_info⚠

get_team_usage_info⚠

get_thread_info⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getdtablesize⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgrent⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgrouplist⚠

getgroups⚠

gethostname⚠

getifaddrs⚠

getitimer⚠

getline⚠

getlogin⚠

getnameinfo⚠

getopt⚠

getpagesize⚠

getpass⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getpriority⚠

getprotobyname⚠

getprotobynumber⚠

getpwent⚠

getpwnam⚠

getpwnam_r⚠

getpwuid⚠

getpwuid_r⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getservbyport⚠

getservent⚠

getsid⚠

getsockname⚠

getsockopt⚠

getspent⚠

getspent_r⚠

getspnam⚠

getspnam_r⚠

gettimeofday⚠

getuid⚠

getusershell⚠

getutxent⚠

getutxid⚠

getutxline⚠

glob⚠

globfree⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

has_data⚠

hcreate⚠

hdestroy⚠

hsearch⚠

hstrerror⚠

if_indextoname⚠

if_nametoindex⚠

initgroups⚠

ioctl⚠

is_computer_on⚠

is_computer_on_fire⚠

isalnum⚠

isalpha⚠

isatty⚠

isblank⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

issetugid⚠

isspace⚠

isupper⚠

isxdigit⚠

jrand48⚠

kill⚠

kill_team⚠

kill_thread⚠

killpg⚠

labs⚠

lchown⚠

lcong48⚠

lfind⚠

link⚠

linkat⚠

listen⚠

load_add_on⚠

load_image⚠

localeconv⚠

localtime⚠

localtime_r⚠

lockf⚠

login_tty⚠

lrand48⚠

lsearch⚠

lseek⚠

lstat⚠

lutimes⚠

madvise⚠

malloc⚠

malloc_usable_size⚠

memalign⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memrchr⚠

memset⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mkfifoat⚠

mknod⚠

mknodat⚠

mkstemp⚠

mkstemps⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

mprotect⚠

mrand48⚠

msgctl⚠

msgget⚠

msgrcv⚠

msgsnd⚠

msync⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

next_dev⚠

nice⚠

nl_langinfo⚠

nrand48⚠

on_exit_thread⚠

open⚠

open_memstream⚠

open_wmemstream⚠

openat⚠

opendir⚠

openlog⚠

openpty⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

poll⚠

popen⚠

port_buffer_size⚠

port_buffer_size_etc⚠

port_count⚠

posix_fadvise⚠

posix_fallocate⚠

posix_madvise⚠

posix_memalign⚠

posix_openpt⚠

posix_spawn⚠

posix_spawn_file_actions_addclose⚠

posix_spawn_file_actions_adddup2⚠

posix_spawn_file_actions_addopen⚠

posix_spawn_file_actions_destroy⚠

posix_spawn_file_actions_init⚠

posix_spawnattr_destroy⚠

posix_spawnattr_getflags⚠

posix_spawnattr_getpgroup⚠

posix_spawnattr_getsigdefault⚠

posix_spawnattr_getsigmask⚠

posix_spawnattr_init⚠

posix_spawnattr_setflags⚠

posix_spawnattr_setpgroup⚠

posix_spawnattr_setsigdefault⚠

posix_spawnattr_setsigmask⚠

posix_spawnp⚠

ppoll⚠

pread⚠

printf⚠

pselect⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_getguardsize⚠

pthread_attr_getstack⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getclock⚠

pthread_condattr_init⚠

pthread_condattr_setclock⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_timedlock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_init⚠

pthread_self⚠

pthread_setspecific⚠

pthread_sigmask⚠

pthread_spin_destroy⚠

pthread_spin_init⚠

pthread_spin_lock⚠

pthread_spin_trylock⚠

pthread_spin_unlock⚠

ptsname⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pututxline⚠

pwrite⚠

qsort⚠

raise⚠

rand⚠

read⚠

read_port⚠

read_port_etc⚠

readdir⚠

readdir_r⚠

The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

readv⚠

real_time_clock⚠

real_time_clock_usecs⚠

realloc⚠

realpath⚠

receive_data⚠

recv⚠

recvfrom⚠

recvmsg⚠

regcomp⚠

regerror⚠

regexec⚠

regfree⚠

release_sem⚠

release_sem_etc⚠

remove⚠

rename⚠

rename_thread⚠

renameat⚠

res_init⚠

resize_area⚠

resume_thread⚠

rewind⚠

rewinddir⚠

rmdir⚠

sbrk⚠

scanf⚠

sched_yield⚠

seed48⚠

seekdir⚠

select⚠

sem_close⚠

sem_destroy⚠

sem_init⚠

sem_open⚠

sem_post⚠

sem_trywait⚠

sem_unlink⚠

sem_wait⚠

semctl⚠

semget⚠

semop⚠

send⚠

send_data⚠

send_signal⚠

sendmsg⚠

sendto⚠

set_alarm⚠

set_area_protection⚠

set_port_owner⚠

set_real_time_clock⚠

set_scheduler_mode⚠

set_sem_owner⚠

set_signal_stack⚠

set_thread_priority⚠

setbuf⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setgrent⚠

setgroups⚠

sethostname⚠

setitimer⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setpriority⚠

setpwent⚠

setregid⚠

setreuid⚠

setrlimit⚠

setservent⚠

setsid⚠

setsockopt⚠

setspent⚠

setuid⚠

setusershell⚠

setutxent⚠

setvbuf⚠

sgetspent⚠

sgetspent_r⚠

shm_open⚠

shm_unlink⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sigtimedwait⚠

sigwait⚠

sigwaitinfo⚠

sleep⚠

snooze⚠

snooze_etc⚠

snooze_until⚠

snprintf⚠

socket⚠

socketpair⚠

spawn_thread⚠

sprintf⚠

srand⚠

srand48⚠

sscanf⚠

stat⚠

statvfs⚠

stpcpy⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsep⚠

strsignal⚠

strspn⚠

strstr⚠

strtod⚠

strtof⚠

strtok⚠

strtok_r⚠

strtol⚠

strtoul⚠

strxfrm⚠

suggest_thread_priority⚠

suspend_thread⚠

switch_sem⚠

switch_sem_etc⚠

symlink⚠

symlinkat⚠

sync⚠

sysconf⚠

syslog⚠

system⚠

system_time⚠

system_time_nsecs⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

telldir⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unload_add_on⚠

unlockpt⚠

unsetenv⚠

usleep⚠

utime⚠

utimensat⚠

utimes⚠

valloc⚠

wait⚠

wait_for_objects⚠

wait_for_objects_etc⚠

wait_for_thread⚠

waitid⚠

waitpid⚠

wcslen⚠

wcstombs⚠

wmemchr⚠

write⚠

write_port⚠

write_port_etc⚠

writev⚠

Type Definitions

ACTION

ENTRY

Elf32_Addr

Elf32_Half

Elf32_Off

Elf32_Sword

Elf32_Word

Elf64_Addr

Elf64_Half

Elf64_Off

Elf64_Sword

Elf64_Sxword

Elf64_Word

Elf64_Xword

Elf_Addr

Elf_Half

Elf_Phdr

area_id

bigtime_t

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

dev_t

fd_mask

fsblkcnt_t

fsfilcnt_t

gid_t

id_t

idtype_t

image_id

in_addr_t

in_port_t

ino_t

int8_tDeprecated

int16_tDeprecated

int32_tDeprecated

int64_tDeprecated

intmax_t

intptr_t

key_t

locale_t

mode_t

msglen_t

msgqnum_t

nanotime_t

nfds_t

nl_item

nlink_t

off_t

perform_code

pid_t

port_id

posix_spawn_file_actions_t

posix_spawnattr_t

pthread_attr_t

pthread_condattr_t

pthread_key_t

pthread_mutexattr_t

pthread_rwlockattr_t

pthread_t

ptrdiff_t

regoff_t

rlim_t

sa_family_t

sem_id

sighandler_t

sigset_t

size_t

socklen_t

speed_t

ssize_t

status_t

suseconds_t

tcflag_t

team_id

thread_func

thread_id

time_t

type_code

uid_t

uint8_tDeprecated

uint16_tDeprecated

uint32_tDeprecated

uint64_tDeprecated

uintmax_t

uintptr_t

useconds_t

wchar_t

Unions

cpuid_info