Crate libc

libc - Raw FFI bindings to platforms’ system libraries

Documentation for other platforms.

Structs

Dl_info

addrinfo

aiocb

arpd_request

arphdr

arpreq

arpreq_old

cmsghdr

cpu_set_t

dirent

dirent64

dqblk

epoll_event

fd_set

flock

flock64

fsid_t

glob_t

group

hostent

if_nameindex

ifaddrs

in6_addr

in6_rtmsg

in_addr

in_pktinfo

iovec

ip_mreq

ip_mreq_source

ip_mreqn

ipc_perm

ipv6_mreq

itimerval

lconv

linger

max_align_t

mmsghdr

mq_attr

msghdr

msginfo

msqid_ds

passwd

pollfd

protoent

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

rlimit

rlimit64

rusage

sched_param

sem_t

sembuf

servent

shmid_ds

sigaction

sigevent

siginfo_t

signalfd_siginfo

sigset_t

sigval

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_ll

sockaddr_storage

sockaddr_un

spwd

stack_t

stat

stat64

statfs

statfs64

statvfs

statvfs64

sysinfo

termios

timespec

timeval

tm

tms

utimbuf

utsname

winsize

Enums

DIR

FILE

c_void

Equivalent to C’s void type when used as a pointer.

fpos64_t

fpos_t

timezone

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

ADDR_COMPAT_LAYOUT

ADDR_LIMIT_3GB

ADDR_LIMIT_32BIT

ADDR_NO_RANDOMIZE

AF_ALG

AF_APPLETALK

AF_ASH

AF_ATMPVC

AF_ATMSVC

AF_AX25

AF_BLUETOOTH

AF_BRIDGE

AF_CAIF

AF_CAN

AF_DECnet

AF_ECONET

AF_IB

AF_IEEE802154

AF_INET

AF_INET6

AF_IPX

AF_IRDA

AF_ISDN

AF_IUCV

AF_KEY

AF_LLC

AF_LOCAL

AF_MPLS

AF_NETBEUI

AF_NETLINK

AF_NETROM

AF_NFC

AF_PACKET

AF_PHONET

AF_PPPOX

AF_RDS

AF_ROSE

AF_ROUTE

AF_RXRPC

AF_SECURITY

AF_SNA

AF_TIPC

AF_UNIX

AF_UNSPEC

AF_VSOCK

AF_WANPIPE

AF_X25

AIO_ALLDONE

AIO_CANCELED

AIO_NOTCANCELED

AI_ADDRCONFIG

AI_ALL

AI_CANONNAME

AI_NUMERICHOST

AI_NUMERICSERV

AI_PASSIVE

AI_V4MAPPED

ALT_DIGITS

AM_STR

ARPD_FLUSH

ARPD_LOOKUP

ARPD_UPDATE

ARPHRD_ADAPT

ARPHRD_APPLETLK

ARPHRD_ARCNET

ARPHRD_ASH

ARPHRD_ATM

ARPHRD_AX25

ARPHRD_BIF

ARPHRD_CAN

ARPHRD_CHAOS

ARPHRD_CISCO

ARPHRD_CSLIP

ARPHRD_CSLIP6

ARPHRD_DDCMP

ARPHRD_DLCI

ARPHRD_ECONET

ARPHRD_EETHER

ARPHRD_ETHER

ARPHRD_EUI64

ARPHRD_FCAL

ARPHRD_FCFABRIC

ARPHRD_FCPL

ARPHRD_FCPP

ARPHRD_FDDI

ARPHRD_FRAD

ARPHRD_HDLC

ARPHRD_HIPPI

ARPHRD_HWX25

ARPHRD_IEEE802

ARPHRD_IEEE802_TR

ARPHRD_IEEE1394

ARPHRD_IEEE80211

ARPHRD_IEEE80211_PRISM

ARPHRD_IEEE80211_RADIOTAP

ARPHRD_IEEE802154

ARPHRD_INFINIBAND

ARPHRD_IPDDP

ARPHRD_IPGRE

ARPHRD_IRDA

ARPHRD_LAPB

ARPHRD_LOCALTLK

ARPHRD_LOOPBACK

ARPHRD_METRICOM

ARPHRD_NETROM

ARPHRD_NONE

ARPHRD_PIMREG

ARPHRD_PPP

ARPHRD_PRONET

ARPHRD_RAWHDLC

ARPHRD_ROSE

ARPHRD_RSRVD

ARPHRD_SIT

ARPHRD_SKIP

ARPHRD_SLIP

ARPHRD_SLIP6

ARPHRD_TUNNEL

ARPHRD_TUNNEL6

ARPHRD_VOID

ARPHRD_X25

ARPOP_InREPLY

ARPOP_InREQUEST

ARPOP_NAK

ARPOP_REPLY

ARPOP_REQUEST

ARPOP_RREPLY

ARPOP_RREQUEST

ATF_COM

ATF_DONTPUB

ATF_MAGIC

ATF_NETMASK

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

AT_EMPTY_PATH

AT_FDCWD

AT_NO_AUTOMOUNT

AT_REMOVEDIR

AT_SYMLINK_FOLLOW

AT_SYMLINK_NOFOLLOW

B0

B50

B75

B110

B134

B150

B200

B300

B600

B1200

B1800

B2400

B4800

B9600

B19200

B38400

B57600

B115200

B230400

B460800

B500000

B576000

B921600

B1000000

B1152000

B1500000

B2000000

B2500000

B3000000

B3500000

B4000000

BRKINT

BS0

BS1

BSDLY

BUFSIZ

BUS_ADRALN

BUS_ADRERR

BUS_MCEERR_AO

BUS_MCEERR_AR

BUS_OBJERR

CBAUD

CBAUDEX

CLD_CONTINUED

CLD_DUMPED

CLD_EXITED

CLD_KILLED

CLD_STOPPED

CLD_TRAPPED

CLOCAL

CLOCK_BOOTTIME

CLOCK_BOOTTIME_ALARM

CLOCK_MONOTONIC

CLOCK_MONOTONIC_COARSE

CLOCK_MONOTONIC_RAW

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_REALTIME_ALARM

CLOCK_REALTIME_COARSE

CLOCK_SGI_CYCLE

CLOCK_TAI

CLOCK_THREAD_CPUTIME_ID

CLONE_CHILD_CLEARTID

CLONE_CHILD_SETTID

CLONE_DETACHED

CLONE_FILES

CLONE_FS

CLONE_IO

CLONE_NEWCGROUP

CLONE_NEWIPC

CLONE_NEWNET

CLONE_NEWNS

CLONE_NEWPID

CLONE_NEWUSER

CLONE_NEWUTS

CLONE_PARENT

CLONE_PARENT_SETTID

CLONE_PTRACE

CLONE_SETTLS

CLONE_SIGHAND

CLONE_SYSVSEM

CLONE_THREAD

CLONE_UNTRACED

CLONE_VFORK

CLONE_VM

CODESET

CPU_SETSIZE

CR0

CR1

CR2

CR3

CRDLY

CREAD

CRNCYSTR

CRTSCTS

CS5

CS6

CS7

CS8

CSIZE

CSTOPB

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

D_FMT

D_T_FMT

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADV

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDOTDOT

EDQUOT

EEXIST

EFAULT

EFBIG

EFD_CLOEXEC

EFD_NONBLOCK

EFD_SEMAPHORE

EHOSTDOWN

EHOSTUNREACH

EHWPOISON

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EISNAM

EKEYEXPIRED

EKEYREJECTED

EKEYREVOKED

EL2HLT

EL2NSYNC

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOOP

EMEDIUMTYPE

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENAVAIL

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOKEY

ENOLCK

ENOLINK

ENOMEDIUM

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTNAM

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENOTUNIQ

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPOLLERR

EPOLLET

EPOLLEXCLUSIVE

EPOLLHUP

EPOLLIN

EPOLLMSG

EPOLLONESHOT

EPOLLOUT

EPOLLPRI

EPOLLRDBAND

EPOLLRDHUP

EPOLLRDNORM

EPOLLWAKEUP

EPOLLWRBAND

EPOLLWRNORM

EPOLL_CLOEXEC

EPOLL_CTL_ADD

EPOLL_CTL_DEL

EPOLL_CTL_MOD

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EREMCHG

EREMOTE

EREMOTEIO

ERESTART

ERFKILL

EROFS

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUCLEAN

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

EXTPROC

FALLOC_FL_KEEP_SIZE

FALLOC_FL_PUNCH_HOLE

FD_CLOEXEC

FD_SETSIZE

FF0

FF1

FFDLY

FILENAME_MAX

FIOCLEX

FIONBIO

FIONREAD

FLUSHO

FOPEN_MAX

F_ADD_SEALS

F_CANCELLK

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLEASE

F_GETLK

F_GETOWN

F_GETPIPE_SZ

F_GET_SEALS

F_LOCK

F_NOTIFY

F_OFD_GETLK

F_OFD_SETLK

F_OFD_SETLKW

F_OK

F_SEAL_GROW

F_SEAL_SEAL

F_SEAL_SHRINK

F_SEAL_WRITE

F_SETFD

F_SETFL

F_SETLEASE

F_SETLK

F_SETLKW

F_SETOWN

F_SETPIPE_SZ

F_TEST

F_TLOCK

F_ULOCK

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRPQUOTA

HUPCL

ICANON

ICRNL

IEXTEN

IFF_ALLMULTI

IFF_AUTOMEDIA

IFF_BROADCAST

IFF_DEBUG

IFF_DYNAMIC

IFF_LOOPBACK

IFF_MASTER

IFF_MULTICAST

IFF_NOARP

IFF_NOTRAILERS

IFF_POINTOPOINT

IFF_PORTSEL

IFF_PROMISC

IFF_RUNNING

IFF_SLAVE

IFF_UP

IFNAMSIZ

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

IMAXBEL

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INLCR

INPCK

INT_MAX

INT_MIN

IPC_CREAT

IPC_EXCL

IPC_INFO

IPC_NOWAIT

IPC_PRIVATE

IPC_RMID

IPC_SET

IPC_STAT

IPDEFTTL

IPOPT_CLASS_MASK

IPOPT_CONTROL

IPOPT_COPY

IPOPT_END

IPOPT_EOL

IPOPT_LSRR

IPOPT_MEASUREMENT

IPOPT_MINOFF

IPOPT_NOOP

IPOPT_NOP

IPOPT_NUMBER_MASK

IPOPT_OFFSET

IPOPT_OLEN

IPOPT_OPTVAL

IPOPT_RA

IPOPT_RESERVED1

IPOPT_RESERVED2

IPOPT_RR

IPOPT_SEC

IPOPT_SID

IPOPT_SSRR

IPOPT_TIMESTAMP

IPOPT_TS

IPOPT_TS_PRESPEC

IPOPT_TS_TSANDADDR

IPOPT_TS_TSONLY

IPPROTO_AH

IP6 Auth Header

IPPROTO_BEETPH

IPPROTO_COMP

IP Payload Comp. Protocol

IPPROTO_DCCP

DCCP

IPPROTO_DSTOPTS

IP6 destination option

IPPROTO_EGP

exterior gateway protocol

IPPROTO_ENCAP

encapsulation header

IPPROTO_ESP

IP6 Encap Sec. Payload

IPPROTO_FRAGMENT

IP6 fragmentation header

IPPROTO_GRE

General Routing Encap.

IPPROTO_HOPOPTS

Hop-by-hop option header

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IDP

xns idp

IPPROTO_IGMP

group mgmt protocol

IPPROTO_IP

IPPROTO_IPIP

for compatibility

IPPROTO_IPV6

IPPROTO_MAX

IPPROTO_MH

IPPROTO_MPLS

IPPROTO_MTP

IPPROTO_NONE

IP6 no next header

IPPROTO_PIM

Protocol indep. multicast

IPPROTO_PUP

pup

IPPROTO_RAW

raw IP packet

IPPROTO_ROUTING

IP6 routing header

IPPROTO_RSVP

resource reservation

IPPROTO_SCTP

SCTP

IPPROTO_TCP

IPPROTO_TP

tp-4 w/ class negotiation

IPPROTO_UDP

IPPROTO_UDPLITE

IPTOS_ECN_CE

IPTOS_ECN_ECT0

IPTOS_ECN_ECT1

IPTOS_ECN_MASK

IPTOS_LOWDELAY

IPTOS_MINCOST

IPTOS_PREC_CRITIC_ECP

IPTOS_PREC_FLASH

IPTOS_PREC_FLASHOVERRIDE

IPTOS_PREC_IMMEDIATE

IPTOS_PREC_INTERNETCONTROL

IPTOS_PREC_NETCONTROL

IPTOS_PREC_PRIORITY

IPTOS_PREC_ROUTINE

IPTOS_RELIABILITY

IPTOS_THROUGHPUT

IPV6_2292DSTOPTS

IPV6_2292HOPLIMIT

IPV6_2292HOPOPTS

IPV6_2292PKTINFO

IPV6_2292PKTOPTIONS

IPV6_2292RTHDR

IPV6_ADDRFORM

IPV6_ADDR_PREFERENCES

IPV6_ADD_MEMBERSHIP

IPV6_AUTHHDR

IPV6_AUTOFLOWLABEL

IPV6_CHECKSUM

IPV6_DONTFRAG

IPV6_DROP_MEMBERSHIP

IPV6_DSTOPTS

IPV6_HDRINCL

IPV6_HOPLIMIT

IPV6_HOPOPTS

IPV6_IPSEC_POLICY

IPV6_JOIN_ANYCAST

IPV6_LEAVE_ANYCAST

IPV6_MINHOPCOUNT

IPV6_MTU

IPV6_MTU_DISCOVER

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_LOOP

IPV6_NEXTHOP

IPV6_ORIGDSTADDR

IPV6_PATHMTU

IPV6_PKTINFO

IPV6_PMTUDISC_DO

IPV6_PMTUDISC_DONT

IPV6_PMTUDISC_INTERFACE

IPV6_PMTUDISC_OMIT

IPV6_PMTUDISC_PROBE

IPV6_PMTUDISC_WANT

IPV6_PREFER_SRC_CGA

IPV6_PREFER_SRC_COA

IPV6_PREFER_SRC_HOME

IPV6_PREFER_SRC_NONCGA

IPV6_PREFER_SRC_PUBLIC

IPV6_PREFER_SRC_PUBTMP_DEFAULT

IPV6_PREFER_SRC_TMP

IPV6_RECVDSTOPTS

IPV6_RECVERR

IPV6_RECVHOPLIMIT

IPV6_RECVHOPOPTS

IPV6_RECVORIGDSTADDR

IPV6_RECVPATHMTU

IPV6_RECVPKTINFO

IPV6_RECVRTHDR

IPV6_RECVTCLASS

IPV6_ROUTER_ALERT

IPV6_RTHDR

IPV6_RTHDRDSTOPTS

IPV6_RTHDR_LOOSE

IPV6_RTHDR_STRICT

IPV6_TCLASS

IPV6_TRANSPARENT

IPV6_UNICAST_HOPS

IPV6_UNICAST_IF

IPV6_V6ONLY

IPV6_XFRM_POLICY

IPVERSION

IP_ADD_MEMBERSHIP

IP_ADD_SOURCE_MEMBERSHIP

IP_BIND_ADDRESS_NO_PORT

IP_BLOCK_SOURCE

IP_CHECKSUM

IP_DEFAULT_MULTICAST_LOOP

IP_DEFAULT_MULTICAST_TTL

IP_DROP_MEMBERSHIP

IP_DROP_SOURCE_MEMBERSHIP

IP_FREEBIND

IP_HDRINCL

IP_IPSEC_POLICY

IP_MINTTL

IP_MSFILTER

IP_MTU

IP_MTU_DISCOVER

IP_MULTICAST_ALL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_NODEFRAG

IP_OPTIONS

IP_ORIGDSTADDR

IP_PASSSEC

IP_PKTINFO

IP_PKTOPTIONS

IP_PMTUDISC_DO

IP_PMTUDISC_DONT

IP_PMTUDISC_INTERFACE

IP_PMTUDISC_OMIT

IP_PMTUDISC_PROBE

IP_PMTUDISC_WANT

IP_RECVERR

IP_RECVOPTS

IP_RECVORIGDSTADDR

IP_RECVTOS

IP_RECVTTL

IP_RETOPTS

IP_ROUTER_ALERT

IP_TOS

IP_TRANSPARENT

IP_TTL

IP_UNBLOCK_SOURCE

IP_UNICAST_IF

IP_XFRM_POLICY

ISIG

ISTRIP

ITIMER_PROF

ITIMER_REAL

ITIMER_VIRTUAL

IXANY

IXOFF

IXON

LC_ALL

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_TIME

LC_TIME_MASK

LIO_NOP

LIO_NOWAIT

LIO_READ

LIO_WAIT

LIO_WRITE

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

LOG_ALERT

LOG_AUTH

LOG_AUTHPRIV

LOG_CONS

LOG_CRIT

LOG_CRON

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_FTP

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NFACILITIES

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PERROR

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

L_tmpnam

MADV_DODUMP

MADV_DOFORK

MADV_DONTDUMP

MADV_DONTFORK

MADV_DONTNEED

MADV_FREE

MADV_HUGEPAGE

MADV_HWPOISON

MADV_MERGEABLE

MADV_NOHUGEPAGE

MADV_NORMAL

MADV_RANDOM

MADV_REMOVE

MADV_SEQUENTIAL

MADV_SOFT_OFFLINE

MADV_UNMERGEABLE

MADV_WILLNEED

MAP_ANON

MAP_ANONYMOUS

MAP_DENYWRITE

MAP_EXECUTABLE

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_GROWSDOWN

MAP_HUGETLB

MAP_LOCKED

MAP_NONBLOCK

MAP_NORESERVE

MAP_POPULATE

MAP_PRIVATE

MAP_SHARED

MAP_STACK

MAP_TYPE

MAXTTL

MAX_ADDR_LEN

MAX_IPOPTLEN

MCAST_BLOCK_SOURCE

MCAST_EXCLUDE

MCAST_INCLUDE

MCAST_JOIN_GROUP

MCAST_JOIN_SOURCE_GROUP

MCAST_LEAVE_GROUP

MCAST_LEAVE_SOURCE_GROUP

MCAST_MSFILTER

MCAST_UNBLOCK_SOURCE

MCL_CURRENT

MCL_FUTURE

MINSIGSTKSZ

MMAP_PAGE_ZERO

MNT_DETACH

MNT_EXPIRE

MNT_FORCE

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

MREMAP_FIXED

MREMAP_MAYMOVE

MSG_CMSG_CLOEXEC

MSG_CONFIRM

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOR

MSG_ERRQUEUE

MSG_EXCEPT

MSG_FASTOPEN

MSG_FIN

MSG_INFO

MSG_MORE

MSG_NOERROR

MSG_NOSIGNAL

MSG_OOB

MSG_PEEK

MSG_RST

MSG_STAT

MSG_SYN

MSG_TRUNC

MSG_WAITALL

MSG_WAITFORONE

MS_ACTIVE

MS_ASYNC

MS_BIND

MS_DIRSYNC

MS_INVALIDATE

MS_I_VERSION

MS_KERNMOUNT

MS_LAZYTIME

MS_MANDLOCK

MS_MGC_MSK

MS_MGC_VAL

MS_MOVE

MS_NOATIME

MS_NODEV

MS_NODIRATIME

MS_NOEXEC

MS_NOSUID

MS_NOUSER

MS_POSIXACL

MS_PRIVATE

MS_RDONLY

MS_REC

MS_RELATIME

MS_REMOUNT

MS_RMT_MASK

MS_SHARED

MS_SILENT

MS_SLAVE

MS_STRICTATIME

MS_SYNC

MS_SYNCHRONOUS

MS_UNBINDABLE

NCCS

NI_DGRAM

NI_MAXHOST

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSERV

NL0

NL1

NLDLY

NOEXPR

NOFLSH

NOSTR

OCRNL

OFDEL

OFILL

OLCUC

ONLCR

ONLRET

ONOCR

OPOST

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECT

O_DIRECTORY

O_DSYNC

O_EXCL

O_EXEC

O_NDELAY

O_NOATIME

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_PATH

O_RDONLY

O_RDWR

O_RSYNC

O_SEARCH

O_SYNC

O_TMPFILE

O_TRUNC

O_WRONLY

PARENB

PARMRK

PARODD

PATH_MAX

PENDIN

PF_ALG

PF_APPLETALK

PF_ASH

PF_ATMPVC

PF_ATMSVC

PF_AX25

PF_BLUETOOTH

PF_BRIDGE

PF_CAIF

PF_CAN

PF_DECnet

PF_ECONET

PF_IB

PF_IEEE802154

PF_INET

PF_INET6

PF_IPX

PF_IRDA

PF_ISDN

PF_IUCV

PF_KEY

PF_LLC

PF_LOCAL

PF_MPLS

PF_NETBEUI

PF_NETLINK

PF_NETROM

PF_NFC

PF_PACKET

PF_PHONET

PF_PPPOX

PF_RDS

PF_ROSE

PF_ROUTE

PF_RXRPC

PF_SECURITY

PF_SNA

PF_TIPC

PF_UNIX

PF_UNSPEC

PF_VSOCK

PF_WANPIPE

PF_X25

PIPE_BUF

PM_STR

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDHUP

POLLRDNORM

POLLWRBAND

POLLWRNORM

POSIX_FADV_DONTNEED

POSIX_FADV_NOREUSE

POSIX_FADV_NORMAL

POSIX_FADV_RANDOM

POSIX_FADV_SEQUENTIAL

POSIX_FADV_WILLNEED

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PROT_EXEC

PROT_GROWSDOWN

PROT_GROWSUP

PROT_NONE

PROT_READ

PROT_WRITE

PR_CAPBSET_DROP

PR_CAPBSET_READ

PR_CAP_AMBIENT

PR_CAP_AMBIENT_CLEAR_ALL

PR_CAP_AMBIENT_IS_SET

PR_CAP_AMBIENT_LOWER

PR_CAP_AMBIENT_RAISE

PR_ENDIAN_BIG

PR_ENDIAN_LITTLE

PR_ENDIAN_PPC_LITTLE

PR_FPEMU_NOPRINT

PR_FPEMU_SIGFPE

PR_FP_EXC_ASYNC

PR_FP_EXC_DISABLED

PR_FP_EXC_DIV

PR_FP_EXC_INV

PR_FP_EXC_NONRECOV

PR_FP_EXC_OVF

PR_FP_EXC_PRECISE

PR_FP_EXC_RES

PR_FP_EXC_SW_ENABLE

PR_FP_EXC_UND

PR_FP_MODE_FR

PR_FP_MODE_FRE

PR_GET_CHILD_SUBREAPER

PR_GET_DUMPABLE

PR_GET_ENDIAN

PR_GET_FPEMU

PR_GET_FPEXC

PR_GET_FP_MODE

PR_GET_KEEPCAPS

PR_GET_NAME

PR_GET_NO_NEW_PRIVS

PR_GET_PDEATHSIG

PR_GET_SECCOMP

PR_GET_SECUREBITS

PR_GET_THP_DISABLE

PR_GET_TID_ADDRESS

PR_GET_TIMERSLACK

PR_GET_TIMING

PR_GET_TSC

PR_GET_UNALIGN

PR_MCE_KILL

PR_MCE_KILL_CLEAR

PR_MCE_KILL_DEFAULT

PR_MCE_KILL_EARLY

PR_MCE_KILL_GET

PR_MCE_KILL_LATE

PR_MCE_KILL_SET

PR_MPX_DISABLE_MANAGEMENT

PR_MPX_ENABLE_MANAGEMENT

PR_SET_CHILD_SUBREAPER

PR_SET_DUMPABLE

PR_SET_ENDIAN

PR_SET_FPEMU

PR_SET_FPEXC

PR_SET_FP_MODE

PR_SET_KEEPCAPS

PR_SET_MM

PR_SET_MM_ARG_END

PR_SET_MM_ARG_START

PR_SET_MM_AUXV

PR_SET_MM_BRK

PR_SET_MM_END_CODE

PR_SET_MM_END_DATA

PR_SET_MM_ENV_END

PR_SET_MM_ENV_START

PR_SET_MM_EXE_FILE

PR_SET_MM_MAP

PR_SET_MM_MAP_SIZE

PR_SET_MM_START_BRK

PR_SET_MM_START_CODE

PR_SET_MM_START_DATA

PR_SET_MM_START_STACK

PR_SET_NAME

PR_SET_NO_NEW_PRIVS

PR_SET_PDEATHSIG

PR_SET_PTRACER

PR_SET_SECCOMP

PR_SET_SECUREBITS

PR_SET_THP_DISABLE

PR_SET_TIMERSLACK

PR_SET_TIMING

PR_SET_TSC

PR_SET_UNALIGN

PR_TASK_PERF_EVENTS_DISABLE

PR_TASK_PERF_EVENTS_ENABLE

PR_TIMING_STATISTICAL

PR_TIMING_TIMESTAMP

PR_TSC_ENABLE

PR_TSC_SIGSEGV

PR_UNALIGN_NOPRINT

PR_UNALIGN_SIGBUS

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

PTRACE_ATTACH

PTRACE_CONT

PTRACE_DETACH

PTRACE_EVENT_CLONE

PTRACE_EVENT_EXEC

PTRACE_EVENT_EXIT

PTRACE_EVENT_FORK

PTRACE_EVENT_SECCOMP

PTRACE_EVENT_VFORK

PTRACE_EVENT_VFORK_DONE

PTRACE_GETEVENTMSG

PTRACE_GETFPREGS

PTRACE_GETFPXREGS

PTRACE_GETREGS

PTRACE_GETREGSET

PTRACE_GETSIGINFO

PTRACE_INTERRUPT

PTRACE_KILL

PTRACE_LISTEN

PTRACE_O_EXITKILL

PTRACE_O_MASK

PTRACE_O_SUSPEND_SECCOMP

PTRACE_O_TRACECLONE

PTRACE_O_TRACEEXEC

PTRACE_O_TRACEEXIT

PTRACE_O_TRACEFORK

PTRACE_O_TRACESECCOMP

PTRACE_O_TRACESYSGOOD

PTRACE_O_TRACEVFORK

PTRACE_O_TRACEVFORKDONE

PTRACE_PEEKDATA

PTRACE_PEEKSIGINFO

PTRACE_PEEKTEXT

PTRACE_PEEKUSER

PTRACE_POKEDATA

PTRACE_POKETEXT

PTRACE_POKEUSER

PTRACE_SEIZE

PTRACE_SETFPREGS

PTRACE_SETFPXREGS

PTRACE_SETOPTIONS

PTRACE_SETREGS

PTRACE_SETREGSET

PTRACE_SETSIGINFO

PTRACE_SINGLESTEP

PTRACE_SYSCALL

PTRACE_TRACEME

P_ALL

P_PGID

P_PID

QFMT_VFS_OLD

QFMT_VFS_V0

QFMT_VFS_V1

QIF_ALL

QIF_BLIMITS

QIF_BTIME

QIF_ILIMITS

QIF_INODES

QIF_ITIME

QIF_LIMITS

QIF_SPACE

QIF_TIMES

QIF_USAGE

Q_GETFMT

Q_GETINFO

Q_GETQUOTA

Q_QUOTAOFF

Q_QUOTAON

Q_SETINFO

Q_SETQUOTA

Q_SYNC

RADIXCHAR

RAND_MAX

RB_AUTOBOOT

RB_DISABLE_CAD

RB_ENABLE_CAD

RB_HALT_SYSTEM

RB_KEXEC

RB_POWER_OFF

RB_SW_SUSPEND

READ_IMPLIES_EXEC

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_LOCKS

RLIMIT_MEMLOCK

RLIMIT_MSGQUEUE

RLIMIT_NICE

RLIMIT_NLIMITS

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_RTPRIO

RLIMIT_SIGPENDING

RLIMIT_STACK

RLIM_INFINITY

RLIM_NLIMITS

RLIM_SAVED_CUR

RLIM_SAVED_MAX

RTLD_DEFAULT

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NEXT

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

RUSAGE_CHILDREN

RUSAGE_SELF

RUSAGE_THREAD

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SCHED_BATCH

SCHED_FIFO

SCHED_IDLE

SCHED_OTHER

SCHED_RR

SCM_CREDENTIALS

SCM_RIGHTS

SCM_TIMESTAMP

SEEK_CUR

SEEK_END

SEEK_SET

SEM_FAILED

SFD_CLOEXEC

SFD_NONBLOCK

SHM_EXEC

SHM_HUGETLB

SHM_LOCK

SHM_NORESERVE

SHM_R

SHM_RDONLY

SHM_REMAP

SHM_RND

SHM_UNLOCK

SHM_W

SHORT_INODE

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPOLL

SIGPROF

SIGPWR

SIGQUIT

SIGSEGV

SIGSTKFLT

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SI_LOAD_SHIFT

SOCK_CLOEXEC

SOCK_DGRAM

SOCK_NONBLOCK

SOCK_RAW

SOCK_RDM

SOCK_SEQPACKET

SOCK_STREAM

SOL_AAL

SOL_ALG

SOL_ATM

SOL_BLUETOOTH

SOL_DCCP

SOL_DECNET

SOL_ICMPV6

SOL_IP

SOL_IPV6

SOL_IRDA

SOL_LLC

SOL_NETBEUI

SOL_NETLINK

SOL_PACKET

SOL_RAW

SOL_SOCKET

SOL_TCP

SOL_TIPC

SOL_UDP

SOL_X25

SOMAXCONN

SO_ACCEPTCONN

SO_BINDTODEVICE

SO_BROADCAST

SO_BUSY_POLL

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_MARK

SO_OOBINLINE

SO_PEEK_OFF

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_RXQ_OVFL

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TYPE

SPLICE_F_GIFT

SPLICE_F_MORE

SPLICE_F_MOVE

SPLICE_F_NONBLOCK

SS_DISABLE

SS_ONSTACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

STICKY_TIMEOUTS

ST_APPEND

ST_IMMUTABLE

ST_MANDLOCK

ST_NOATIME

ST_NODEV

ST_NODIRATIME

ST_NOEXEC

ST_NOSUID

ST_RDONLY

ST_SYNCHRONOUS

ST_WRITE

SYNC_FILE_RANGE_WAIT_AFTER

SYNC_FILE_RANGE_WAIT_BEFORE

SYNC_FILE_RANGE_WRITE

SYS_gettid

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TAB0

TAB1

TAB2

TAB3

TABDLY

TCFLSH

TCGETA

TCGETS

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_CC_INFO

TCP_CONGESTION

TCP_CORK

TCP_DEFER_ACCEPT

TCP_FASTOPEN

TCP_INFO

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_LINGER2

TCP_MAXSEG

TCP_MD5SIG

TCP_NODELAY

TCP_NOTSENT_LOWAT

TCP_QUEUE_SEQ

TCP_QUICKACK

TCP_REPAIR

TCP_REPAIR_OPTIONS

TCP_REPAIR_QUEUE

TCP_SAVED_SYN

TCP_SAVE_SYN

TCP_SYNCNT

TCP_THIN_DUPACK

TCP_THIN_LINEAR_TIMEOUTS

TCP_TIMESTAMP

TCP_USER_TIMEOUT

TCP_WINDOW_CLAMP

TCSADRAIN

TCSAFLUSH

TCSANOW

TCSBRK

TCSETA

TCSETAF

TCSETAW

TCSETS

TCSETSF

TCSETSW

TCXONC

THOUSEP

TIMER_ABSTIME

TIOCCONS

TIOCEXCL

TIOCGPGRP

TIOCGSERIAL

TIOCGSOFTCAR

TIOCGWINSZ

TIOCINQ

TIOCLINUX

TIOCMBIC

TIOCMBIS

TIOCMGET

TIOCMSET

TIOCM_CAR

TIOCM_CD

TIOCM_CTS

TIOCM_DSR

TIOCM_DTR

TIOCM_LE

TIOCM_RI

TIOCM_RNG

TIOCM_RTS

TIOCM_SR

TIOCM_ST

TIOCNXCL

TIOCOUTQ

TIOCSCTTY

TIOCSPGRP

TIOCSSOFTCAR

TIOCSTI

TIOCSWINSZ

TMP_MAX

TOSTOP

T_FMT

T_FMT_AMPM

UIO_MAXIOV

UMOUNT_NOFOLLOW

USRQUOTA

UTIME_NOW

UTIME_OMIT

VDISCARD

VEOF

VEOL

VEOL2

VERASE

VINTR

VKILL

VLNEXT

VMIN

VQUIT

VREPRINT

VSTART

VSTOP

VSUSP

VSWTC

VT0

VT1

VTDLY

VTIME

VWERASE

WCONTINUED

WEXITED

WHOLE_SECONDS

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

W_OK

XATTR_CREATE

XATTR_REPLACE

XTABS

X_OK

YESEXPR

YESSTR

_IOFBF

_IOLBF

_IONBF

_PC_2_SYMLINKS

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SOCK_MAXBUF

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_POSIX_VDISABLE

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_NZERO

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SS_REPL_MAX

_SC_STREAMS

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_ROBUST_PRIO_INHERIT

_SC_THREAD_ROBUST_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_UIO_MAXIOV

_SC_V6_ILP32_OFF32

_SC_V6_ILP32_OFFBIG

_SC_V6_LP64_OFF64

_SC_V6_LPBIG_OFFBIG

_SC_V7_ILP32_OFF32

_SC_V7_ILP32_OFFBIG

_SC_V7_LP64_OFF64

_SC_V7_LPBIG_OFFBIG

_SC_VERSION

_SC_XBS5_ILP32_OFF32

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_LP64_OFF64

_SC_XBS5_LPBIG_OFFBIG

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_SC_XOPEN_XPG2

_SC_XOPEN_XPG3

_SC_XOPEN_XPG4

__SIZEOF_PTHREAD_CONDATTR_T

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCKATTR_T

__SIZEOF_PTHREAD_RWLOCK_T

__WALL

__WCLONE

__WNOTHREAD

Functions

CMSG_DATA⚠

CMSG_FIRSTHDR⚠

CMSG_LEN⚠

CMSG_NXTHDR⚠

CMSG_SPACE⚠

CPU_CLR⚠

CPU_EQUAL⚠

CPU_ISSET⚠

CPU_SET⚠

CPU_ZERO⚠

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

IPOPT_CLASS

IPOPT_COPIED

IPOPT_NUMBER

IPTOS_ECN

QCMD

SIGRTMAX

SIGRTMIN

WCOREDUMP

WEXITSTATUS

WIFCONTINUED

WIFEXITED

WIFSIGNALED

WIFSTOPPED

WSTOPSIG

WTERMSIG

W_EXITCODE

W_STOPCODE

__errno_location⚠

_exit⚠

abort⚠

abs⚠

accept⚠

accept4⚠

access⚠

acct⚠

alarm⚠

atexit⚠

atof⚠

atoi⚠

bind⚠

brk⚠

bsearch⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clearenv⚠

clearerr⚠

clock_getcpuclockid⚠

clock_getres⚠

clock_gettime⚠

clock_settime⚠

close⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

creat64⚠

difftime⚠

dirfd⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

dup3⚠

duplocale⚠

endpwent⚠

endservent⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

execvpe⚠

exit⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdatasync⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fexecve⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgetpos64⚠

fgets⚠

fileno⚠

flock⚠

fmemopen⚠

fopen⚠

fopen64⚠

fork⚠

forkpty⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freeifaddrs⚠

freelocale⚠

freopen⚠

freopen64⚠

fscanf⚠

fseek⚠

fseeko⚠

fseeko64⚠

fsetpos⚠

fsetpos64⚠

fstat⚠

fstat64⚠

fstatat⚠

fstatat64⚠

fstatfs⚠

fstatfs64⚠

fstatvfs⚠

fstatvfs64⚠

fsync⚠

ftell⚠

ftello⚠

ftello64⚠

ftruncate⚠

ftruncate64⚠

futimens⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getdomainname⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgroups⚠

gethostname⚠

getifaddrs⚠

getline⚠

getloadavg⚠

getlogin⚠

getnameinfo⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getpriority⚠

getprotobyname⚠

getprotobynumber⚠

getpwent⚠

getpwnam⚠

getpwuid⚠

getresgid⚠

getresuid⚠

getrlimit⚠

getrlimit64⚠

getrusage⚠

getservbyname⚠

getservbyport⚠

getservent⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

glob⚠

globfree⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

hstrerror⚠

if_freenameindex⚠

if_indextoname⚠

if_nameindex⚠

if_nametoindex⚠

ioctl⚠

isalnum⚠

isalpha⚠

isatty⚠

isblank⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kill⚠

killpg⚠

labs⚠

lchown⚠

link⚠

linkat⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lockf⚠

login_tty⚠

lseek⚠

lseek64⚠

lstat⚠

lstat64⚠

madvise⚠

major⚠

makedev

malloc⚠

memalign⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memrchr⚠

memset⚠

mincore⚠

minor⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mkfifoat⚠

mknod⚠

mknodat⚠

mkostemp⚠

mkostemps⚠

mkstemp⚠

mkstemps⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

mmap64⚠

mprotect⚠

mremap⚠

msync⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

newlocale⚠

nice⚠

nl_langinfo⚠

nl_langinfo_l⚠

open⚠

open64⚠

open_memstream⚠

open_wmemstream⚠

openat⚠

openat64⚠

opendir⚠

openlog⚠

openpty⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

pipe2⚠

poll⚠

posix_fadvise⚠

posix_fadvise64⚠

posix_fallocate⚠

posix_madvise⚠

posix_memalign⚠

posix_openpt⚠

pread⚠

pread64⚠

preadv⚠

preadv64⚠

printf⚠

pselect⚠

pthread_attr_destroy⚠

pthread_attr_getstack⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getclock⚠

pthread_condattr_init⚠

pthread_condattr_setclock⚠

pthread_condattr_setpshared⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_getattr_np⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_setpshared⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_getpshared⚠

pthread_rwlockattr_init⚠

pthread_rwlockattr_setpshared⚠

pthread_self⚠

pthread_setspecific⚠

ptsname⚠

ptsname_r⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pwrite⚠

pwrite64⚠

pwritev⚠

pwritev64⚠

qsort⚠

raise⚠

rand⚠

read⚠

readdir⚠

readdir64⚠

readdir64_r⚠

readdir_r⚠

The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

readv⚠

realloc⚠

realpath⚠

recv⚠

recvfrom⚠

recvmmsg⚠

recvmsg⚠

remove⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

sbrk⚠

scanf⚠

sched_yield⚠

seekdir⚠

select⚠

sem_destroy⚠

sem_init⚠

sem_post⚠

sem_trywait⚠

sem_wait⚠

send⚠

sendmmsg⚠

sendmsg⚠

sendto⚠

setbuf⚠

setdomainname⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setgroups⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setpriority⚠

setpwent⚠

setregid⚠

setresgid⚠

setresuid⚠

setreuid⚠

setrlimit⚠

setrlimit64⚠

setservent⚠

setsid⚠

setsockopt⚠

setuid⚠

setvbuf⚠

shm_open⚠

shm_unlink⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

srand⚠

sscanf⚠

stat⚠

stat64⚠

statfs⚠

statfs64⚠

statvfs⚠

statvfs64⚠

stpcpy⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsignal⚠

strspn⚠

strstr⚠

strtod⚠

strtof⚠

strtok⚠

strtok_r⚠

strtol⚠

strtoul⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sync⚠

sysconf⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

telldir⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpfile64⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

truncate64⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

uselocale⚠

usleep⚠

utime⚠

utimensat⚠

utimes⚠

vfork⚠Deprecated

wait⚠

wait4⚠

waitid⚠

waitpid⚠

wcslen⚠

wcstombs⚠

wmemchr⚠

write⚠

writev⚠

Type Definitions

blkcnt64_t

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

dev_t

fsblkcnt_t

fsfilcnt_t

gid_t

id_t

idtype_t

in_addr_t

in_port_t

ino64_t

ino_t

int8_tDeprecated

int16_tDeprecated

int32_tDeprecated

int64_tDeprecated

intmax_t

intptr_t

key_t

locale_t

loff_t

mode_t

mqd_t

msglen_t

msgqnum_t

nfds_t

nl_item

nlink_t

off64_t

off_t

pid_t

pthread_key_t

pthread_t

ptrdiff_t

rlim64_t

rlim_t

sa_family_t

shmatt_t

sighandler_t

size_t

socklen_t

speed_t

ssize_t

suseconds_t

tcflag_t

time_t

timer_t

uid_t

uint8_tDeprecated

uint16_tDeprecated

uint32_tDeprecated

uint64_tDeprecated

uintmax_t

uintptr_t

useconds_t

wchar_t