Crate libc

libc - Raw FFI bindings to platforms’ system libraries

Documentation for other platforms.

Structs

Dl_info

Elf32_Auxinfo

Elf32_Phdr

Elf64_Phdr

__c_anonymous__timer

__c_anonymous_filestat

__c_anonymous_ph

__c_anonymous_pthread_barrier

__c_anonymous_pthread_barrierattr

__c_anonymous_pthread_spinlock

__c_anonymous_sem

__c_anonymous_stailq_entry_devstat

_sem

_umtx_time

accept_filter_arg

addrinfo

aiocb

arphdr

au_mask_t

au_tid_t

auditinfo_t

bintime

bpf_dltlist

bpf_hdr

bpf_insn

bpf_program

bpf_stat

bpf_version

cap_rights_t

clockinfo

cmsgcred

cmsghdr

cpuset_t

device_selection

devinfo

devstat

devstat_match

devstat_match_table

dirent

dl_phdr_info

eui64

fd_set

fhandle

fid

filestat

filestat_list

flock

fsid_t

glob_t

group

hostent

if_announcemsghdr

if_clonereq

if_data

if_msghdr

if_msghdrl

if_nameindex

ifa_msghdr

ifa_msghdrl

ifaddrs

ifaliasreq

ifconf

ifdownreason

ifdrv

ifi2creq

ifma_msghdr

ifmediareq

ifmib_iso_8802_3

ifmibdata

ifreq

ifreq_buffer

ifrsshash

ifrsskey

ifstat

in6_addr

in6_pktinfo

in_addr

iovec

ip_mreq

ip_mreq_source

ip_mreqn

ipc_perm

ipv6_mreq

itimerspec

itimerval

jail

kevent

kinfo_proc

kinfo_vmentry

kvm_nlist

kvm_swap

lconv

linger

mmsghdr

mq_attr

msghdr

msqid_ds

nlist

ntptimeval

oifaliasreq

9.x compat

pargs

Used to hold a copy of the command line, if it had a sane length.

passwd

pollfd

priority

procstat

protoent

ptrace_coredump

ptrace_io_desc

ptrace_lwpinfo

ptrace_sc_ret

ptrace_vm_entry

ptsstat

regex_t

regmatch_t

rlimit

rtprio

rusage

rusage_ext

sched_param

sembuf

semid_ds

servent

sf_hdtr

shm_largepage_conf

shmid_ds

shmstat

sigaction

sigevent

siginfo_t

sigset_t

sigval

sockaddr

sockaddr_dl

sockaddr_in

sockaddr_in6

sockaddr_storage

sockaddr_un

sockcred

sockcred2

sockstat

spacectl_range

stack_t

stat

statfs

statvfs

tcp_fastopen

tcp_function_set

termios

timespec

timeval

timex

tm

tms

ucond

umutex

utimbuf

utmpx

utsname

uuid

vmtotal

vnstat

winsize

xucred

Enums

DIR

FILE

c_void

Equivalent to C’s void type when used as a pointer.

devstat_match_flags

devstat_metric

devstat_priority

devstat_select_mode

devstat_support_flags

devstat_tag_type

devstat_trans_flags

devstat_type_flags

dot3Vendors

fpos_t

timezone

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

AF_APPLETALK

AF_ARP

AF_ATM

AF_BLUETOOTH

AF_CCITT

AF_CHAOS

AF_CNT

AF_COIP

AF_DATAKIT

AF_DECnet

AF_DLI

AF_E164

AF_ECMA

AF_HYLINK

AF_IEEE80211

AF_IMPLINK

AF_INET

AF_INET6

AF_INET6_SDP

AF_INET_SDP

AF_IPX

AF_ISDN

AF_ISO

AF_LAT

AF_LINK

AF_LOCAL

AF_NATM

AF_NETBIOS

AF_NETGRAPH

AF_OSI

AF_PUP

AF_ROUTE

AF_SCLUSTER

AF_SIP

AF_SLOW

AF_SNA

AF_UNIX

AF_UNSPEC

AIO_ALLDONE

AIO_CANCELED

AIO_LISTIO_MAXDeprecated

AIO_NOTCANCELED

AI_ADDRCONFIG

AI_ALL

AI_CANONNAME

AI_NUMERICHOST

AI_NUMERICSERV

AI_PASSIVE

AI_V4MAPPED

ALTMON_1

ALTMON_2

ALTMON_3

ALTMON_4

ALTMON_5

ALTMON_6

ALTMON_7

ALTMON_8

ALTMON_9

ALTMON_10

ALTMON_11

ALTMON_12

ALTWERASE

ALT_DIGITS

AM_STR

ARG_MAX

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

AT_BASE

AT_EACCESS

AT_EGID

AT_ENTRY

AT_EUID

AT_EXECFD

AT_EXECPATH

AT_FDCWD

AT_FLAGS

AT_GID

AT_IGNORE

AT_NOTELF

AT_NULL

AT_PAGESZ

AT_PHDR

AT_PHENT

AT_PHNUM

AT_REMOVEDIR

AT_SYMLINK_FOLLOW

AT_SYMLINK_NOFOLLOW

AT_UID

B0

B50

B75

B110

B134

B150

B200

B300

B600

B1200

B1800

B2400

B4800

B7200

B9600

B14400

B19200

B28800

B38400

B57600

B76800

B115200

B230400

B460800

B921600

BC_BASE_MAX

max ibase/obase values in bc(1)

BC_DIM_MAX

max array elements in bc(1)

BC_SCALE_MAX

max scale value in bc(1)

BC_STRING_MAX

max const string length in bc(1)

BIOCFLUSH

BIOCGBLEN

BIOCGDLT

BIOCGDLTLIST

BIOCGETIF

BIOCGHDRCMPLT

BIOCGRSIG

BIOCGRTIMEOUT

BIOCGSEESENT

BIOCGSTATS

BIOCIMMEDIATE

BIOCPROMISC

BIOCSBLEN

BIOCSDLT

BIOCSETF

BIOCSETFNR

BIOCSETIF

BIOCSHDRCMPLT

BIOCSRSIG

BIOCSRTIMEOUT

BIOCSSEESENT

BIOCVERSION

BOOT_TIME

BPF_ALIGNMENT

BRKINT

BUFSIZ

BUS_ADRALN

BUS_ADRERR

BUS_OBJERR

CCAR_OFLOW

CCTS_OFLOW

CDSR_OFLOW

CDTR_IFLOW

CHARCLASS_NAME_MAX

max character class name size

CHILD_MAX

CIGNORE

CLD_CONTINUED

CLD_DUMPED

CLD_EXITED

CLD_KILLED

CLD_STOPPED

CLD_TRAPPED

CLOCAL

CLOCK_MONOTONIC

CLOCK_MONOTONIC_FAST

CLOCK_MONOTONIC_PRECISE

CLOCK_PROCESS_CPUTIME_ID

CLOCK_PROF

CLOCK_REALTIME

CLOCK_REALTIME_FAST

CLOCK_REALTIME_PRECISE

CLOCK_SECOND

CLOCK_THREAD_CPUTIME_ID

CLOCK_UPTIME

CLOCK_UPTIME_FAST

CLOCK_UPTIME_PRECISE

CLOCK_VIRTUAL

CMGROUP_MAX

CODESET

COLL_WEIGHTS_MAX

max weights for order keyword

COMMLEN

size of returned ki_comm name

CPUCLOCK_WHICH_PID

CPUCLOCK_WHICH_TID

CPUSTATES

CPU_LEVEL_CPUSET

CPU_LEVEL_ROOT

CPU_LEVEL_WHICH

CPU_SETSIZE

CPU_WHICH_CPUSET

CPU_WHICH_IRQ

CPU_WHICH_JAIL

CPU_WHICH_PID

CPU_WHICH_TID

CP_IDLE

CP_INTR

CP_NICE

CP_SYS

CP_USER

CREAD

CRNCYSTR

CRTSCTS

CRTS_IFLOW

CS5

CS6

CS7

CS8

CSIZE

CSTOPB

CTLFLAG_ANYBODY

CTLFLAG_CAPRD

CTLFLAG_CAPRW

CTLFLAG_CAPWR

CTLFLAG_DORMANT

CTLFLAG_DYING

CTLFLAG_DYN

CTLFLAG_MPSAFE

CTLFLAG_NEEDGIANT

CTLFLAG_NOFETCH

CTLFLAG_PRISON

CTLFLAG_RD

CTLFLAG_RDTUN

CTLFLAG_RW

CTLFLAG_RWTUN

CTLFLAG_SECURE

CTLFLAG_SECURE1

CTLFLAG_SECURE2

CTLFLAG_SECURE3

CTLFLAG_SKIP

CTLFLAG_STATS

CTLFLAG_TUN

CTLFLAG_VNET

CTLFLAG_WR

CTLMASK_SECURE

CTLSHIFT_SECURE

CTLTYPE

CTLTYPE_INT

CTLTYPE_LONG

CTLTYPE_NODE

CTLTYPE_OPAQUE

CTLTYPE_S8

CTLTYPE_S16

CTLTYPE_S32

CTLTYPE_S64

CTLTYPE_STRING

CTLTYPE_STRUCT

CTLTYPE_U8

CTLTYPE_U16

CTLTYPE_U32

CTLTYPE_U64

CTLTYPE_UINT

CTLTYPE_ULONG

CTL_DEBUG

CTL_HW

CTL_KERN

CTL_MACHDEP

CTL_MAXNAME

CTL_NET

CTL_P1003_1B

CTL_P1003_1B_AIO_LISTIO_MAX

CTL_P1003_1B_AIO_MAX

CTL_P1003_1B_AIO_PRIO_DELTA_MAX

CTL_P1003_1B_ASYNCHRONOUS_IO

CTL_P1003_1B_DELAYTIMER_MAX

CTL_P1003_1B_FSYNC

CTL_P1003_1B_MAPPED_FILES

CTL_P1003_1B_MEMLOCK

CTL_P1003_1B_MEMLOCK_RANGE

CTL_P1003_1B_MEMORY_PROTECTION

CTL_P1003_1B_MESSAGE_PASSING

CTL_P1003_1B_MQ_OPEN_MAX

CTL_P1003_1B_PAGESIZE

CTL_P1003_1B_PRIORITIZED_IO

CTL_P1003_1B_PRIORITY_SCHEDULING

CTL_P1003_1B_REALTIME_SIGNALS

CTL_P1003_1B_RTSIG_MAX

CTL_P1003_1B_SEMAPHORES

CTL_P1003_1B_SEM_NSEMS_MAX

CTL_P1003_1B_SEM_VALUE_MAX

CTL_P1003_1B_SHARED_MEMORY_OBJECTS

CTL_P1003_1B_SIGQUEUE_MAX

CTL_P1003_1B_SYNCHRONIZED_IO

CTL_P1003_1B_TIMERS

CTL_P1003_1B_TIMER_MAX

CTL_SYSCTL

CTL_SYSCTL_DEBUG

CTL_SYSCTL_NAME

CTL_SYSCTL_NAME2OID

CTL_SYSCTL_NEXT

CTL_SYSCTL_NEXTNOSKIP

CTL_SYSCTL_OIDDESCR

CTL_SYSCTL_OIDFMT

CTL_SYSCTL_OIDLABEL

CTL_USER

CTL_VFS

CTL_VM

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

DEAD_PROCESS

DEVSTAT_NAME_LEN

DEVSTAT_N_TRANS_FLAGS

DOT3COMPLIANCE_COLLS

DOT3COMPLIANCE_STATS

DST_AUST

Australian style dst

DST_CAN

Canada

DST_EET

Eastern European dst

DST_MET

Middle European dst

DST_NONE

not on dst

DST_USA

USA style dst

DST_WET

Western European dst

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

D_FMT

D_MD_ORDER

D_T_FMT

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EAUTH

EBADF

EBADMSG

EBADRPC

EBUSY

ECANCELED

ECAPMODE

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDESTADDRREQ

EDOM

EDOOFUS

EDQUOT

EEXIST

EFAULT

EFBIG

EFD_CLOEXEC

EFD_NONBLOCK

EFD_SEMAPHORE

EFTYPE

EHOSTDOWN

EHOSTUNREACH

EIDRM

EILSEQ

EINPROGRESS

EINTEGRITY

EINTR

EINVAL

EIO

EISCONN

EISDIR

ELAST

ELOOP

EMFILE

EMLINK

EMPTY

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENEEDAUTH

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOATTR

ENOBUFS

ENODEV

ENOENT

ENOEXEC

ENOLCK

ENOLINK

ENOMEM

ENOMSG

ENOPROTOOPT

ENOSPC

ENOSYS

ENOTBLK

ENOTCAPABLE

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPROCLIM

EPROCUNAVAIL

EPROGMISMATCH

EPROGUNAVAIL

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EREMOTE

EROFS

ERPCMISMATCH

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESTALE

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUI64_LEN

EUSERS

EVFILT_AIO

EVFILT_EMPTY

EVFILT_FS

EVFILT_LIO

EVFILT_PROC

EVFILT_PROCDESC

EVFILT_READ

EVFILT_SENDFILE

EVFILT_SIGNAL

EVFILT_TIMER

EVFILT_USER

EVFILT_VNODE

EVFILT_WRITE

EV_ADD

EV_CLEAR

EV_DELETE

EV_DISABLE

EV_DISPATCH

EV_DROP

EV_ENABLE

EV_EOF

EV_ERROR

EV_FLAG1

EV_ONESHOT

EV_RECEIPT

EV_SYSFLAGS

EWOULDBLOCK

EXDEV

EXIT_FAILURE

EXIT_SUCCESS

EXPR_NEST_MAX

max expressions nested in expr(1)

EXTA

EXTATTR_NAMESPACE_EMPTY

EXTATTR_NAMESPACE_SYSTEM

EXTATTR_NAMESPACE_USER

EXTB

EXTPROC

FD_CLOEXEC

FD_SETSIZE

FILENAME_MAX

FIOASYNC

FIOCLEX

FIODGNAME

FIODTYPE

FIOGETLBA

FIOGETOWN

FIONBIO

FIONCLEX

FIONREAD

FIONSPACE

FIONWRITE

FIOSEEKDATA

FIOSEEKHOLE

FIOSETOWN

FIOSSHMLPGCNF

FLUSHO

FOPEN_MAX

F_ADD_SEALS

F_DUP2FD

F_DUP2FD_CLOEXEC

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLK

F_GETOWN

F_GET_SEALS

F_KINFO

F_LOCK

F_OGETLK

F_OK

F_OSETLK

F_OSETLKW

F_RDAHEAD

F_RDLCK

F_READAHEAD

F_SEAL_GROW

F_SEAL_SEAL

F_SEAL_SHRINK

F_SEAL_WRITE

F_SETFD

F_SETFL

F_SETLK

F_SETLKW

F_SETLK_REMOTE

F_SETOWN

F_TEST

F_TLOCK

F_ULOCK

F_UNLCK

F_WRLCK

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRND_INSECURE

GRND_NONBLOCK

GRND_RANDOM

GRPQUOTA

H4DISC

HOSTNAME_FOUND

HOSTNAME_INCORRECTNAME

HOSTNAME_INVALIDADDR

HOSTNAME_INVALIDNAME

HUPCL

HW_BYTEORDER

HW_DISKNAMES

HW_DISKSTATS

HW_FLOATINGPT

HW_MACHINE

HW_MACHINE_ARCH

HW_MODEL

HW_NCPU

HW_PAGESIZE

HW_PHYSMEM

HW_REALMEM

HW_USERMEM

ICANON

ICRNL

IEXTEN

IFAN_ARRIVAL

IFAN_DEPARTURE

IFCAP_CANTCHANGE

IFCAP_HWCSUM

IFCAP_HWCSUM_IPV6

IFCAP_HWRXTSTMP

hardware rx timestamping

IFCAP_HWSTATS

manages counters internally

IFCAP_JUMBO_MTU

9000 byte MTU supported

IFCAP_LINKSTATE

the runtime link state is dynamic

IFCAP_LRO

can do Large Receive Offload

IFCAP_MEXTPG

understands M_EXTPG mbufs

IFCAP_NETCONS

can be a network console

IFCAP_NETMAP

netmap mode supported/enabled

IFCAP_NV

can do SIOCGIFCAPNV/SIOCSIFCAPNV

IFCAP_POLLING

driver supports polling

IFCAP_RXCSUM

can offload checksum on RX

IFCAP_RXCSUM_IPV6

can offload checksum on IPv6 RX

IFCAP_TOE

IFCAP_TOE4

interface can offload TCP

IFCAP_TOE6

interface can offload TCP6

IFCAP_TSO

IFCAP_TSO4

can do TCP Segmentation Offload

IFCAP_TSO6

can do TCP6 Segmentation Offload

IFCAP_TXCSUM

can offload checksum on TX

IFCAP_TXCSUM_IPV6

can offload checksum on IPv6 TX

IFCAP_TXRTLMT

hardware supports TX rate limiting

IFCAP_TXTLS

IFCAP_TXTLS4

can do TLS encryption and segmentation for TCP

IFCAP_TXTLS6

can do TLS encryption and segmentation for TCP6

IFCAP_TXTLS_RTLMT

can do TLS with rate limiting

IFCAP_VLAN_HWCSUM

can do IFCAP_HWCSUM on VLANs

IFCAP_VLAN_HWFILTER

interface hw can filter vlan tag

IFCAP_VLAN_HWTAGGING

hardware VLAN tag support

IFCAP_VLAN_HWTSO

can do IFCAP_TSO on VLANs

IFCAP_VLAN_MTU

VLAN-compatible MTU

IFCAP_VXLAN_HWCSUM

can do IFCAN_HWCSUM on VXLANs

IFCAP_VXLAN_HWTSO

can do IFCAP_TSO on VXLANs

IFCAP_WOL

IFCAP_WOL_MAGIC

wake on any Magic Packet

IFCAP_WOL_MCAST

wake on any multicast frame

IFCAP_WOL_UCAST

wake on any unicast frame

IFDATA_DRIVERNAME

driver name and unit

IFDATA_GENERAL

generic stats for all kinds of ifaces

IFDATA_LINKSPECIFIC

specific to the type of interface

IFDR_MSG_SIZE

IFDR_REASON_MSG

IFDR_REASON_VENDOR

IFF_ALLMULTI

(n) receive all multicast packets

IFF_ALTPHYS

use alternate physical connection

IFF_BROADCAST

(i) broadcast address valid

IFF_CANTCONFIG

(i) unconfigurable using ioctl(2)

IFF_DEBUG

(n) turn on debugging

IFF_DYING

(n) interface is winding down

IFF_KNOWSEPOCH

(i) calls if_input in net epoch

IFF_LINK0

per link layer defined bit

IFF_LINK1

per link layer defined bit

IFF_LINK2

per link layer defined bit

IFF_LOOPBACK

(i) is a loopback net

IFF_MONITOR

(n) user-requested monitor mode

IFF_MULTICAST

(i) supports multicast

IFF_NOARP

(n) no address resolution protocol

IFF_NOGROUP

interface is not part of any groups

IFF_OACTIVE

(d) tx hardware queue is full

IFF_POINTOPOINT

(i) is a point-to-point link

IFF_PPROMISC

(n) user-requested promisc mode

IFF_PROMISC

(n) receive all packets

IFF_RENAMING

(n) interface is being renamed

IFF_RUNNING

(d) resources allocated

IFF_SIMPLEX

(i) can’t hear own transmissions

IFF_STATICARP

(n) static ARP

IFF_UP

(n) interface is up

IFMIB_IFCOUNT

number of interfaces configured

IFMIB_IFDATA

per-interface data table

IFMIB_SYSTEM

non-interface-specific

IFNAMSIZ

IFNET_PCP_NONE

IFNET_SLOWHZ

IFQ_MAXLEN

IFSTATMAX

IF_MAXUNIT

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

IMAXBEL

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INIT_PROCESS

INLCR

INPCK

INT_MAX

INT_MIN

IOV_MAX

IPC_CREAT

IPC_EXCL

IPC_INFO

IPC_M

IPC_NOWAIT

IPC_PRIVATE

IPC_R

IPC_RMID

IPC_SET

IPC_STAT

IPC_W

IPPROTO_3PC

Third Party Connect

IPPROTO_ADFS

Any distributed FS

IPPROTO_AH

IP6 Auth Header

IPPROTO_AHIP

any host internal protocol

IPPROTO_APES

any private encr. scheme

IPPROTO_ARGUS

Argus

IPPROTO_AX25

AX.25 Frames

IPPROTO_BHA

BHA

IPPROTO_BLT

Bulk Data Transfer

IPPROTO_BRSATMON

BackRoom SATNET Monitoring

IPPROTO_CARP

CARP

IPPROTO_CFTP

CFTP

IPPROTO_CHAOS

Chaos

IPPROTO_CMTP

Control Message Transport

IPPROTO_CPHB

Comp. Prot. HeartBeat

IPPROTO_CPNX

Comp. Prot. Net. Executive

IPPROTO_DCCP

Datagram Congestion Control Protocol

IPPROTO_DDP

Datagram Delivery

IPPROTO_DGP

dissimilar gateway prot.

IPPROTO_DIVERT

divert pseudo-protocol

IPPROTO_DONE

last return value of *_input(), meaning “all job for this pkt is done”.

IPPROTO_DSTOPTS

IP6 destination option

IPPROTO_EGP

exterior gateway protocol

IPPROTO_EMCON

EMCON

IPPROTO_ENCAP

encapsulation header

IPPROTO_EON

ISO cnlp

IPPROTO_ESP

IP6 Encap Sec. Payload

IPPROTO_ETHERIP

Ethernet IP encapsulation

IPPROTO_FRAGMENT

IP6 fragmentation header

IPPROTO_GGP

gateway^2 (deprecated)

IPPROTO_GMTP

GMTP

IPPROTO_GRE

General Routing Encap.

IPPROTO_HELLO

“hello” routing protocol

IPPROTO_HIP

IP6 Host Identity Protocol

IPPROTO_HMP

Host Monitoring

IPPROTO_HOPOPTS

IP6 hop-by-hop options

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IDP

xns idp

IPPROTO_IDPR

InterDomain Policy Routing

IPPROTO_IDRP

InterDomain Routing

IPPROTO_IGMP

group mgmt protocol

IPPROTO_IGP

NSFNET-IGP

IPPROTO_IGRP

Cisco/GXS IGRP

IPPROTO_IL

IL transport protocol

IPPROTO_INLSP

Integ. Net Layer Security

IPPROTO_INP

Merit Internodal

IPPROTO_IP

IPPROTO_IPCOMP

payload compression (IPComp)

IPPROTO_IPCV

Packet Core Utility

IPPROTO_IPEIP

IP encapsulated in IP

IPPROTO_IPIP

for compatibility

IPPROTO_IPPC

Pluribus Packet Core

IPPROTO_IPV6

IPPROTO_IRTP

Reliable Transaction

IPPROTO_KRYPTOLAN

Kryptolan

IPPROTO_LARP

Locus Address Resoloution

IPPROTO_LEAF1

Leaf-1

IPPROTO_LEAF2

Leaf-2

IPPROTO_MAX

IPPROTO_MEAS

DCN Measurement Subsystems

IPPROTO_MH

IPv6 Mobility Header

IPPROTO_MHRP

Mobile Host Routing

IPPROTO_MICP

Mobile Int.ing control

IPPROTO_MOBILE

IP Mobility

IPPROTO_MPLS

MPLS-in-IP

IPPROTO_MTP

Multicast Transport

IPPROTO_MUX

Multiplexing

IPPROTO_ND

Sun net disk proto (temp.)

IPPROTO_NHRP

Next Hop Resolution

IPPROTO_NONE

IP6 no next header

IPPROTO_NSP

Network Services

IPPROTO_NVPII

network voice protocol

IPPROTO_OLD_DIVERT

OLD divert pseudo-proto

IPPROTO_OSPFIGP

OSPFIGP

IPPROTO_PFSYNC

PFSYNC

IPPROTO_PGM

PGM

IPPROTO_PIGP

private interior gateway

IPPROTO_PIM

Protocol Independent Mcast

IPPROTO_PRM

Packet Radio Measurement

IPPROTO_PUP

pup

IPPROTO_PVP

Packet Video Protocol

IPPROTO_RAW

IPPROTO_RCCMON

BBN RCC Monitoring

IPPROTO_RDP

Reliable Data

IPPROTO_ROUTING

IP6 routing header

IPPROTO_RSVP

resource reservation

IPPROTO_RVD

Remote Virtual Disk

IPPROTO_SATEXPAK

SATNET/Backroom EXPAK

IPPROTO_SATMON

Satnet Monitoring

IPPROTO_SCCSP

Semaphore Comm. security

IPPROTO_SCTP

SCTP

IPPROTO_SDRP

Source Demand Routing

IPPROTO_SEND

SeND pseudo-protocol

IPPROTO_SHIM6

IP6 Shim6 Protocol

IPPROTO_SKIP

SKIP

IPPROTO_SRPC

Strite RPC protocol

IPPROTO_ST

Stream protocol II.

IPPROTO_SVMTP

Secure VMTP

IPPROTO_SWIPE

IP with encryption

IPPROTO_TCF

TCF

IPPROTO_TCP

IPPROTO_TLSP

Transport Layer Security

IPPROTO_TP

tp-4 w/ class negotiation

IPPROTO_TPXX

TP++ Transport

IPPROTO_TRUNK1

Trunk-1

IPPROTO_TRUNK2

Trunk-2

IPPROTO_TTP

TTP

IPPROTO_UDP

IPPROTO_UDPLITE

UDP-Lite

IPPROTO_VINES

Banyon VINES

IPPROTO_VISA

VISA Protocol

IPPROTO_VMTP

VMTP

IPPROTO_WBEXPAK

WIDEBAND EXPAK

IPPROTO_WBMON

WIDEBAND Monitoring

IPPROTO_WSN

Wang Span Network

IPPROTO_XNET

Cross Net Debugger

IPPROTO_XTP

XTP

IPTOS_ECN_CE

IPTOS_ECN_ECT0

IPTOS_ECN_ECT1

IPTOS_ECN_MASK

IPTOS_ECN_NOTECT

IPV6_BINDANY

IPV6_CHECKSUM

IPV6_DONTFRAG

IPV6_HOPLIMIT

IPV6_JOIN_GROUP

IPV6_LEAVE_GROUP

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_LOOP

IPV6_ORIGDSTADDR

IPV6_PKTINFO

IPV6_RECVORIGDSTADDR

IPV6_RECVPKTINFO

IPV6_RECVTCLASS

IPV6_TCLASS

IPV6_UNICAST_HOPS

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_ADD_SOURCE_MEMBERSHIP

IP_BINDANY

IP_BINDMULTI

IP_BLOCK_SOURCE

IP_DONTFRAG

IP_DROP_MEMBERSHIP

IP_DROP_SOURCE_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_ORIGDSTADDR

IP_RECVDSTADDR

IP_RECVIF

IP_RECVORIGDSTADDR

IP_RECVTOS

IP_RSS_LISTEN_BUCKET

IP_SENDSRCADDR

IP_TOS

IP_TTL

IP_UNBLOCK_SOURCE

ISIG

ISTRIP

ITIMER_PROF

ITIMER_REAL

ITIMER_VIRTUAL

IXANY

IXOFF

IXON

JAIL_API_VERSION

JAIL_ATTACH

JAIL_CREATE

JAIL_DYING

JAIL_GET_MASK

JAIL_SET_MASK

JAIL_SYS_DISABLE

JAIL_SYS_INHERIT

JAIL_SYS_NEW

JAIL_UPDATE

KERN_ARGMAX

KERN_ARND

KERN_BOOTFILE

KERN_BOOTTIME

KERN_CLOCKRATE

KERN_DUMMY

KERN_DUMPDEV

KERN_FILE

KERN_HOSTID

KERN_HOSTNAME

KERN_HOSTUUID

KERN_IOV_MAX

KERN_IPC

KERN_JOB_CONTROL

KERN_LOGSIGEXIT

KERN_MAXFILES

KERN_MAXFILESPERPROC

KERN_MAXPHYS

KERN_MAXPROC

KERN_MAXPROCPERUID

KERN_MAXVNODES

KERN_NGROUPS

KERN_NISDOMAINNAME

KERN_NTP_PLL

KERN_OSRELDATE

KERN_OSRELEASE

KERN_OSREV

KERN_OSTYPE

KERN_POSIX1

KERN_PROC

KERN_PROC_ALL

KERN_PROC_ARGS

KERN_PROC_AUXV

KERN_PROC_CWD

KERN_PROC_ENV

KERN_PROC_FILEDESC

KERN_PROC_GID

KERN_PROC_GROUPS

KERN_PROC_INC_THREAD

KERN_PROC_KSTACK

KERN_PROC_NFDS

KERN_PROC_OFILEDESC

KERN_PROC_OSREL

KERN_PROC_OVMMAP

KERN_PROC_PATHNAME

KERN_PROC_PGRP

KERN_PROC_PID

KERN_PROC_PROC

KERN_PROC_PS_STRINGS

KERN_PROC_RGID

KERN_PROC_RLIMIT

KERN_PROC_RUID

KERN_PROC_SESSION

KERN_PROC_SIGFASTBLK

KERN_PROC_SIGTRAMP

KERN_PROC_SV_NAME

KERN_PROC_TTY

KERN_PROC_UID

KERN_PROC_UMASK

KERN_PROC_VMMAP

KERN_PROF

KERN_PS_STRINGS

KERN_SAVED_IDS

KERN_SECURELVL

KERN_UPDATEINTERVAL

KERN_USRSTACK

KERN_VERSION

KERN_VNODE

KF_ATTR_VALID

KF_FD_TYPE_CTTY

Controlling terminal

KF_FD_TYPE_CWD

Current working directory

KF_FD_TYPE_JAIL

Jail directory

KF_FD_TYPE_ROOT

Root directory

KF_FD_TYPE_TEXT

KF_FD_TYPE_TRACE

Ktrace vnode

KF_FLAG_APPEND

KF_FLAG_ASYNC

KF_FLAG_CREAT

KF_FLAG_DIRECT

KF_FLAG_EXCL

KF_FLAG_EXEC

KF_FLAG_EXLOCK

KF_FLAG_FSYNC

KF_FLAG_HASLOCK

KF_FLAG_NOFOLLOW

KF_FLAG_NONBLOCK

KF_FLAG_READ

KF_FLAG_SHLOCK

KF_FLAG_TRUNC

KF_FLAG_WRITE

KF_TYPE_DEV

KF_TYPE_FIFO

KF_TYPE_KQUEUE

KF_TYPE_MQUEUE

KF_TYPE_NONE

KF_TYPE_PIPE

KF_TYPE_PROCDESC

KF_TYPE_PTS

KF_TYPE_SEM

KF_TYPE_SHM

KF_TYPE_SOCKET

KF_TYPE_UNKNOWN

KF_TYPE_VNODE

KF_VTYPE_UNKNOWN

KF_VTYPE_VBAD

KF_VTYPE_VBLK

KF_VTYPE_VCHR

KF_VTYPE_VDIR

KF_VTYPE_VFIFO

KF_VTYPE_VLNK

KF_VTYPE_VNON

KF_VTYPE_VREG

KF_VTYPE_VSOCK

KIPC_MAXSOCKBUF

KIPC_MAX_DATALEN

KIPC_MAX_HDR

KIPC_MAX_LINKHDR

KIPC_MAX_PROTOHDR

KIPC_SOCKBUF_WASTE

KIPC_SOMAXCONN

KI_CRF_CAPABILITY_MODE

Flags for the process credential.

KI_CRF_GRP_OVERFLOW

Steal a bit from ki_cr_flags to indicate that the cred had more than KI_NGROUPS groups.

KI_CTTY

controlling tty vnode active

KI_EMULNAMELEN

size of returned ki_emul

KI_LOCKBLOCK

proc blocked on lock ki_lockname

KI_NGROUPS

number of groups in ki_groups

KI_NSPARE_INT

KI_NSPARE_LONG

KI_NSPARE_PTR

KI_SLEADER

session leader

KKST_MAXLEN

KKST_STATE_RUNNING

KKST_STATE_STACKOK

Stack is valid.

KKST_STATE_SWAPPED

Stack swapped out.

KVME_FLAG_COW

KVME_FLAG_GROWS_DOWN

KVME_FLAG_GROWS_UP

KVME_FLAG_NEEDS_COPY

KVME_FLAG_NOCOREDUMP

KVME_FLAG_SUPER

KVME_FLAG_USER_WIRED

KVME_PROT_EXEC

KVME_PROT_READ

KVME_PROT_WRITE

KVME_TYPE_DEAD

KVME_TYPE_DEFAULT

KVME_TYPE_DEVICE

KVME_TYPE_MGTDEVICE

KVME_TYPE_NONE

KVME_TYPE_PHYS

KVME_TYPE_SG

KVME_TYPE_SWAP

KVME_TYPE_UNKNOWN

KVME_TYPE_VNODE

LC_ALL

LC_ALL_MASK

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_TIME

LC_TIME_MASK

LINE_MAX

max bytes in an input line

LINK_STATE_DOWN

link is down

LINK_STATE_UNKNOWN

link invalid/unknown

LINK_STATE_UP

link is up

LIO_NOP

LIO_NOWAIT

LIO_READ

LIO_READV

LIO_VECTORED

LIO_WAIT

LIO_WRITE

LIO_WRITEV

LOCAL_CONNWAIT

LOCAL_CREDS

LOCAL_CREDS_PERSISTENT

LOCAL_PEERCRED

LOCAL_VENDOR

LOCKNAMELEN

size of returned lock name

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

LOGINCLASSLEN

size of returned ki_loginclass

LOGIN_PROCESS

LOGNAMELEN

size of returned ki_login

LOG_ALERT

LOG_AUTH

LOG_AUTHPRIV

LOG_CONS

LOG_CONSOLE

LOG_CRIT

LOG_CRON

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_FTP

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NFACILITIES

LOG_NOTICE

LOG_NOWAIT

LOG_NTP

LOG_ODELAY

LOG_PERROR

LOG_PID

LOG_PRIMASK

LOG_SECURITY

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

L_tmpnam

MADV_AUTOSYNC

MADV_CORE

MADV_DONTNEED

MADV_FREE

MADV_NOCORE

MADV_NORMAL

MADV_NOSYNC

MADV_PROTECT

MADV_RANDOM

MADV_SEQUENTIAL

MADV_WILLNEED

MALLOCX_ZERO

MAP_ALIGNED_SUPER

MAP_ANON

MAP_ANONYMOUS

MAP_COPY

MAP_EXCL

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_GUARD

MAP_HASSEMAPHORE

MAP_NOCORE

MAP_NOSYNC

MAP_PREFAULT_READ

MAP_PRIVATE

MAP_SHARED

MAP_STACK

MAXCOMLEN

max command name remembered

MAXFIDSZ

File identifier. These are unique per filesystem on a single machine.

MAXFREQ

MAXHOSTNAMELEN

max hostname size

MAXINTERP

max interpreter file name length

MAXLOGNAME

max login name length (incl. NUL)

MAXPATHLEN

MAXPHASE

MAXSEC

MAXSECFLAVORS

MAXSYMLINKS

MAXTC

MAXUPRC

max simultaneous processes

MAX_CANON

max bytes in term canon input line

MAX_INPUT

max bytes in terminal input

MCL_CURRENT

MCL_FUTURE

MDMBUF

MFD_ALLOW_SEALING

MFD_CLOEXEC

MFD_HUGETLB

MFD_HUGE_1GB

MFD_HUGE_1MB

MFD_HUGE_2GB

MFD_HUGE_2MB

MFD_HUGE_8MB

MFD_HUGE_16GB

MFD_HUGE_16MB

MFD_HUGE_32MB

MFD_HUGE_64KB

MFD_HUGE_256MB

MFD_HUGE_512KB

MFD_HUGE_512MB

MFD_HUGE_MASK

MFSNAMELEN

Length of type name including null.

MINCORE_INCORE

MINCORE_MODIFIED

MINCORE_MODIFIED_OTHER

MINCORE_REFERENCED

MINCORE_REFERENCED_OTHER

MINCORE_SUPER

MINSEC

MINSIGSTKSZ

MNAMELEN

Size of on/from name bufs.

MNT_ACLS

MNT_ASYNC

MNT_AUTOMOUNTED

Mounted by automountd(8).

MNT_BYFSID

MNT_DEFERRED

Unmount in async context.

MNT_DEFEXPORTED

MNT_DELEXPORT

MNT_EMPTYDIR

Only mount on empty dir.

MNT_EXKERB

MNT_EXPORTANON

MNT_EXPORTED

MNT_EXPUBLIC

MNT_EXRDONLY

MNT_EXTLS

Require TLS.

MNT_EXTLSCERT

Require TLS with client cert.

MNT_EXTLSCERTUSER

Require TLS with user cert.

MNT_FORCE

MNT_GJOURNAL

MNT_IGNORE

Do not show entry in df.

MNT_LAZY

Push data not written by filesystem syncer.

MNT_LOCAL

Filesystem is stored locally.

MNT_MULTILABEL

MNT_NFS4ACLS

MNT_NOATIME

MNT_NOCLUSTERR

MNT_NOCLUSTERW

MNT_NOCOVER

Do not cover a mount point.

MNT_NOEXEC

MNT_NONBUSY

MNT_NOSUID

MNT_NOSYMFOLLOW

MNT_NOWAIT

Start all I/O, but do not wait for it.

MNT_QUOTA

Quotas are enabled on fs.

MNT_RDONLY

MNT_RECURSE

Recursively unmount uppers.

MNT_RELOAD

MNT_ROOTFS

Identifies the root fs.

MNT_SNAPSHOT

MNT_SOFTDEP

MNT_SUIDDIR

MNT_SUJ

Using journaled soft updates.

MNT_SUSPEND

Suspend file system after sync.

MNT_SYNCHRONOUS

MNT_UNION

MNT_UNTRUSTED

Filesys metadata untrusted.

MNT_UPDATE

MNT_USER

Mounted by a user.

MNT_VERIFIED

Filesystem is verified.

MNT_WAIT

Synchronously wait for I/O to complete.

MOD_CLKA

MOD_CLKB

MOD_ESTERROR

MOD_FREQUENCY

MOD_MAXERROR

MOD_MICRO

MOD_NANO

MOD_OFFSET

MOD_PPSMAX

MOD_STATUS

MOD_TAI

MOD_TIMECONST

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

MSG_CMSG_CLOEXEC

MSG_COMPAT

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOF

MSG_EOR

MSG_NBIO

MSG_NOERROR

MSG_NOSIGNAL

MSG_NOTIFICATION

MSG_OOB

MSG_PEEK

MSG_TRUNC

MSG_WAITALL

MS_ASYNC

MS_INVALIDATE

MS_SYNC

NAME_MAX

max bytes in a file name

NANOSECOND

NCARGS

max bytes for an exec function

NCCS

NETGRAPHDISC

NETLINK_GENERIC

functions not specific to a type of iface

NET_RT_DUMP

NET_RT_FLAGS

NET_RT_IFLIST

NET_RT_IFLISTL

NET_RT_IFMALIST

NEW_TIME

NGROUPS

/* max number groups

NGROUPS_MAX

max supplemental group id’s

NI_DGRAM

NI_MAXHOST

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSCOPE

NI_NUMERICSERV

NOEXPR

NOFILE

max open files per process

NOFLSH

NOGROUP

marker for empty group set member

NOKERNINFO

NOSTR

NOTE_ATTRIB

NOTE_CHILD

NOTE_DELETE

NOTE_EXEC

NOTE_EXIT

NOTE_EXTEND

NOTE_FFAND

NOTE_FFCOPY

NOTE_FFCTRLMASK

NOTE_FFLAGSMASK

NOTE_FFNOP

NOTE_FFOR

NOTE_FORK

NOTE_LINK

NOTE_LOWAT

NOTE_MSECONDS

NOTE_NSECONDS

NOTE_PCTRLMASK

NOTE_PDATAMASK

NOTE_RENAME

NOTE_REVOKE

NOTE_SECONDS

NOTE_TRACK

NOTE_TRACKERR

NOTE_TRIGGER

NOTE_USECONDS

NOTE_WRITE

NTP_API

NZERO

OCRNL

OID_AUTO

OLD_TIME

ONLCR

ONLRET

ONOCR

ONOEOT

OPEN_MAX

max open files per process

OPOST

OXTABS

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECT

O_DIRECTORY

O_EXCL

O_EXEC

O_EXLOCK

O_FSYNC

O_NDELAY

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_RDONLY

O_RDWR

O_SHLOCK

O_SYNC

O_TRUNC

O_TTY_INIT

O_WRONLY

P2_ASLR_DISABLE

P2_ASLR_ENABLE

P2_ASLR_IGNSTART

P2_AST_SU

P2_INHERIT_PROTECTED

P2_NOTRACE

P2_NOTRACE_EXEC

P2_PROTMAX_DISABLE

P2_PROTMAX_ENABLE

P2_PTRACE_FSTP

P2_STKGAP_DISABLE

P2_STKGAP_DISABLE_EXEC

P2_TRAPCAP

PARENB

PARMRK

PARODD

PATH_MAX

PC_ALL

PC_COMPRESS

PD_ALLOWED_AT_FORK

PD_CLOEXEC

PD_DAEMON

PENDIN

PF_APPLETALK

PF_ARP

PF_ATM

PF_BLUETOOTH

PF_CCITT

PF_CHAOS

PF_CNT

PF_COIP

PF_DATAKIT

PF_DECnet

PF_DLI

PF_ECMA

PF_HYLINK

PF_IEEE80211

PF_IMPLINK

PF_INET

PF_INET6

PF_INET6_SDP

PF_INET_SDP

PF_IPX

PF_ISDN

PF_ISO

PF_KEY

PF_LAT

PF_LINK

PF_LOCAL

PF_NATM

PF_NETBIOS

PF_NETGRAPH

PF_OSI

PF_PIP

PF_PUP

PF_ROUTE

PF_RTIP

PF_SCLUSTER

PF_SIP

PF_SLOW

PF_SNA

PF_UNIX

PF_UNSPEC

PF_XTP

PINODDeprecated

PIOD_READ_D

PIOD_READ_I

PIOD_WRITE_D

PIOD_WRITE_I

PIPE_BUF

PI_AVDeprecated

PI_DISKDeprecated

PI_DULLDeprecated

PI_NETDeprecated

PI_REALTIME

PI_SOFTDeprecated

PI_TTYDeprecated

PLOCKDeprecated

PL_EVENT_NONE

PL_EVENT_SIGNAL

PL_FLAG_BORN

PL_FLAG_BOUND

PL_FLAG_CHILD

PL_FLAG_EXEC

PL_FLAG_EXITED

PL_FLAG_FORKED

PL_FLAG_SA

PL_FLAG_SCE

PL_FLAG_SCX

PL_FLAG_SI

PL_FLAG_VFORKED

PL_FLAG_VFORK_DONE

PM_STR

POLLERR

POLLHUP

POLLIN

POLLINIGNEOF

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

POLLSTANDARD

POLLWRBAND

POLLWRNORM

POSIX_FADV_DONTNEED

POSIX_FADV_NOREUSE

POSIX_FADV_NORMAL

POSIX_FADV_RANDOM

POSIX_FADV_SEQUENTIAL

POSIX_FADV_WILLNEED

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

POSIX_SPAWN_RESETIDS

POSIX_SPAWN_SETPGROUP

POSIX_SPAWN_SETSCHEDPARAM

POSIX_SPAWN_SETSCHEDULER

POSIX_SPAWN_SETSIGDEF

POSIX_SPAWN_SETSIGMASK

PPAUSEDeprecated

PPPDISC

PPROT_CLEAR

PPROT_DESCEND

PPROT_INHERIT

PPROT_SET

PRIBIODeprecated

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PRI_MAX

PRI_MAX_IDLE

PRI_MAX_ITHDDeprecated

PRI_MAX_KERNDeprecated

PRI_MAX_REALTIMEDeprecated

PRI_MAX_TIMESHARE

PRI_MIN

PRI_MIN_IDLE

PRI_MIN_ITHD

PRI_MIN_KERNDeprecated

PRI_MIN_REALTIMEDeprecated

PRI_MIN_TIMESHAREDeprecated

PROC_ASLR_ACTIVE

PROC_ASLR_CTL

PROC_ASLR_FORCE_DISABLE

PROC_ASLR_FORCE_ENABLE

PROC_ASLR_NOFORCE

PROC_ASLR_STATUS

PROC_PDEATHSIG_CTL

PROC_PDEATHSIG_STATUS

PROC_PROCCTL_MD_MIN

PROC_PROTMAX_ACTIVE

PROC_PROTMAX_CTL

PROC_PROTMAX_FORCE_DISABLE

PROC_PROTMAX_FORCE_ENABLE

PROC_PROTMAX_NOFORCE

PROC_PROTMAX_STATUS

PROC_REAP_ACQUIRE

PROC_REAP_GETPIDS

PROC_REAP_KILL

PROC_REAP_RELEASE

PROC_REAP_STATUS

PROC_SPROTECT

PROC_STACKGAP_CTL

PROC_STACKGAP_DISABLE

PROC_STACKGAP_DISABLE_EXEC

PROC_STACKGAP_ENABLE

PROC_STACKGAP_ENABLE_EXEC

PROC_STACKGAP_STATUS

PROC_TRACE_CTL

PROC_TRACE_CTL_DISABLE

PROC_TRACE_CTL_DISABLE_EXEC

PROC_TRACE_CTL_ENABLE

PROC_TRACE_STATUS

PROC_TRAPCAP_CTL

PROC_TRAPCAP_CTL_DISABLE

PROC_TRAPCAP_CTL_ENABLE

PROC_TRAPCAP_STATUS

PROT_EXEC

PROT_NONE

PROT_READ

PROT_WRITE

PSOCKDeprecated

PSWPDeprecated

PS_FST_FFLAG_APPEND

PS_FST_FFLAG_ASYNC

PS_FST_FFLAG_CREAT

PS_FST_FFLAG_DIRECT

PS_FST_FFLAG_EXCL

PS_FST_FFLAG_EXEC

PS_FST_FFLAG_EXLOCK

PS_FST_FFLAG_HASLOCK

PS_FST_FFLAG_NOFOLLOW

PS_FST_FFLAG_NONBLOCK

PS_FST_FFLAG_READ

PS_FST_FFLAG_SHLOCK

PS_FST_FFLAG_SYNC

PS_FST_FFLAG_TRUNC

PS_FST_FFLAG_WRITE

PS_FST_TYPE_DEV

PS_FST_TYPE_EVENTFD

PS_FST_TYPE_FIFO

PS_FST_TYPE_KQUEUE

PS_FST_TYPE_MQUEUE

PS_FST_TYPE_NONE

PS_FST_TYPE_PIPE

PS_FST_TYPE_PROCDESC

PS_FST_TYPE_PTS

PS_FST_TYPE_SEM

PS_FST_TYPE_SHM

PS_FST_TYPE_SOCKET

PS_FST_TYPE_UNKNOWN

PS_FST_TYPE_VNODE

PS_FST_UFLAG_CDIR

PS_FST_UFLAG_CTTY

PS_FST_UFLAG_JAIL

PS_FST_UFLAG_MMAP

PS_FST_UFLAG_RDIR

PS_FST_UFLAG_TEXT

PS_FST_UFLAG_TRACE

PS_FST_VTYPE_UNKNOWN

PS_FST_VTYPE_VBAD

PS_FST_VTYPE_VBLK

PS_FST_VTYPE_VCHR

PS_FST_VTYPE_VDIR

PS_FST_VTYPE_VFIFO

PS_FST_VTYPE_VLNK

PS_FST_VTYPE_VNON

PS_FST_VTYPE_VREG

PS_FST_VTYPE_VSOCK

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_MUTEX_ADAPTIVE_NP

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_MUTEX_ROBUST

PTHREAD_MUTEX_STALLED

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

PTRACE_DEFAULT

PTRACE_EXEC

PTRACE_FORK

PTRACE_LWP

PTRACE_SCE

PTRACE_SCX

PTRACE_SYSCALL

PTRACE_VFORK

PT_ATTACH

PT_CLEARSTEP

PT_CONTINUE

PT_COREDUMP

PT_DETACH

PT_FIRSTMACH

PT_FOLLOW_FORK

PT_GETDBREGS

PT_GETFPREGS

PT_GETLWPLIST

PT_GETNUMLWPS

PT_GETREGS

PT_GETREGSET

PT_GET_EVENT_MASK

PT_GET_SC_ARGS

PT_GET_SC_RET

PT_IO

PT_KILL

PT_LWPINFO

PT_LWP_EVENTS

PT_READ_D

PT_READ_I

PT_RESUME

PT_SETDBREGS

PT_SETFPREGS

PT_SETREGS

PT_SETREGSET

PT_SETSTEP

PT_SET_EVENT_MASK

PT_STEP

PT_SUSPEND

PT_SYSCALL

PT_TO_SCE

PT_TO_SCX

PT_TRACE_ME

PT_VM_ENTRY

PT_VM_TIMESTAMP

PT_WRITE_D

PT_WRITE_I

PUSERDeprecated

PVFSDeprecated

PVMDeprecated

PWAITDeprecated

PZERODeprecated

P_ADVLOCK

P_ALL

P_CONTINUED

P_CONTROLT

P_EXEC

P_HADTHREADS

P_HWPMC

P_INEXEC

P_INMEM

P_JAILED

P_KPROC

P_MAGIC

P_PGID

P_PID

P_PPTRACE

P_PPWAIT

P_PROFIL

P_PROTECTED

P_SIGEVENT

P_SINGLE_BOUNDARY

P_SINGLE_EXIT

P_STATCHILD

P_STOPPED

P_STOPPED_SIG

P_STOPPED_SINGLE

P_STOPPED_TRACE

P_STOPPROF

P_SUGID

P_SWAPPINGIN

P_SWAPPINGOUT

P_SYSTEM

P_TOTAL_STOP

P_TRACED

P_TREE_FIRST_ORPHAN

P_TREE_GRPEXITED

P_TREE_ORPHANED

P_TREE_REAPER

P_UNUSED3

P_WAITED

P_WEXIT

P_WKILLED

Q_GETQUOTA

Q_QUOTAOFF

Q_QUOTAON

Q_SETQUOTA

Q_SYNC

RADIXCHAR

RAND_MAX

REG_ASSERT

REG_ATOI

REG_BACKR

REG_BADBR

REG_BADPAT

REG_BADRPT

REG_BASIC

REG_DUMP

REG_EBRACE

REG_EBRACK

REG_ECOLLATE

REG_ECTYPE

REG_EESCAPE

REG_EMPTY

REG_ENOSYS

REG_EPAREN

REG_ERANGE

REG_ESPACE

REG_ESUBREG

REG_EXTENDED

REG_ICASE

REG_ILLSEQ

REG_INVARG

REG_ITOA

REG_LARGE

REG_NEWLINE

REG_NOMATCH

REG_NOSPEC

REG_NOSUB

REG_NOTBOL

REG_NOTEOL

REG_PEND

REG_STARTEND

REG_TRACE

RE_DUP_MAX

max RE’s in interval notation

RFCFDG

RFFDG

RFLINUXTHPN

RFMEM

RFNOWAIT

RFPROC

RFSPAWN

RFTHREAD

RFTSIGZMB

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_KQUEUES

RLIMIT_MEMLOCK

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_NPTS

RLIMIT_RSS

RLIMIT_SBSIZE

RLIMIT_STACK

RLIMIT_SWAP

RLIMIT_UMTXP

RLIMIT_VMEM

RLIM_INFINITY

RLIM_NLIMITSDeprecated

RLIM_SAVED_CUR

RLIM_SAVED_MAX

RSS_FUNC_NONE

RSS_FUNC_PRIVATE

RSS_FUNC_TOEPLITZ

RSS_KEYLEN

RSS_TYPE_IPV4

RSS_TYPE_IPV6

RSS_TYPE_IPV6_EX

RSS_TYPE_TCP_IPV4

RSS_TYPE_TCP_IPV6

RSS_TYPE_TCP_IPV6_EX

RSS_TYPE_UDP_IPV4

RSS_TYPE_UDP_IPV6

RSS_TYPE_UDP_IPV6_EX

RTLD_DEFAULT

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NEXT

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

RTLD_SELF

RTP_LOOKUP

RTP_PRIO_IDLE

RTP_PRIO_MAX

RTP_PRIO_MIN

RTP_PRIO_NORMAL

RTP_PRIO_REALTIME

RTP_SET

RUSAGE_CHILDREN

RUSAGE_SELF

RUSAGE_THREAD

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SCALE_PPM

SCHED_FIFO

SCHED_OTHER

SCHED_RR

SCM_CREDS

SCM_CREDS2

SCM_RIGHTS

SCM_TIMESTAMP

SEEK_CUR

SEEK_DATA

SEEK_END

SEEK_HOLE

SEEK_SET

SEM_FAILED

SF_APPEND

SF_ARCHIVED

SF_IMMUTABLE

SF_MNOWAIT

SF_NOCACHE

SF_NODISKIO

SF_NOUNLINK

SF_SETTABLE

SF_SNAPSHOT

SF_SYNC

SF_USER_READAHEAD

SHM_ANON

SHM_INFO

SHM_LARGEPAGE_ALLOC_DEFAULT

SHM_LARGEPAGE_ALLOC_HARD

SHM_LARGEPAGE_ALLOC_NOWAIT

SHM_LOCK

SHM_R

SHM_RDONLY

SHM_RENAME_EXCHANGE

SHM_RENAME_NOREPLACE

SHM_RND

SHM_STAT

SHM_UNLOCK

SHM_W

SHUTDOWN_TIME

SHUT_RD

SHUT_RDWR

SHUT_WR

SIDL

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEMT

SIGEV_KEVENT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGEV_THREAD_ID

SIGFPE

SIGHUP

SIGILL

SIGINFO

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPROF

SIGQUIT

SIGSEGV

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SIOCGIFADDR

SLIPDISC

SLOCK

SOCK_CLOEXEC

SOCK_DGRAM

SOCK_MAXADDRLEN

SOCK_NONBLOCK

SOCK_RAW

SOCK_RDM

SOCK_SEQPACKET

SOCK_STREAM

SOL_SOCKET

SOMAXCONN

SO_ACCEPTCONN

SO_ACCEPTFILTER

SO_BINTIME

SO_BROADCAST

SO_DEBUG

SO_DOMAIN

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LABEL

SO_LINGER

SO_LISTENINCQLEN

SO_LISTENQLEN

SO_LISTENQLIMIT

SO_NOSIGPIPE

SO_NO_DDP

SO_NO_OFFLOAD

SO_OOBINLINE

SO_PEERLABEL

SO_PROTOCOL

SO_PROTOTYPE

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_REUSEPORT_LB

SO_SETFIB

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TYPE

SO_USELOOPBACK

SO_USER_COOKIE

SO_VENDOR

SPACECTL_DEALLOC

SPECNAMELEN

max length of devicename

SRUN

SSLEEP

SSTOP

SS_DISABLE

SS_ONSTACK

STA_CLK

STA_CLOCKERR

STA_DEL

STA_FLL

STA_FREQHOLD

STA_INS

STA_MODE

STA_NANO

STA_PLL

STA_PPSERROR

STA_PPSFREQ

STA_PPSJITTER

STA_PPSSIGNAL

STA_PPSTIME

STA_PPSWANDER

STA_RONLY

STA_UNSYNC

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

ST_NOSUID

ST_RDONLY

SWAIT

SZOMB

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TAB0

TAB3

TABDLY

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_CCALGOOPT

TCP_CONGESTION

TCP_DELACK

TCP_FASTOPEN

TCP_FASTOPEN_PSK_LEN

TCP_FIN_IS_RST

TCP_FUNCTION_NAME_LEN_MAX

TCP_IDLE_REDUCE

TCP_INFO

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINIT

TCP_KEEPINTVL

TCP_LOG_LIMIT

TCP_LRD

TCP_MAXPEAKRATE

TCP_MAXSEG

TCP_MAXUNACKTIME

TCP_MD5SIG

TCP_NODELAY

TCP_NOOPT

TCP_NOPUSH

TCP_PCAP_IN

TCP_PCAP_OUT

TCP_PERF_INFO

TCP_PROC_ACCOUNTING

TCP_REMOTE_UDP_ENCAPS_PORT

TCP_SHARED_CWND_ALLOWED

TCP_USE_CMP_ACKS

TCSADRAIN

TCSAFLUSH

TCSANOW

TDB_BORN

TDB_CHILD

TDB_EXEC

TDB_EXIT

TDB_FORK

TDB_FSTP

TDB_SCE

TDB_SCX

TDB_STEP

TDB_STOPATFORK

TDB_SUSPEND

TDB_USERWR

TDB_VFORK

TDB_XSIG

TDF_ALLPROCSUSP

TDF_ALRMPENDDeprecated

TDF_ASTPENDINGDeprecated

TDF_BORROWING

TDF_BOUNDARY

TDF_CANSWAP

TDF_IDLETD

TDF_INMEM

TDF_INPANIC

TDF_KTH_SUSP

TDF_MACPENDDeprecated

TDF_NEEDRESCHEDDeprecated

TDF_NEEDSIGCHKDeprecated

TDF_NEEDSUSPCHKDeprecated

TDF_NOLOAD

TDF_PROFPENDDeprecated

TDF_SBDRY

TDF_SCHED0

TDF_SCHED1

TDF_SCHED2

TDF_SCHED3

TDF_SEINTR

TDF_SERESTART

TDF_SINTR

TDF_SWAPINREQ

TDF_THRWAKEUP

TDF_TIMEOUT

TDF_UNUSED23Deprecated

TDF_UPIBLOCKED

TDI_IWAIT

TDI_LOCK

TDI_SLEEPING

TDI_SUSPENDED

TDI_SWAPPED

TDNAMLEN

size of returned thread name

TDP2_COMPAT32RB

TDP2_SBPAGES

TDP_ALTSTACK

TDP_AUDITREC

TDP_BUFNEED

TDP_CALLCHAIN

TDP_COWINPROGRESS

TDP_DEADLKTREAT

TDP_EXECVMSPC

TDP_GEOM

TDP_IGNSUSP

TDP_INBDFLUSH

TDP_INKTR

TDP_INKTRACE

TDP_ITHREAD

TDP_KTHREAD

TDP_NERRNO

TDP_NOFAULTING

TDP_NORUNNINGBUF

TDP_OLDMASK

TDP_OWEUPC

TDP_RESETSPUR

TDP_RFPPWAIT

TDP_SCHED1

TDP_SCHED2

TDP_SCHED3

TDP_SCHED4

TDP_SIGFASTBLOCK

TDP_SIGFASTPENDING

TDP_SOFTDEP

TDP_SYNCIO

TDP_UIOHELD

TDP_WAKEUP

THOUSEP

TIMER_ABSTIME

TIME_DEL

TIME_ERROR

TIME_INS

TIME_OK

TIME_OOP

TIME_WAIT

TIOCCBRK

TIOCCDTR

TIOCCONS

TIOCDRAIN

TIOCEXCL

TIOCEXT

TIOCFLUSH

TIOCGDRAINWAIT

TIOCGETA

TIOCGETD

TIOCGPGRP

TIOCGPTN

TIOCGSID

TIOCGWINSZ

TIOCMBIC

TIOCMBIS

TIOCMGDTRWAIT

TIOCMGET

TIOCMSDTRWAIT

TIOCMSET

TIOCM_CAR

TIOCM_CD

TIOCM_CTS

TIOCM_DCD

TIOCM_DSR

TIOCM_DTR

TIOCM_LE

TIOCM_RI

TIOCM_RNG

TIOCM_RTS

TIOCM_SR

TIOCM_ST

TIOCNOTTY

TIOCNXCL

TIOCOUTQ

TIOCPKT

TIOCPKT_DATA

TIOCPKT_DOSTOP

TIOCPKT_FLUSHREAD

TIOCPKT_FLUSHWRITE

TIOCPKT_IOCTL

TIOCPKT_NOSTOP

TIOCPKT_START

TIOCPKT_STOP

TIOCPTMASTER

TIOCSBRK

TIOCSCTTY

TIOCSDRAINWAIT

TIOCSDTR

TIOCSETA

TIOCSETAF

TIOCSETAW

TIOCSETD

TIOCSIG

TIOCSPGRP

TIOCSTART

TIOCSTAT

TIOCSTI

TIOCSTOP

TIOCSWINSZ

TIOCTIMESTAMP

TIOCUCNTL

TMP_MAX

TOSTOP

TTYDISC

T_FMT

T_FMT_AMPM

UF_APPEND

UF_ARCHIVE

UF_HIDDEN

UF_IMMUTABLE

UF_NODUMP

UF_NOUNLINK

UF_OFFLINE

UF_OPAQUE

UF_READONLY

UF_REPARSE

UF_SETTABLE

UF_SPARSE

UF_SYSTEM

UMTX_ABSTIME

UMTX_OP_CV_BROADCAST

UMTX_OP_CV_SIGNAL

UMTX_OP_CV_WAIT

UMTX_OP_MUTEX_LOCK

UMTX_OP_MUTEX_TRYLOCK

UMTX_OP_MUTEX_UNLOCK

UMTX_OP_MUTEX_WAIT

UMTX_OP_MUTEX_WAKE2

UMTX_OP_NWAKE_PRIVATE

UMTX_OP_ROBUST_LISTS

UMTX_OP_RW_RDLOCK

UMTX_OP_RW_UNLOCK

UMTX_OP_RW_WRLOCK

UMTX_OP_SEM2_WAIT

UMTX_OP_SEM2_WAKE

UMTX_OP_SET_CEILING

UMTX_OP_SHM

UMTX_OP_WAIT

UMTX_OP_WAIT_UINT

UMTX_OP_WAIT_UINT_PRIVATE

UMTX_OP_WAKE

UMTX_OP_WAKE_PRIVATE

USER_BC_BASE_MAX

USER_BC_DIM_MAX

USER_BC_SCALE_MAX

USER_BC_STRING_MAX

USER_COLL_WEIGHTS_MAX

USER_CS_PATH

USER_EXPR_NEST_MAX

USER_LINE_MAX

USER_LOCALBASE

USER_POSIX2_CHAR_TERM

USER_POSIX2_C_BIND

USER_POSIX2_C_DEV

USER_POSIX2_FORT_DEV

USER_POSIX2_FORT_RUN

USER_POSIX2_LOCALEDEF

USER_POSIX2_SW_DEV

USER_POSIX2_UPE

USER_POSIX2_VERSION

USER_PROCESS

USER_RE_DUP_MAX

USER_STREAM_MAX

USER_TZNAME_MAX

USRQUOTA

UTIME_NOW

UTIME_OMIT

UTXDB_ACTIVE

UTXDB_LASTLOGIN

UTXDB_LOG

VDISCARD

VDSUSP

VEOF

VEOL

VEOL2

VERASE

VERASE2

VFCF_DELEGADMIN

Supports delegated administration.

VFCF_JAIL

Can be mounted from within a jail.

VFCF_LOOPBACK

Aliases some other mounted FS.

VFCF_NETWORK

May get data over the network.

VFCF_READONLY

Writes are not implemented.

VFCF_SBDRY

Stop at Boundary: defer stop requests to kernel->user (AST) transition.

VFCF_STATIC

Statically compiled into kernel.

VFCF_SYNTHETIC

Data does not represent real files.

VFCF_UNICODE

Stores file names as Unicode.

VFS_CONF

struct: vfsconf for filesystem given as next argument.

VFS_GENERIC

Generic filesystem information.

VFS_MAXTYPENUM

int: highest defined filesystem type.

VFS_VFSCONF

Get configured filesystems.

VINTR

VKILL

VLNEXT

VMIN

VM_TOTAL

VQUIT

VREPRINT

VSTART

VSTATUS

VSTOP

VSUSP

VTIME

VWERASE

WCONTINUED

WEXITED

WMESGLEN

size of returned wchan message

WNOHANG

WNOWAIT

WSTOPPED

WTRAPPED

WUNTRACED

W_OK

XUCRED_VERSION

XU_NGROUPS

X_OK

YESEXPR

YESSTR

_IOFBF

_IOLBF

_IONBF

_PC_ACL_EXTENDED

_PC_ACL_NFS4

_PC_ACL_PATH_MAX

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CAP_PRESENT

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_INF_PRESENT

_PC_LINK_MAX

_PC_MAC_PRESENT

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_MIN_HOLE_SIZE

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_POSIX2_BC_BASE_MAX

_POSIX2_BC_DIM_MAX

_POSIX2_BC_SCALE_MAX

_POSIX2_BC_STRING_MAX

_POSIX2_CHARCLASS_NAME_MAX

_POSIX2_COLL_WEIGHTS_MAX

_POSIX2_EQUIV_CLASS_MAX

_POSIX2_EXPR_NEST_MAX

_POSIX2_LINE_MAX

_POSIX2_RE_DUP_MAX

_POSIX_ARG_MAX

_POSIX_LINK_MAX

_POSIX_MAX_CANON

_POSIX_MAX_INPUT

_POSIX_NAME_MAX

_POSIX_PIPE_BUF

_POSIX_SSIZE_MAX

_POSIX_STREAM_MAX

_POSIX_VDISABLE

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUSET_SIZE

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FILE_LOCKING

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_V6_ILP32_OFF32

_SC_V6_ILP32_OFFBIG

_SC_V6_LP64_OFF64

_SC_V6_LPBIG_OFFBIG

_SC_VERSION

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_UUID_NODE_LEN

dot3ChipSetAMD79C940

dot3ChipSetAMD7990

dot3ChipSetAMD79900

dot3ChipSetDigitalDC21040

dot3ChipSetDigitalDC21041

dot3ChipSetDigitalDC21140

dot3ChipSetDigitalDC21140A

dot3ChipSetDigitalDC21142

dot3ChipSetFujitsu86950

dot3ChipSetIntel82557

dot3ChipSetIntel82586

dot3ChipSetIntel82596

dot3ChipSetNational8390

dot3ChipSetNationalSonic

dot3ChipSetWesternDigital83C690

dot3ChipSetWesternDigital83C790

pseudo_AF_HDRCMPLT

pseudo_AF_KEY

pseudo_AF_PIP

pseudo_AF_RTIP

pseudo_AF_XTP

Functions

CMSG_DATA⚠

CMSG_FIRSTHDR⚠

CMSG_LEN⚠

CMSG_NXTHDR⚠

CMSG_SPACE⚠

CPU_CLR⚠

CPU_COUNT⚠

CPU_FILL⚠

CPU_ISSET⚠

CPU_SET⚠

CPU_ZERO⚠

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

MALLOCX_ALIGN⚠

MALLOCX_ARENA⚠

MALLOCX_TCACHE⚠

QCMD

SOCKCRED2SIZE⚠

SOCKCREDSIZE⚠

WCOREDUMP

WEXITSTATUS

WIFCONTINUED

WIFEXITED

WIFSIGNALED

WIFSTOPPED

WSTOPSIG

WTERMSIG

__cap_rights_clear⚠

__cap_rights_init⚠

__cap_rights_is_set⚠

__cap_rights_set⚠

__error⚠

__xuname⚠

_exit⚠

_umtx_op⚠

abort⚠

abs⚠

accept⚠

accept4⚠

access⚠

acct⚠

adjtime⚠

aio_cancel⚠

aio_error⚠

aio_fsync⚠

aio_read⚠

aio_readv⚠

aio_return⚠

aio_suspend⚠

aio_waitcomplete⚠

aio_write⚠

aio_writev⚠

alarm⚠

arc4random⚠

arc4random_buf⚠

arc4random_uniform⚠

atexit⚠

atof⚠

atoi⚠

backtrace⚠

backtrace_symbols⚠

backtrace_symbols_fd⚠

bind⚠

bsearch⚠

calloc⚠

cap_enter⚠

cap_getmode⚠

cap_rights_contains⚠

cap_rights_is_valid⚠

cap_rights_limit⚠

cap_rights_merge⚠

cap_rights_remove⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfmakesane⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chflags⚠

chflagsat⚠

chmod⚠

chown⚠

chroot⚠

clearerr⚠

clock_getcpuclockid⚠

clock_getcpuclockid2⚠

clock_getres⚠

clock_gettime⚠

clock_settime⚠

close⚠

closedir⚠

closelog⚠

connect⚠

cpuset⚠

cpuset_getaffinity⚠

cpuset_getid⚠

cpuset_setaffinity⚠

cpuset_setid⚠

creat⚠

daemon⚠

dallocx⚠

devname_r⚠

devstat_buildmatch⚠

devstat_checkversion⚠

devstat_getgeneration⚠

devstat_getnumdevs⚠

devstat_getversion⚠

devstat_selectdevs⚠

difftime⚠

dirfd⚠

dl_iterate_phdr⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

drand48⚠

dup⚠

dup2⚠

dup3⚠

duplocale⚠

endgrent⚠

endpwent⚠

endservent⚠

endutxent⚠

erand48⚠

eui64_aton⚠

eui64_hostton⚠

eui64_ntoa⚠

eui64_ntohost⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

exit⚠

explicit_bzero⚠

extattr_delete_fd⚠

extattr_delete_file⚠

extattr_delete_link⚠

extattr_get_fd⚠

extattr_get_file⚠

extattr_get_link⚠

extattr_list_fd⚠

extattr_list_file⚠

extattr_list_link⚠

extattr_namespace_to_string⚠

extattr_set_fd⚠

extattr_set_file⚠

extattr_set_link⚠

extattr_string_to_namespace⚠

faccessat⚠

fchdir⚠

fchflags⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdatasync⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fexecve⚠

fflush⚠

ffs⚠

ffsl⚠

ffsll⚠

fgetc⚠

fgetpos⚠

fgets⚠

fhopen⚠

fhstat⚠

fhstatfs⚠

fileno⚠

flock⚠

flopen⚠

flopenat⚠

fls⚠

flsl⚠

flsll⚠

fmemopen⚠

fopen⚠

fork⚠

forkpty⚠

fparseln⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freeifaddrs⚠

freelocale⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fspacectl⚠

fstat⚠

fstatat⚠

fstatfs⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftok⚠

ftruncate⚠

futimens⚠

futimes⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getdomainname⚠

getdtablesize⚠

getegid⚠

getenv⚠

geteuid⚠

getfh⚠

getfsstat⚠

getgid⚠

getgrent⚠

getgrent_r⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgrouplist⚠

getgroups⚠

gethostid⚠

gethostname⚠

getifaddrs⚠

getitimer⚠

getline⚠

getloadavg⚠

getlocalbase⚠

getlogin⚠

getmntinfo⚠

getnameinfo⚠

getopt⚠

getpagesize⚠

getpagesizes⚠

getpeereid⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getpriority⚠

getprogname⚠

getprotobyname⚠

getprotobynumber⚠

getpwent⚠

getpwent_r⚠

getpwnam⚠

getpwnam_r⚠

getpwuid⚠

getpwuid_r⚠

getresgid⚠

getresuid⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getservbyport⚠

getservent⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

getutxent⚠

getutxid⚠

getutxline⚠

getutxuser⚠

glob⚠

globfree⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

hexdump⚠

hstrerror⚠

humanize_number⚠

iconv⚠

iconv_close⚠

iconv_open⚠

if_freenameindex⚠

if_indextoname⚠

if_nameindex⚠

if_nametoindex⚠

initgroups⚠

ioctl⚠

isalnum⚠

isalpha⚠

isatty⚠

isblank⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

jail⚠

jail_attach⚠

jail_get⚠

jail_remove⚠

jail_set⚠

jrand48⚠

kevent⚠

kill⚠

killpg⚠

kinfo_getvmmap⚠

kld_isloaded⚠

kld_load⚠

kqueue⚠

kvm_close⚠

kvm_counter_u64_fetch⚠

kvm_dpcpu_setcpu⚠

kvm_getargv⚠

kvm_getcptime⚠

kvm_getenvv⚠

kvm_geterr⚠

kvm_getloadavg⚠

kvm_getmaxcpu⚠

kvm_getncpus⚠

kvm_getpcpu⚠

kvm_getprocs⚠

kvm_getswapinfo⚠

kvm_native⚠

kvm_nlist⚠

kvm_nlist2⚠

kvm_open⚠

kvm_openfiles⚠

kvm_read⚠

kvm_read2⚠

kvm_read_zpcpu⚠

kvm_write⚠

labs⚠

lchflags⚠

lchown⚠

lcong48⚠

lgetfh⚠

link⚠

linkat⚠

lio_listio⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lockf⚠

login_tty⚠

lrand48⚠

lseek⚠

lstat⚠

lutimes⚠

madvise⚠

makedev

mallctl⚠

mallctlbymib⚠

mallctlnametomib⚠

malloc⚠

malloc_stats_print⚠

malloc_usable_size⚠

mallocx⚠

memchr⚠

memcmp⚠

memcpy⚠

memfd_create⚠

memmem⚠

memmove⚠

memrchr⚠

memset⚠

memset_s⚠

mincore⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mkfifoat⚠

mknod⚠

mknodat⚠

mkostemp⚠

mkostemps⚠

mkstemp⚠

mkstemps⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

mount⚠

mprotect⚠

mq_close⚠

mq_getattr⚠

mq_getfd_np⚠

mq_notify⚠

mq_open⚠

mq_receive⚠

mq_send⚠

mq_setattr⚠

mq_timedreceive⚠

mq_timedsend⚠

mq_unlink⚠

mrand48⚠

msgctl⚠

msgget⚠

msgrcv⚠

msgsnd⚠

msync⚠

munlock⚠

munlockall⚠

munmap⚠

nallocx⚠

nanosleep⚠

newlocale⚠

nice⚠

nl_langinfo⚠

nl_langinfo_l⚠

nmount⚠

nrand48⚠

ntp_adjtime⚠

ntp_gettime⚠

open⚠

open_memstream⚠

open_wmemstream⚠

openat⚠

opendir⚠

openlog⚠

openpty⚠

pathconf⚠

pause⚠

pclose⚠

pdfork⚠

pdgetpid⚠

pdkill⚠

perror⚠

pipe⚠

pipe2⚠

poll⚠

popen⚠

posix_fadvise⚠

posix_fallocate⚠

posix_madvise⚠

posix_memalign⚠

posix_openpt⚠

posix_spawn⚠

posix_spawn_file_actions_addclose⚠

posix_spawn_file_actions_adddup2⚠

posix_spawn_file_actions_addopen⚠

posix_spawn_file_actions_destroy⚠

posix_spawn_file_actions_init⚠

posix_spawnattr_destroy⚠

posix_spawnattr_getflags⚠

posix_spawnattr_getpgroup⚠

posix_spawnattr_getschedparam⚠

posix_spawnattr_getschedpolicy⚠

posix_spawnattr_getsigdefault⚠

posix_spawnattr_getsigmask⚠

posix_spawnattr_init⚠

posix_spawnattr_setflags⚠

posix_spawnattr_setpgroup⚠

posix_spawnattr_setschedparam⚠

posix_spawnattr_setschedpolicy⚠

posix_spawnattr_setsigdefault⚠

posix_spawnattr_setsigmask⚠

posix_spawnp⚠

ppoll⚠

pread⚠

preadv⚠

printf⚠

procctl⚠

procstat_close⚠

procstat_freeargv⚠

procstat_freeenvv⚠

procstat_freefiles⚠

procstat_freegroups⚠

procstat_freeprocs⚠

procstat_freeptlwpinfo⚠

procstat_freevmmap⚠

procstat_get_pts_info⚠

procstat_get_shm_info⚠

procstat_get_socket_info⚠

procstat_get_vnode_info⚠

procstat_getargv⚠

procstat_getenvv⚠

procstat_getfiles⚠

procstat_getgroups⚠

procstat_getosrel⚠

procstat_getpathname⚠

procstat_getprocs⚠

procstat_getrlimit⚠

procstat_getumask⚠

procstat_getvmmap⚠

procstat_open_core⚠

procstat_open_kvm⚠

procstat_open_sysctl⚠

pselect⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_get_np⚠

pthread_attr_getguardsize⚠

pthread_attr_getstack⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_barrier_destroy⚠

pthread_barrier_init⚠

pthread_barrier_wait⚠

pthread_barrierattr_destroy⚠

pthread_barrierattr_getpshared⚠

pthread_barrierattr_init⚠

pthread_barrierattr_setpshared⚠

pthread_cancel⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getclock⚠

pthread_condattr_getpshared⚠

pthread_condattr_init⚠

pthread_condattr_setclock⚠

pthread_condattr_setpshared⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_get_name_np⚠

pthread_getaffinity_np⚠

pthread_getcpuclockid⚠

pthread_getschedparam⚠

pthread_getspecific⚠

pthread_getthreadid_np⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_main_np⚠

pthread_mutex_consistent⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_timedlock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_getpshared⚠

pthread_mutexattr_getrobust⚠

pthread_mutexattr_init⚠

pthread_mutexattr_setpshared⚠

pthread_mutexattr_setrobust⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_getpshared⚠

pthread_rwlockattr_init⚠

pthread_rwlockattr_setpshared⚠

pthread_self⚠

pthread_set_name_np⚠

pthread_setaffinity_np⚠

pthread_setschedparam⚠

pthread_setspecific⚠

pthread_sigmask⚠

pthread_spin_destroy⚠

pthread_spin_init⚠

pthread_spin_lock⚠

pthread_spin_trylock⚠

pthread_spin_unlock⚠

ptrace⚠

ptsname⚠

ptsname_r⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pututxline⚠

pwrite⚠

pwritev⚠

qsort⚠

querylocale⚠

raise⚠

rallocx⚠

rand⚠

read⚠

readdir⚠

readdir_r⚠

The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

readv⚠

realhostname⚠

realhostname_sa⚠

realloc⚠

reallocarray⚠

realpath⚠

recv⚠

recvfrom⚠

recvmmsg⚠

recvmsg⚠

regcomp⚠

regerror⚠

regexec⚠

regfree⚠

remove⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rfork⚠

rmdir⚠

rtprio⚠

rtprio_thread⚠

sallocx⚠

scanf⚠

sched_get_priority_max⚠

sched_get_priority_min⚠

sched_getaffinity⚠

sched_getcpu⚠

sched_getparam⚠

sched_getscheduler⚠

sched_rr_get_interval⚠

sched_setaffinity⚠

sched_setparam⚠

sched_setscheduler⚠

sched_yield⚠

sdallocx⚠

seed48⚠

seekdir⚠

select⚠

sem_close⚠

sem_destroy⚠

sem_getvalue⚠

sem_init⚠

sem_open⚠

sem_post⚠

sem_timedwait⚠

sem_trywait⚠

sem_unlink⚠

sem_wait⚠

semctl⚠

semget⚠

semop⚠

send⚠

sendfile⚠

sendmmsg⚠

sendmsg⚠

sendto⚠

setaudit⚠

setbuf⚠

setdomainname⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setgrent⚠

setgroups⚠

sethostid⚠

sethostname⚠

setitimer⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setpriority⚠

setproctitle⚠

setprogname⚠

setpwent⚠

setregid⚠

setresgid⚠

setresuid⚠

setreuid⚠

setrlimit⚠

setservent⚠

setsid⚠

setsockopt⚠

settimeofday⚠

setuid⚠

setutxdb⚠

setutxent⚠

setvbuf⚠

shm_create_largepage⚠

shm_open⚠

shm_rename⚠

shm_unlink⚠

shmat⚠

shmctl⚠

shmdt⚠

shmget⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sigtimedwait⚠

sigwait⚠

sigwaitinfo⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

srand⚠

srand48⚠

sscanf⚠

stat⚠

statfs⚠

statvfs⚠

stpcpy⚠

stpncpy⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strndup⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strsignal⚠

strspn⚠

strstr⚠

strtod⚠

strtof⚠

strtok⚠

strtok_r⚠

strtol⚠

strtoul⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sync⚠

syscall⚠

sysconf⚠

sysctl⚠

sysctlbyname⚠

sysctlnametomib⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

telldir⚠

thr_kill⚠

thr_kill2⚠

thr_self⚠

time⚠

timegm⚠

timer_create⚠

timer_delete⚠

timer_getoverrun⚠

timer_gettime⚠

timer_settime⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unmount⚠

unsetenv⚠

uselocale⚠

usleep⚠

utime⚠

utimensat⚠

utimes⚠

utrace⚠

uuidgen⚠

wait⚠

wait4⚠

waitid⚠

waitpid⚠

wcslen⚠

wcstombs⚠

wmemchr⚠

write⚠

writev⚠

xallocx⚠

Type Definitions

Elf32_Addr

Elf32_Half

Elf32_Lword

Elf32_Off

Elf32_Sword

Elf32_Word

Elf64_Addr

Elf64_Half

Elf64_Lword

Elf64_Off

Elf64_Sword

Elf64_Sxword

Elf64_Word

Elf64_Xword

__fixpt_t

__lwpid_t

au_asid_t

au_id_t

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

caddr_t

cc_t

clock_t

clockid_t

cpulevel_t

cpusetid_t

cpuwhich_t

dev_t

fflags_t

fhandle_t

fixpt_t

fsblkcnt_t

fsfilcnt_t

gid_t

iconv_t

id_t

idtype_t

in_addr_t

in_port_t

ino_t

int8_tDeprecated

int16_tDeprecated

int32_tDeprecated

int64_tDeprecated

intmax_t

intptr_t

key_t

kvaddr_t

kvm_t

locale_t

lwpid_t

mode_t

mqd_t

msglen_t

msgqnum_t

nfds_t

nl_item

nlink_t

off_t

pid_t

posix_spawn_file_actions_t

posix_spawnattr_t

pthread_attr_t

pthread_barrier_t

pthread_barrierattr_t

pthread_cond_t

pthread_condattr_t

pthread_key_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

pthread_spinlock_t

pthread_t

ptrdiff_t

register_t

regoff_t

rlim_t

sa_family_t

segsz_t

sem_t

sighandler_t

size_t

socklen_t

speed_t

ssize_t

suseconds_t

tcflag_t

time_t

timer_t

u_char

u_int

u_long

u_short

uid_t

uint8_tDeprecated

uint16_tDeprecated

uint32_tDeprecated

uint64_tDeprecated

uintmax_t

uintptr_t

useconds_t

uuid_t

vm_prot_t

vm_size_t

wchar_t

Unions

__c_anonymous_cr_pid

__c_anonymous_elf32_auxv_union

__c_anonymous_ifc_ifcu

__c_anonymous_ifi_epoch

__c_anonymous_ifi_lastchange

__c_anonymous_ifr_ifru