Crate libc

libc - Raw FFI bindings to platforms' system libraries

Documentation for other platforms.

Structs

Dl_info

addrinfo

dirent

fd_set

group

hostent

in6_addr

in_addr

iovec

ip_mreq

ipv6_mreq

itimerval

lconv

linger

passwd

pollfd

protoent

pte_handle_t

rlimit

rusage

sched_param

sem_t

servent

sigaction

sigval

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_storage

sockaddr_un

stat

statvfs

termios

timespec

timeval

tm

tms

utimbuf

utsname

winsize

Enums

DIR

FILE

c_void
Equivalent to C's void type when used as a pointer.

fpos_t

Constants

AF_INET

AF_INET6

AF_UNIX

AF_UNSPEC

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

CLOCK_MONOTONIC

CLOCK_REALTIME

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADV

EAFNOSUPPORT

EAGAIN

EAI_SYSTEM

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDOTDOT

EDQUOT

EEXIST

EFAULT

EFBIG

EHOSTDOWN

EHOSTUNREACH

EHWPOISON

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EISNAM

EKEYEXPIRED

EKEYREJECTED

EKEYREVOKED

EL2NSYNC

EL2HLT

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOOP

EMEDIUMTYPE

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENAVAIL

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOKEY

ENOLCK

ENOLINK

ENOMEDIUM

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTNAM

ENOTRECOVERABLE

ENOTSOCK

ENOTTY

ENOTUNIQ

ENXIO

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERANGE

EREMCHG

EREMOTE

EREMOTEIO

ERESTART

ERFKILL

EROFS

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUCLEAN

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

FD_CLOEXEC

FD_SETSIZE

FIOCLEX

FIONBIO

FIONREAD

F_CNVT

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLK

F_GETOWN

F_RGETLK

F_RSETLK

F_RSETLKW

F_SETFD

F_SETFL

F_SETLK

F_SETLKW

F_SETOWN

GRPQUOTA

IFNAMSIZ

IF_NAMESIZE

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INT_MAX

INT_MIN

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_TCP

IPPROTO_UDP

IPV6_JOIN_GROUP

IPV6_ADD_MEMBERSHIP

IPV6_LEAVE_GROUP

IPV6_DROP_MEMBERSHIP

IPV6_MULTICAST_LOOP

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TOS

IP_TTL

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

MSG_DONTWAIT

MSG_MORE

MSG_OOB

MSG_PEEK

MSG_WAITALL

O_ACCMODE

O_APPEND

O_CLOEXEC

O_CREAT

O_EXCL

O_NOCTTY

O_NONBLOCK

O_RDONLY

O_RDWR

O_TRUNC

O_WRONLY

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PTHREAD_COND_INITIALIZER

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

RTLD_DEFAULT

SEEK_CUR

SEEK_END

SEEK_SET

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGEMT

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIOT

SIGKILL

SIGPIPE

SIGQUIT

SIGSEGV

SIGTERM

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SOCK_DGRAM

SOCK_RAW

SOCK_STREAM

SOL_SOCKET

SO_ACCEPTCONN

SO_BROADCAST

SO_CONTIMEO

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_NO_CHECK

SO_OOBINLINE

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TYPE

SO_USELOOPBACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

S_BLKSIZE

S_ENFMT

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TCP_KEEPALIVE

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_NODELAY

USRQUOTA

WNOHANG

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LEVEL1_ICACHE_SIZE

_SC_LEVEL1_ICACHE_ASSOC

_SC_LEVEL1_ICACHE_LINESIZE

_SC_LEVEL1_DCACHE_SIZE

_SC_LEVEL1_DCACHE_ASSOC

_SC_LEVEL1_DCACHE_LINESIZE

_SC_LEVEL2_CACHE_SIZE

_SC_LEVEL2_CACHE_ASSOC

_SC_LEVEL2_CACHE_LINESIZE

_SC_LEVEL3_CACHE_SIZE

_SC_LEVEL3_CACHE_ASSOC

_SC_LEVEL3_CACHE_LINESIZE

_SC_LEVEL4_CACHE_SIZE

_SC_LEVEL4_CACHE_ASSOC

_SC_LEVEL4_CACHE_LINESIZE

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SS_REPL_MAX

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_CEILING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_ROBUST_PRIO_INHERIT

_SC_THREAD_ROBUST_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_V6_LPBIG_OFFBIG

_SC_V7_LPBIG_OFFBIG

_SC_V6_ILP32_OFFBIG

_SC_V6_ILP32_OFF32

_SC_V6_LP64_OFF64

_SC_V7_ILP32_OFFBIG

_SC_V7_ILP32_OFF32

_SC_V7_LP64_OFF64

_SC_VERSION

_SC_XBS5_LPBIG_OFFBIG

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_ILP32_OFF32

_SC_XBS5_LP64_OFF64

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_UUCP

_SC_XOPEN_VERSION

Functions

WEXITSTATUS⚠

WIFEXITED⚠

WTERMSIG⚠

_exit⚠

abort⚠

abs⚠

accept⚠

access⚠

alarm⚠

atexit⚠

atof⚠

atoi⚠

bind⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clock_gettime⚠

close⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

difftime⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

exit⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgets⚠

fileno⚠

flock⚠

fopen⚠

fork⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fstat⚠

fstatat⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftruncate⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgroups⚠

gethostname⚠

getline⚠

getlogin⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getprotobyname⚠

getprotobynumber⚠

getpwnam⚠

getpwuid⚠

getpwuid_r⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

if_indextoname⚠

if_nametoindex⚠

ioctl⚠

isalnum⚠

isalpha⚠

isatty⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kill⚠

killpg⚠

labs⚠

lchown⚠

link⚠

linkat⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lseek⚠

lstat⚠

malloc⚠

memalign⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memset⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mknod⚠

mkstemp⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

nice⚠

open⚠

openat⚠

opendir⚠

openlog⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

poll⚠

posix_memalign⚠

posix_openpt⚠

pread⚠

printf⚠

pselect⚠

pthread_attr_destroy⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_init⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_init⚠

pthread_self⚠

pthread_setspecific⚠

pthread_sigmask⚠

ptsname⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pwrite⚠

raise⚠

rand⚠

read⚠

readdir⚠

readdir_r⚠
The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

realloc⚠

realpath⚠

recv⚠

recvfrom⚠

remove⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

scanf⚠

sched_yield⚠

select⚠

sem_destroy⚠

sem_init⚠

sem_post⚠

sem_trywait⚠

sem_wait⚠

send⚠

sendto⚠

setbuf⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setgroups⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setrlimit⚠

setsid⚠

setsockopt⚠

setuid⚠

setvbuf⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

srand⚠

sscanf⚠

stat⚠

statvfs⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strspn⚠

strstr⚠

strtod⚠

strtok⚠

strtol⚠

strtoul⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sysconf⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

usleep⚠

utime⚠

utimes⚠

wait⚠

waitpid⚠

wcslen⚠

wcstombs⚠

write⚠

Type Definitions

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

dev_t

gid_t

in_addr_t

in_port_t

ino_t

int8_t	Deprecated

int16_t	Deprecated

int32_t	Deprecated

int64_t	Deprecated

intmax_t

intptr_t

locale_t

mode_t

nfds_t

nlink_t

off_t

pid_t

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_key_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

pthread_t

ptrdiff_t

rlim_t

sa_family_t

sighandler_t

sigset_t

size_t

socklen_t

speed_t

ssize_t

stat64

suseconds_t

time_t

uid_t

uint8_t	Deprecated

uint16_t	Deprecated

uint32_t	Deprecated

uint64_t	Deprecated

uintmax_t

uintptr_t

wchar_t