Crate libc

libc - Raw FFI bindings to platforms' system libraries

Documentation for other platforms.

Structs

Dl_info

addrinfo

cmsghdr

dirent

door_arg_t

door_desc_t

door_desc_t__d_data__d_desc

epoll_event

fd_set

flock

glob_t

group

hostent

if_nameindex

ifaddrs

in6_addr

in_addr

iovec

ip_mreq

ipv6_mreq

itimerval

lconv

linger

mq_attr

msghdr

passwd

pollfd

port_event

protoent

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

rlimit

rusage

sched_param

sem_t

servent

sigaction

sigevent

siginfo_t

sigset_t

sigval

sockaddr

sockaddr_dl

sockaddr_in

sockaddr_in6

sockaddr_storage

sockaddr_un

stack_t

stat

statvfs

termios

timespec

timeval

tm

tms

utimbuf

utsname

winsize

Enums

DIR

FILE

c_void
Equivalent to C's void type when used as a pointer.

fpos_t

timezone

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

AF_802

AF_APPLETALK

AF_CCITT

AF_CHAOS

AF_DATAKIT

AF_DECnet

AF_DLI

AF_ECMA

AF_FILE

AF_GOSIP

AF_HYLINK

AF_IMPLINK

AF_INET

AF_INET6

AF_INET_OFFLOAD

AF_IPX

AF_KEY

AF_LAT

AF_LINK

AF_LOCAL

AF_LX_NETLINK

AF_NBS

AF_NCA

AF_NIT

AF_NS

AF_OSI

AF_OSINET

AF_PACKET

AF_POLICY

AF_PUP

AF_ROUTE

AF_SNA

AF_TRILL

AF_UNIX

AF_UNSPEC

AF_X25

ALT_DIGITS

AM_STR

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

AT_FDCWD

AT_SYMLINK_NOFOLLOW

B0

B50

B75

B110

B134

B150

B200

B300

B600

B1200

B1800

B2400

B4800

B9600

B19200

B38400

B57600

B76800

B115200

B153600

B230400

B307200

B460800

B921600

BRKINT

BUFSIZ

CLOCAL

CLOCK_MONOTONIC

CLOCK_REALTIME

CODESET

CREAD

CRNCYSTR

CRTSCTS

CS5

CS6

CS7

CS8

CSIZE

CSTART

CSTOP

CSTOPB

CSWTCH

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

D_FMT

D_T_FMT

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADI

EADV

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NODATA

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDQUOT

EEXIST

EFAULT

EFBIG

EHOSTDOWN

EHOSTUNREACH

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EL2NSYNC

EL2HLT

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOCKUNMAPPED

ELOOP

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOLCK

ENOLINK

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTACTIVE

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENOTUNIQ

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPOLLERR

EPOLLET

EPOLLEXCLUSIVE

EPOLLHUP

EPOLLIN

EPOLLMSG

EPOLLONESHOT

EPOLLOUT

EPOLLPRI

EPOLLRDBAND

EPOLLRDHUP

EPOLLRDNORM

EPOLLWRBAND

EPOLLWRNORM

EPOLL_CLOEXEC

EPOLL_CTL_ADD

EPOLL_CTL_DEL

EPOLL_CTL_MOD

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EREMCHG

EREMOTE

ERESTART

EROFS

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

EXTPROC

FD_CLOEXEC

FD_SETSIZE

FILENAME_MAX

FIOASYNC

FIOCLEX

FIOGETOWN

FIONBIO

FIONCLEX

FIONREAD

FIOSETOWN

FLUSHO

FOPEN_MAX

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLK

F_LOCK

F_OK

F_RDLCK

F_SETFD

F_SETFL

F_SETLK

F_SETLKW

F_TEST

F_TLOCK

F_ULOCK

F_UNLCK

F_WRLCK

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRPQUOTA

HUPCL

ICANON

ICRNL

IEXTEN

IFF_ADDRCONF

IFF_ALLMULTI

IFF_ANYCAST

IFF_BROADCAST

IFF_COS_ENABLED

IFF_DEBUG

IFF_DEPRECATED

IFF_DHCPRUNNING

IFF_DUPLICATE

IFF_FAILED

IFF_FIXEDMTU

IFF_INACTIVE

IFF_INTELLIGENT

IFF_IPMP

IFF_IPV4

IFF_IPV6

IFF_LOOPBACK

IFF_MULTICAST

IFF_MULTI_BCAST

IFF_NOARP

IFF_NOFAILOVER

IFF_NOLOCAL

IFF_NONUD

IFF_NORTEXCH

IFF_NOTRAILERS

IFF_NOXMIT

IFF_OFFLINE

IFF_POINTOPOINT

IFF_PREFERRED

IFF_PRIVATE

IFF_PROMISC

IFF_ROUTER

IFF_RUNNING

IFF_STANDBY

IFF_TEMPORARY

IFF_UNNUMBERED

IFF_UP

IFF_VIRTUAL

IFNAMSIZ

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

IMAXBEL

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INLCR

INPCK

INT_MAX

INT_MIN

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_RAW

IPPROTO_TCP

IPPROTO_UDP

IPV6_UNICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_LOOP

IPV6_JOIN_GROUP

IPV6_LEAVE_GROUP

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TTL

ISIG

ISTRIP

IXANY

IXOFF

IXON

LC_ALL

LC_ALL_MASK

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_TIME

LC_TIME_MASK

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_CRON

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

L_tmpnam

MADV_DONTNEED

MADV_FREE

MADV_NORMAL

MADV_RANDOM

MADV_SEQUENTIAL

MADV_WILLNEED

MAP_ALIGN

MAP_ANON

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_INITDATA

MAP_NORESERVE

MAP_PRIVATE

MAP_RENAME

MAP_SHARED

MAP_TEXT

MAXSTRMSG

MCL_CURRENT

MCL_FUTURE

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_DUPCTRL

MSG_EOR

MSG_MAXIOVLEN

MSG_NOSIGNAL

MSG_NOTIFICATION

MSG_OOB

MSG_PEEK

MSG_TRUNC

MSG_WAITALL

MSG_XPG4_2

MS_ASYNC

MS_INVALCURPROC

MS_INVALIDATE

MS_SYNC

NCCS

NI_MAXHOST

NOEXPR

NOFLSH

NOSTR

OCRNL

ONLCR

ONLRET

ONOCR

OPOST

O_ACCMODE

O_APPEND

O_CLOEXEC

O_CREAT

O_DSYNC

O_EXCL

O_EXEC

O_NDELAY

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_RDONLY

O_RDWR

O_SEARCH

O_SYNC

O_TRUNC

O_WRONLY

PARENB

PARMRK

PARODD

PATH_MAX

PENDIN

PM_STR

POLLERR

POLLHUP

POLLIN

POLLNORM

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

POLLWRBAND

POLLWRNORM

PORT_SOURCE_AIO

PORT_SOURCE_ALERT

PORT_SOURCE_FD

PORT_SOURCE_FILE

PORT_SOURCE_MQ

PORT_SOURCE_POSTWAIT

PORT_SOURCE_SIGNAL

PORT_SOURCE_TIMER

PORT_SOURCE_USER

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PROT_EXEC

PROT_NONE

PROT_READ

PROT_WRITE

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

P_ALL

P_CID

P_CPUID

P_CTID

P_GID

P_LWPID

P_PGID

P_PID

P_POOLID

P_PPID

P_PROJID

P_PSETID

P_SID

P_TASKID

P_UID

P_ZONEID

RADIXCHAR

RAND_MAX

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_NOFILE

RLIMIT_STACK

RLIMIT_VMEM

RLIM_INFINITY

RLIM_NLIMITS

RTLD_CONFGEN

RTLD_DEFAULT

RTLD_FIRST

RTLD_GLOBAL

RTLD_GROUP

RTLD_LAZY

RTLD_LOCAL

RTLD_NEXT

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

RTLD_PARENT

RTLD_PROBE

RTLD_SELF

RTLD_WORLD

RUSAGE_CHILDREN

RUSAGE_SELF

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SCM_RIGHTS

SCM_TIMESTAMP

SCM_UCRED

SEEK_CUR

SEEK_END

SEEK_SET

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEMT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGFPE

SIGHUP

SIGILL

SIGINFO

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPOLL

SIGPROF

SIGPWR

SIGQUIT

SIGSEGV

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SOCK_DGRAM

SOCK_RAW

SOCK_RDM

SOCK_SEQPACKET

SOCK_STREAM

SOL_SOCKET

SO_ACCEPTCONN

SO_BROADCAST

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_OOBINLINE

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TYPE

SO_USELOOPBACK

SS_DISABLE

SS_ONSTACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

ST_NOSUID

ST_RDONLY

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TCDSET

TCFLSH

TCGETA

TCGETS

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_KEEPIDLE

TCP_NODELAY

TCSADRAIN

TCSAFLUSH

TCSANOW

TCSBRK

TCSETA

TCSETAF

TCSETAW

TCSETS

TCSETSF

TCSETSW

TCXONC

THOUSEP

TIMER_ABSTIME

TIMER_RELTIME

TIOC

TIOCCBRK

TIOCCDTR

TIOCCILOOP

TIOCEXCL

TIOCFLUSH

TIOCGETC

TIOCGETD

TIOCGETLD

TIOCGETP

TIOCGLTC

TIOCGPGRP

TIOCGPPS

TIOCGPPSEV

TIOCGSID

TIOCGSOFTCAR

TIOCGWINSZ

TIOCHPCL

TIOCKBOF

TIOCKBON

TIOCLBIC

TIOCLBIS

TIOCLGET

TIOCLSET

TIOCMBIC

TIOCMBIS

TIOCMGET

TIOCMSET

TIOCNOTTY

TIOCNXCL

TIOCOUTQ

TIOCREMOTE

TIOCSBRK

TIOCSCTTY

TIOCSDTR

TIOCSETC

TIOCSETD

TIOCSETLD

TIOCSETN

TIOCSETP

TIOCSIGNAL

TIOCSILOOP

TIOCSLTC

TIOCSPGRP

TIOCSPPS

TIOCSSOFTCAR

TIOCSTART

TIOCSTI

TIOCSTOP

TIOCSWINSZ

TMP_MAX

TOSTOP

T_FMT

T_FMT_AMPM

USRQUOTA

VDISCARD

VDSUSP

VEOF

VEOL

VEOL2

VERASE

VERASE2

VINTR

VKILL

VLNEXT

VMIN

VQUIT

VREPRINT

VSTART

VSTATUS

VSTOP

VSUSP

VSWTCH

VTIME

VWERASE

WCONTINUED

WEXITED

WNOHANG

WNOWAIT

WSTOPPED

WTRAPPED

WUNTRACED

W_OK

X_OK

YESEXPR

YESSTR

_COND_MAGIC

_DATE_FMT

_IOFBF

_IOLBF

_IONBF

_MUTEX_MAGIC

_PC_2_SYMLINKS

_PC_ACCESS_FILTERING

_PC_ACL_ENABLED

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CASE_BEHAVIOR

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LAST

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_MIN_HOLE_SIZE

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SATTR_ENABLED

_PC_SATTR_EXISTS

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_TIMESTAMP_RESOLUTION

_PC_VDISABLE

_PC_XATTR_ENABLED

_PC_XATTR_EXISTS

_RWL_MAGIC

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_C_VERSION

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_2_CHAR_TERM

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COHER_BLKSZ

_SC_COLL_WEIGHTS_MAX

_SC_CPUID_MAX

_SC_CPUTIME

_SC_DCACHE_ASSOC

_SC_DCACHE_BLKSZ

_SC_DCACHE_LINESZ

_SC_DCACHE_SZ

_SC_DCACHE_TBLKSZ

_SC_DELAYTIMER_MAX

_SC_EPHID_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_ICACHE_ASSOC

_SC_ICACHE_BLKSZ

_SC_ICACHE_LINESZ

_SC_ICACHE_SZ

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_LOGNAME_MAX

_SC_MAPPED_FILES

_SC_MAXPID

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_MAX

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SIGRT_MAX

_SC_SIGRT_MIN

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPLIT_CACHE

_SC_SPORADIC_SERVER

_SC_SS_REPL_MAX

_SC_STACK_PROT

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_T_IOV_MAX

_SC_V6_LPBIG_OFFBIG

_SC_V6_ILP32_OFFBIG

_SC_V6_ILP32_OFF32

_SC_V6_LP64_OFF64

_SC_VERSION

_SC_XBS5_LPBIG_OFFBIG

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_ILP32_OFF32

_SC_XBS5_LP64_OFF64

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

Functions

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

WCOREDUMP⚠

WEXITSTATUS⚠

WIFCONTINUED⚠

WIFEXITED⚠

WIFSIGNALED⚠

WIFSTOPPED⚠

WSTOPSIG⚠

WTERMSIG⚠

___errno⚠

_exit⚠

abort⚠

abs⚠

accept⚠

access⚠

acct⚠

alarm⚠

atexit⚠

atof⚠

atoi⚠

bind⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

clock_getres⚠

clock_gettime⚠

clock_nanosleep⚠

clock_settime⚠

close⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

daemon⚠

difftime⚠

dirfd⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

door_call⚠

door_create⚠

door_return⚠

dup⚠

dup2⚠

dup3⚠

duplocale⚠

endgrent⚠

endpwent⚠

epoll_create⚠

epoll_create1⚠

epoll_ctl⚠

epoll_pwait⚠

epoll_wait⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

exit⚠

fattach⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdatasync⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fexecve⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgets⚠

fileno⚠

flock⚠

fopen⚠

fork⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freeifaddrs⚠

freelocale⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fstat⚠

fstatat⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftruncate⚠

futimesat⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getdtablesize⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgrent⚠

getgrent_r⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgroups⚠

gethostname⚠

getifaddrs⚠

getline⚠

getloadavg⚠

getlogin⚠

getnameinfo⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getpriority⚠

getprogname⚠

getprotobyname⚠

getprotobynumber⚠

getpwent⚠

getpwent_r⚠

getpwnam⚠

getpwnam_r⚠

getpwuid⚠

getpwuid_r⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

glob⚠

globfree⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

if_freenameindex⚠

if_indextoname⚠

if_nameindex⚠

if_nametoindex⚠

initgroups⚠

ioctl⚠

isalnum⚠

isalpha⚠

isatty⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kill⚠

killpg⚠

labs⚠

lchown⚠

link⚠

linkat⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lseek⚠

lstat⚠

madvise⚠

malloc⚠

memalign⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memset⚠

mincore⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mkfifoat⚠

mknod⚠

mknodat⚠

mkstemp⚠

mkstemps⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

mprotect⚠

mq_close⚠

mq_getattr⚠

mq_open⚠

mq_receive⚠

mq_send⚠

mq_setattr⚠

mq_timedreceive⚠

mq_timedsend⚠

mq_unlink⚠

msync⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

newlocale⚠

nice⚠

nl_langinfo⚠

nl_langinfo_l⚠

open⚠

openat⚠

opendir⚠

openlog⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

pipe2⚠

poll⚠

popen⚠

port_associate⚠

port_create⚠

port_dissociate⚠

port_get⚠

port_getn⚠

posix_madvise⚠

posix_memalign⚠

posix_openpt⚠

pread⚠

printf⚠

pselect⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getclock⚠

pthread_condattr_init⚠

pthread_condattr_setclock⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_timedlock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_init⚠

pthread_self⚠

pthread_setspecific⚠

pthread_sigmask⚠

ptsname⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pwrite⚠

raise⚠

rand⚠

read⚠

readdir⚠

readdir_r⚠
The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

readv⚠

realloc⚠

realpath⚠

recv⚠

recvfrom⚠

recvmsg⚠

remove⚠

rename⚠

renameat⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

scanf⚠

sched_yield⚠

seekdir⚠

select⚠

sem_close⚠

sem_destroy⚠

sem_getvalue⚠

sem_init⚠

sem_open⚠

sem_post⚠

sem_timedwait⚠

sem_trywait⚠

sem_unlink⚠

sem_wait⚠

send⚠

sendmsg⚠

sendto⚠

setbuf⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setgrent⚠

setgroups⚠

sethostname⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setpriority⚠

setprogname⚠

setpwent⚠

setrlimit⚠

setsid⚠

setsockopt⚠

setuid⚠

setvbuf⚠

shm_open⚠

shm_unlink⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sigwait⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

srand⚠

sscanf⚠

stack_getbounds⚠

stat⚠

statvfs⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strspn⚠

strstr⚠

strtod⚠

strtok⚠

strtol⚠

strtoul⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sysconf⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

telldir⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

uselocale⚠

usleep⚠

utime⚠

utimensat⚠

utimes⚠

wait⚠

waitid⚠

waitpid⚠

wcslen⚠

wcstombs⚠

write⚠

writev⚠

Type Definitions

blkcnt_t

blksize_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

daddr_t

dev_t

door_attr_t

door_id_t

fsblkcnt_t

fsfilcnt_t

gid_t

id_t

idtype_t

in_addr_t

in_port_t

ino_t

int8_t	Deprecated

int16_t	Deprecated

int32_t	Deprecated

int64_t	Deprecated

intmax_t

intptr_t

key_t

locale_t

major_t

minor_t

mode_t

mqd_t

nfds_t

nl_item

nlink_t

off_t

pid_t

pthread_key_t

pthread_t

ptrdiff_t

rlim_t

sa_family_t

sighandler_t

size_t

socklen_t

speed_t

ssize_t

suseconds_t

tcflag_t

time_t

uid_t

uint8_t	Deprecated

uint16_t	Deprecated

uint32_t	Deprecated

uint64_t	Deprecated

uintmax_t

uintptr_t

useconds_t

wchar_t

Unions

door_desc_t__d_data