Crate libc

libc - Raw FFI bindings to platforms' system libraries

Documentation for other platforms.

Structs

Dl_info

addrinfo

aiocb

arphdr

bpf_hdr

cmsghdr

dirent

dqblk

fd_set

flock

fsid_t

fstore_t

glob_t

group

hostent

if_data

if_msghdr

if_nameindex

ifaddrs

in6_pktinfo

in6_addr

in_addr

in_pktinfo

iovec

ip_mreq

ipc_perm

ipv6_mreq

itimerval

kevent

kevent64_s

lconv

linger

load_command

mach_header	Deprecated

mach_header_64	Deprecated

mach_timebase_info	Deprecated

max_align_t

msghdr

passwd

pollfd

proc_bsdinfo

proc_taskallinfo

proc_taskinfo

proc_threadinfo

protoent

pthread_attr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

radvisory

rlimit

rusage

segment_command

segment_command_64

sembuf

semid_ds

servent

sf_hdtr

shmid_ds

sigaction

sigevent

siginfo_t

sigval

sockaddr

sockaddr_ctl

sockaddr_dl

sockaddr_in

sockaddr_in6

sockaddr_inarp

sockaddr_storage

sockaddr_un

stack_t

stat

statfs

statvfs

termios

timespec

timeval

timeval32

tm

tms

utimbuf

utmpx

utsname

winsize

xsw_usage

xucred

Enums

DIR

FILE

c_void
Equivalent to C's void type when used as a pointer.

fpos_t

timezone

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

ACCOUNTING

AF_APPLETALK

AF_CCITT

AF_CHAOS

AF_CNT

AF_COIP

AF_DATAKIT

AF_DECnet

AF_DLI

AF_E164

AF_ECMA

AF_HYLINK

AF_IMPLINK

AF_INET

AF_INET6

AF_IPX

AF_ISDN

AF_ISO

AF_LAT

AF_LINK

AF_LOCAL

AF_NATM

AF_NETBIOS

AF_NS

AF_OSI

AF_PPP

AF_PUP

AF_ROUTE

AF_SIP

AF_SNA

AF_SYSTEM

AF_SYS_CONTROL

AF_UNIX

AF_UNSPEC

AIO_ALLDONE

AIO_CANCELED

AIO_LISTIO_MAX	Deprecated

AIO_NOTCANCELED

AI_ADDRCONFIG

AI_ALL

AI_CANONNAME

AI_DEFAULT

AI_MASK

AI_NUMERICHOST

AI_NUMERICSERV

AI_PASSIVE

AI_UNUSABLE

AI_V4MAPPED_CFG

AI_V4MAPPED

ALTWERASE

ALT_DIGITS

AM_STR

ARPOP_REPLY

ARPOP_REQUEST

ATF_COM

ATF_PERM

ATF_PUBL

ATF_USETRAILERS

AT_EACCESS

AT_FDCWD

AT_REMOVEDIR

AT_SYMLINK_FOLLOW

AT_SYMLINK_NOFOLLOW

B0

B50

B75

B110

B134

B150

B200

B300

B600

B1200

B1800

B2400

B4800

B7200

B9600

B14400

B19200

B28800

B38400

B57600

B76800

B115200

B230400

BIOCFLUSH

BIOCGBLEN

BIOCGDLT

BIOCGDLTLIST

BIOCGETIF

BIOCGHDRCMPLT

BIOCGRSIG

BIOCGRTIMEOUT

BIOCGSEESENT

BIOCGSTATS

BIOCIMMEDIATE

BIOCPROMISC

BIOCSBLEN

BIOCSDLT

BIOCSETF

BIOCSETFNR

BIOCSETIF

BIOCSHDRCMPLT

BIOCSRSIG

BIOCSRTIMEOUT

BIOCSSEESENT

BIOCVERSION

BOOT_TIME

BPF_ALIGNMENT

BRKINT

BS0

BS1

BSDLY

BUFSIZ

CIGNORE

CLOCAL

CLOCK_MONOTONIC

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_THREAD_CPUTIME_ID

CODESET

CR0

CR1

CR2

CR3

CRDLY

CREAD

CRNCYSTR

CRTSCTS

CS5

CS6

CS7

CS8

CSIZE

CSTOPB

CTLFLAG_ANYBODY

CTLFLAG_KERN

CTLFLAG_LOCKED

CTLFLAG_MASKED

CTLFLAG_NOAUTO

CTLFLAG_NOLOCK

CTLFLAG_OID2

CTLFLAG_RD

CTLFLAG_RW

CTLFLAG_SECURE

CTLFLAG_WR

CTLTYPE

CTLTYPE_INT

CTLTYPE_NODE

CTLTYPE_OPAQUE

CTLTYPE_QUAD

CTLTYPE_STRING

CTLTYPE_STRUCT

CTL_DEBUG

CTL_DEBUG_MAXID

CTL_DEBUG_NAME

CTL_DEBUG_VALUE

CTL_HW

CTL_KERN

CTL_MACHDEP

CTL_MAXID

CTL_NET

CTL_UNSPEC

CTL_USER

CTL_VFS

CTL_VM

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

DEAD_PROCESS

DLT_ARCNET

DLT_ATM_RFC1483

DLT_AX25

DLT_CHAOS

DLT_EN3MB

DLT_EN10MB

DLT_FDDI

DLT_IEEE802

DLT_LOOP

DLT_NULL

DLT_PPP

DLT_PRONET

DLT_RAW

DLT_SLIP

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

DT_UNKNOWN

D_FMT

D_MD_ORDER

D_T_FMT

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EAFNOSUPPORT

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NODATA

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EALREADY

EAUTH

EBADARCH

EBADEXEC

EBADF

EBADMACHO

EBADMSG

EBADRPC

EBUSY

ECANCELED

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDESTADDRREQ

EDEVERR

EDOM

EDQUOT

EEXIST

EFAULT

EFBIG

EFTYPE

EHOSTDOWN

EHOSTUNREACH

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

ELAST

ELOOP

EMFILE

EMLINK

EMPTY

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENEEDAUTH

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOATTR

ENOBUFS

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOLCK

ENOLINK

ENOMEM

ENOMSG

ENOPOLICY

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTRECOVERABLE

ENOTSOCK

ENOTSUP

ENOTTY

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPROCLIM

EPROCUNAVAIL

EPROGMISMATCH

EPROGUNAVAIL

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

EPWROFF

EQFULL

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EREMOTE

EROFS

ERPCMISMATCH

ESHLIBVERS

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESTALE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUSERS

EVFILT_AIO

EVFILT_FS

EVFILT_MACHPORT

EVFILT_PROC

EVFILT_READ

EVFILT_SIGNAL

EVFILT_TIMER

EVFILT_USER

EVFILT_VM

EVFILT_VNODE

EVFILT_WRITE

EV_ADD

EV_CLEAR

EV_DELETE

EV_DISABLE

EV_DISPATCH

EV_ENABLE

EV_EOF

EV_ERROR

EV_FLAG0

EV_FLAG1

EV_ONESHOT

EV_OOBAND

EV_POLL

EV_RECEIPT

EV_SYSFLAGS

EWOULDBLOCK

EXDEV

EXIT_FAILURE

EXIT_SUCCESS

EXTA

EXTB

EXTPROC

FD_CLOEXEC

FD_SETSIZE

FF0

FF1

FFDLY

FILENAME_MAX

FIOASYNC

FIOCLEX

FIODTYPE

FIOGETOWN

FIONBIO

FIONCLEX

FIONREAD

FIOSETOWN

FLUSHO

FOPEN_MAX

F_ALLOCATEALL

F_ALLOCATECONTIG

F_DUPFD

F_DUPFD_CLOEXEC

F_FREEZE_FS

F_FULLFSYNC

F_GETFD

F_GETFL

F_GETLK

F_GETOWN

F_GETPATH

F_GLOBAL_NOCACHE

F_LOCK

F_NOCACHE

F_NODIRECT

F_OK

F_PEOFPOSMODE

F_PREALLOCATE

F_RDADVISE

F_RDAHEAD

F_RDLCK

F_SETFD

F_SETFL

F_SETLK

F_SETLKW

F_SETOWN

F_TEST

F_THAW_FS

F_TLOCK

F_ULOCK

F_UNLCK

F_VOLPOSMODE

F_WRLCK

GETALL

GETNCNT

GETPID

GETVAL

GETZCNT

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRPQUOTA

HUPCL

HW_AVAILCPU

HW_BUS_FREQ

HW_BYTEORDER

HW_CACHELINE

HW_CPU_FREQ

HW_DISKNAMES

HW_DISKSTATS

HW_EPOCH

HW_FLOATINGPT

HW_L1ICACHESIZE

HW_L1DCACHESIZE

HW_L2SETTINGS

HW_L2CACHESIZE

HW_L3SETTINGS

HW_L3CACHESIZE

HW_MACHINE

HW_MACHINE_ARCH

HW_MAXID

HW_MEMSIZE

HW_MODEL

HW_NCPU

HW_PAGESIZE

HW_PHYSMEM

HW_TB_FREQ

HW_USERMEM

HW_VECTORUNIT

ICANON

ICRNL

IEXTEN

IFF_ALLMULTI

IFF_ALTPHYS

IFF_BROADCAST

IFF_DEBUG

IFF_LINK0

IFF_LINK1

IFF_LINK2

IFF_LOOPBACK

IFF_MULTICAST

IFF_NOARP

IFF_NOTRAILERS

IFF_OACTIVE

IFF_POINTOPOINT

IFF_PROMISC

IFF_RUNNING

IFF_SIMPLEX

IFF_UP

IFNAMSIZ

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

IMAXBEL

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

INIT_PROCESS

INLCR

INPCK

INT_MAX

INT_MIN

IPC_CREAT

IPC_EXCL

IPC_M

IPC_NOWAIT

IPC_PRIVATE

IPC_R

IPC_RMID

IPC_SET

IPC_STAT

IPC_W

IPPROTO_3PC
Third Party Connect

IPPROTO_ADFS
Any distributed FS

IPPROTO_AH
IP6 Auth Header

IPPROTO_AHIP
any host internal protocol

IPPROTO_APES
any private encr. scheme

IPPROTO_ARGUS
Argus

IPPROTO_AX25
AX.25 Frames

IPPROTO_BHA
BHA

IPPROTO_BLT
Bulk Data Transfer

IPPROTO_BRSATMON
BackRoom SATNET Monitoring

IPPROTO_CFTP
CFTP

IPPROTO_CHAOS
Chaos

IPPROTO_CMTP
Control Message Transport

IPPROTO_CPHB
Comp. Prot. HeartBeat

IPPROTO_CPNX
Comp. Prot. Net. Executive

IPPROTO_DDP
Datagram Delivery

IPPROTO_DGP
dissimilar gateway prot.

IPPROTO_DIVERT
divert pseudo-protocol

IPPROTO_DONE
last return value of *_input(), meaning "all job for this pkt is done".

IPPROTO_DSTOPTS
IP6 destination option

IPPROTO_EGP
exterior gateway protocol

IPPROTO_EMCON
EMCON

IPPROTO_ENCAP
encapsulation header

IPPROTO_EON
ISO cnlp

IPPROTO_ESP
IP6 Encap Sec. Payload

IPPROTO_ETHERIP
Ethernet IP encapsulation

IPPROTO_FRAGMENT
IP6 fragmentation header

IPPROTO_GGP
gateway2 (deprecated)

IPPROTO_GMTP
GMTP

IPPROTO_GRE
General Routing Encap.

IPPROTO_HELLO
"hello" routing protocol

IPPROTO_HMP
Host Monitoring

IPPROTO_HOPOPTS
IP6 hop-by-hop options

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IDP
xns idp

IPPROTO_IDPR
InterDomain Policy Routing

IPPROTO_IDRP
InterDomain Routing

IPPROTO_IGMP
group mgmt protocol

IPPROTO_IGP
NSFNET-IGP

IPPROTO_IGRP
Cisco/GXS IGRP

IPPROTO_IL
IL transport protocol

IPPROTO_INLSP
Integ. Net Layer Security

IPPROTO_INP
Merit Internodal

IPPROTO_IP

IPPROTO_IPCOMP
payload compression (IPComp)

IPPROTO_IPCV
Packet Core Utility

IPPROTO_IPEIP
IP encapsulated in IP

IPPROTO_IPIP
for compatibility

IPPROTO_IPPC
Pluribus Packet Core

IPPROTO_IPV6

IPPROTO_IRTP
Reliable Transaction

IPPROTO_KRYPTOLAN
Kryptolan

IPPROTO_LARP
Locus Address Resoloution

IPPROTO_LEAF1
Leaf-1

IPPROTO_LEAF2
Leaf-2

IPPROTO_MAX

IPPROTO_MEAS
DCN Measurement Subsystems

IPPROTO_MHRP
Mobile Host Routing

IPPROTO_MICP
Mobile Int.ing control

IPPROTO_MTP
Multicast Transport

IPPROTO_MUX
Multiplexing

IPPROTO_ND
Sun net disk proto (temp.)

IPPROTO_NHRP
Next Hop Resolution

IPPROTO_NONE
IP6 no next header

IPPROTO_NSP
Network Services

IPPROTO_NVPII
network voice protocol

IPPROTO_OSPFIGP
OSPFIGP

IPPROTO_PGM
PGM

IPPROTO_PIGP
private interior gateway

IPPROTO_PIM
Protocol Independent Mcast

IPPROTO_PRM
Packet Radio Measurement

IPPROTO_PUP
pup

IPPROTO_PVP
Packet Video Protocol

IPPROTO_RAW
raw IP packet

IPPROTO_RCCMON
BBN RCC Monitoring

IPPROTO_RDP
Reliable Data

IPPROTO_ROUTING
IP6 routing header

IPPROTO_RSVP
resource reservation

IPPROTO_RVD
Remote Virtual Disk

IPPROTO_SATEXPAK
SATNET/Backroom EXPAK

IPPROTO_SATMON
Satnet Monitoring

IPPROTO_SCCSP
Semaphore Comm. security

IPPROTO_SCTP
SCTP

IPPROTO_SDRP
Source Demand Routing

IPPROTO_SEP
Sequential Exchange

IPPROTO_SRPC
Strite RPC protocol

IPPROTO_ST
Stream protocol II.

IPPROTO_SVMTP
Secure VMTP

IPPROTO_SWIPE
IP with encryption

IPPROTO_TCF
TCF

IPPROTO_TCP

IPPROTO_TP
tp-4 w/ class negotiation

IPPROTO_TPXX
TP++ Transport

IPPROTO_TRUNK1
Trunk-1

IPPROTO_TRUNK2
Trunk-2

IPPROTO_TTP
TTP

IPPROTO_UDP

IPPROTO_VINES
Banyon VINES

IPPROTO_VISA
VISA Protocol

IPPROTO_VMTP
VMTP

IPPROTO_WBEXPAK
WIDEBAND EXPAK

IPPROTO_WBMON
WIDEBAND Monitoring

IPPROTO_WSN
Wang Span Network

IPPROTO_XNET
Cross Net Debugger

IPPROTO_XTP
XTP

IPTOS_ECN_CE

IPTOS_ECN_ECT0

IPTOS_ECN_ECT1

IPTOS_ECN_MASK

IPTOS_ECN_NOTECT

IPV6_UNICAST_HOPS

IPV6_MULTICAST_IF

IPV6_MULTICAST_HOPS

IPV6_MULTICAST_LOOP

IPV6_JOIN_GROUP

IPV6_LEAVE_GROUP

IPV6_RECVTCLASS

IPV6_TCLASS

IPV6_PKTINFO

IPV6_RECVPKTINFO

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_IF

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_PKTINFO

IP_RECVDSTADDR

IP_RECVIF

IP_RECVTOS

IP_TOS

IP_TTL

ISIG

ISTRIP

IUTF8

IXANY

IXOFF

IXON

KERN_AFFINITY

KERN_AIOMAX

KERN_AIOPROCMAX

KERN_AIOTHREADS

KERN_ARGMAX

KERN_BOOTFILE

KERN_BOOTTIME

KERN_CHECKOPENEVT

KERN_CLASSIC

KERN_CLASSICHANDLER

KERN_CLOCKRATE

KERN_COREDUMP

KERN_COREFILE

KERN_DOMAINNAME

KERN_DUMMY

KERN_DUMPDEV

KERN_EXEC

KERN_FILE

KERN_HOSTID

KERN_HOSTNAME

KERN_IPC

KERN_JOB_CONTROL

KERN_KDBUFWAIT

KERN_KDCPUMAP

KERN_KDDFLAGS

KERN_KDEBUG

KERN_KDEFLAGS

KERN_KDENABLE

KERN_KDGETBUF

KERN_KDGETENTROPY

KERN_KDGETREG

KERN_KDPIDEX

KERN_KDPIDTR

KERN_KDREADCURTHRMAP

KERN_KDREADTR

KERN_KDREMOVE

KERN_KDSETBUF

KERN_KDSETREG

KERN_KDSETRTCDEC

KERN_KDSETUP

KERN_KDSET_TYPEFILTER

KERN_KDTHRMAP

KERN_KDWRITEMAP

KERN_KDWRITETR

KERN_LOGSIGEXIT

KERN_LOW_PRI_DELAY

KERN_LOW_PRI_WINDOW

KERN_MAXFILES

KERN_MAXFILESPERPROC

KERN_MAXID

KERN_MAXPARTITIONS

KERN_MAXPROC

KERN_MAXPROCPERUID

KERN_MAXVNODES

KERN_NETBOOT

KERN_NGROUPS

KERN_NISDOMAINNAME

KERN_NTP_PLL

KERN_NX_PROTECTION

KERN_OPENEVT_PROC

KERN_OSRELDATE

KERN_OSRELEASE

KERN_OSREV

KERN_OSTYPE

KERN_OSVERSION

KERN_POSIX

KERN_POSIX1

KERN_PROC

KERN_PROCARGS

KERN_PROCARGS2

KERN_PROCDELAYTERM

KERN_PROCNAME

KERN_PROC_ALL

KERN_PROC_LCID

KERN_PROC_PGRP

KERN_PROC_PID

KERN_PROC_RUID

KERN_PROC_SESSION

KERN_PROC_TTY

KERN_PROC_UID

KERN_PROF

KERN_PS_STRINGS

KERN_RAGEVNODE

KERN_RAGE_PROC

KERN_RAGE_THREAD

KERN_SAFEBOOT

KERN_SAVED_IDS

KERN_SECURELVL

KERN_SHREG_PRIVATIZABLE

KERN_SPECULATIVE_READS

KERN_SUGID_COREDUMP

KERN_SYMFILE

KERN_SYSV

KERN_TFP

KERN_TFP_POLICY

KERN_TFP_POLICY_DEFAULT

KERN_TFP_POLICY_DENY

KERN_THALTSTACK

KERN_THREADNAME

KERN_TRANSLATE

KERN_TTY

KERN_UNOPENEVT_PROC

KERN_UNRAGE_PROC

KERN_UNRAGE_THREAD

KERN_UPDATEINTERVAL

KERN_USRSTACK32

KERN_USRSTACK64

KERN_VERSION

KERN_VNODE

KIPC_MAXSOCKBUF

KIPC_MAX_DATALEN

KIPC_MAX_HDR

KIPC_MAX_LINKHDR

KIPC_MAX_PROTOHDR

KIPC_MBSTAT

KIPC_NMBCLUSTERS

KIPC_SOCKBUF_WASTE

KIPC_SOMAXCONN

KIPC_SOQLIMITCOMPAT

LC_ALL

LC_ALL_MASK

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_SEGMENT

LC_SEGMENT_64

LC_TIME

LC_TIME_MASK

LIO_NOP

LIO_NOWAIT

LIO_READ

LIO_WAIT

LIO_WRITE

LOCAL_PEERCRED

LOCAL_PEEREPID

LOCAL_PEEREUUID

LOCAL_PEERPID

LOCAL_PEERUUID

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

LOGIN_PROCESS

LOG_ALERT

LOG_AUTH

LOG_AUTHPRIV

LOG_CONS

LOG_CRIT

LOG_CRON

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_FTP

LOG_INFO

LOG_INSTALL

LOG_KERN

LOG_LAUNCHD

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NETINFO

LOG_NEWS

LOG_NFACILITIES

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PERROR

LOG_PID

LOG_PRIMASK

LOG_RAS

LOG_REMOTEAUTH

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

L_tmpnam

MADV_CAN_REUSE

MADV_DONTNEED

MADV_FREE

MADV_FREE_REUSABLE

MADV_FREE_REUSE

MADV_NORMAL

MADV_RANDOM

MADV_SEQUENTIAL

MADV_WILLNEED

MADV_ZERO_WIRED_PAGES

MAP_ANON

MAP_ANONYMOUS

MAP_COPY

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_HASSEMAPHORE

MAP_JIT

MAP_NOCACHE

MAP_NOEXTEND

MAP_NORESERVE

MAP_PRIVATE

MAP_RENAME

MAP_SHARED

MAXCOMLEN

MAXTHREADNAMESIZE

MCL_CURRENT

MCL_FUTURE

MDMBUF

MH_MAGIC

MH_MAGIC_64

MINCORE_INCORE

MINCORE_MODIFIED

MINCORE_MODIFIED_OTHER

MINCORE_REFERENCED

MINCORE_REFERENCED_OTHER

MNT_FORCE

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

MSG_CTRUNC

MSG_DONTROUTE

MSG_DONTWAIT

MSG_EOF

MSG_EOR

MSG_FLUSH

MSG_HAVEMORE

MSG_HOLD

MSG_OOB

MSG_PEEK

MSG_RCVMORE

MSG_SEND

MSG_TRUNC

MSG_WAITALL

MS_ASYNC

MS_DEACTIVATE

MS_INVALIDATE

MS_KILLPAGES

MS_SYNC

NCCS

NET_RT_DUMP

NET_RT_FLAGS

NET_RT_IFLIST

NET_RT_IFLIST2

NEW_TIME

NI_DGRAM

NI_MAXHOST

NI_MAXSERV

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSCOPE

NI_NUMERICSERV

NL0

NL1

NLDLY

NOEXPR

NOFLSH

NOKERNINFO

NOSTR

NOTE_ABSOLUTE

NOTE_ATTRIB

NOTE_BACKGROUND

NOTE_CHILD

NOTE_CRITICAL

NOTE_DELETE

NOTE_EXEC

NOTE_EXIT

NOTE_EXITSTATUS

NOTE_EXIT_CSERROR

NOTE_EXIT_DECRYPTFAIL

NOTE_EXIT_DETAIL

NOTE_EXIT_DETAIL_MASK

NOTE_EXIT_MEMORY

NOTE_EXTEND

NOTE_FFAND

NOTE_FFCOPY

NOTE_FFCTRLMASK

NOTE_FFLAGSMASK

NOTE_FFNOP

NOTE_FFOR

NOTE_FORK

NOTE_LEEWAY

NOTE_LINK

NOTE_LOWAT

NOTE_NONE

NOTE_NSECONDS

NOTE_PCTRLMASK

NOTE_PDATAMASK

NOTE_RENAME

NOTE_REVOKE

NOTE_SECONDS

NOTE_SIGNAL

NOTE_TRACK

NOTE_TRACKERR

NOTE_TRIGGER

NOTE_USECONDS

NOTE_VM_ERROR

NOTE_VM_PRESSURE

NOTE_VM_PRESSURE_SUDDEN_TERMINATE

NOTE_VM_PRESSURE_TERMINATE

NOTE_WRITE

OCRNL

OFDEL

OFILL

OLD_TIME

ONLCR

ONLRET

ONOCR

ONOEOT

OPOST

OXTABS

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECTORY

O_DSYNC

O_EXCL

O_EXLOCK

O_FSYNC

O_NDELAY

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_RDONLY

O_RDWR

O_SHLOCK

O_SYNC

O_TRUNC

O_WRONLY

PARENB

PARMRK

PARODD

PATH_MAX

PENDIN

PF_APPLETALK

PF_CCITT

PF_CHAOS

PF_CNT

PF_COIP

PF_DATAKIT

PF_DECnet

PF_DLI

PF_ECMA

PF_HYLINK

PF_IMPLINK

PF_INET

PF_INET6

PF_IPX

PF_ISDN

PF_ISO

PF_KEY

PF_LAT

PF_LINK

PF_LOCAL

PF_NATM

PF_NETBIOS

PF_NS

PF_OSI

PF_PIP

PF_PPP

PF_PUP

PF_ROUTE

PF_RTIP

PF_SIP

PF_SNA

PF_SYSTEM

PF_UNIX

PF_UNSPEC

PF_XTP

PIPE_BUF

PM_STR

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

POLLRDBAND

POLLRDNORM

POLLWRBAND

POLLWRNORM

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

POSIX_SPAWN_CLOEXEC_DEFAULT

POSIX_SPAWN_RESETIDS

POSIX_SPAWN_SETEXEC

POSIX_SPAWN_SETPGROUP

POSIX_SPAWN_SETSIGDEF

POSIX_SPAWN_SETSIGMASK

POSIX_SPAWN_START_SUSPENDED

PRIO_DARWIN_BG

PRIO_DARWIN_NONUI

PRIO_DARWIN_PROCESS

PRIO_DARWIN_THREAD

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

PROC_PIDTASKALLINFO

PROC_PIDTASKINFO

PROC_PIDTHREADINFO

PROT_EXEC

PROT_NONE

PROT_READ

PROT_WRITE

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

PT_ATTACH

PT_ATTACHEXC

PT_CONTINUE

PT_DENY_ATTACH

PT_DETACH

PT_FIRSTMACH

PT_FORCEQUOTA

PT_KILL

PT_READ_D

PT_READ_I

PT_READ_U

PT_SIGEXC

PT_STEP

PT_THUPDATE

PT_TRACE_ME

PT_WRITE_D

PT_WRITE_I

PT_WRITE_U

P_ALL

P_PGID

P_PID

Q_GETQUOTA

Q_QUOTAOFF

Q_QUOTAON

Q_SETQUOTA

Q_SYNC

RADIXCHAR

RAND_MAX

RENAME_EXCL

RENAME_SWAP

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_MEMLOCK

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_STACK

RLIM_INFINITY

RLIM_NLIMITS	Deprecated

RTAX_AUTHOR

RTAX_BRD

RTAX_DST

RTAX_GATEWAY

RTAX_GENMASK

RTAX_IFA

RTAX_IFP

RTAX_MAX

RTAX_NETMASK

RTA_AUTHOR

RTA_BRD

RTA_DST

RTA_GATEWAY

RTA_GENMASK

RTA_IFA

RTA_IFP

RTA_NETMASK

RTF_BLACKHOLE

RTF_BROADCAST

RTF_CLONING

RTF_CONDEMNED

RTF_DELCLONE

RTF_DONE

RTF_DYNAMIC

RTF_GATEWAY

RTF_HOST

RTF_IFREF

RTF_IFSCOPE

RTF_LLINFO

RTF_LOCAL

RTF_MODIFIED

RTF_MULTICAST

RTF_NOIFREF

RTF_PINNED

RTF_PRCLONING

RTF_PROTO1

RTF_PROTO2

RTF_PROTO3

RTF_PROXY

RTF_REJECT

RTF_ROUTER

RTF_STATIC

RTF_UP

RTF_WASCLONED

RTF_XRESOLVE

RTLD_DEFAULT

RTLD_FIRST

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NEXT

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

RTLD_SELF

RTM_ADD

RTM_CHANGE

RTM_DELADDR

RTM_DELETE

RTM_DELMADDR

RTM_GET

RTM_GET2

RTM_IFINFO

RTM_IFINFO2

RTM_LOCK

RTM_LOSING

RTM_MISS

RTM_NEWADDR

RTM_NEWMADDR

RTM_NEWMADDR2

RTM_OLDADD

RTM_OLDDEL

RTM_REDIRECT

RTM_RESOLVE

RTM_VERSION

RTV_EXPIRE

RTV_HOPCOUNT

RTV_MTU

RTV_RPIPE

RTV_RTT

RTV_RTTVAR

RTV_SPIPE

RTV_SSTHRESH

RUN_LVL

RUSAGE_CHILDREN

RUSAGE_SELF

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SCM_CREDS

SCM_RIGHTS

SCM_TIMESTAMP

SEEK_CUR

SEEK_END

SEEK_SET

SEM_FAILED

SEM_UNDO

SETALL

SETVAL

SF_APPEND

SF_ARCHIVED

SF_IMMUTABLE

SF_SETTABLE

SHMLBA

SHM_R

SHM_RDONLY

SHM_RND

SHM_W

SHUTDOWN_TIME

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGEMT

SIGEV_NONE

SIGEV_SIGNAL

SIGEV_THREAD

SIGFPE

SIGHUP

SIGILL

SIGINFO

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGNATURE

SIGPIPE

SIGPROF

SIGQUIT

SIGSEGV

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SIOCGIFADDR

SOCK_DGRAM

SOCK_MAXADDRLEN

SOCK_RAW

SOCK_RDM

SOCK_SEQPACKET

SOCK_STREAM

SOL_LOCAL

SOL_SOCKET

SOMAXCONN

SO_ACCEPTCONN

SO_BROADCAST

SO_DEBUG

SO_DONTROUTE

SO_DONTTRUNC

SO_ERROR

SO_KEEPALIVE

SO_LABEL

SO_LINGER

SO_NKE

SO_NOADDRERR

SO_NOSIGPIPE

SO_NOTIFYCONFLICT

SO_NP_EXTENSIONS

SO_NREAD

SO_NWRITE

SO_OOBINLINE

SO_PEERLABEL

SO_RANDOMPORT

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_REUSESHAREUID

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TIMESTAMP

SO_TIMESTAMP_MONOTONIC

SO_TYPE

SO_USELOOPBACK

SO_WANTMORE

SO_WANTOOBFLAG

SS_DISABLE

SS_ONSTACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

ST_NOSUID

ST_RDONLY

SUPERPAGE_NONE	Deprecated

SUPERPAGE_SIZE_2MB	Deprecated

SUPERPAGE_SIZE_ANY	Deprecated

SYSPROTO_CONTROL

SYSPROTO_EVENT

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TAB0

TAB1

TAB2

TAB3

TABDLY

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_KEEPALIVE

TCP_MAXSEG

TCP_NODELAY

TCP_NOOPT

TCP_NOPUSH

TCSADRAIN

TCSAFLUSH

TCSANOW

THOUSEP

TIOCCBRK

TIOCCDTR

TIOCCONS

TIOCDCDTIMESTAMP

TIOCDRAIN

TIOCDSIMICROCODE

TIOCEXCL

TIOCEXT

TIOCFLUSH

TIOCGDRAINWAIT

TIOCGETD

TIOCGPGRP

TIOCGWINSZ

TIOCIXOFF

TIOCIXON

TIOCMBIC

TIOCMBIS

TIOCMGDTRWAIT

TIOCMGET

TIOCMODG

TIOCMODS

TIOCMSDTRWAIT

TIOCMSET

TIOCM_CAR

TIOCM_CD

TIOCM_CTS

TIOCM_DSR

TIOCM_DTR

TIOCM_LE

TIOCM_RI

TIOCM_RNG

TIOCM_RTS

TIOCM_SR

TIOCM_ST

TIOCNOTTY

TIOCNXCL

TIOCOUTQ

TIOCPKT

TIOCPKT_DATA

TIOCPKT_DOSTOP

TIOCPKT_FLUSHREAD

TIOCPKT_FLUSHWRITE

TIOCPKT_IOCTL

TIOCPKT_NOSTOP

TIOCPKT_START

TIOCPKT_STOP

TIOCPTYGNAME

TIOCPTYGRANT

TIOCPTYUNLK

TIOCREMOTE

TIOCSBRK

TIOCSCONS

TIOCSCTTY

TIOCSDRAINWAIT

TIOCSDTR

TIOCSETD

TIOCSIG

TIOCSPGRP

TIOCSTART

TIOCSTAT

TIOCSTI

TIOCSTOP

TIOCSWINSZ

TIOCTIMESTAMP

TIOCUCNTL

TMP_MAX

TOSTOP

T_FMT

T_FMT_AMPM

UF_APPEND

UF_COMPRESSED

UF_HIDDEN

UF_IMMUTABLE

UF_NODUMP

UF_OPAQUE

UF_SETTABLE

UF_TRACKED

USER_BC_BASE_MAX

USER_BC_DIM_MAX

USER_BC_SCALE_MAX

USER_BC_STRING_MAX

USER_COLL_WEIGHTS_MAX

USER_CS_PATH

USER_EXPR_NEST_MAX

USER_LINE_MAX

USER_MAXID

USER_POSIX2_VERSION

USER_POSIX2_C_BIND

USER_POSIX2_C_DEV

USER_POSIX2_CHAR_TERM

USER_POSIX2_FORT_DEV

USER_POSIX2_FORT_RUN

USER_POSIX2_LOCALEDEF

USER_POSIX2_SW_DEV

USER_POSIX2_UPE

USER_PROCESS

USER_RE_DUP_MAX

USER_STREAM_MAX

USER_TZNAME_MAX

USRQUOTA

UTIME_NOW

UTIME_OMIT

UTUN_OPT_FLAGS

UTUN_OPT_IFNAME

VDISCARD

VDSUSP

VEOF

VEOL

VEOL2

VERASE

VINTR

VKILL

VLNEXT

VMIN

VM_FLAGS_ALIAS_MASK	Deprecated

VM_FLAGS_ANYWHERE	Deprecated

VM_FLAGS_FIXED	Deprecated

VM_FLAGS_NO_CACHE	Deprecated

VM_FLAGS_OVERWRITE	Deprecated

VM_FLAGS_PURGABLE	Deprecated

VM_FLAGS_RANDOM_ADDR	Deprecated

VM_FLAGS_RESILIENT_CODESIGN	Deprecated

VM_FLAGS_RESILIENT_MEDIA	Deprecated

VM_FLAGS_RETURN_4K_DATA_ADDR	Deprecated

VM_FLAGS_RETURN_DATA_ADDR	Deprecated

VM_FLAGS_SUPERPAGE_MASK	Deprecated

VM_FLAGS_SUPERPAGE_NONE	Deprecated

VM_FLAGS_SUPERPAGE_SHIFT	Deprecated

VM_FLAGS_SUPERPAGE_SIZE_2MB	Deprecated

VM_FLAGS_SUPERPAGE_SIZE_ANY	Deprecated

VM_FLAGS_USER_ALLOCATE	Deprecated

VM_FLAGS_USER_MAP	Deprecated

VM_FLAGS_USER_REMAP	Deprecated

VM_LOADAVG

VM_MACHFACTOR

VM_MAXID

VM_MEMORY_ACCELERATE	Deprecated

VM_MEMORY_ANALYSIS_TOOL	Deprecated

VM_MEMORY_APPKIT	Deprecated

VM_MEMORY_APPLICATION_SPECIFIC_1	Deprecated

VM_MEMORY_APPLICATION_SPECIFIC_16	Deprecated

VM_MEMORY_ASL	Deprecated

VM_MEMORY_ASSETSD	Deprecated

VM_MEMORY_ATS	Deprecated

VM_MEMORY_CARBON	Deprecated

VM_MEMORY_CGIMAGE	Deprecated

VM_MEMORY_COREDATA	Deprecated

VM_MEMORY_COREDATA_OBJECTIDS	Deprecated

VM_MEMORY_COREGRAPHICS	Deprecated

VM_MEMORY_COREGRAPHICS_BACKINGSTORES	Deprecated

VM_MEMORY_COREGRAPHICS_DATA	Deprecated

VM_MEMORY_COREGRAPHICS_FRAMEBUFFERS	Deprecated

VM_MEMORY_COREGRAPHICS_MISC	Deprecated

VM_MEMORY_COREGRAPHICS_SHARED	Deprecated

VM_MEMORY_COREGRAPHICS_XALLOC	Deprecated

VM_MEMORY_COREIMAGE	Deprecated

VM_MEMORY_COREPROFILE	Deprecated

VM_MEMORY_CORESERVICES	Deprecated

VM_MEMORY_COREUI	Deprecated

VM_MEMORY_COREUIFILE	Deprecated

VM_MEMORY_CORPSEINFO	Deprecated

VM_MEMORY_DHMM	Deprecated

VM_MEMORY_DYLD	Deprecated

VM_MEMORY_DYLD_MALLOC	Deprecated

VM_MEMORY_DYLIB	Deprecated

VM_MEMORY_FOUNDATION	Deprecated

VM_MEMORY_GENEALOGY	Deprecated

VM_MEMORY_GLSL	Deprecated

VM_MEMORY_GUARD	Deprecated

VM_MEMORY_IMAGEIO	Deprecated

VM_MEMORY_IOKIT	Deprecated

VM_MEMORY_JAVA	Deprecated

VM_MEMORY_JAVASCRIPT_CORE	Deprecated

VM_MEMORY_JAVASCRIPT_JIT_EXECUTABLE_ALLOCATOR	Deprecated

VM_MEMORY_JAVASCRIPT_JIT_REGISTER_FILE	Deprecated

VM_MEMORY_LAYERKIT	Deprecated

VM_MEMORY_LIBDISPATCH	Deprecated

VM_MEMORY_MACH_MSG	Deprecated

VM_MEMORY_MALLOC	Deprecated

VM_MEMORY_MALLOC_HUGE	Deprecated

VM_MEMORY_MALLOC_LARGE	Deprecated

VM_MEMORY_MALLOC_LARGE_REUSABLE	Deprecated

VM_MEMORY_MALLOC_LARGE_REUSED	Deprecated

VM_MEMORY_MALLOC_NANO	Deprecated

VM_MEMORY_MALLOC_SMALL	Deprecated

VM_MEMORY_MALLOC_TINY	Deprecated

VM_MEMORY_OBJC_DISPATCHERS	Deprecated

VM_MEMORY_OPENCL	Deprecated

VM_MEMORY_OS_ALLOC_ONCE	Deprecated

VM_MEMORY_RAWCAMERA	Deprecated

VM_MEMORY_REALLOC	Deprecated

VM_MEMORY_SBRK	Deprecated

VM_MEMORY_SCENEKIT	Deprecated

VM_MEMORY_SHARED_PMAP	Deprecated

VM_MEMORY_SKYWALK	Deprecated

VM_MEMORY_SQLITE	Deprecated

VM_MEMORY_STACK	Deprecated

VM_MEMORY_SWIFT_METADATA	Deprecated

VM_MEMORY_SWIFT_RUNTIME	Deprecated

VM_MEMORY_TCMALLOC	Deprecated

VM_MEMORY_UNSHARED_PMAP	Deprecated

VM_MEMORY_WEBCORE_PURGEABLE_BUFFERS	Deprecated

VM_METER

VM_SWAPUSAGE

VQUIT

VREPRINT

VSTART

VSTATUS

VSTOP

VSUSP

VT0

VT1

VTDLY

VTIME

VWERASE

WCONTINUED

WEXITED

WNOHANG

WNOWAIT

WSTOPPED

WUNTRACED

W_OK

XATTR_CREATE

XATTR_NODEFAULT

XATTR_NOFOLLOW

XATTR_NOSECURITY

XATTR_REPLACE

XATTR_SHOWCOMPRESSION

XUCRED_VERSION

X_OK

YESEXPR

YESSTR

_IOFBF

_IOLBF

_IONBF

_PC_CHOWN_RESTRICTED

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_VDISABLE

_POSIX_VDISABLE

_PTHREAD_COND_SIG_init

_PTHREAD_MUTEX_SIG_init

_PTHREAD_RWLOCK_SIG_init

_RLIMIT_POSIX_FLAG

_SC_2_VERSION

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_CHAR_TERM

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FILE_LOCKING

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SS_REPL_MAX

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_V6_LPBIG_OFFBIG

_SC_V6_ILP32_OFFBIG

_SC_V6_ILP32_OFF32

_SC_V6_LP64_OFF64

_SC_VERSION

_SC_XBS5_LPBIG_OFFBIG

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_ILP32_OFF32

_SC_XBS5_LP64_OFF64

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_UTX_HOSTSIZE

_UTX_IDSIZE

_UTX_LINESIZE

_UTX_USERSIZE

_WSTOPPED

__PTHREAD_CONDATTR_SIZE__

__PTHREAD_COND_SIZE__

__PTHREAD_MUTEX_SIZE__

__PTHREAD_RWLOCKATTR_SIZE__

__PTHREAD_RWLOCK_SIZE__

pseudo_AF_HDRCMPLT

pseudo_AF_KEY

pseudo_AF_PIP

pseudo_AF_RTIP

pseudo_AF_XTP

Functions

CMSG_DATA⚠

CMSG_FIRSTHDR⚠

CMSG_LEN⚠

CMSG_NXTHDR⚠

CMSG_SPACE⚠

FD_CLR⚠

FD_ISSET⚠

FD_SET⚠

FD_ZERO⚠

QCMD⚠

WCOREDUMP⚠

WEXITSTATUS⚠

WIFCONTINUED⚠

WIFEXITED⚠

WIFSIGNALED⚠

WIFSTOPPED⚠

WSTOPSIG⚠

WTERMSIG⚠

_WSTATUS⚠

__error⚠

_dyld_get_image_header⚠	Deprecated

_dyld_get_image_name⚠	Deprecated

_dyld_get_image_vmaddr_slide⚠	Deprecated

_dyld_image_count⚠	Deprecated

_exit⚠

abort⚠

abs⚠

accept⚠

access⚠

acct⚠

aio_cancel⚠

aio_error⚠

aio_fsync⚠

aio_read⚠

aio_return⚠

aio_suspend⚠

aio_write⚠

alarm⚠

atexit⚠

atof⚠

atoi⚠

backtrace⚠

bind⚠

brk⚠

calloc⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chflags⚠

chmod⚠

chown⚠

chroot⚠

clock_getres⚠

clock_gettime⚠

close⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

difftime⚠

dirfd⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

duplocale⚠

endgrent⚠

endpwent⚠

endutxent⚠

exchangedata⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

exit⚠

faccessat⚠

fchdir⚠

fchflags⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fclose⚠

fcntl⚠

fdopen⚠

fdopendir⚠

feof⚠

ferror⚠

fflush⚠

fgetc⚠

fgetpos⚠

fgets⚠

fgetxattr⚠

fileno⚠

flistxattr⚠

flock⚠

fopen⚠

fork⚠

forkpty⚠

fpathconf⚠

fprintf⚠

fputc⚠

fputs⚠

fread⚠

free⚠

freeaddrinfo⚠

freeifaddrs⚠

freelocale⚠

fremovexattr⚠

freopen⚠

fscanf⚠

fseek⚠

fseeko⚠

fsetpos⚠

fsetxattr⚠

fstat⚠

fstatat⚠

fstatfs⚠

fstatvfs⚠

fsync⚠

ftell⚠

ftello⚠

ftok⚠

ftruncate⚠

futimens⚠

futimes⚠

fwrite⚠

gai_strerror⚠

getaddrinfo⚠

getchar⚠

getchar_unlocked⚠

getcwd⚠

getdomainname⚠

getdtablesize⚠

getegid⚠

getenv⚠

geteuid⚠

getgid⚠

getgrent⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgrouplist⚠

getgroups⚠

gethostname⚠

getifaddrs⚠

getline⚠

getloadavg⚠

getlogin⚠

getnameinfo⚠

getopt⚠

getpeereid⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getpriority⚠

getprogname⚠

getprotobyname⚠

getprotobynumber⚠

getpwent⚠

getpwnam⚠

getpwnam_r⚠

getpwuid⚠

getpwuid_r⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

getutxent⚠

getutxid⚠

getutxline⚠

getxattr⚠

glob⚠

globfree⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

if_freenameindex⚠

if_indextoname⚠

if_nameindex⚠

if_nametoindex⚠

initgroups⚠

ioctl⚠

isalnum⚠

isalpha⚠

isatty⚠

isblank⚠

iscntrl⚠

isdigit⚠

isgraph⚠

islower⚠

isprint⚠

ispunct⚠

isspace⚠

isupper⚠

isxdigit⚠

kevent⚠

kevent64⚠

kill⚠

killpg⚠

kqueue⚠

labs⚠

lchown⚠

link⚠

linkat⚠

lio_listio⚠

listen⚠

listxattr⚠

localeconv⚠

localeconv_l⚠

localtime⚠

localtime_r⚠

login_tty⚠

lseek⚠

lstat⚠

lutimes⚠

mach_absolute_time⚠	Deprecated

mach_timebase_info⚠	Deprecated

madvise⚠

malloc⚠

memchr⚠

memcmp⚠

memcpy⚠

memmove⚠

memset⚠

mincore⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mknod⚠

mkstemp⚠

mkstemps⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

mount⚠

mprotect⚠

msync⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

newlocale⚠

nice⚠

nl_langinfo⚠

open⚠

openat⚠

opendir⚠

openlog⚠

openpty⚠

pathconf⚠

pause⚠

pclose⚠

perror⚠

pipe⚠

poll⚠

popen⚠

posix_madvise⚠

posix_memalign⚠

posix_openpt⚠

posix_spawn⚠

posix_spawn_file_actions_addclose⚠

posix_spawn_file_actions_adddup2⚠

posix_spawn_file_actions_addopen⚠

posix_spawn_file_actions_destroy⚠

posix_spawn_file_actions_init⚠

posix_spawnattr_destroy⚠

posix_spawnattr_getflags⚠

posix_spawnattr_getpgroup⚠

posix_spawnattr_getsigdefault⚠

posix_spawnattr_getsigmask⚠

posix_spawnattr_init⚠

posix_spawnattr_setflags⚠

posix_spawnattr_setpgroup⚠

posix_spawnattr_setsigdefault⚠

posix_spawnattr_setsigmask⚠

posix_spawnp⚠

pread⚠

printf⚠

pselect⚠

pthread_atfork⚠

pthread_attr_destroy⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cancel⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_getpshared⚠

pthread_condattr_init⚠

pthread_condattr_setpshared⚠

pthread_create⚠

pthread_detach⚠

pthread_exit⚠

pthread_get_stackaddr_np⚠

pthread_get_stacksize_np⚠

pthread_getname_np⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_kill⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_getpshared⚠

pthread_mutexattr_init⚠

pthread_mutexattr_setpshared⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_getpshared⚠

pthread_rwlockattr_init⚠

pthread_rwlockattr_setpshared⚠

pthread_self⚠

pthread_setname_np⚠

pthread_setspecific⚠

pthread_sigmask⚠

ptrace⚠

ptsname⚠

putchar⚠

putchar_unlocked⚠

putenv⚠

puts⚠

pututxline⚠

pwrite⚠

querylocale⚠

quotactl⚠

raise⚠

rand⚠

read⚠

readdir⚠

readdir_r⚠
The 64-bit libc on Solaris and illumos only has readdir_r. If a 32-bit Solaris or illumos target is ever created, it should use __posix_readdir_r. See libc(3LIB) on Solaris or illumos: https://illumos.org/man/3lib/libc https://docs.oracle.com/cd/E36784_01/html/E36873/libc-3lib.html https://www.unix.com/man-page/opensolaris/3LIB/libc/

readlink⚠

readlinkat⚠

readv⚠

realloc⚠

realpath⚠

recv⚠

recvfrom⚠

recvmsg⚠

remove⚠

removexattr⚠

rename⚠

renameat⚠

renameatx_np⚠

renamex_np⚠

res_init⚠

rewind⚠

rewinddir⚠

rmdir⚠

sbrk⚠

scanf⚠

sched_yield⚠

seekdir⚠

select⚠

sem_close⚠

sem_open⚠

sem_post⚠

sem_trywait⚠

sem_unlink⚠

sem_wait⚠

semctl⚠

semget⚠

semop⚠

send⚠

sendfile⚠

sendmsg⚠

sendto⚠

setbuf⚠

setdomainname⚠

setegid⚠

setenv⚠

seteuid⚠

setgid⚠

setgrent⚠

setgroups⚠

sethostname⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setpriority⚠

setprogname⚠

setpwent⚠

setrlimit⚠

setsid⚠

setsockopt⚠

settimeofday⚠

setuid⚠

setutxent⚠

setvbuf⚠

setxattr⚠

shm_open⚠

shm_unlink⚠

shmat⚠

shmctl⚠

shmdt⚠

shmget⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigaltstack⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sigwait⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

srand⚠

sscanf⚠

stat⚠

statfs⚠

statvfs⚠

strcasecmp⚠

strcasestr⚠

strcat⚠

strchr⚠

strcmp⚠

strcoll⚠

strcpy⚠

strcspn⚠

strdup⚠

strerror⚠

strerror_r⚠

strlen⚠

strncasecmp⚠

strncat⚠

strncmp⚠

strncpy⚠

strnlen⚠

strpbrk⚠

strrchr⚠

strspn⚠

strstr⚠

strtod⚠

strtok⚠

strtol⚠

strtoul⚠

strxfrm⚠

symlink⚠

symlinkat⚠

sync⚠

syscall⚠

sysconf⚠

sysctl⚠

sysctlbyname⚠

sysctlnametomib⚠

syslog⚠

system⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

telldir⚠

time⚠

timegm⚠

times⚠

tmpfile⚠

tmpnam⚠

tolower⚠

toupper⚠

truncate⚠

ttyname⚠

ttyname_r⚠

umask⚠

uname⚠

ungetc⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unmount⚠

unsetenv⚠

uselocale⚠

usleep⚠

utime⚠

utimensat⚠

utimes⚠

utmpxname⚠

wait⚠

waitid⚠

waitpid⚠

wcslen⚠

wcstombs⚠

write⚠

writev⚠

Type Definitions

blkcnt_t

blksize_t

boolean_t

c_char

c_double

c_float

c_int

c_long

c_longlong

c_schar

c_short

c_uchar

c_uint

c_ulong

c_ulonglong

c_ushort

cc_t

clock_t

clockid_t

cpu_subtype_t

cpu_type_t

dev_t

fsblkcnt_t

fsfilcnt_t

gid_t

id_t

idtype_t

in_addr_t

in_port_t

ino_t

int8_t	Deprecated

int16_t	Deprecated

int32_t	Deprecated

int64_t	Deprecated

integer_t

intmax_t

intptr_t

key_t

locale_t

mach_timebase_info_data_t	Deprecated

mode_t

nfds_t

nl_item

nlink_t

off_t

pid_t

posix_spawn_file_actions_t

posix_spawnattr_t

pthread_key_t

pthread_t

ptrdiff_t

rlim_t

sa_family_t

sem_t

shmatt_t

sighandler_t

sigset_t

size_t

socklen_t

speed_t

ssize_t

suseconds_t

tcflag_t

time_t

uid_t

uint8_t	Deprecated

uint16_t	Deprecated

uint32_t	Deprecated

uint64_t	Deprecated

uintmax_t

uintptr_t

useconds_t

vm_prot_t	Deprecated

vm_size_t	Deprecated

wchar_t

Unions

semun