libc
Module macros

Macros

__item 🔒

cfg_if 🔒

A macro for defining #cfg if-else statements.

const_fn 🔒

A nonpublic function that is const as long as libc_const_extern_fn is enabled.

deprecated_mach 🔒

e 🔒

Implement Clone and Copy for an enum, as well as Debug, Eq, Hash, and PartialEq if the extra_traits feature is enabled.

f 🔒

Define an unsafe function that is const as long as libc_const_extern_fn is enabled.

missing 🔒

Specify that an enum should have no traits that aren’t specified in the macro invocation, i.e. no Clone or Copy.

prelude 🔒

Create an internal crate prelude with core reexports and common types.

s 🔒

Implement Clone and Copy for a struct, as well as Debug, Eq, Hash, and PartialEq if the extra_traits feature is enabled.

s_no_extra_traits 🔒

Implement Clone and Copy for a struct with no extra_traits feature.

s_paren 🔒

Implement Clone and Copy for a tuple struct, as well as Debug, Eq, Hash, and PartialEq if the extra_traits feature is enabled.

safe_f 🔒

Define a safe function that is const as long as libc_const_extern_fn is enabled.