libc

Module macros

Macros

__item 🔒

c_enum 🔒

Represent a C enum as Rust constants and a type.

cfg_if 🔒

A macro for defining #cfg if-else statements.

deprecated_mach 🔒

e 🔒

Implement Clone and Copy for an enum, as well as Debug, Eq, Hash, and PartialEq if the extra_traits feature is enabled.

extern_ty 🔒

Create an uninhabited type that can’t be constructed. It implements Debug, Clone, and Copy, but these aren’t meaningful for extern types so they should eventually be removed.

f 🔒

Define a unsafe function.

offset_of 🔒

Polyfill for std’s offset_of.

prelude 🔒

Create an internal crate prelude with core reexports and common types.

s 🔒

Implement Clone and Copy for a struct, as well as Debug, Eq, Hash, and PartialEq if the extra_traits feature is enabled.

s_no_extra_traits 🔒

Implement Clone, Copy, and Debug since those can be derived, but exclude PartialEq, Eq, and Hash.

s_paren 🔒

Implement Clone and Copy for a tuple struct, as well as Debug, Eq, Hash, and PartialEq if the extra_traits feature is enabled.

safe_f 🔒

Define a safe function.