libc

Module new

This module contains the future directory structure. If possible, new definitions should get added here.

Eventually everything should be moved over, and we will move this directory to the top level in src.

Basic structure

Each child module here represents a library or group of libraries that we are binding. Each of these has several submodules, representing either a directory or a header file in that library.

#includes turn into pub use ...*; statements. Then at the root level (here), we choose which top-level headers we want to reexport the definitions for.

All modules are only crate-public since we don’t reexport this structure.

Re-exports

pub use linux::can::bcm::*;

pub use linux::can::j1939::*;

pub use linux::can::raw::*;

pub use linux::can::*;

pub use linux::keyctl::*;

pub use linux::membarrier::*;

pub use linux::netlink::*;

pub use net::route::*;

pub use pthread::*;

pub use unistd::*;

Modules

common 🔒

Interfaces that are common across multiple platforms

glibc 🔒

GNU libc.

linux_uapi 🔒

This directory maps to include/uapi in the Linux source tree.