libc
Module prelude

Frequently-used types that are available on all platforms

We need to reexport the core types so this works with rust-dep-of-std.

Modules

fmt 🔒

Utilities for formatting and printing strings.

hash 🔒

Generic hashing support.

iter 🔒

Composable external iteration.

mem 🔒

Basic functions for dealing with memory.

Enums

Option 🔒

The Option type. See the module level documentation for more.

c_void 🔒

Equivalent to C’s void type when used as a pointer.

Traits

Clone 🔒

A common trait for the ability to explicitly duplicate an object.

Copy 🔒

Types whose values can be duplicated simply by copying bits.

Send 🔒

Types that can be transferred across thread boundaries.

Sync 🔒

Types for which it is safe to share references between threads.

Derive Macros

Clone 🔒

Derive macro generating an impl of the trait Clone.

Copy 🔒

Derive macro generating an impl of the trait Copy.