libc

Module prelude

Frequently-used types that are available on all platforms

We need to reexport the core types so this works with rust-dep-of-std.

Modules

fmt 🔒

Utilities for formatting and printing strings.

hash 🔒

Generic hashing support.

iter 🔒

Composable external iteration.

mem 🔒

Basic functions for dealing with memory.

ptr 🔒

Manually manage memory through raw pointers.

Macros

cfg 🔒

Evaluates boolean combinations of configuration flags at compile-time.

Enums

Option 🔒

The Option type. See the module level documentation for more.

c_void 🔒

Equivalent to C’s void type when used as a pointer.

Traits

Clone 🔒

A common trait that allows explicit creation of a duplicate value.

Copy 🔒

Types whose values can be duplicated simply by copying bits.

Debug 🔒

? formatting.

Default 🔒

A trait for giving a type a useful default value.

Send 🔒

Types that can be transferred across thread boundaries.

Sync 🔒

Types for which it is safe to share references between threads.

Functions

align_of 🔒

Returns the ABI-required minimum alignment of a type in bytes.

align_of_val 🔒

Returns the ABI-required minimum alignment of the type of the value that val points to in bytes.

size_of 🔒

Returns the size of a type in bytes.

size_of_val 🔒

Returns the size of the pointed-to value in bytes.

Attribute Macros

derive 🔒

Attribute macro used to apply derive macros.

Derive Macros

Clone 🔒

Derive macro generating an impl of the trait Clone.

Copy 🔒

Derive macro generating an impl of the trait Copy.

Debug 🔒

Derive macro generating an impl of the trait Debug.

Default 🔒

Derive macro generating an impl of the trait Default.