Module libc::unix

Definitions found commonly among almost all Unix derivatives

More functions and definitions can be found in the more specific modules according to the platform in question.

Reexports

pub use self::notbsd::*;

Modules

notbsd

Structs

group

hostent

in6_addr

in_addr

iovec

ip_mreq

ipv6_mreq

linger

pollfd

rlimit

rusage

timespec

timeval

utimbuf

winsize

Enums

DIR

locale_t

Constants

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

FD_CLOEXEC

GRPQUOTA

IF_NAMESIZE

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

RTLD_LAZY

SIGIOT

SIG_DFL

SIG_ERR

SIG_IGN

S_ISGID

S_ISUID

S_ISVTX

USRQUOTA

Functions

accept

access

alarm

bind

cfgetispeed

cfgetospeed

cfsetispeed

cfsetospeed

chdir

chmod

chown

chroot

close

closedir

closelog

connect

creat

daemon

dladdr

dlclose

dlerror

dlopen

dlsym

dup

dup2

endpwent

execv

execve

execvp

fchmod

fcntl

fdopen

fileno

flock

fork

fpathconf

fprintf

freeaddrinfo

fscanf

fseeko

fstat

fstatvfs

fsync

ftello

ftruncate

futimes

gai_strerror

getaddrinfo

getchar_unlocked

getcwd

getdtablesize

getegid

geteuid

getgid

getgrgid

getgrnam

getgroups

gethostname

getlogin

getopt

getpeername

getpgid

getpgrp

getpid

getppid

getpwnam

getpwuid

getrlimit

getrusage

getsid

getsockname

getsockopt

gettimeofday

getuid

glob

globfree

gmtime_r

if_indextoname

if_nametoindex

isatty

kill

lchown

link

listen

localeconv

localtime

localtime_r

lseek

lstat

madvise

mkdir

mkdtemp

mkfifo

mknod

mkstemp

mkstemps

mktime

mlock

mlockall

mmap

msync

munlock

munlockall

munmap

nanosleep

nice

nl_langinfo

open

opendir

openlog

pathconf

pause

pclose

pipe

poll

popen

posix_madvise

posix_memalign

pread

printf

pselect

pthread_attr_destroy

pthread_attr_init

pthread_attr_setdetachstate

pthread_attr_setstacksize

pthread_cond_broadcast

pthread_cond_destroy

pthread_cond_init

pthread_cond_signal

pthread_cond_timedwait

pthread_cond_wait

pthread_condattr_destroy

pthread_condattr_init

pthread_create

pthread_detach

pthread_getspecific

pthread_join

pthread_key_create

pthread_key_delete

pthread_kill

pthread_mutex_destroy

pthread_mutex_init

pthread_mutex_lock

pthread_mutex_trylock

pthread_mutex_unlock

pthread_mutexattr_destroy

pthread_mutexattr_init

pthread_mutexattr_settype

pthread_rwlock_destroy

pthread_rwlock_rdlock

pthread_rwlock_tryrdlock

pthread_rwlock_trywrlock

pthread_rwlock_unlock

pthread_rwlock_wrlock

pthread_self

pthread_setspecific

pthread_sigmask

putchar_unlocked

putenv

pwrite

raise

read

readdir_r

readlink

readv

realpath

recv

recvfrom

recvmsg

rewinddir

rmdir

scanf

sched_yield

seekdir

select

sem_close

sem_destroy

sem_init

sem_open

sem_post

sem_trywait

sem_unlink

sem_wait

send

sendmsg

sendto

setenv

setgid

setlocale

setlogmask

setpgid

setrlimit

setsid

setsockopt

setuid

shm_unlink

shutdown

sigaction

sigaddset

sigaltstack

sigdelset

sigemptyset

sigfillset

sigismember

signal

sigwait

sleep

snprintf

socket

socketpair

sprintf

sscanf

stat

statvfs

strerror_r

symlink

sysconf

syslog

tcdrain

tcflow

tcflush

tcgetattr

tcgetpgrp

tcsendbreak

tcsetattr

telldir

time

timegm

ttyname

umask

uname

unlink

unsetenv

usleep

utime

utimes

wait

waitpid

write

writev

Type Definitions

cc_t

gid_t

in_addr_t

in_port_t

pid_t

sighandler_t

uid_t