Module libc::unix

Definitions found commonly among almost all Unix derivatives

More functions and definitions can be found in the more specific modules according to the platform in question.

Modules

b64
64-bit specific definitions for linux-like values

linux
Linux-specific definitions for linux-like values

notbsd

other

x86_64
x86_64-specific definitions for 64-bit linux-like values

Structs

Dl_info

addrinfo

cpu_set_t

dirent

dirent64

dqblk

epoll_event

fd_set

flock

fsid_t

glob64_t

glob_t

hostent

ifaddrs

in6_addr

in_addr

iovec

ip_mreq

ipc_perm

ipv6_mreq

lconv

mcontext_t

mq_attr

msghdr

passwd

pollfd

pthread_attr_t

pthread_cond_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

rlimit

rlimit64

rusage

sched_param

shmid_ds

sigaction

siginfo_t

signalfd_siginfo

sigset_t

sockaddr

sockaddr_in

sockaddr_in6

sockaddr_ll

sockaddr_nl

sockaddr_storage

sockaddr_un

stack_t

stat

stat64

statfs

statvfs

termios

timespec

timeval

tm

ucontext_t

ucred

utimbuf

utsname

winsize

Enums

DIR

fpos64_t

locale_t

timezone

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_10

ABMON_11

ABMON_12

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ADFS_SUPER_MAGIC

AFFS_SUPER_MAGIC

AF_INET

AF_INET6

AF_NETLINK

AF_PACKET

AF_UNIX

ALT_DIGITS

AM_STR

AT_FDCWD

AT_SYMLINK_NOFOLLOW

BRKINT

BS0

BS1

BUFSIZ

CBAUD

CLOCAL

CLOCK_BOOTTIME

CLOCK_BOOTTIME_ALARM

CLOCK_MONOTONIC

CLOCK_MONOTONIC_COARSE

CLOCK_MONOTONIC_RAW

CLOCK_PROCESS_CPUTIME_ID

CLOCK_REALTIME

CLOCK_REALTIME_ALARM

CLOCK_REALTIME_COARSE

CLOCK_THREAD_CPUTIME_ID

CLONE_CHILD_CLEARTID

CLONE_CHILD_SETTID

CLONE_DETACHED

CLONE_FILES

CLONE_FS

CLONE_IO

CLONE_NEWIPC

CLONE_NEWNET

CLONE_NEWNS

CLONE_NEWPID

CLONE_NEWUSER

CLONE_NEWUTS

CLONE_PARENT

CLONE_PARENT_SETTID

CLONE_PTRACE

CLONE_SETTLS

CLONE_SIGHAND

CLONE_SYSVSEM

CLONE_THREAD

CLONE_UNTRACED

CLONE_VFORK

CLONE_VM

CODA_SUPER_MAGIC

CODESET

CPU_SETSIZE

CR0

CR1

CR2

CR3

CRAMFS_MAGIC

CREAD

CRNCYSTR

CRTSCTS

CS5

CS6

CS7

CS8

CSIZE

CSTOPB

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

D_FMT

D_T_FMT

E2BIG

EACCES

EADDRINUSE

EADDRNOTAVAIL

EADV

EAFNOSUPPORT

EAGAIN

EALREADY

EBADE

EBADF

EBADFD

EBADMSG

EBADR

EBADRQC

EBADSLT

EBFONT

EBUSY

ECANCELED

ECHILD

ECHO

ECHOCTL

ECHOE

ECHOK

ECHOKE

ECHONL

ECHOPRT

ECHRNG

ECOMM

ECONNABORTED

ECONNREFUSED

ECONNRESET

EDEADLK

EDEADLOCK

EDESTADDRREQ

EDOM

EDOTDOT

EDQUOT

EEXIST

EFAULT

EFBIG

EFD_CLOEXEC

EFD_NONBLOCK

EFD_SEMAPHORE

EFS_SUPER_MAGIC

EHOSTDOWN

EHOSTUNREACH

EHWPOISON

EIDRM

EILSEQ

EINPROGRESS

EINTR

EINVAL

EIO

EISCONN

EISDIR

EISNAM

EKEYEXPIRED

EKEYREJECTED

EKEYREVOKED

EL2HLT

EL2NSYNC

EL3HLT

EL3RST

ELIBACC

ELIBBAD

ELIBEXEC

ELIBMAX

ELIBSCN

ELNRNG

ELOOP

EMEDIUMTYPE

EMFILE

EMLINK

EMSGSIZE

EMULTIHOP

ENAMETOOLONG

ENAVAIL

ENETDOWN

ENETRESET

ENETUNREACH

ENFILE

ENOANO

ENOBUFS

ENOCSI

ENODATA

ENODEV

ENOENT

ENOEXEC

ENOKEY

ENOLCK

ENOLINK

ENOMEDIUM

ENOMEM

ENOMSG

ENONET

ENOPKG

ENOPROTOOPT

ENOSPC

ENOSR

ENOSTR

ENOSYS

ENOTBLK

ENOTCONN

ENOTDIR

ENOTEMPTY

ENOTNAM

ENOTRECOVERABLE

ENOTSOCK

ENOTTY

ENOTUNIQ

ENXIO

EOF

EOPNOTSUPP

EOVERFLOW

EOWNERDEAD

EPERM

EPFNOSUPPORT

EPIPE

EPOLLERR

EPOLLET

EPOLLHUP

EPOLLIN

EPOLLMSG

EPOLLONESHOT

EPOLLOUT

EPOLLPRI

EPOLLRDBAND

EPOLLRDHUP

EPOLLRDNORM

EPOLLWAKEUP

EPOLLWRBAND

EPOLLWRNORM

EPOLL_CTL_ADD

EPOLL_CTL_DEL

EPOLL_CTL_MOD

EPROTO

EPROTONOSUPPORT

EPROTOTYPE

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EREMCHG

EREMOTE

EREMOTEIO

ERESTART

ERFKILL

EROFS

ESHUTDOWN

ESOCKTNOSUPPORT

ESPIPE

ESRCH

ESRMNT

ESTALE

ESTRPIPE

ETIME

ETIMEDOUT

ETOOMANYREFS

ETXTBSY

EUCLEAN

EUNATCH

EUSERS

EWOULDBLOCK

EXDEV

EXFULL

EXIT_FAILURE

EXIT_SUCCESS

EXT2_SUPER_MAGIC

EXT3_SUPER_MAGIC

EXT4_SUPER_MAGIC

EXTPROC

FALLOC_FL_KEEP_SIZE

FALLOC_FL_PUNCH_HOLE

FD_CLOEXEC

FD_SETSIZE

FF0

FF1

FILENAME_MAX

FIOCLEX

FIONBIO

FIONREAD

FLUSHO

FOPEN_MAX

F_DUPFD

F_DUPFD_CLOEXEC

F_GETFD

F_GETFL

F_GETLEASE

F_GETLK

F_GETOWN

F_GETPIPE_SZ

F_LOCK

F_NOTIFY

F_OK

F_SETFD

F_SETFL

F_SETLEASE

F_SETLK

F_SETLKW

F_SETOWN

F_SETPIPE_SZ

F_TEST

F_TLOCK

F_ULOCK

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRPQUOTA

HPFS_SUPER_MAGIC

HUGETLBFS_MAGIC

HUPCL

ICANON

ICRNL

IEXTEN

IFF_LOOPBACK

IF_NAMESIZE

IGNBRK

IGNCR

IGNPAR

IMAXBEL

INLCR

INPCK

IPC_CREAT

IPC_EXCL

IPC_INFO

IPC_NOWAIT

IPC_RMID

IPC_SET

IPC_STAT

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_RAW

IPPROTO_TCP

IPV6_ADD_MEMBERSHIP

IPV6_DROP_MEMBERSHIP

IPV6_MULTICAST_LOOP

IPV6_V6ONLY

IP_ADD_MEMBERSHIP

IP_DROP_MEMBERSHIP

IP_HDRINCL

IP_MULTICAST_LOOP

IP_MULTICAST_TTL

IP_TRANSPARENT

IP_TTL

ISIG

ISOFS_SUPER_MAGIC

ISTRIP

IUTF8

IXANY

IXOFF

IXON

JFFS2_SUPER_MAGIC

LC_ADDRESS

LC_ADDRESS_MASK

LC_ALL

LC_ALL_MASK

LC_COLLATE

LC_COLLATE_MASK

LC_CTYPE

LC_CTYPE_MASK

LC_IDENTIFICATION

LC_IDENTIFICATION_MASK

LC_MEASUREMENT

LC_MEASUREMENT_MASK

LC_MESSAGES

LC_MESSAGES_MASK

LC_MONETARY

LC_MONETARY_MASK

LC_NAME

LC_NAME_MASK

LC_NUMERIC

LC_NUMERIC_MASK

LC_PAPER

LC_PAPER_MASK

LC_TELEPHONE

LC_TELEPHONE_MASK

LC_TIME

LC_TIME_MASK

LOCK_EX

LOCK_NB

LOCK_SH

LOCK_UN

L_tmpnam

MADV_DODUMP

MADV_DOFORK

MADV_DONTDUMP

MADV_DONTFORK

MADV_DONTNEED

MADV_HUGEPAGE

MADV_HWPOISON

MADV_MERGEABLE

MADV_NOHUGEPAGE

MADV_NORMAL

MADV_RANDOM

MADV_REMOVE

MADV_SEQUENTIAL

MADV_UNMERGEABLE

MADV_WILLNEED

MAP_32BIT

MAP_ANON

MAP_ANONYMOUS

MAP_DENYWRITE

MAP_EXECUTABLE

MAP_FAILED

MAP_FILE

MAP_FIXED

MAP_GROWSDOWN

MAP_HUGETLB

MAP_LOCKED

MAP_NONBLOCK

MAP_NORESERVE

MAP_POPULATE

MAP_PRIVATE

MAP_SHARED

MAP_STACK

MAP_TYPE

MCL_CURRENT

MCL_FUTURE

MINIX2_SUPER_MAGIC

MINIX2_SUPER_MAGIC2

MINIX_SUPER_MAGIC

MINIX_SUPER_MAGIC2

MNT_DETACH

MNT_EXPIRE

MNT_FORCE

MON_1

MON_10

MON_11

MON_12

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MSDOS_SUPER_MAGIC

MS_ACTIVE

MS_ASYNC

MS_BIND

MS_DIRSYNC

MS_INVALIDATE

MS_I_VERSION

MS_KERNMOUNT

MS_MANDLOCK

MS_MGC_MSK

MS_MGC_VAL

MS_MOVE

MS_NOATIME

MS_NODEV

MS_NODIRATIME

MS_NOEXEC

MS_NOSUID

MS_NOUSER

MS_POSIXACL

MS_PRIVATE

MS_RDONLY

MS_REC

MS_RELATIME

MS_REMOUNT

MS_RMT_MASK

MS_SHARED

MS_SILENT

MS_SLAVE

MS_STRICTATIME

MS_SYNC

MS_SYNCHRONOUS

MS_UNBINDABLE

NCCS

NCP_SUPER_MAGIC

NFS_SUPER_MAGIC

NI_MAXHOST

NL0

NL1

NOEXPR

NOFLSH

NOSTR

ONLCR

OPENPROM_SUPER_MAGIC

OPOST

O_ACCMODE

O_APPEND

O_ASYNC

O_CLOEXEC

O_CREAT

O_DIRECT

O_DIRECTORY

O_DSYNC

O_EXCL

O_FSYNC

O_NDELAY

O_NOCTTY

O_NOFOLLOW

O_NONBLOCK

O_RDONLY

O_RDWR

O_RSYNC

O_SYNC

O_TRUNC

O_WRONLY

PARENB

PARMRK

PARODD

PATH_MAX

PENDIN

PM_STR

POLLERR

POLLHUP

POLLIN

POLLNVAL

POLLOUT

POLLPRI

POSIX_FADV_DONTNEED

POSIX_FADV_NOREUSE

POSIX_FADV_NORMAL

POSIX_FADV_RANDOM

POSIX_FADV_SEQUENTIAL

POSIX_FADV_WILLNEED

POSIX_MADV_DONTNEED

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

PROC_SUPER_MAGIC

PROT_EXEC

PROT_GROWSDOWN

PROT_GROWSUP

PROT_NONE

PROT_READ

PROT_WRITE

PTHREAD_COND_INITIALIZER

PTHREAD_CREATE_DETACHED

PTHREAD_CREATE_JOINABLE

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_RECURSIVE

PTHREAD_RWLOCK_INITIALIZER

PTHREAD_STACK_MIN

PTRACE_ATTACH

PTRACE_CONT

PTRACE_DETACH

PTRACE_GETEVENTMSG

PTRACE_GETFPREGS

PTRACE_GETFPXREGS

PTRACE_GETREGS

PTRACE_GETREGSET

PTRACE_GETSIGINFO

PTRACE_INTERRUPT

PTRACE_KILL

PTRACE_LISTEN

PTRACE_PEEKDATA

PTRACE_PEEKSIGINFO

PTRACE_PEEKTEXT

PTRACE_PEEKUSER

PTRACE_POKEDATA

PTRACE_POKETEXT

PTRACE_POKEUSER

PTRACE_SEIZE

PTRACE_SETFPREGS

PTRACE_SETFPXREGS

PTRACE_SETOPTIONS

PTRACE_SETREGS

PTRACE_SETREGSET

PTRACE_SETSIGINFO

PTRACE_SINGLESTEP

PTRACE_SYSCALL

PTRACE_TRACEME

QFMT_VFS_OLD

QFMT_VFS_V0

QFMT_VFS_V1

QIF_ALL

QIF_BLIMITS

QIF_BTIME

QIF_ILIMITS

QIF_INODES

QIF_ITIME

QIF_LIMITS

QIF_SPACE

QIF_TIMES

QIF_USAGE

QNX4_SUPER_MAGIC

Q_GETFMT

Q_GETINFO

Q_GETQUOTA

Q_QUOTAOFF

Q_QUOTAON

Q_SETINFO

Q_SETQUOTA

Q_SYNC

RADIXCHAR

RAND_MAX

REISERFS_SUPER_MAGIC

RLIMIT_AS

RLIMIT_CORE

RLIMIT_CPU

RLIMIT_DATA

RLIMIT_FSIZE

RLIMIT_LOCKS

RLIMIT_MEMLOCK

RLIMIT_MSGQUEUE

RLIMIT_NICE

RLIMIT_NLIMITS

RLIMIT_NOFILE

RLIMIT_NPROC

RLIMIT_RSS

RLIMIT_RTPRIO

RLIMIT_RTTIME

RLIMIT_SIGPENDING

RLIMIT_STACK

RLIM_INFINITY

RLIM_SAVED_CUR

RLIM_SAVED_MAX

RTLD_DEEPBIND

RTLD_DEFAULT

RTLD_GLOBAL

RTLD_LAZY

RTLD_LOCAL

RTLD_NEXT

RTLD_NODELETE

RTLD_NOLOAD

RTLD_NOW

RUSAGE_CHILDREN

RUSAGE_SELF

RUSAGE_THREAD

R_OK

SA_NOCLDSTOP

SA_NOCLDWAIT

SA_NODEFER

SA_ONSTACK

SA_RESETHAND

SA_RESTART

SA_SIGINFO

SCHED_BATCH

SCHED_FIFO

SCHED_IDLE

SCHED_OTHER

SCHED_RR

SEEK_CUR

SEEK_END

SEEK_SET

SFD_CLOEXEC

SFD_NONBLOCK

SHM_EXEC

SHM_HUGETLB

SHM_LOCK

SHM_NORESERVE

SHM_R

SHM_RDONLY

SHM_REMAP

SHM_RND

SHM_UNLOCK

SHM_W

SHUT_RD

SHUT_RDWR

SHUT_WR

SIGABRT

SIGALRM

SIGBUS

SIGCHLD

SIGCONT

SIGFPE

SIGHUP

SIGILL

SIGINT

SIGIO

SIGIOT

SIGKILL

SIGPIPE

SIGPOLL

SIGPROF

SIGPWR

SIGQUIT

SIGSEGV

SIGSTKFLT

SIGSTKSZ

SIGSTOP

SIGSYS

SIGTERM

SIGTRAP

SIGTSTP

SIGTTIN

SIGTTOU

SIGUNUSED

SIGURG

SIGUSR1

SIGUSR2

SIGVTALRM

SIGWINCH

SIGXCPU

SIGXFSZ

SIG_BLOCK

SIG_DFL

SIG_ERR

SIG_IGN

SIG_SETMASK

SIG_UNBLOCK

SMB_SUPER_MAGIC

SOCK_CLOEXEC

SOCK_DGRAM

SOCK_NONBLOCK

SOCK_RAW

SOCK_STREAM

SOL_SOCKET

SO_ACCEPTCONN

SO_BROADCAST

SO_DEBUG

SO_DONTROUTE

SO_ERROR

SO_KEEPALIVE

SO_LINGER

SO_OOBINLINE

SO_PEERCRED

SO_RCVBUF

SO_RCVLOWAT

SO_RCVTIMEO

SO_REUSEADDR

SO_REUSEPORT

SO_SNDBUF

SO_SNDLOWAT

SO_SNDTIMEO

SO_TYPE

SPLICE_F_GIFT

SPLICE_F_MORE

SPLICE_F_MOVE

SPLICE_F_NONBLOCK

SS_DISABLE

SS_ONSTACK

STDERR_FILENO

STDIN_FILENO

STDOUT_FILENO

ST_APPEND

ST_IMMUTABLE

ST_MANDLOCK

ST_NOATIME

ST_NODEV

ST_NODIRATIME

ST_NOEXEC

ST_NOSUID

ST_RDONLY

ST_RELATIME

ST_SYNCHRONOUS

ST_WRITE

SYS_gettid

SYS_perf_event_open

S_IEXEC

S_IFBLK

S_IFCHR

S_IFDIR

S_IFIFO

S_IFLNK

S_IFMT

S_IFREG

S_IFSOCK

S_IREAD

S_IRGRP

S_IROTH

S_IRUSR

S_IRWXG

S_IRWXO

S_IRWXU

S_ISGID

S_ISUID

S_ISVTX

S_IWGRP

S_IWOTH

S_IWRITE

S_IWUSR

S_IXGRP

S_IXOTH

S_IXUSR

TAB0

TAB1

TAB2

TAB3

TCFLSH

TCGETA

TCGETS

TCIFLUSH

TCIOFF

TCIOFLUSH

TCION

TCOFLUSH

TCOOFF

TCOON

TCP_CONGESTION

TCP_COOKIE_TRANSACTIONS

TCP_CORK

TCP_DEFER_ACCEPT

TCP_FASTOPEN

TCP_INFO

TCP_KEEPCNT

TCP_KEEPIDLE

TCP_KEEPINTVL

TCP_LINGER2

TCP_MAXSEG

TCP_MD5SIG

TCP_NODELAY

TCP_QUEUE_SEQ

TCP_QUICKACK

TCP_REPAIR

TCP_REPAIR_OPTIONS

TCP_REPAIR_QUEUE

TCP_SYNCNT

TCP_THIN_DUPACK

TCP_THIN_LINEAR_TIMEOUTS

TCP_TIMESTAMP

TCP_USER_TIMEOUT

TCP_WINDOW_CLAMP

TCSADRAIN

TCSAFLUSH

TCSANOW

TCSBRK

TCSETA

TCSETAF

TCSETAW

TCSETS

TCSETSF

TCSETSW

TCXONC

THOUSEP

TIOCCONS

TIOCEXCL

TIOCGPGRP

TIOCGSERIAL

TIOCGSOFTCAR

TIOCGWINSZ

TIOCINQ

TIOCLINUX

TIOCMBIC

TIOCMBIS

TIOCMGET

TIOCMSET

TIOCNXCL

TIOCOUTQ

TIOCSCTTY

TIOCSPGRP

TIOCSSOFTCAR

TIOCSTI

TIOCSWINSZ

TMPFS_MAGIC

TMP_MAX

TOSTOP

T_FMT

T_FMT_AMPM

ULONG_SIZE

USBDEVICE_SUPER_MAGIC

USRQUOTA

VDISCARD

VEOF

VEOL

VEOL2

VERASE

VINTR

VKILL

VLNEXT

VMIN

VQUIT

VREPRINT

VSTART

VSTOP

VSUSP

VT0

VT1

VTIME

VWERASE

WNOHANG

W_OK

X_OK

YESEXPR

YESSTR

_IOFBF

_IOLBF

_IONBF

_PC_NAME_MAX

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_C_VERSION

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_COLL_WEIGHTS_MAX

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_IOV_MAX

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_ONLN

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_REALTIME_SIGNALS

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SIGQUEUE_MAX

_SC_STREAM_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMERS

_SC_TIMER_MAX

_SC_TTY_NAME_MAX

_SC_TZNAME_MAX

_SC_VERSION

_SC_XBS5_ILP32_OFF32

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_LPBIG_OFFBIG

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

__SIZEOF_PTHREAD_COND_T

__SIZEOF_PTHREAD_MUTEXATTR_T

__SIZEOF_PTHREAD_MUTEX_T

__SIZEOF_PTHREAD_RWLOCK_T

Functions

CPU_EQUAL

CPU_ISSET

CPU_SET

CPU_ZERO

FD_CLR

FD_ISSET

FD_SET

FD_ZERO

WEXITSTATUS

WIFEXITED

WTERMSIG

__errno_location

accept

access

alarm

backtrace

bind

cfgetispeed

cfgetospeed

cfsetispeed

cfsetospeed

chdir

chmod

chown

chroot

clock_getres

clock_gettime

clone

close

closedir

connect

creat

creat64

daemon

dladdr

dlclose

dlerror

dlopen

dlsym

dup

dup2

dup3

duplocale

epoll_create

epoll_ctl

epoll_pwait

epoll_wait

eventfd

execv

execve

execvp

fallocate

fchmod

fcntl

fdatasync

fdopen

fgetpos64

fgetxattr

fileno

flistxattr

flock

fopen64

fork

forkpty

fpathconf

fprintf

freeaddrinfo

freeifaddrs

freelocale

fremovexattr

freopen64

fscanf

fseeko

fseeko64

fsetpos64

fsetxattr

fstat

fstat64

fstatfs

fstatvfs

fsync

ftello

ftello64

ftruncate

ftruncate64

futimens

futimes

gai_strerror

getaddrinfo

getcontext

getcwd

getdtablesize

getegid

geteuid

getgid

getgroups

gethostname

getifaddrs

getlogin

getnameinfo

getopt

getpeername

getpgrp

getpid

getppid

getpwuid_r

getrlimit

getrlimit64

getrusage

getsid

getsockname

getsockopt

gettimeofday

getuid

getxattr

glob

glob64

globfree

globfree64

gmtime_r

if_indextoname

if_nametoindex

ioctl

isatty

kill

lgetxattr

link

listen

listxattr

llistxattr

localeconv

localtime_r

lremovexattr

lseek

lseek64

lsetxattr

lstat

lstat64

madvise

makecontext

memalign

memrchr

mincore

mkdir

mkdtemp

mkfifo

mknod

mkostemp

mkostemps

mkstemp

mkstemps

mktime

mlock

mlockall

mmap

mmap64

mount

mprotect

mq_close

mq_getattr

mq_open

mq_receive

mq_send

mq_setattr

mq_unlink

msync

munlock

munlockall

munmap

nanosleep

newlocale

nl_langinfo

nl_langinfo_l

open

open64

opendir

openpty

pathconf

pause

pclose

pipe

pipe2

poll

popen

posix_fadvise

posix_fallocate

posix_madvise

posix_memalign

prctl

pread

pread64

preadv

printf

prlimit

prlimit64

pselect

pthread_attr_destroy

pthread_attr_getaffinity_np

pthread_attr_getguardsize

pthread_attr_getstack

pthread_attr_init

pthread_attr_setaffinity_np

pthread_attr_setdetachstate

pthread_attr_setstacksize

pthread_cond_broadcast

pthread_cond_destroy

pthread_cond_signal

pthread_cond_timedwait

pthread_cond_wait

pthread_create

pthread_detach

pthread_getattr_np

pthread_getspecific

pthread_join

pthread_key_create

pthread_key_delete

pthread_kill

pthread_mutex_destroy

pthread_mutex_init

pthread_mutex_lock

pthread_mutex_trylock

pthread_mutex_unlock

pthread_mutexattr_destroy

pthread_mutexattr_init

pthread_mutexattr_settype

pthread_rwlock_destroy

pthread_rwlock_rdlock

pthread_rwlock_tryrdlock

pthread_rwlock_trywrlock

pthread_rwlock_unlock

pthread_rwlock_wrlock

pthread_self

pthread_setspecific

pthread_sigmask

ptrace

putenv

pwrite

pwrite64

pwritev

quotactl

raise

read

readahead

readdir64_r

readdir_r

readlink

readv

realpath

recv

recvfrom

recvmsg

removexattr

rewinddir

rmdir

scanf

sched_get_priority_max

sched_get_priority_min

sched_getaffinity

sched_getscheduler

sched_setaffinity

sched_setscheduler

sched_yield

seekdir

select

send

sendfile

sendmsg

sendto

setcontext

setenv

setgid

setgroups

sethostname

setlocale

setns

setpgid

setrlimit

setrlimit64

setsid

setsockopt

setuid

setxattr

shm_open

shm_unlink

shmat

shmctl

shmdt

shmget

shutdown

sigaction

sigaddset

sigaltstack

sigdelset

sigemptyset

sigfillset

sigismember

signal

signalfd

sigtimedwait

sigwait

sigwaitinfo

sleep

snprintf

socket

socketpair

splice

sprintf

sscanf

stat

stat64

statfs

statvfs

strerror_r

swapcontext

symlink

syscall

sysconf

sysctl

tcdrain

tcflow

tcflush

tcgetattr

tcgetpgrp

tcsendbreak

tcsetattr

tee

telldir

timegm

tmpfile64

ttyname

umask

umount

umount2

uname

unlink

unsetenv

unshare

uselocale

usleep

utime

utimensat

utimes

vmsplice

wait

waitpid

write

writev

Type Definitions

__fsword_t

blkcnt64_t

blkcnt_t

blksize_t

c_char

c_long

c_ulong

cc_t

clock_t

clockid_t

dev_t

fsblkcnt_t

fsfilcnt_t

gid_t

in_addr_t

in_port_t

ino64_t

ino_t

key_t

loff_t

mode_t

mqd_t

nfds_t

nl_item

nlink_t

off64_t

off_t

pid_t

pthread_key_t

pthread_t

rlim64_t

rlim_t

sa_family_t

shmatt_t

sighandler_t

socklen_t

speed_t

suseconds_t

tcflag_t

time_t

uid_t

useconds_t

wchar_t