Module libc::unix

Definitions found commonly among almost all Unix derivatives

More functions and definitions can be found in the more specific modules according to the platform in question.

Re-exports

pub use self::notbsd::*;

Modules

notbsd

Structs

group

hostent

in6_addr

in_addr

iovec

ip_mreq

ipv6_mreq

itimerval

linger

pollfd

protoent

rlimit

rusage

servent

sigval

timespec

timeval

tms

utimbuf

winsize

Enums

DIR

locale_t

Constants

DT_BLK

DT_CHR

DT_DIR

DT_FIFO

DT_LNK

DT_REG

DT_SOCK

FD_CLOEXEC

GRPQUOTA

IFNAMSIZ

IF_NAMESIZE

INADDR_ANY

INADDR_BROADCAST

INADDR_LOOPBACK

INADDR_NONE

IPPROTO_ICMP

IPPROTO_ICMPV6

IPPROTO_IP

IPPROTO_IPV6

IPPROTO_TCP

IPPROTO_UDP

LOG_ALERT

LOG_AUTH

LOG_CONS

LOG_CRIT

LOG_DAEMON

LOG_DEBUG

LOG_EMERG

LOG_ERR

LOG_FACMASK

LOG_INFO

LOG_KERN

LOG_LOCAL0

LOG_LOCAL1

LOG_LOCAL2

LOG_LOCAL3

LOG_LOCAL4

LOG_LOCAL5

LOG_LOCAL6

LOG_LOCAL7

LOG_LPR

LOG_MAIL

LOG_NDELAY

LOG_NEWS

LOG_NOTICE

LOG_NOWAIT

LOG_ODELAY

LOG_PID

LOG_PRIMASK

LOG_SYSLOG

LOG_USER

LOG_UUCP

LOG_WARNING

PRIO_MAX

PRIO_MIN

PRIO_PGRP

PRIO_PROCESS

PRIO_USER

SIGIOT

SIG_DFL

SIG_ERR

SIG_IGN

S_ISGID

S_ISUID

S_ISVTX

USRQUOTA

Functions

accept⚠

access⚠

alarm⚠

cfgetispeed⚠

cfgetospeed⚠

cfmakeraw⚠

cfsetispeed⚠

cfsetospeed⚠

cfsetspeed⚠

chdir⚠

chmod⚠

chown⚠

chroot⚠

close⚠

closedir⚠

closelog⚠

connect⚠

creat⚠

difftime⚠

dladdr⚠

dlclose⚠

dlerror⚠

dlopen⚠

dlsym⚠

dup⚠

dup2⚠

execl⚠

execle⚠

execlp⚠

execv⚠

execve⚠

execvp⚠

fchdir⚠

fchmod⚠

fchmodat⚠

fchown⚠

fchownat⚠

fcntl⚠

fdopen⚠

fileno⚠

flock⚠

fork⚠

fpathconf⚠

fprintf⚠

freeaddrinfo⚠

fscanf⚠

fseeko⚠

fstat⚠

fstatat⚠

fstatvfs⚠

fsync⚠

ftello⚠

ftruncate⚠

gai_strerror⚠

getaddrinfo⚠

getchar_unlocked⚠

getcwd⚠

getegid⚠

geteuid⚠

getgid⚠

getgroups⚠

gethostname⚠

getlogin⚠

getopt⚠

getpeername⚠

getpgid⚠

getpgrp⚠

getpid⚠

getppid⚠

getprotobyname⚠

getprotobynumber⚠

getpwnam⚠

getpwuid⚠

getrlimit⚠

getrusage⚠

getservbyname⚠

getsid⚠

getsockname⚠

getsockopt⚠

gettimeofday⚠

getuid⚠

gmtime⚠

gmtime_r⚠

grantpt⚠

if_indextoname⚠

if_nametoindex⚠

isatty⚠

kill⚠

lchown⚠

link⚠

linkat⚠

listen⚠

localeconv⚠

localtime⚠

localtime_r⚠

lseek⚠

lstat⚠

mkdir⚠

mkdirat⚠

mkdtemp⚠

mkfifo⚠

mknod⚠

mkstemp⚠

mktime⚠

mlock⚠

mlockall⚠

mmap⚠

munlock⚠

munlockall⚠

munmap⚠

nanosleep⚠

nice⚠

open⚠

openat⚠

opendir⚠

openlog⚠

pathconf⚠

pause⚠

pclose⚠

pipe⚠

poll⚠

posix_memalign⚠

posix_openpt⚠

pread⚠

printf⚠

pselect⚠

pthread_attr_destroy⚠

pthread_attr_init⚠

pthread_attr_setdetachstate⚠

pthread_attr_setstacksize⚠

pthread_cond_broadcast⚠

pthread_cond_destroy⚠

pthread_cond_init⚠

pthread_cond_signal⚠

pthread_cond_timedwait⚠

pthread_cond_wait⚠

pthread_condattr_destroy⚠

pthread_condattr_init⚠

pthread_detach⚠

pthread_exit⚠

pthread_getspecific⚠

pthread_join⚠

pthread_key_create⚠

pthread_key_delete⚠

pthread_mutex_destroy⚠

pthread_mutex_init⚠

pthread_mutex_lock⚠

pthread_mutex_trylock⚠

pthread_mutex_unlock⚠

pthread_mutexattr_destroy⚠

pthread_mutexattr_init⚠

pthread_mutexattr_settype⚠

pthread_rwlock_destroy⚠

pthread_rwlock_init⚠

pthread_rwlock_rdlock⚠

pthread_rwlock_tryrdlock⚠

pthread_rwlock_trywrlock⚠

pthread_rwlock_unlock⚠

pthread_rwlock_wrlock⚠

pthread_rwlockattr_destroy⚠

pthread_rwlockattr_init⚠

pthread_self⚠

pthread_setspecific⚠

ptsname⚠

putchar_unlocked⚠

putenv⚠

pwrite⚠

raise⚠

read⚠

readdir⚠

readdir_r⚠

readlink⚠

readlinkat⚠

realpath⚠

recv⚠

renameat⚠

res_init⚠

rewinddir⚠

rmdir⚠

scanf⚠

sched_yield⚠

select⚠

sem_destroy⚠

sem_init⚠

sem_post⚠

sem_trywait⚠

sem_wait⚠

send⚠

sendto⚠

setenv⚠

seteuid⚠

setgid⚠

setlocale⚠

setlogmask⚠

setpgid⚠

setrlimit⚠

setsid⚠

setsockopt⚠

setuid⚠

shutdown⚠

sigaction⚠

sigaddset⚠

sigdelset⚠

sigemptyset⚠

sigfillset⚠

sigismember⚠

signal⚠

sigpending⚠

sigprocmask⚠

sleep⚠

snprintf⚠

socket⚠

socketpair⚠

sprintf⚠

sscanf⚠

stat⚠

statvfs⚠

strerror_r⚠

symlink⚠

symlinkat⚠

sysconf⚠

syslog⚠

tcdrain⚠

tcflow⚠

tcflush⚠

tcgetattr⚠

tcgetpgrp⚠

tcgetsid⚠

tcsendbreak⚠

tcsetattr⚠

tcsetpgrp⚠

time⚠

timegm⚠

times⚠

tmpnam⚠

ttyname⚠

umask⚠

uname⚠

unlink⚠

unlinkat⚠

unlockpt⚠

unsetenv⚠

usleep⚠

utime⚠

utimes⚠

wait⚠

waitpid⚠

write⚠

Type Definitions

cc_t

gid_t

in_addr_t

in_port_t

pid_t

sighandler_t

uid_t