libc::unix::linux_like

Module linux

Linux-specific definitions for linux-like values

Re-exports

pub use self::gnu::*;

pub use self::arch::*;

Modules

arch 🔒

gnu 🔒

Structs

__c_anonymous__kernel_fsid_t

af_alg_ivDeprecated

WARNING: The PartialEq, Eq and Hash implementations of this type are unsound and will be removed in the future.

dmabuf_cmsg

dmabuf_token

dqblk

epoll_params

fanotify_event_info_fid

fanotify_event_info_header

fanotify_event_metadata

fanotify_response

fanout_args

ff_condition_effect

ff_constant_effect

ff_effect

ff_envelope

ff_periodic_effect

ff_ramp_effect

ff_replay

ff_rumble_effect

ff_trigger

genlmsghdr

hwtstamp_config

in6_ifreq

inotify_event

input_absinfo

input_event

input_id

input_keymap_entry

input_mask

iw_discarded

iw_encode_ext

iw_event

iw_freq

iw_michaelmicfailure

iw_missed

iw_mlme

iw_param

iw_pmkid_cand

iw_pmksa

iw_point

iw_priv_args

iw_quality

iw_range

iw_scan_req

iw_statistics

iw_thrspy

iwreq

mnt_ns_info

mount_attr

mq_attr

msginfo

open_how

pidfd_info

posix_spawn_file_actions_t

posix_spawnattr_t

pthread_barrier_t

pthread_barrierattr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

ptp_clock_caps

ptp_clock_time

ptp_extts_event

ptp_extts_request

ptp_perout_request

ptp_pin_desc

ptp_sys_offset

ptp_sys_offset_extended

ptp_sys_offset_precise

sched_attr

sctp_authinfo

sctp_initmsg

sctp_nxtinfo

sctp_prinfo

sctp_rcvinfo

sctp_sndinfo

sctp_sndrcvinfo

seccomp_data

seccomp_notif

seccomp_notif_addfd

seccomp_notif_resp

seccomp_notif_sizes

signalfd_siginfo

sock_extended_err

sock_txtime

sockaddr_alg

sockaddr_pktDeprecated

sockaddr_vm

sockaddr_xdp

tls12_crypto_info_aes_ccm_128

tls12_crypto_info_aes_gcm_128

tls12_crypto_info_aes_gcm_256

tls12_crypto_info_aria_gcm_128

tls12_crypto_info_aria_gcm_256

tls12_crypto_info_chacha20_poly1305

tls12_crypto_info_sm4_ccm

tls12_crypto_info_sm4_gcm

tls_crypto_info

tpacket2_hdr

tpacket3_hdr

tpacket_auxdata

tpacket_bd_ts

tpacket_block_desc

tpacket_hdr

tpacket_hdr_v1

tpacket_hdr_variant1

tpacket_req

tpacket_req3

tpacket_rollover_stats

tpacket_stats

tpacket_stats_v3

uinput_abs_setup

uinput_ff_erase

uinput_ff_upload

uinput_setup

uinput_user_dev

xdp_desc

xdp_mmap_offsets

xdp_mmap_offsets_v1

xdp_options

xdp_ring_offset

xdp_ring_offset_v1

xdp_statistics

xdp_statistics_v1

xdp_umem_reg

xdp_umem_reg_v1

xsk_tx_metadata

xsk_tx_metadata_completion

xsk_tx_metadata_request

Enums

tpacket_versions

Constants

ABS_CNT

ABS_MAX

ALG_OP_DECRYPT

ALG_OP_ENCRYPT

ALG_SET_AEAD_ASSOCLEN

ALG_SET_AEAD_AUTHSIZE

ALG_SET_DRBG_ENTROPY

ALG_SET_IV

ALG_SET_KEY

ALG_SET_KEY_BY_KEY_SERIAL

ALG_SET_OP

AT_EXECVE_CHECK

BPF_A

BPF_ABS

BPF_ADD

BPF_ALU

BPF_AND

BPF_B

BPF_DIV

BPF_H

BPF_IMM

BPF_IND

BPF_JA

BPF_JEQ

BPF_JGE

BPF_JGT

BPF_JMP

BPF_JSET

BPF_K

BPF_LD

BPF_LDX

BPF_LEN

BPF_LL_OFF

BPF_LSH

BPF_MAXINSNS

BPF_MEM

BPF_MEMWORDS

BPF_MISC

BPF_MOD

BPF_MSH

BPF_MUL

BPF_NEG

BPF_NET_OFF

BPF_OR

BPF_RET

BPF_RSH

BPF_ST

BPF_STX

BPF_SUB

BPF_TAX

BPF_TXA

BPF_W

BPF_X

BPF_XOR

CLONE_PIDFD

CLOSE_RANGE_CLOEXEC

CLOSE_RANGE_UNSHARE

CN_DST_IDX

CN_DST_VAL

CN_IDX_BB

CN_IDX_CIFS

CN_IDX_DM

CN_IDX_DRBD

CN_IDX_PROC

CN_IDX_V86D

CN_KVP_IDX

CN_KVP_VAL

CN_VAL_CIFS

CN_VAL_DM_USERSPACE_LOG

CN_VAL_DRBD

CN_VAL_PROC

CN_VAL_V86D_UVESAFB

CN_VSS_IDX

CN_VSS_VAL

CN_W1_IDX

CN_W1_VAL

CTL_ABI

CTL_BUS

CTL_BUS_ISA

CTL_CPU

CTL_DEBUG

CTL_DEV

CTL_FS

CTL_KERN

CTL_NET

CTL_VM

CTRL_ATTR_FAMILY_ID

CTRL_ATTR_FAMILY_NAME

CTRL_ATTR_HDRSIZE

CTRL_ATTR_MAXATTR

CTRL_ATTR_MCAST_GROUPS

CTRL_ATTR_MCAST_GRP_ID

CTRL_ATTR_MCAST_GRP_NAME

CTRL_ATTR_MCAST_GRP_UNSPEC

CTRL_ATTR_OPS

CTRL_ATTR_OP_FLAGS

CTRL_ATTR_OP_ID

CTRL_ATTR_OP_UNSPEC

CTRL_ATTR_UNSPEC

CTRL_ATTR_VERSION

CTRL_CMD_DELFAMILY

CTRL_CMD_DELMCAST_GRP

CTRL_CMD_DELOPS

CTRL_CMD_GETFAMILY

CTRL_CMD_GETMCAST_GRP

CTRL_CMD_GETOPS

CTRL_CMD_NEWFAMILY

CTRL_CMD_NEWMCAST_GRP

CTRL_CMD_NEWOPS

CTRL_CMD_UNSPEC

DCCP_SERVICE_LIST_MAX_LEN

maximum number of services provided on the same listening port

DCCP_SOCKOPT_AVAILABLE_CCIDS

DCCP_SOCKOPT_CCID

DCCP_SOCKOPT_CCID_RX_INFO

DCCP_SOCKOPT_CCID_TX_INFO

DCCP_SOCKOPT_CHANGE_L

DCCP_SOCKOPT_CHANGE_R

DCCP_SOCKOPT_GET_CUR_MPS

DCCP_SOCKOPT_PACKET_SIZE

DCCP_SOCKOPT_QPOLICY_ID

DCCP_SOCKOPT_QPOLICY_TXQLEN

DCCP_SOCKOPT_RECV_CSCOV

DCCP_SOCKOPT_RX_CCID

DCCP_SOCKOPT_SEND_CSCOV

DCCP_SOCKOPT_SERVER_TIMEWAIT

DCCP_SOCKOPT_SERVICE

DCCP_SOCKOPT_TX_CCID

EFD_SEMAPHORE

ENOATTRDeprecated

EPIOCGPARAMS

EPIOCSPARAMS

ETH_ALEN

ETH_DATA_LEN

ETH_FCS_LEN

ETH_FRAME_LEN

ETH_HLEN

ETH_P_802_2

ETH_P_802_3

ETH_P_802_3_MIN

ETH_P_802_EX1

ETH_P_1588

ETH_P_8021AD

ETH_P_8021AH

ETH_P_8021Q

ETH_P_80221

ETH_P_AARP

ETH_P_AF_IUCV

ETH_P_ALL

ETH_P_AOE

ETH_P_ARCNET

ETH_P_ARP

ETH_P_ATALK

ETH_P_ATMFATE

ETH_P_ATMMPOA

ETH_P_AX25

ETH_P_BATMAN

ETH_P_BPQ

ETH_P_CAIF

ETH_P_CANFD

ETH_P_CONTROL

ETH_P_CUST

ETH_P_DDCMP

ETH_P_DEC

ETH_P_DIAG

ETH_P_DNA_DL

ETH_P_DNA_RC

ETH_P_DNA_RT

ETH_P_DSA

ETH_P_ECONET

ETH_P_EDSA

ETH_P_FCOE

ETH_P_FIP

ETH_P_HDLC

ETH_P_IEEE802154

ETH_P_IEEEPUP

ETH_P_IEEEPUPAT

ETH_P_IP

ETH_P_IPV6

ETH_P_IPX

ETH_P_IRDA

ETH_P_LAT

ETH_P_LINK_CTL

ETH_P_LOCALTALK

ETH_P_LOOP

ETH_P_LOOPBACK

ETH_P_MACSEC

ETH_P_MOBITEX

ETH_P_MPLS_MC

ETH_P_MPLS_UC

ETH_P_MVRP

ETH_P_PAE

ETH_P_PAUSE

ETH_P_PHONET

ETH_P_PPPTALK

ETH_P_PPP_DISC

ETH_P_PPP_MP

ETH_P_PPP_SES

ETH_P_PRP

ETH_P_PUP

ETH_P_PUPAT

ETH_P_QINQ1

ETH_P_QINQ2

ETH_P_QINQ3

ETH_P_RARP

ETH_P_SCA

ETH_P_SLOW

ETH_P_SNAP

ETH_P_TDLS

ETH_P_TEB

ETH_P_TIPC

ETH_P_TRAILER

ETH_P_TR_802_2

ETH_P_WAN_PPP

ETH_P_WCCP

ETH_P_X25

ETH_ZLEN

EV_CNT

EV_MAX

FALLOC_FL_COLLAPSE_RANGE

FALLOC_FL_INSERT_RANGE

FALLOC_FL_KEEP_SIZE

FALLOC_FL_PUNCH_HOLE

FALLOC_FL_UNSHARE_RANGE

FALLOC_FL_ZERO_RANGE

FANOTIFY_METADATA_VERSION

FAN_ACCESS

FAN_ACCESS_PERM

FAN_ALLOW

FAN_ATTRIB

FAN_AUDIT

FAN_CLASS_CONTENT

FAN_CLASS_NOTIF

FAN_CLASS_PRE_CONTENT

FAN_CLOEXEC

FAN_CLOSE

FAN_CLOSE_NOWRITE

FAN_CLOSE_WRITE

FAN_CREATE

FAN_DELETE

FAN_DELETE_SELF

FAN_DENY

FAN_ENABLE_AUDIT

FAN_EPIDFD

FAN_EVENT_INFO_TYPE_DFID

FAN_EVENT_INFO_TYPE_DFID_NAME

FAN_EVENT_INFO_TYPE_ERROR

FAN_EVENT_INFO_TYPE_FID

FAN_EVENT_INFO_TYPE_NEW_DFID_NAME

FAN_EVENT_INFO_TYPE_OLD_DFID_NAME

FAN_EVENT_INFO_TYPE_PIDFD

FAN_EVENT_ON_CHILD

FAN_FS_ERROR

FAN_INFO

FAN_MARK_ADD

FAN_MARK_DONT_FOLLOW

FAN_MARK_EVICTABLE

FAN_MARK_FILESYSTEM

FAN_MARK_FLUSH

FAN_MARK_IGNORE

FAN_MARK_IGNORED_MASK

FAN_MARK_IGNORED_SURV_MODIFY

FAN_MARK_IGNORE_SURV

FAN_MARK_INODE

FAN_MARK_MOUNT

FAN_MARK_ONLYDIR

FAN_MARK_REMOVE

FAN_MODIFY

FAN_MOVE

FAN_MOVED_FROM

FAN_MOVED_TO

FAN_MOVE_SELF

FAN_NOFD

FAN_NONBLOCK

FAN_NOPIDFD

FAN_ONDIR

FAN_OPEN

FAN_OPEN_EXEC

FAN_OPEN_EXEC_PERM

FAN_OPEN_PERM

FAN_Q_OVERFLOW

FAN_RENAME

FAN_REPORT_DFID_NAME

FAN_REPORT_DFID_NAME_TARGET

FAN_REPORT_DIR_FID

FAN_REPORT_FID

FAN_REPORT_NAME

FAN_REPORT_PIDFD

FAN_REPORT_TARGET_FID

FAN_REPORT_TID

FAN_RESPONSE_INFO_AUDIT_RULE

FAN_RESPONSE_INFO_NONE

FAN_UNLIMITED_MARKS

FAN_UNLIMITED_QUEUE

FF_CNT

FF_MAX

FUTEX_BITSET_MATCH_ANY

FUTEX_CLOCK_REALTIME

FUTEX_CMD_MASK

FUTEX_CMP_REQUEUE

FUTEX_CMP_REQUEUE_PI

FUTEX_FD

FUTEX_LOCK_PI

FUTEX_LOCK_PI2

FUTEX_OP_ADD

FUTEX_OP_ANDN

FUTEX_OP_CMP_EQ

FUTEX_OP_CMP_GE

FUTEX_OP_CMP_GT

FUTEX_OP_CMP_LE

FUTEX_OP_CMP_LT

FUTEX_OP_CMP_NE

FUTEX_OP_OPARG_SHIFT

FUTEX_OP_OR

FUTEX_OP_SET

FUTEX_OP_XOR

FUTEX_OWNER_DIED

FUTEX_PRIVATE_FLAG

FUTEX_REQUEUE

FUTEX_TID_MASK

FUTEX_TRYLOCK_PI

FUTEX_UNLOCK_PI

FUTEX_WAIT

FUTEX_WAITERS

FUTEX_WAIT_BITSET

FUTEX_WAIT_REQUEUE_PI

FUTEX_WAKE

FUTEX_WAKE_BITSET

FUTEX_WAKE_OP

F_SEAL_EXEC

F_SEAL_FUTURE_WRITE

GENL_ADMIN_PERM

GENL_CMD_CAP_DO

GENL_CMD_CAP_DUMP

GENL_CMD_CAP_HASPOL

GENL_ID_CTRL

GENL_MAX_ID

GENL_MIN_ID

GENL_NAMSIZ

GETALL

GETNCNT

GETPID

GETVAL

GETZCNT

GRND_INSECURE

GRND_NONBLOCK

GRND_RANDOM

HWTSTAMP_FILTER_ALL

HWTSTAMP_FILTER_NONE

HWTSTAMP_FILTER_NTP_ALL

HWTSTAMP_FILTER_PTP_V1_L4_DELAY_REQ

HWTSTAMP_FILTER_PTP_V1_L4_EVENT

HWTSTAMP_FILTER_PTP_V1_L4_SYNC

HWTSTAMP_FILTER_PTP_V2_DELAY_REQ

HWTSTAMP_FILTER_PTP_V2_EVENT

HWTSTAMP_FILTER_PTP_V2_L2_DELAY_REQ

HWTSTAMP_FILTER_PTP_V2_L2_EVENT

HWTSTAMP_FILTER_PTP_V2_L2_SYNC

HWTSTAMP_FILTER_PTP_V2_L4_DELAY_REQ

HWTSTAMP_FILTER_PTP_V2_L4_EVENT

HWTSTAMP_FILTER_PTP_V2_L4_SYNC

HWTSTAMP_FILTER_PTP_V2_SYNC

HWTSTAMP_FILTER_SOME

HWTSTAMP_TX_OFF

HWTSTAMP_TX_ON

HWTSTAMP_TX_ONESTEP_P2P

HWTSTAMP_TX_ONESTEP_SYNC

IFA_ADDRESS

IFA_ANYCAST

IFA_BROADCAST

IFA_CACHEINFO

IFA_FLAGS

IFA_F_DADFAILED

IFA_F_DEPRECATED

IFA_F_HOMEADDRESS

IFA_F_MANAGETEMPADDR

IFA_F_MCAUTOJOIN

IFA_F_NODAD

IFA_F_NOPREFIXROUTE

IFA_F_OPTIMISTIC

IFA_F_PERMANENT

IFA_F_SECONDARY

IFA_F_STABLE_PRIVACY

IFA_F_TEMPORARY

IFA_F_TENTATIVE

IFA_LABEL

IFA_LOCAL

IFA_MULTICAST

IFA_UNSPEC

IFF_DORMANT

IFF_ECHO

IFF_LOWER_UP

IFLA_ADDRESS

IFLA_AF_SPEC

IFLA_ALLMULTI

IFLA_ALT_IFNAME

IFLA_BROADCAST

IFLA_CARRIER

IFLA_CARRIER_CHANGES

IFLA_CARRIER_DOWN_COUNT

IFLA_CARRIER_UP_COUNT

IFLA_COST

IFLA_EVENT

IFLA_EXT_MASK

IFLA_GROUP

IFLA_GRO_MAX_SIZE

IFLA_GSO_MAX_SEGS

IFLA_GSO_MAX_SIZE

IFLA_IFALIAS

IFLA_IFNAME

IFLA_IF_NETNSID

IFLA_INFO_DATA

IFLA_INFO_KIND

IFLA_INFO_SLAVE_DATA

IFLA_INFO_SLAVE_KIND

IFLA_INFO_UNSPEC

IFLA_INFO_XSTATS

IFLA_LINK

IFLA_LINKINFO

IFLA_LINKMODE

IFLA_LINK_NETNSID

IFLA_MAP

IFLA_MASTER

IFLA_MAX_MTU

IFLA_MIN_MTU

IFLA_MTU

IFLA_NET_NS_FD

IFLA_NET_NS_PID

IFLA_NEW_IFINDEX

IFLA_NEW_NETNSID

IFLA_NUM_RX_QUEUES

IFLA_NUM_TX_QUEUES

IFLA_NUM_VF

IFLA_OPERSTATE

IFLA_PAD

IFLA_PARENT_DEV_BUS_NAME

IFLA_PARENT_DEV_NAME

IFLA_PERM_ADDRESS

IFLA_PHYS_PORT_ID

IFLA_PHYS_PORT_NAME

IFLA_PHYS_SWITCH_ID

IFLA_PORT_SELF

IFLA_PRIORITY

IFLA_PROMISCUITY

IFLA_PROP_LIST

IFLA_PROTINFO

IFLA_PROTO_DOWN

IFLA_PROTO_DOWN_REASON

IFLA_QDISC

IFLA_STATS

IFLA_STATS64

IFLA_TARGET_NETNSID

IFLA_TSO_MAX_SEGS

IFLA_TSO_MAX_SIZE

IFLA_TXQLEN

IFLA_UNSPEC

IFLA_VFINFO_LIST

IFLA_VF_PORTS

IFLA_WEIGHT

IFLA_WIRELESS

IFLA_XDP

IF_LINK_MODE_DEFAULT

IF_LINK_MODE_DORMANT

IF_LINK_MODE_TESTING

IF_OPER_DORMANT

IF_OPER_DOWN

IF_OPER_LOWERLAYERDOWN

IF_OPER_NOTPRESENT

IF_OPER_TESTING

IF_OPER_UNKNOWN

IF_OPER_UP

INOTIFY_MAX_QUEUED_EVENTS

INOTIFY_MAX_USER_INSTANCES

INOTIFY_MAX_USER_WATCHES

INPUT_PROP_ACCELEROMETER

INPUT_PROP_BUTTONPAD

INPUT_PROP_CNT

INPUT_PROP_DIRECT

INPUT_PROP_MAX

INPUT_PROP_POINTER

INPUT_PROP_POINTING_STICK

INPUT_PROP_SEMI_MT

INPUT_PROP_TOPBUTTONPAD

IN_ACCESS

IN_ALL_EVENTS

IN_ATTRIB

IN_CLOEXEC

IN_CLOSE

IN_CLOSE_NOWRITE

IN_CLOSE_WRITE

IN_CREATE

IN_DELETE

IN_DELETE_SELF

IN_DONT_FOLLOW

IN_EXCL_UNLINK

IN_IGNORED

IN_ISDIR

IN_MASK_ADD

IN_MASK_CREATE

IN_MODIFY

IN_MOVE

IN_MOVED_FROM

IN_MOVED_TO

IN_MOVE_SELF

IN_NONBLOCK

IN_ONESHOT

IN_ONLYDIR

IN_OPEN

IN_Q_OVERFLOW

IN_UNMOUNT

IP6T_SO_ORIGINAL_DST

IPV6_FLOWINFO

IPV6_FLOWINFO_FLOWLABEL

IPV6_FLOWINFO_PRIORITY

IPV6_FLOWINFO_SEND

IPV6_FLOWLABEL_MGR

IPV6_FREEBIND

IPV6_RECVFRAGSIZE

IP_RECVFRAGSIZE

IWEVASSOCREQIE

IWEVASSOCRESPIE

IWEVCUSTOM

IWEVEXPIRED

IWEVFIRST

IWEVGENIE

IWEVMICHAELMICFAILURE

IWEVPMKIDCAND

IWEVQUAL

IWEVREGISTERED

IWEVTXDROP

IW_AUTH_80211_AUTH_ALG

IW_AUTH_ALG_LEAP

IW_AUTH_ALG_OPEN_SYSTEM

IW_AUTH_ALG_SHARED_KEY

IW_AUTH_CIPHER_AES_CMAC

IW_AUTH_CIPHER_CCMP

IW_AUTH_CIPHER_GROUP

IW_AUTH_CIPHER_GROUP_MGMT

IW_AUTH_CIPHER_NONE

IW_AUTH_CIPHER_PAIRWISE

IW_AUTH_CIPHER_TKIP

IW_AUTH_CIPHER_WEP40

IW_AUTH_CIPHER_WEP104

IW_AUTH_DROP_UNENCRYPTED

IW_AUTH_FLAGS

IW_AUTH_INDEX

IW_AUTH_KEY_MGMT

IW_AUTH_KEY_MGMT_802_1X

IW_AUTH_KEY_MGMT_PSK

IW_AUTH_MFP

IW_AUTH_MFP_DISABLED

IW_AUTH_MFP_OPTIONAL

IW_AUTH_MFP_REQUIRED

IW_AUTH_PRIVACY_INVOKED

IW_AUTH_ROAMING_CONTROL

IW_AUTH_ROAMING_DISABLE

IW_AUTH_ROAMING_ENABLE

IW_AUTH_RX_UNENCRYPTED_EAPOL

IW_AUTH_TKIP_COUNTERMEASURES

IW_AUTH_WPA_ENABLED

IW_AUTH_WPA_VERSION

IW_AUTH_WPA_VERSION_DISABLED

IW_AUTH_WPA_VERSION_WPA

IW_AUTH_WPA_VERSION_WPA2

IW_CUSTOM_MAX

IW_ENCODE_ALG_AES_CMAC

IW_ENCODE_ALG_CCMP

IW_ENCODE_ALG_NONE

IW_ENCODE_ALG_PMK

IW_ENCODE_ALG_TKIP

IW_ENCODE_ALG_WEP

IW_ENCODE_DISABLED

IW_ENCODE_ENABLED

IW_ENCODE_EXT_GROUP_KEY

IW_ENCODE_EXT_RX_SEQ_VALID

IW_ENCODE_EXT_SET_TX_KEY

IW_ENCODE_EXT_TX_SEQ_VALID

IW_ENCODE_FLAGS

IW_ENCODE_INDEX

IW_ENCODE_MODE

IW_ENCODE_NOKEY

IW_ENCODE_OPEN

IW_ENCODE_RESTRICTED

IW_ENCODE_SEQ_MAX_SIZE

IW_ENCODE_TEMP

IW_ENCODING_TOKEN_MAX

IW_ENC_CAPA_4WAY_HANDSHAKE

IW_ENC_CAPA_CIPHER_CCMP

IW_ENC_CAPA_CIPHER_TKIP

IW_ENC_CAPA_WPA

IW_ENC_CAPA_WPA2

IW_ESSID_MAX_SIZE

IW_EVENT_CAPA_K_0

IW_EVENT_CAPA_K_1

IW_EV_ADDR_PK_LEN

IW_EV_CHAR_PK_LEN

IW_EV_FREQ_PK_LEN

IW_EV_LCP_PK_LEN

IW_EV_PARAM_PK_LEN

IW_EV_POINT_PK_LEN

IW_EV_QUAL_PK_LEN

IW_EV_UINT_PK_LEN

IW_FREQ_AUTO

IW_FREQ_FIXED

IW_GENERIC_IE_MAX

IW_MAX_AP

IW_MAX_BITRATES

IW_MAX_ENCODING_SIZES

IW_MAX_FREQUENCIES

IW_MAX_SPY

IW_MAX_TXPOWER

IW_MICFAILURE_COUNT

IW_MICFAILURE_GROUP

IW_MICFAILURE_KEY_ID

IW_MICFAILURE_PAIRWISE

IW_MICFAILURE_STAKEY

IW_MLME_ASSOC

IW_MLME_AUTH

IW_MLME_DEAUTH

IW_MLME_DISASSOC

IW_MODE_ADHOC

IW_MODE_AUTO

IW_MODE_INFRA

IW_MODE_MASTER

IW_MODE_MESH

IW_MODE_MONITOR

IW_MODE_REPEAT

IW_MODE_SECOND

IW_PMKID_CAND_PREAUTH

IW_PMKID_LEN

IW_PMKSA_ADD

IW_PMKSA_FLUSH

IW_PMKSA_REMOVE

IW_POWER_ALL_R

IW_POWER_FORCE_S

IW_POWER_MAX

IW_POWER_MIN

IW_POWER_MODE

IW_POWER_MODIFIER

IW_POWER_MULTICAST_R

IW_POWER_ON

IW_POWER_PERIOD

IW_POWER_RELATIVE

IW_POWER_REPEATER

IW_POWER_TIMEOUT

IW_POWER_TYPE

IW_POWER_UNICAST_R

IW_PRIV_SIZE_FIXED

IW_PRIV_SIZE_MASK

IW_PRIV_TYPE_ADDR

IW_PRIV_TYPE_BYTE

IW_PRIV_TYPE_CHAR

IW_PRIV_TYPE_FLOAT

IW_PRIV_TYPE_INT

IW_PRIV_TYPE_MASK

IW_PRIV_TYPE_NONE

IW_QUAL_ALL_INVALID

IW_QUAL_ALL_UPDATED

IW_QUAL_DBM

IW_QUAL_LEVEL_INVALID

IW_QUAL_LEVEL_UPDATED

IW_QUAL_NOISE_INVALID

IW_QUAL_NOISE_UPDATED

IW_QUAL_QUAL_INVALID

IW_QUAL_QUAL_UPDATED

IW_QUAL_RCPI

IW_RETRY_LIFETIME

IW_RETRY_LIMIT

IW_RETRY_LONG

IW_RETRY_MAX

IW_RETRY_MIN

IW_RETRY_MODIFIER

IW_RETRY_ON

IW_RETRY_RELATIVE

IW_RETRY_SHORT

IW_RETRY_TYPE

IW_SCAN_ALL_ESSID

IW_SCAN_ALL_FREQ

IW_SCAN_ALL_MODE

IW_SCAN_ALL_RATE

IW_SCAN_CAPA_BSSID

IW_SCAN_CAPA_CHANNEL

IW_SCAN_CAPA_ESSID

IW_SCAN_CAPA_MODE

IW_SCAN_CAPA_NONE

IW_SCAN_CAPA_RATE

IW_SCAN_CAPA_TIME

IW_SCAN_CAPA_TYPE

IW_SCAN_DEFAULT

IW_SCAN_MAX_DATA

IW_SCAN_THIS_ESSID

IW_SCAN_THIS_FREQ

IW_SCAN_THIS_MODE

IW_SCAN_THIS_RATE

IW_SCAN_TYPE_ACTIVE

IW_SCAN_TYPE_PASSIVE

IW_TXPOW_DBM

IW_TXPOW_MWATT

IW_TXPOW_RANGE

IW_TXPOW_RELATIVE

IW_TXPOW_TYPE

KERN_ACCT

KERN_ACPI_VIDEO_FLAGS

KERN_BOOTLOADER_TYPE

KERN_CADPID

KERN_COMPAT_LOG

KERN_CORE_PATTERN

KERN_CORE_USES_PID

KERN_CTLALTDEL

KERN_DOMAINNAME

KERN_HOTPLUG

KERN_HPPA_PWRSW

KERN_HPPA_UNALIGNED

KERN_HZ_TIMER

KERN_IA64_UNALIGNED

KERN_IEEE_EMULATION_WARNINGS

KERN_MAX_LOCK_DEPTH

KERN_MAX_THREADS

KERN_MODPROBE

KERN_MSGMAX

KERN_MSGMNB

KERN_MSGMNI

KERN_MSGPOOL

KERN_NAMETRANS

KERN_NGROUPS_MAX

KERN_NMI_WATCHDOG

KERN_NODENAME

KERN_OSRELEASE

KERN_OSREV

KERN_OSTYPE

KERN_OVERFLOWGID

KERN_OVERFLOWUID

KERN_PANIC

KERN_PANIC_ON_NMI

KERN_PANIC_ON_OOPS

KERN_PIDMAX

KERN_PPC_HTABRECLAIM

KERN_PPC_L2CR

KERN_PPC_POWERSAVE_NAP

KERN_PPC_ZEROPAGED

KERN_PRINTK

KERN_PRINTK_RATELIMIT

KERN_PRINTK_RATELIMIT_BURST

KERN_PROF

KERN_PTY

KERN_RANDOM

KERN_RANDOMIZE

KERN_REALROOTDEV

KERN_RTSIGMAX

KERN_RTSIGNR

KERN_S390_USER_DEBUG_LOGGING

KERN_SECUREMASK

KERN_SEM

KERN_SETUID_DUMPABLE

KERN_SG_BIG_BUFF

KERN_SHMALL

KERN_SHMMAX

KERN_SHMMNI

KERN_SHMPATH

KERN_SPARC_REBOOT

KERN_SPARC_SCONS_PWROFF

KERN_SPARC_STOP_A

KERN_SPIN_RETRY

KERN_SYSRQ

KERN_TAINTED

KERN_UNKNOWN_NMI_PANIC

KERN_VERSION

KEXEC_ARCH_MASK

KEXEC_FILE_NO_INITRAMFS

KEXEC_FILE_ON_CRASH

KEXEC_FILE_UNLOAD

KEXEC_ON_CRASH

KEXEC_PRESERVE_CONTEXT

KEY_CNT

KEY_MAX

LED_CNT

LED_MAX

LINUX_REBOOT_CMD_CAD_OFF

LINUX_REBOOT_CMD_CAD_ON

LINUX_REBOOT_CMD_HALT

LINUX_REBOOT_CMD_KEXEC

LINUX_REBOOT_CMD_POWER_OFF

LINUX_REBOOT_CMD_RESTART

LINUX_REBOOT_CMD_RESTART2

LINUX_REBOOT_CMD_SW_SUSPEND

LINUX_REBOOT_MAGIC1

LINUX_REBOOT_MAGIC2

LINUX_REBOOT_MAGIC2A

LINUX_REBOOT_MAGIC2B

LINUX_REBOOT_MAGIC2C

MAP_DROPPABLE

MAP_SHARED_VALIDATE

MNT_NS_INFO_SIZE_VER0

MODULE_INIT_IGNORE_MODVERSIONS

MODULE_INIT_IGNORE_VERMAGIC

MOUNT_ATTR_IDMAP

MOUNT_ATTR_NOATIME

MOUNT_ATTR_NODEV

MOUNT_ATTR_NODIRATIME

MOUNT_ATTR_NOEXEC

MOUNT_ATTR_NOSUID

MOUNT_ATTR_NOSYMFOLLOW

MOUNT_ATTR_RDONLY

MOUNT_ATTR_RELATIME

MOUNT_ATTR_SIZE_VER0

MOUNT_ATTR_STRICTATIME

MOUNT_ATTR__ATIME

MPOL_BIND

MPOL_DEFAULT

MPOL_F_NUMA_BALANCING

MPOL_F_RELATIVE_NODES

MPOL_F_STATIC_NODES

MPOL_INTERLEAVE

MPOL_LOCAL

MPOL_PREFERRED

MREMAP_DONTUNMAP

MREMAP_FIXED

MREMAP_MAYMOVE

MSC_CNT

MSC_MAX

MSG_EXCEPT

MSG_INFO

MSG_NOERROR

MSG_NOTIFICATION

MSG_STAT

MSG_ZEROCOPY

NDA_CACHEINFO

NDA_DST

NDA_IFINDEX

NDA_LLADDR

NDA_PORT

NDA_PROBES

NDA_UNSPEC

NDA_VLAN

NDA_VNI

NET_802

NET_ATALK

NET_AX25

NET_BRIDGE

NET_CORE

NET_DCCP

NET_DECNET

NET_ECONET

NET_ETHER

NET_IPV4

NET_IPV6

NET_IPX

NET_IRDA

NET_LLC

NET_NETFILTER

NET_NETROM

NET_ROSE

NET_SCTP

NET_TR

NET_UNIX

NET_X25

NFNETLINK_V0

NFNLGRP_ACCT_QUOTA

NFNLGRP_CONNTRACK_DESTROY

NFNLGRP_CONNTRACK_EXP_DESTROY

NFNLGRP_CONNTRACK_EXP_NEW

NFNLGRP_CONNTRACK_EXP_UPDATE

NFNLGRP_CONNTRACK_NEW

NFNLGRP_CONNTRACK_UPDATE

NFNLGRP_NFTABLES

NFNLGRP_NFTRACE

NFNLGRP_NONE

NFNL_BATCH_GENID

NFNL_BATCH_UNSPEC

NFNL_MSG_BATCH_BEGIN

NFNL_MSG_BATCH_END

NFNL_SUBSYS_ACCT

NFNL_SUBSYS_COUNT

NFNL_SUBSYS_CTHELPER

NFNL_SUBSYS_CTNETLINK

NFNL_SUBSYS_CTNETLINK_EXP

NFNL_SUBSYS_CTNETLINK_TIMEOUT

NFNL_SUBSYS_HOOK

NFNL_SUBSYS_IPSET

NFNL_SUBSYS_NFTABLES

NFNL_SUBSYS_NFT_COMPAT

NFNL_SUBSYS_NONE

NFNL_SUBSYS_OSF

NFNL_SUBSYS_QUEUE

NFNL_SUBSYS_ULOG

NFPROTO_ARP

NFPROTO_BRIDGE

NFPROTO_DECNET

NFPROTO_INET

NFPROTO_IPV4

NFPROTO_IPV6

NFPROTO_NETDEV

NFPROTO_NUMPROTO

NFPROTO_UNSPEC

NFQA_CAP_LEN

NFQA_CFG_CMD

NFQA_CFG_FLAGS

NFQA_CFG_F_CONNTRACK

NFQA_CFG_F_FAIL_OPEN

NFQA_CFG_F_GSO

NFQA_CFG_F_MAX

NFQA_CFG_F_SECCTX

NFQA_CFG_F_UID_GID

NFQA_CFG_MASK

NFQA_CFG_PARAMS

NFQA_CFG_QUEUE_MAXLEN

NFQA_CFG_UNSPEC

NFQA_CT

NFQA_CT_INFO

NFQA_EXP

NFQA_GID

NFQA_HWADDR

NFQA_IFINDEX_INDEV

NFQA_IFINDEX_OUTDEV

NFQA_IFINDEX_PHYSINDEV

NFQA_IFINDEX_PHYSOUTDEV

NFQA_L2HDR

NFQA_MARK

NFQA_PACKET_HDR

NFQA_PAYLOAD

NFQA_PRIORITY

NFQA_SECCTX

NFQA_SKB_CSUMNOTREADY

NFQA_SKB_CSUM_NOTVERIFIED

NFQA_SKB_GSO

NFQA_SKB_INFO

NFQA_TIMESTAMP

NFQA_UID

NFQA_UNSPEC

NFQA_VERDICT_HDR

NFQA_VLAN

NFQA_VLAN_PROTO

NFQA_VLAN_TCI

NFQA_VLAN_UNSPEC

NFQNL_CFG_CMD_BIND

NFQNL_CFG_CMD_NONE

NFQNL_CFG_CMD_PF_BIND

NFQNL_CFG_CMD_PF_UNBIND

NFQNL_CFG_CMD_UNBIND

NFQNL_COPY_META

NFQNL_COPY_NONE

NFQNL_COPY_PACKET

NFQNL_MSG_CONFIG

NFQNL_MSG_PACKET

NFQNL_MSG_VERDICT

NFQNL_MSG_VERDICT_BATCH

NFT_BREAK

NFT_BYTEORDER_HTON

NFT_BYTEORDER_NTOH

NFT_CHAIN_MAXNAMELEN

NFT_CMP_EQ

NFT_CMP_GT

NFT_CMP_GTE

NFT_CMP_LT

NFT_CMP_LTE

NFT_CMP_NEQ

NFT_CONTINUE

NFT_CT_AVGPKT

NFT_CT_BYTES

NFT_CT_DIRECTION

NFT_CT_DST

NFT_CT_DST_IP

NFT_CT_DST_IP6

NFT_CT_EVENTMASK

NFT_CT_EXPIRATION

NFT_CT_HELPER

NFT_CT_L3PROTOCOL

NFT_CT_LABELS

NFT_CT_MARK

NFT_CT_PKTS

NFT_CT_PROTOCOL

NFT_CT_PROTO_DST

NFT_CT_PROTO_SRC

NFT_CT_SECMARK

NFT_CT_SRC

NFT_CT_SRC_IP

NFT_CT_SRC_IP6

NFT_CT_STATE

NFT_CT_STATUS

NFT_CT_ZONE

NFT_DATA_RESERVED_MASK

NFT_DATA_VALUE

NFT_DATA_VALUE_MAXLEN

NFT_DATA_VERDICT

NFT_DYNSET_F_INV

NFT_DYNSET_OP_ADD

NFT_DYNSET_OP_UPDATE

NFT_GOTO

NFT_JUMP

NFT_LIMIT_F_INV

NFT_LIMIT_PKTS

NFT_LIMIT_PKT_BYTES

NFT_LOOKUP_F_INV

NFT_META_BRI_IIFNAME

NFT_META_BRI_OIFNAME

NFT_META_CGROUP

NFT_META_CPU

NFT_META_IIF

NFT_META_IIFGROUP

NFT_META_IIFNAME

NFT_META_IIFTYPE

NFT_META_L4PROTO

NFT_META_LEN

NFT_META_MARK

NFT_META_NFPROTO

NFT_META_NFTRACE

NFT_META_OIF

NFT_META_OIFGROUP

NFT_META_OIFNAME

NFT_META_OIFTYPE

NFT_META_PKTTYPE

NFT_META_PRANDOM

NFT_META_PRIORITY

NFT_META_PROTOCOL

NFT_META_RTCLASSID

NFT_META_SECMARK

NFT_META_SKGID

NFT_META_SKUID

NFT_MSG_DELCHAIN

NFT_MSG_DELOBJ

NFT_MSG_DELRULE

NFT_MSG_DELSET

NFT_MSG_DELSETELEM

NFT_MSG_DELTABLE

NFT_MSG_GETCHAIN

NFT_MSG_GETGEN

NFT_MSG_GETOBJ

NFT_MSG_GETOBJ_RESET

NFT_MSG_GETRULE

NFT_MSG_GETSET

NFT_MSG_GETSETELEM

NFT_MSG_GETTABLE

NFT_MSG_MAX

NFT_MSG_NEWCHAIN

NFT_MSG_NEWGEN

NFT_MSG_NEWOBJ

NFT_MSG_NEWRULE

NFT_MSG_NEWSET

NFT_MSG_NEWSETELEM

NFT_MSG_NEWTABLE

NFT_MSG_TRACE

NFT_NAT_DNAT

NFT_NAT_SNAT

NFT_NG_INCREMENTAL

NFT_NG_RANDOM

NFT_OBJ_MAXNAMELEN

NFT_PAYLOAD_CSUM_INET

NFT_PAYLOAD_CSUM_NONE

NFT_PAYLOAD_LL_HEADER

NFT_PAYLOAD_NETWORK_HEADER

NFT_PAYLOAD_TRANSPORT_HEADER

NFT_QUEUE_FLAG_BYPASS

NFT_QUEUE_FLAG_CPU_FANOUT

NFT_QUEUE_FLAG_MASK

NFT_QUOTA_F_INV

NFT_RANGE_EQ

NFT_RANGE_NEQ

NFT_REG32_00

NFT_REG32_01

NFT_REG32_02

NFT_REG32_03

NFT_REG32_04

NFT_REG32_05

NFT_REG32_06

NFT_REG32_07

NFT_REG32_08

NFT_REG32_09

NFT_REG32_10

NFT_REG32_11

NFT_REG32_12

NFT_REG32_13

NFT_REG32_14

NFT_REG32_15

NFT_REG32_SIZE

NFT_REG_1

NFT_REG_2

NFT_REG_3

NFT_REG_4

NFT_REG_SIZE

NFT_REG_VERDICT

NFT_REJECT_ICMPX_ADMIN_PROHIBITED

NFT_REJECT_ICMPX_HOST_UNREACH

NFT_REJECT_ICMPX_NO_ROUTE

NFT_REJECT_ICMPX_PORT_UNREACH

NFT_REJECT_ICMPX_UNREACH

NFT_REJECT_ICMP_UNREACH

NFT_REJECT_TCP_RST

NFT_RETURN

NFT_SET_ANONYMOUS

NFT_SET_CONSTANT

NFT_SET_ELEM_INTERVAL_END

NFT_SET_EVAL

NFT_SET_INTERVAL

NFT_SET_MAP

NFT_SET_MAXNAMELEN

NFT_SET_POL_MEMORY

NFT_SET_POL_PERFORMANCE

NFT_SET_TIMEOUT

NFT_TABLE_MAXNAMELEN

NFT_TRACETYPE_POLICY

NFT_TRACETYPE_RETURN

NFT_TRACETYPE_RULE

NFT_TRACETYPE_UNSPEC

NFT_USERDATA_MAXLEN

NFULA_CFG_CMD

NFULA_CFG_FLAGS

NFULA_CFG_MODE

NFULA_CFG_NLBUFSIZ

NFULA_CFG_QTHRESH

NFULA_CFG_TIMEOUT

NFULA_CFG_UNSPEC

NFULA_CT

NFULA_CT_INFO

NFULA_GID

NFULA_HWADDR

NFULA_HWHEADER

NFULA_HWLEN

NFULA_HWTYPE

NFULA_IFINDEX_INDEV

NFULA_IFINDEX_OUTDEV

NFULA_IFINDEX_PHYSINDEV

NFULA_IFINDEX_PHYSOUTDEV

NFULA_L2HDR

NFULA_MARK

NFULA_PACKET_HDR

NFULA_PAYLOAD

NFULA_PREFIX

NFULA_SEQ

NFULA_SEQ_GLOBAL

NFULA_TIMESTAMP

NFULA_UID

NFULA_UNSPEC

NFULA_VLAN

NFULA_VLAN_PROTO

NFULA_VLAN_TCI

NFULA_VLAN_UNSPEC

NFULNL_CFG_CMD_BIND

NFULNL_CFG_CMD_NONE

NFULNL_CFG_CMD_PF_BIND

NFULNL_CFG_CMD_PF_UNBIND

NFULNL_CFG_CMD_UNBIND

NFULNL_CFG_F_CONNTRACK

NFULNL_CFG_F_SEQ

NFULNL_CFG_F_SEQ_GLOBAL

NFULNL_COPY_META

NFULNL_COPY_NONE

NFULNL_COPY_PACKET

NFULNL_MSG_CONFIG

NFULNL_MSG_PACKET

NF_ACCEPT

NF_ARP

NF_ARP_FORWARD

NF_ARP_IN

NF_ARP_NUMHOOKS

NF_ARP_OUT

NF_BR_BROUTING

NF_BR_FORWARD

NF_BR_LOCAL_IN

NF_BR_LOCAL_OUT

NF_BR_NUMHOOKS

NF_BR_POST_ROUTING

NF_BR_PRE_ROUTING

NF_BR_PRI_BRNF

NF_BR_PRI_FILTER_BRIDGED

NF_BR_PRI_FILTER_OTHER

NF_BR_PRI_FIRST

NF_BR_PRI_LAST

NF_BR_PRI_NAT_DST_BRIDGED

NF_BR_PRI_NAT_DST_OTHER

NF_BR_PRI_NAT_SRC

NF_DROP

NF_INET_FORWARD

NF_INET_INGRESS

NF_INET_LOCAL_IN

NF_INET_LOCAL_OUT

NF_INET_NUMHOOKS

NF_INET_POST_ROUTING

NF_INET_PRE_ROUTING

NF_IP6_FORWARD

NF_IP6_LOCAL_IN

NF_IP6_LOCAL_OUT

NF_IP6_NUMHOOKS

NF_IP6_POST_ROUTING

NF_IP6_PRE_ROUTING

NF_IP6_PRI_CONNTRACK

NF_IP6_PRI_CONNTRACK_DEFRAG

NF_IP6_PRI_CONNTRACK_HELPER

NF_IP6_PRI_FILTER

NF_IP6_PRI_FIRST

NF_IP6_PRI_LAST

NF_IP6_PRI_MANGLE

NF_IP6_PRI_NAT_DST

NF_IP6_PRI_NAT_SRC

NF_IP6_PRI_RAW

NF_IP6_PRI_RAW_BEFORE_DEFRAG

NF_IP6_PRI_SECURITY

NF_IP6_PRI_SELINUX_FIRST

NF_IP6_PRI_SELINUX_LAST

NF_IP_FORWARD

NF_IP_LOCAL_IN

NF_IP_LOCAL_OUT

NF_IP_NUMHOOKS

NF_IP_POST_ROUTING

NF_IP_PRE_ROUTING

NF_IP_PRI_CONNTRACK

NF_IP_PRI_CONNTRACK_CONFIRM

NF_IP_PRI_CONNTRACK_DEFRAG

NF_IP_PRI_CONNTRACK_HELPER

NF_IP_PRI_FILTER

NF_IP_PRI_FIRST

NF_IP_PRI_LAST

NF_IP_PRI_MANGLE

NF_IP_PRI_NAT_DST

NF_IP_PRI_NAT_SRC

NF_IP_PRI_RAW

NF_IP_PRI_RAW_BEFORE_DEFRAG

NF_IP_PRI_SECURITY

NF_IP_PRI_SELINUX_FIRST

NF_IP_PRI_SELINUX_LAST

NF_MAX_VERDICT

NF_NETDEV_EGRESS

NF_NETDEV_INGRESS

NF_NETDEV_NUMHOOKS

NF_QUEUE

NF_REPEAT

NF_STOLEN

NF_STOP

NF_VERDICT_BITS

NF_VERDICT_FLAG_QUEUE_BYPASS

NF_VERDICT_MASK

NF_VERDICT_QBITS

NF_VERDICT_QMASK

NLM_F_BULK

NSIO 🔒

NS_GET_MNTNS_ID

NS_GET_NSTYPE

NS_GET_OWNER_UID

NS_GET_PARENT

NS_GET_PID_FROM_PIDNS

NS_GET_PID_IN_PIDNS

NS_GET_TGID_FROM_PIDNS

NS_GET_TGID_IN_PIDNS

NS_GET_USERNS

NS_MNT_GET_INFO

NS_MNT_GET_NEXT

NS_MNT_GET_PREV

NTF_MASTER

NTF_PROXY

NTF_ROUTER

NTF_SELF

NTF_USE

NUD_DELAY

NUD_FAILED

NUD_INCOMPLETE

NUD_NOARP

NUD_NONE

NUD_PERMANENT

NUD_PROBE

NUD_REACHABLE

NUD_STALE

OPEN_TREE_CLOEXEC

OPEN_TREE_CLONE

PACKET_FANOUT

PACKET_FANOUT_CBPF

PACKET_FANOUT_CPU

PACKET_FANOUT_DATA

PACKET_FANOUT_EBPF

PACKET_FANOUT_FLAG_DEFRAG

PACKET_FANOUT_FLAG_IGNORE_OUTGOING

PACKET_FANOUT_FLAG_ROLLOVER

PACKET_FANOUT_FLAG_UNIQUEID

PACKET_FANOUT_HASH

PACKET_FANOUT_LB

PACKET_FANOUT_QM

PACKET_FANOUT_RND

PACKET_FANOUT_ROLLOVER

PACKET_IGNORE_OUTGOING

PACKET_QDISC_BYPASS

PACKET_ROLLOVER_STATS

PACKET_TX_HAS_OFF

PACKET_VNET_HDR_SZ

PF_BLOCK_TS

Plug has ts that needs updating.

PF_DUMPCORE

Dumped core.

PF_EXITING

Getting shut down.

PF_FORKNOEXEC

Forked but didn’t exec.

PF_IDLE

I am an IDLE thread.

PF_IO_WORKER

Task is an IO worker.

PF_KSWAPD

I am kswapd.

PF_KTHREAD

I am a kernel thread.

PF_LOCAL_THROTTLE

Throttle writes only against the bdi I write to, I am cleaning dirty pages from some other bdi.

PF_MCE_EARLY

Early kill for mce process policy.

PF_MCE_PROCESS

Process policy on mce errors.

PF_MEMALLOC

Allocating memory to free memory.

PF_MEMALLOC_NOFS

All allocations inherit GFP_NOFS.

PF_MEMALLOC_NOIO

All allocations inherit GFP_NOIO.

PF_MEMALLOC_PIN

Allocations constrained to zones which allow long term pinning.

PF_NOFREEZE

This thread should not be frozen.

PF_NO_SETAFFINITY

Userland is not allowed to meddle with cpus_mask.

PF_NPROC_EXCEEDED

set_user() noticed that RLIMIT_NPROC was exceeded.

PF_POSTCOREDUMP

Coredumps should ignore this task.

PF_RANDOMIZE

Randomize virtual address space.

PF_SIGNALED

Killed by a signal.

PF_SUPERPRIV

Used super-user privileges.

PF_SUSPEND_TASK

This thread called freeze_processes() and should not be frozen.

PF_SUSPEND_TASK_UINT 🔒

PF_USED_MATH

If unset the fpu must be initialized before use.

PF_USER_WORKER

Kernel thread cloned from userspace thread.

PF_VCPU

I’m a virtual CPU.

PF_WQ_WORKER

I’m a workqueue worker.

PIDFD_GET_CGROUP_NAMESPACE

PIDFD_GET_INFO

PIDFD_GET_IPC_NAMESPACE

PIDFD_GET_MNT_NAMESPACE

PIDFD_GET_NET_NAMESPACE

PIDFD_GET_PID_FOR_CHILDREN_NAMESPACE

PIDFD_GET_PID_NAMESPACE

PIDFD_GET_TIME_FOR_CHILDREN_NAMESPACE

PIDFD_GET_TIME_NAMESPACE

PIDFD_GET_USER_NAMESPACE

PIDFD_GET_UTS_NAMESPACE

PIDFD_INFO_CGROUPID

PIDFD_INFO_CREDS

PIDFD_INFO_EXIT

PIDFD_INFO_PID

PIDFD_INFO_SIZE_VER0

PIDFD_NONBLOCK

PIDFD_SIGNAL_PROCESS_GROUP

PIDFD_SIGNAL_THREAD

PIDFD_SIGNAL_THREAD_GROUP

PIDFD_THREAD

PIDFS_IOCTL_MAGIC 🔒

PIDTYPE_MAX

PIDTYPE_PGID

PIDTYPE_PID

PIDTYPE_SID

PIDTYPE_TGID

POSIX_SPAWN_RESETIDS

POSIX_SPAWN_SETPGROUP

POSIX_SPAWN_SETSCHEDPARAM

POSIX_SPAWN_SETSCHEDULER

POSIX_SPAWN_SETSID

POSIX_SPAWN_SETSIGDEF

POSIX_SPAWN_SETSIGMASK

POSIX_SPAWN_USEVFORK

PROC_CN_MCAST_IGNORE

PROC_CN_MCAST_LISTEN

PROC_EVENT_COMM

PROC_EVENT_COREDUMP

PROC_EVENT_EXEC

PROC_EVENT_EXIT

PROC_EVENT_FORK

PROC_EVENT_GID

PROC_EVENT_NONE

PROC_EVENT_NONZERO_EXIT

PROC_EVENT_PTRACE

PROC_EVENT_SID

PROC_EVENT_UID

PR_GET_MDWE

PR_MDWE_NO_INHERIT

PR_MDWE_REFUSE_EXEC_GAIN

PR_SET_MDWE

PTHREAD_COND_INITIALIZER

PTHREAD_MUTEX_INITIALIZER

PTHREAD_RWLOCK_INITIALIZER

PTP_CLK_MAGIC 🔒

PTP_CLOCK_GETCAPS

PTP_CLOCK_GETCAPS2

PTP_ENABLE_PPS

PTP_ENABLE_PPS2

PTP_EXTTS_REQUEST

PTP_EXTTS_REQUEST2

PTP_MAX_SAMPLES

PTP_PEROUT_REQUEST

PTP_PEROUT_REQUEST2

PTP_PF_EXTTS

PTP_PF_NONE

PTP_PF_PEROUT

PTP_PF_PHYSYNC

PTP_PIN_GETFUNC

PTP_PIN_GETFUNC2

PTP_PIN_SETFUNC

PTP_PIN_SETFUNC2

PTP_SYS_OFFSET

PTP_SYS_OFFSET2

PTP_SYS_OFFSET_EXTENDED

PTP_SYS_OFFSET_EXTENDED2

PTP_SYS_OFFSET_PRECISE

PTP_SYS_OFFSET_PRECISE2

QFMT_VFS_OLD

QFMT_VFS_V0

QFMT_VFS_V1

RB_AUTOBOOT

RB_DISABLE_CAD

RB_ENABLE_CAD

RB_HALT_SYSTEM

RB_KEXEC

RB_POWER_OFF

RB_SW_SUSPEND

REL_CNT

REL_MAX

RENAME_EXCHANGE

RENAME_NOREPLACE

RENAME_WHITEOUT

REP_CNT

REP_MAX

RESOLVE_BENEATH

RESOLVE_CACHED

RESOLVE_IN_ROOT

RESOLVE_NO_MAGICLINKS

RESOLVE_NO_SYMLINKS

RESOLVE_NO_XDEV

RTA_CACHEINFO

RTA_DST

RTA_FLOW

RTA_GATEWAY

RTA_IIF

RTA_MARK

RTA_METRICS

RTA_MFC_STATS

RTA_MP_ALGO

RTA_MULTIPATH

RTA_OIF

RTA_PREFSRC

RTA_PRIORITY

RTA_PROTOINFO

RTA_SESSION

RTA_SRC

RTA_TABLE

RTA_UNSPEC

RTEXT_FILTER_BRVLAN

RTEXT_FILTER_BRVLAN_COMPRESSED

RTEXT_FILTER_CFM_CONFIG

RTEXT_FILTER_CFM_STATUS

RTEXT_FILTER_MRP

RTEXT_FILTER_SKIP_STATS

RTEXT_FILTER_VF

RTMGRP_DECnet_IFADDR

RTMGRP_DECnet_ROUTE

RTMGRP_IPV4_IFADDR

RTMGRP_IPV4_MROUTE

RTMGRP_IPV4_ROUTE

RTMGRP_IPV4_RULE

RTMGRP_IPV6_IFADDR

RTMGRP_IPV6_IFINFO

RTMGRP_IPV6_MROUTE

RTMGRP_IPV6_PREFIX

RTMGRP_IPV6_ROUTE

RTMGRP_LINK

RTMGRP_NEIGH

RTMGRP_NOTIFY

RTMGRP_TC

RTMSG_AR_FAILED

RTMSG_CONTROL

RTMSG_DELDEVICE

RTMSG_DELROUTE

RTMSG_DELRULE

RTMSG_NEWDEVICE

RTMSG_NEWROUTE

RTMSG_NEWRULE

RTMSG_OVERRUN

RTM_DELACTION

RTM_DELADDR

RTM_DELADDRLABEL

RTM_DELLINK

RTM_DELMDB

RTM_DELNEIGH

RTM_DELNSID

RTM_DELQDISC

RTM_DELROUTE

RTM_DELRULE

RTM_DELTCLASS

RTM_DELTFILTER

RTM_F_CLONED

RTM_F_EQUALIZE

RTM_F_NOTIFY

RTM_F_PREFIX

RTM_GETACTION

RTM_GETADDR

RTM_GETADDRLABEL

RTM_GETANYCAST

RTM_GETDCB

RTM_GETLINK

RTM_GETMDB

RTM_GETMULTICAST

RTM_GETNEIGH

RTM_GETNEIGHTBL

RTM_GETNETCONF

RTM_GETNSID

RTM_GETQDISC

RTM_GETROUTE

RTM_GETRULE

RTM_GETTCLASS

RTM_GETTFILTER

RTM_NEWACTION

RTM_NEWADDR

RTM_NEWADDRLABEL

RTM_NEWLINK

RTM_NEWMDB

RTM_NEWNDUSEROPT

RTM_NEWNEIGH

RTM_NEWNEIGHTBL

RTM_NEWNETCONF

RTM_NEWNSID

RTM_NEWPREFIX

RTM_NEWQDISC

RTM_NEWROUTE

RTM_NEWRULE

RTM_NEWTCLASS

RTM_NEWTFILTER

RTM_SETDCB

RTM_SETLINK

RTM_SETNEIGHTBL

RTNLGRP_BRVLAN

RTNLGRP_DCB

RTNLGRP_DECnet_IFADDR

RTNLGRP_DECnet_ROUTE

RTNLGRP_DECnet_RULE

RTNLGRP_IPV4_IFADDR

RTNLGRP_IPV4_MROUTE

RTNLGRP_IPV4_MROUTE_R

RTNLGRP_IPV4_NETCONF

RTNLGRP_IPV4_ROUTE

RTNLGRP_IPV4_RULE

RTNLGRP_IPV6_IFADDR

RTNLGRP_IPV6_IFINFO

RTNLGRP_IPV6_MROUTE

RTNLGRP_IPV6_MROUTE_R

RTNLGRP_IPV6_NETCONF

RTNLGRP_IPV6_PREFIX

RTNLGRP_IPV6_ROUTE

RTNLGRP_IPV6_RULE

RTNLGRP_LINK

RTNLGRP_MCTP_IFADDR

RTNLGRP_MDB

RTNLGRP_MPLS_NETCONF

RTNLGRP_MPLS_ROUTE

RTNLGRP_ND_USEROPT

RTNLGRP_NEIGH

RTNLGRP_NEXTHOP

RTNLGRP_NONE

RTNLGRP_NOP2

RTNLGRP_NOP4

RTNLGRP_NOTIFY

RTNLGRP_NSID

RTNLGRP_PHONET_IFADDR

RTNLGRP_PHONET_ROUTE

RTNLGRP_STATS

RTNLGRP_TC

RTNLGRP_TUNNEL

RTN_ANYCAST

RTN_BLACKHOLE

RTN_BROADCAST

RTN_LOCAL

RTN_MULTICAST

RTN_NAT

RTN_PROHIBIT

RTN_THROW

RTN_UNICAST

RTN_UNREACHABLE

RTN_UNSPEC

RTN_XRESOLVE

RTPROT_BOOT

RTPROT_KERNEL

RTPROT_REDIRECT

RTPROT_STATIC

RTPROT_UNSPEC

RT_SCOPE_HOST

RT_SCOPE_LINK

RT_SCOPE_NOWHERE

RT_SCOPE_SITE

RT_SCOPE_UNIVERSE

RT_TABLE_COMPAT

RT_TABLE_DEFAULT

RT_TABLE_LOCAL

RT_TABLE_MAIN

RT_TABLE_UNSPEC

RWF_APPEND

RWF_ATOMIC

RWF_DONTCACHE

RWF_DSYNC

RWF_HIPRI

RWF_NOAPPEND

RWF_NOWAIT

RWF_SYNC

SCHED_FLAG_ALL

SCHED_FLAG_DL_OVERRUN

SCHED_FLAG_KEEP_ALL

SCHED_FLAG_KEEP_PARAMS

SCHED_FLAG_KEEP_POLICY

SCHED_FLAG_RECLAIM

SCHED_FLAG_RESET_ON_FORK

SCHED_FLAG_UTIL_CLAMP

SCHED_FLAG_UTIL_CLAMP_MAX

SCHED_FLAG_UTIL_CLAMP_MIN

SCTP_ABORT

SCTP_ADAPTATION_LAYER

SCTP_ADDR_OVER

SCTP_ALL_ASSOC

SCTP_ASSOCINFO

SCTP_AUTHINFO

SCTP_AUTH_ACTIVE_KEY

SCTP_AUTH_CHUNK

SCTP_AUTH_DEACTIVATE_KEY

SCTP_AUTH_DELETE_KEY

SCTP_AUTH_KEY

SCTP_AUTOCLOSE

SCTP_AUTO_ASCONF

SCTP_CONTEXT

SCTP_CURRENT_ASSOC

SCTP_DEFAULT_SEND_PARAM

SCTP_DEFAULT_SNDINFO

SCTP_DELAYED_ACK

SCTP_DELAYED_ACK_TIME

SCTP_DELAYED_SACK

SCTP_DISABLE_FRAGMENTS

SCTP_DSTADDRV4

SCTP_DSTADDRV6

SCTP_ENABLE_CHANGE_ASSOC_REQ

SCTP_ENABLE_RESET_ASSOC_REQ

SCTP_ENABLE_RESET_STREAM_REQ

SCTP_ENABLE_STRRESET_MASK

SCTP_EOF

SCTP_EVENTS

SCTP_FRAGMENT_INTERLEAVE

SCTP_FUTURE_ASSOC

SCTP_GET_ASSOC_ID_LIST

SCTP_GET_ASSOC_NUMBER

SCTP_GET_PEER_ADDR_INFO

SCTP_HMAC_IDENT

SCTP_INIT

SCTP_INITMSG

SCTP_I_WANT_MAPPED_V4_ADDR

SCTP_LOCAL_AUTH_CHUNKS

SCTP_MAXSEG

SCTP_MAX_BURST

SCTP_NODELAY

SCTP_NOTIFICATION

SCTP_NXTINFO

SCTP_PARTIAL_DELIVERY_POINT

SCTP_PEER_ADDR_PARAMS

SCTP_PEER_ADDR_THLDS

SCTP_PEER_ADDR_THLDS_V2

SCTP_PEER_AUTH_CHUNKS

SCTP_PRIMARY_ADDR

SCTP_PRINFO

SCTP_PR_SCTP_ALL

SCTP_PR_SCTP_MASK

SCTP_PR_SCTP_MAX

SCTP_PR_SCTP_NONE

SCTP_PR_SCTP_PRIO

SCTP_PR_SCTP_RTX

SCTP_PR_SCTP_TTL

SCTP_RCVINFO

SCTP_RECVNXTINFO

SCTP_RECVRCVINFO

SCTP_REUSE_PORT

SCTP_RTOINFO

SCTP_SACK_IMMEDIATELY

SCTP_SENDALL

SCTP_SET_PEER_PRIMARY_ADDR

SCTP_SNDINFO

SCTP_SNDRCV

SCTP_STATUS

SCTP_STREAM_RESET_INCOMING

SCTP_STREAM_RESET_OUTGOING

SCTP_UNORDERED

SECBIT_EXEC_DENY_INTERACTIVE

SECBIT_EXEC_DENY_INTERACTIVE_LOCKED

SECBIT_EXEC_RESTRICT_FILE

SECBIT_EXEC_RESTRICT_FILE_LOCKED

SECBIT_KEEP_CAPS

SECBIT_KEEP_CAPS_LOCKED

SECBIT_NOROOT

SECBIT_NOROOT_LOCKED

SECBIT_NO_CAP_AMBIENT_RAISE

SECBIT_NO_CAP_AMBIENT_RAISE_LOCKED

SECBIT_NO_SETUID_FIXUP

SECBIT_NO_SETUID_FIXUP_LOCKED

SECCOMP_ADDFD_FLAG_SEND

SECCOMP_ADDFD_FLAG_SETFD

SECCOMP_FILTER_FLAG_LOG

SECCOMP_FILTER_FLAG_NEW_LISTENER

SECCOMP_FILTER_FLAG_SPEC_ALLOW

SECCOMP_FILTER_FLAG_TSYNC

SECCOMP_FILTER_FLAG_TSYNC_ESRCH

SECCOMP_FILTER_FLAG_WAIT_KILLABLE_RECV

SECCOMP_GET_ACTION_AVAIL

SECCOMP_GET_NOTIF_SIZES

SECCOMP_MODE_DISABLED

SECCOMP_MODE_FILTER

SECCOMP_MODE_STRICT

SECCOMP_RET_ACTION

SECCOMP_RET_ACTION_FULL

SECCOMP_RET_ALLOW

SECCOMP_RET_DATA

SECCOMP_RET_ERRNO

SECCOMP_RET_KILL

SECCOMP_RET_KILL_PROCESS

SECCOMP_RET_KILL_THREAD

SECCOMP_RET_LOG

SECCOMP_RET_TRACE

SECCOMP_RET_TRAP

SECCOMP_RET_USER_NOTIF

SECCOMP_SET_MODE_FILTER

SECCOMP_SET_MODE_STRICT

SECCOMP_USER_NOTIF_FLAG_CONTINUE

SECUREBITS_DEFAULT

SECURE_ALL_BITS

SECURE_ALL_LOCKS

SECURE_ALL_UNPRIVILEGED

SECURE_EXEC_DENY_INTERACTIVE 🔒

SECURE_EXEC_DENY_INTERACTIVE_LOCKED 🔒

SECURE_EXEC_RESTRICT_FILE 🔒

SECURE_EXEC_RESTRICT_FILE_LOCKED 🔒

SECURE_KEEP_CAPS 🔒

SECURE_KEEP_CAPS_LOCKED 🔒

SECURE_NOROOT 🔒

SECURE_NOROOT_LOCKED 🔒

SECURE_NO_CAP_AMBIENT_RAISE 🔒

SECURE_NO_CAP_AMBIENT_RAISE_LOCKED 🔒

SECURE_NO_SETUID_FIXUP 🔒

SECURE_NO_SETUID_FIXUP_LOCKED 🔒

SEEK_DATA

SEEK_HOLE

SEM_INFO

SEM_STAT

SEM_STAT_ANY

SEM_UNDO

SETALL

SETVAL

SIOCGHWTSTAMP

SIOCGIWAP

SIOCGIWAPLIST

SIOCGIWAUTH

SIOCGIWENCODE

SIOCGIWENCODEEXT

SIOCGIWESSID

SIOCGIWFRAG

SIOCGIWFREQ

SIOCGIWGENIE

SIOCGIWMODE

SIOCGIWNAME

SIOCGIWNICKN

SIOCGIWNWID

SIOCGIWPOWER

SIOCGIWPRIV

SIOCGIWRANGE

SIOCGIWRATE

SIOCGIWRETRY

SIOCGIWRTS

SIOCGIWSCAN

SIOCGIWSENS

SIOCGIWSPY

SIOCGIWSTATS

SIOCGIWTHRSPY

SIOCGIWTXPOW

SIOCIWFIRST

SIOCIWFIRSTPRIV

SIOCIWLAST

SIOCIWLASTPRIV

SIOCSHWTSTAMP

SIOCSIWAP

SIOCSIWAUTH

SIOCSIWCOMMIT

SIOCSIWENCODE

SIOCSIWENCODEEXT

SIOCSIWESSID

SIOCSIWFRAG

SIOCSIWFREQ

SIOCSIWGENIE

SIOCSIWMLME

SIOCSIWMODE

SIOCSIWNICKN

SIOCSIWNWID

SIOCSIWPMKSA

SIOCSIWPOWER

SIOCSIWPRIV

SIOCSIWRANGE

SIOCSIWRATE

SIOCSIWRETRY

SIOCSIWRTS

SIOCSIWSCAN

SIOCSIWSENS

SIOCSIWSPY

SIOCSIWSTATS

SIOCSIWTHRSPY

SIOCSIWTXPOW

SI_DETHREAD

SKF_AD_ALU_XOR_X

SKF_AD_CPU

SKF_AD_HATYPE

SKF_AD_IFINDEX

SKF_AD_MARK

SKF_AD_MAX

SKF_AD_NLATTR

SKF_AD_NLATTR_NEST

SKF_AD_OFF

SKF_AD_PAY_OFFSET

SKF_AD_PKTTYPE

SKF_AD_PROTOCOL

SKF_AD_QUEUE

SKF_AD_RANDOM

SKF_AD_RXHASH

SKF_AD_VLAN_TAG

SKF_AD_VLAN_TAG_PRESENT

SKF_AD_VLAN_TPID

SKF_LL_OFF

SKF_NET_OFF

SK_MEMINFO_BACKLOG

SK_MEMINFO_DROPS

SK_MEMINFO_FWD_ALLOC

SK_MEMINFO_OPTMEM

SK_MEMINFO_RCVBUF

SK_MEMINFO_RMEM_ALLOC

SK_MEMINFO_SNDBUF

SK_MEMINFO_WMEM_ALLOC

SK_MEMINFO_WMEM_QUEUED

SND_CNT

SND_MAX

SOF_TIMESTAMPING_BIND_PHC

SOF_TIMESTAMPING_OPT_CMSG

SOF_TIMESTAMPING_OPT_ID

SOF_TIMESTAMPING_OPT_ID_TCP

SOF_TIMESTAMPING_OPT_PKTINFO

SOF_TIMESTAMPING_OPT_RX_FILTER

SOF_TIMESTAMPING_OPT_STATS

SOF_TIMESTAMPING_OPT_TSONLY

SOF_TIMESTAMPING_OPT_TX_SWHW

SOF_TIMESTAMPING_RAW_HARDWARE

SOF_TIMESTAMPING_RX_HARDWARE

SOF_TIMESTAMPING_RX_SOFTWARE

SOF_TIMESTAMPING_SOFTWARE

SOF_TIMESTAMPING_SYS_HARDWARE

SOF_TIMESTAMPING_TX_ACK

SOF_TIMESTAMPING_TX_HARDWARE

SOF_TIMESTAMPING_TX_SCHED

SOF_TIMESTAMPING_TX_SOFTWARE

SOF_TXTIME_DEADLINE_MODE

SOF_TXTIME_REPORT_ERRORS

SOL_TLS

SOL_XDP

SO_EE_ORIGIN_ICMP

SO_EE_ORIGIN_ICMP6

SO_EE_ORIGIN_LOCAL

SO_EE_ORIGIN_NONE

SO_EE_ORIGIN_TIMESTAMPING

SO_EE_ORIGIN_TXSTATUS

SO_ORIGINAL_DST

SW_CNT

SW_MAX

SYNC_FILE_RANGE_WAIT_AFTER

SYNC_FILE_RANGE_WAIT_BEFORE

SYNC_FILE_RANGE_WRITE

SYN_CNT

SYN_MAX

TCA_FCNT

TCA_KIND

TCA_OPTIONS

TCA_RATE

TCA_STAB

TCA_STATS

TCA_STATS2

TCA_UNSPEC

TCA_XSTATS

TFD_CLOEXEC

TFD_NONBLOCK

TFD_TIMER_ABSTIME

TFD_TIMER_CANCEL_ON_SET

TLS_1_2_VERSION

TLS_1_2_VERSION_MAJOR

TLS_1_2_VERSION_MINOR

TLS_1_3_VERSION

TLS_1_3_VERSION_MAJOR

TLS_1_3_VERSION_MINOR

TLS_CIPHER_AES_CCM_128

TLS_CIPHER_AES_CCM_128_IV_SIZE

TLS_CIPHER_AES_CCM_128_KEY_SIZE

TLS_CIPHER_AES_CCM_128_REC_SEQ_SIZE

TLS_CIPHER_AES_CCM_128_SALT_SIZE

TLS_CIPHER_AES_CCM_128_TAG_SIZE

TLS_CIPHER_AES_GCM_128

TLS_CIPHER_AES_GCM_256

TLS_CIPHER_AES_GCM_128_IV_SIZE

TLS_CIPHER_AES_GCM_128_KEY_SIZE

TLS_CIPHER_AES_GCM_128_REC_SEQ_SIZE

TLS_CIPHER_AES_GCM_128_SALT_SIZE

TLS_CIPHER_AES_GCM_128_TAG_SIZE

TLS_CIPHER_AES_GCM_256_IV_SIZE

TLS_CIPHER_AES_GCM_256_KEY_SIZE

TLS_CIPHER_AES_GCM_256_REC_SEQ_SIZE

TLS_CIPHER_AES_GCM_256_SALT_SIZE

TLS_CIPHER_AES_GCM_256_TAG_SIZE

TLS_CIPHER_ARIA_GCM_128

TLS_CIPHER_ARIA_GCM_256

TLS_CIPHER_ARIA_GCM_128_IV_SIZE

TLS_CIPHER_ARIA_GCM_128_KEY_SIZE

TLS_CIPHER_ARIA_GCM_128_REC_SEQ_SIZE

TLS_CIPHER_ARIA_GCM_128_SALT_SIZE

TLS_CIPHER_ARIA_GCM_128_TAG_SIZE

TLS_CIPHER_ARIA_GCM_256_IV_SIZE

TLS_CIPHER_ARIA_GCM_256_KEY_SIZE

TLS_CIPHER_ARIA_GCM_256_REC_SEQ_SIZE

TLS_CIPHER_ARIA_GCM_256_SALT_SIZE

TLS_CIPHER_ARIA_GCM_256_TAG_SIZE

TLS_CIPHER_CHACHA20_POLY1305

TLS_CIPHER_CHACHA20_POLY1305_IV_SIZE

TLS_CIPHER_CHACHA20_POLY1305_KEY_SIZE

TLS_CIPHER_CHACHA20_POLY1305_REC_SEQ_SIZE

TLS_CIPHER_CHACHA20_POLY1305_SALT_SIZE

TLS_CIPHER_CHACHA20_POLY1305_TAG_SIZE

TLS_CIPHER_SM4_CCM

TLS_CIPHER_SM4_CCM_IV_SIZE

TLS_CIPHER_SM4_CCM_KEY_SIZE

TLS_CIPHER_SM4_CCM_REC_SEQ_SIZE

TLS_CIPHER_SM4_CCM_SALT_SIZE

TLS_CIPHER_SM4_CCM_TAG_SIZE

TLS_CIPHER_SM4_GCM

TLS_CIPHER_SM4_GCM_IV_SIZE

TLS_CIPHER_SM4_GCM_KEY_SIZE

TLS_CIPHER_SM4_GCM_REC_SEQ_SIZE

TLS_CIPHER_SM4_GCM_SALT_SIZE

TLS_CIPHER_SM4_GCM_TAG_SIZE

TLS_CONF_BASE

TLS_CONF_HW

TLS_CONF_HW_RECORD

TLS_CONF_SW

TLS_GET_RECORD_TYPE

TLS_INFO_CIPHER

TLS_INFO_RXCONF

TLS_INFO_RX_NO_PAD

TLS_INFO_TXCONF

TLS_INFO_UNSPEC

TLS_INFO_VERSION

TLS_INFO_ZC_RO_TX

TLS_RX

TLS_RX_EXPECT_NO_PAD

TLS_SET_RECORD_TYPE

TLS_TX

TLS_TX_ZEROCOPY_RO

TPACKET2_HDRLEN

TPACKET3_HDRLEN

TPACKET_ALIGNMENT

TPACKET_HDRLEN

TP_FT_REQ_FILL_RXHASH

TP_STATUS_AVAILABLE

TP_STATUS_BLK_TMO

TP_STATUS_COPY

TP_STATUS_CSUMNOTREADY

TP_STATUS_CSUM_VALID

TP_STATUS_KERNEL

TP_STATUS_LOSING

TP_STATUS_SENDING

TP_STATUS_SEND_REQUEST

TP_STATUS_TS_RAW_HARDWARE

TP_STATUS_TS_SOFTWARE

TP_STATUS_TS_SYS_HARDWARE

TP_STATUS_USER

TP_STATUS_VLAN_TPID_VALID

TP_STATUS_VLAN_VALID

TP_STATUS_WRONG_FORMAT

TRAP_PERF

UINPUT_MAX_NAME_SIZE

UINPUT_VERSION

VMADDR_CID_ANY

VMADDR_CID_HOST

VMADDR_CID_HYPERVISOR

VMADDR_CID_LOCAL

VMADDR_CID_RESERVEDDeprecated

VMADDR_PORT_ANY

VM_BLOCK_DUMP

VM_DIRTY_BACKGROUND

VM_DIRTY_EXPIRE_CS

VM_DIRTY_RATIO

VM_DIRTY_WB_CS

VM_DROP_PAGECACHE

VM_HUGETLB_GROUP

VM_HUGETLB_PAGES

VM_LAPTOP_MODE

VM_LEGACY_VA_LAYOUT

VM_LOWMEM_RESERVE_RATIO

VM_MAX_MAP_COUNT

VM_MIN_FREE_KBYTES

VM_MIN_SLAB

VM_MIN_UNMAPPED

VM_NR_PDFLUSH_THREADS

VM_OVERCOMMIT_MEMORY

VM_OVERCOMMIT_RATIO

VM_PAGEBUF

VM_PAGE_CLUSTER

VM_PANIC_ON_OOM

VM_PERCPU_PAGELIST_FRACTION

VM_SWAPPINESS

VM_SWAP_TOKEN_TIMEOUT

VM_VDSO_ENABLED

VM_VFS_CACHE_PRESSURE

VM_ZONE_RECLAIM_MODE

WIRELESS_EXT

XDP_COPY

XDP_MMAP_OFFSETS

XDP_OPTIONS

XDP_OPTIONS_ZEROCOPY

XDP_PGOFF_RX_RING

XDP_PGOFF_TX_RING

XDP_PKT_CONTD

XDP_RING_NEED_WAKEUP

XDP_RX_RING

XDP_SHARED_UMEM

XDP_STATISTICS

XDP_TXMD_FLAGS_CHECKSUM

XDP_TXMD_FLAGS_TIMESTAMP

XDP_TX_METADATA

XDP_TX_RING

XDP_UMEM_COMPLETION_RING

XDP_UMEM_FILL_RING

XDP_UMEM_PGOFF_COMPLETION_RING

XDP_UMEM_PGOFF_FILL_RING

XDP_UMEM_REG

XDP_UMEM_TX_METADATA_LEN

XDP_UMEM_TX_SW_CSUM

XDP_UMEM_UNALIGNED_CHUNK_FLAG

XDP_USE_NEED_WAKEUP

XDP_USE_SG

XDP_ZEROCOPY

XSK_UNALIGNED_BUF_ADDR_MASK

XSK_UNALIGNED_BUF_OFFSET_SHIFT

__NFT_REG_MAX

Functions

BPF_CLASS⚠

BPF_JUMP⚠

BPF_MISCOP⚠

BPF_MODE⚠

BPF_OP⚠

BPF_RVAL⚠

BPF_SIZE⚠

BPF_SRC⚠

BPF_STMT⚠

FUTEX_OP

SCTP_PR_INDEX⚠

SCTP_PR_POLICY⚠

SCTP_PR_PRIO_ENABLED

SCTP_PR_RTX_ENABLED

SCTP_PR_SET_POLICY⚠

SCTP_PR_TTL_ENABLED

SO_EE_OFFENDER⚠

SUN_LEN⚠

TPACKET_ALIGN⚠

accept4⚠

clock_nanosleep⚠

clone⚠

dup3⚠

epoll_create⚠

epoll_create1⚠

epoll_ctl⚠

epoll_pwait⚠

epoll_wait⚠

eventfd⚠

eventfd_read⚠

eventfd_write⚠

fallocate⚠

fallocate64⚠

fanotify_init⚠

fgetpos64⚠

fgetxattr⚠

flistxattr⚠

fopen64⚠

fread_unlocked⚠

fremovexattr⚠

fsetxattr⚠

ftok⚠

getdtablesize⚠

getgrouplist⚠

gethostid⚠

getspnam_r⚠

getxattr⚠

inotify_add_watch⚠

inotify_init⚠

inotify_init1⚠

inotify_rm_watch⚠

issecure_mask 🔒

klogctl⚠

lcong48⚠

lgetxattr⚠

listxattr⚠

llistxattr⚠

lremovexattr⚠

lsetxattr⚠

lutimes⚠

mkfifoat⚠

mkstemps⚠

mq_close⚠

mq_getattr⚠

mq_open⚠

mq_receive⚠

mq_send⚠

mq_setattr⚠

mq_timedreceive⚠

mq_timedsend⚠

mq_unlink⚠

mrand48⚠

msgctl⚠

msgget⚠

msgrcv⚠

msgsnd⚠

personality⚠

posix_fallocate⚠

posix_fallocate64⚠

posix_madvise⚠

posix_spawn⚠

posix_spawn_file_actions_addclose⚠

posix_spawn_file_actions_adddup2⚠

posix_spawn_file_actions_addopen⚠

posix_spawn_file_actions_destroy⚠

posix_spawn_file_actions_init⚠

posix_spawnattr_destroy⚠

posix_spawnattr_getflags⚠

posix_spawnattr_getpgroup⚠

posix_spawnattr_getschedparam⚠

posix_spawnattr_getschedpolicy⚠

posix_spawnattr_getsigdefault⚠

posix_spawnattr_getsigmask⚠

posix_spawnattr_init⚠

posix_spawnattr_setflags⚠

posix_spawnattr_setpgroup⚠

posix_spawnattr_setschedparam⚠

posix_spawnattr_setschedpolicy⚠

posix_spawnattr_setsigdefault⚠

posix_spawnattr_setsigmask⚠

posix_spawnp⚠

quotactl⚠

readahead⚠

reboot⚠

remap_file_pages⚠

removexattr⚠

sched_getparam⚠

sched_getscheduler⚠

sched_rr_get_interval⚠

sched_setaffinity⚠

sched_setparam⚠

sched_setscheduler⚠

seed48⚠

semctl⚠

semget⚠

semop⚠

sendfile⚠

sendfile64⚠

setfsgid⚠

setfsuid⚠

setns⚠

setxattr⚠

shm_open⚠

shm_unlink⚠

sigaltstack⚠

signalfd⚠

sigtimedwait⚠

sigwaitinfo⚠

splice⚠

swapoff⚠

swapon⚠

sync⚠

sync_file_range⚠

syncfs⚠

syscall⚠

tee⚠

timerfd_create⚠

timerfd_gettime⚠

timerfd_settime⚠

tmpfile64⚠

umount⚠

umount2⚠

unshare⚠

vhangup⚠

vmsplice⚠

Type Aliases

__kernel_clockid_t

__kernel_fsid_t

__s16

__s32

__u8

__u16

__u32

blkcnt64_t

dev_t

eventfd_t

idtype_t

ino64_t

loff_t

mode_t

mqd_t

nfds_t

nl_item

off64_t

pid_type

proc_cn_event

proc_cn_mcast_op

pthread_key_t

pthread_once_t

pthread_spinlock_t

rlim64_t

sctp_assoc_t

socklen_t

Unions

__c_anonymous_iwreq

__c_anonymous_ptp_perout_request_1

__c_anonymous_ptp_perout_request_2

__c_anonymous_xsk_tx_metadata_union

iwreq_data

tpacket_bd_header_u

tpacket_req_u