Module libc::unix::linux_like::linux

Linux-specific definitions for linux-like values

Re-exports

pub use self::gnu::*;

pub use self::arch::*;

pub use self::non_exhaustive::*;

Modules

align

arch

gnu

non_exhaustive

Structs

Elf32_Ehdr

Elf32_Phdr

Elf32_Shdr

Elf32_Sym

Elf64_Ehdr

Elf64_Phdr

Elf64_Shdr

Elf64_Sym

__c_anonymous_sockaddr_can_j1939

__c_anonymous_sockaddr_can_tp

af_alg_ivDeprecated

WARNING: The PartialEq, Eq and Hash implementations of this type are unsound and will be removed in the future.

arpd_request

can_filter

can_frame

canfd_frame

cpu_set_t

dirent

dirent64

dl_phdr_info

dqblk

fanotify_event_metadata

fanotify_response

ff_condition_effect

ff_constant_effect

ff_effect

ff_envelope

ff_periodic_effect

ff_ramp_effect

ff_replay

ff_rumble_effect

ff_trigger

fsid_t

genlmsghdr

glob_t

if_nameindex

in6_pktinfo

inotify_event

input_absinfo

input_event

input_id

input_keymap_entry

input_mask

itimerspec

mntent

mq_attr

msginfo

nlattr

nlmsgerr

nlmsghdr

packet_mreq

passwd

posix_spawn_file_actions_t

posix_spawnattr_t

pthread_cond_t

pthread_condattr_t

pthread_mutex_t

pthread_mutexattr_t

pthread_rwlock_t

pthread_rwlockattr_t

regmatch_t

rlimit64

seccomp_data

sembuf

signalfd_siginfo

sock_extended_err

sock_filter

sock_fprog

sock_txtime

sockaddr_alg

sockaddr_can

sockaddr_nl

sockaddr_vm

spwd

ucred

uinput_abs_setup

uinput_ff_erase

uinput_ff_upload

uinput_setup

uinput_user_dev

Enums

fpos64_t

Constants

ABDAY_1

ABDAY_2

ABDAY_3

ABDAY_4

ABDAY_5

ABDAY_6

ABDAY_7

ABMON_1

ABMON_2

ABMON_3

ABMON_4

ABMON_5

ABMON_6

ABMON_7

ABMON_8

ABMON_9

ABMON_10

ABMON_11

ABMON_12

ABS_CNT

ABS_MAX

AIO_ALLDONE

AIO_CANCELED

AIO_NOTCANCELED

AI_ADDRCONFIG

AI_ALL

AI_CANONNAME

AI_NUMERICHOST

AI_NUMERICSERV

AI_PASSIVE

AI_V4MAPPED

ALG_OP_DECRYPT

ALG_OP_ENCRYPT

ALG_SET_AEAD_ASSOCLEN

ALG_SET_AEAD_AUTHSIZE

ALG_SET_IV

ALG_SET_KEY

ALG_SET_OP

ALT_DIGITS

AM_STR

ARPD_FLUSH

ARPD_LOOKUP

ARPD_UPDATE

ATF_MAGIC

AT_BASE

AT_BASE_PLATFORM

AT_CLKTCK

AT_EACCESS

AT_EGID

AT_ENTRY

AT_EUID

AT_EXECFD

AT_EXECFN

AT_FLAGS

AT_GID

AT_HWCAP

AT_HWCAP2

AT_IGNORE

AT_NOTELF

AT_NULL

AT_PAGESZ

AT_PHDR

AT_PHENT

AT_PHNUM

AT_PLATFORM

AT_RANDOM

AT_SECURE

AT_UID

CANFD_BRS

CANFD_ESI

CANFD_MAX_DLC

CANFD_MAX_DLEN

CANFD_MTU

CAN_BCM

CAN_EFF_FLAG

CAN_EFF_ID_BITS

CAN_EFF_MASK

CAN_ERR_FLAG

CAN_ERR_MASK

CAN_INV_FILTER

CAN_ISOTP

CAN_J1939

CAN_MAX_DLC

CAN_MAX_DLEN

CAN_MCNET

CAN_MTU

CAN_NPROTO

CAN_RAW

CAN_RAW_ERR_FILTER

CAN_RAW_FD_FRAMES

CAN_RAW_FILTER

CAN_RAW_FILTER_MAX

CAN_RAW_JOIN_FILTERS

CAN_RAW_LOOPBACK

CAN_RAW_RECV_OWN_MSGS

CAN_RTR_FLAG

CAN_SFF_ID_BITS

CAN_SFF_MASK

CAN_TP16

CAN_TP20

CLONE_PIDFD

CLOSE_RANGE_CLOEXEC

CLOSE_RANGE_UNSHARE

CMSPAR

CODESET

CRNCYSTR

CTRL_ATTR_FAMILY_ID

CTRL_ATTR_FAMILY_NAME

CTRL_ATTR_HDRSIZE

CTRL_ATTR_MAXATTR

CTRL_ATTR_MCAST_GROUPS

CTRL_ATTR_MCAST_GRP_ID

CTRL_ATTR_MCAST_GRP_NAME

CTRL_ATTR_MCAST_GRP_UNSPEC

CTRL_ATTR_OPS

CTRL_ATTR_OP_FLAGS

CTRL_ATTR_OP_ID

CTRL_ATTR_OP_UNSPEC

CTRL_ATTR_UNSPEC

CTRL_ATTR_VERSION

CTRL_CMD_DELFAMILY

CTRL_CMD_DELMCAST_GRP

CTRL_CMD_DELOPS

CTRL_CMD_GETFAMILY

CTRL_CMD_GETMCAST_GRP

CTRL_CMD_GETOPS

CTRL_CMD_NEWFAMILY

CTRL_CMD_NEWMCAST_GRP

CTRL_CMD_NEWOPS

CTRL_CMD_UNSPEC

DAY_1

DAY_2

DAY_3

DAY_4

DAY_5

DAY_6

DAY_7

D_FMT

D_T_FMT

E2BIG

EACCES

EAGAIN

EAI_AGAIN

EAI_BADFLAGS

EAI_FAIL

EAI_FAMILY

EAI_MEMORY

EAI_NODATA

EAI_NONAME

EAI_OVERFLOW

EAI_SERVICE

EAI_SOCKTYPE

EAI_SYSTEM

EBADF

EBUSY

ECHILD

EDOM

EEXIST

EFAULT

EFBIG

EFD_SEMAPHORE

EINTR

EINVAL

EIO

EISDIR

EMFILE

EMLINK

ENFILE

ENOATTRDeprecated

ENODEV

ENOENT

ENOEXEC

ENOMEM

ENOSPC

ENOTBLK

ENOTDIR

ENOTTY

ENXIO

EPERM

EPIPE

ERA

ERANGE

ERA_D_FMT

ERA_D_T_FMT

ERA_T_FMT

EROFS

ESPIPE

ESRCH

ETH_ALEN

ETH_DATA_LEN

ETH_FCS_LEN

ETH_FRAME_LEN

ETH_HLEN

ETH_P_802_2

ETH_P_802_3

ETH_P_802_3_MIN

ETH_P_802_EX1

ETH_P_1588

ETH_P_8021AD

ETH_P_8021AH

ETH_P_8021Q

ETH_P_80221

ETH_P_AARP

ETH_P_AF_IUCV

ETH_P_ALL

ETH_P_AOE

ETH_P_ARCNET

ETH_P_ARP

ETH_P_ATALK

ETH_P_ATMFATE

ETH_P_ATMMPOA

ETH_P_AX25

ETH_P_BATMAN

ETH_P_BPQ

ETH_P_CAIF

ETH_P_CANFD

ETH_P_CONTROL

ETH_P_CUST

ETH_P_DDCMP

ETH_P_DEC

ETH_P_DIAG

ETH_P_DNA_DL

ETH_P_DNA_RC

ETH_P_DNA_RT

ETH_P_DSA

ETH_P_ECONET

ETH_P_EDSA

ETH_P_FCOE

ETH_P_FIP

ETH_P_HDLC

ETH_P_IEEE802154

ETH_P_IEEEPUP

ETH_P_IEEEPUPAT

ETH_P_IP

ETH_P_IPV6

ETH_P_IPX

ETH_P_IRDA

ETH_P_LAT

ETH_P_LINK_CTL

ETH_P_LOCALTALK

ETH_P_LOOP

ETH_P_LOOPBACK

ETH_P_MACSEC

ETH_P_MOBITEX

ETH_P_MPLS_MC

ETH_P_MPLS_UC

ETH_P_MVRP

ETH_P_PAE

ETH_P_PAUSE

ETH_P_PHONET

ETH_P_PPPTALK

ETH_P_PPP_DISC

ETH_P_PPP_MP

ETH_P_PPP_SES

ETH_P_PRP

ETH_P_PUP

ETH_P_PUPAT

ETH_P_QINQ1

ETH_P_QINQ2

ETH_P_QINQ3

ETH_P_RARP

ETH_P_SCA

ETH_P_SLOW

ETH_P_SNAP

ETH_P_TDLS

ETH_P_TEB

ETH_P_TIPC

ETH_P_TRAILER

ETH_P_TR_802_2

ETH_P_WAN_PPP

ETH_P_WCCP

ETH_P_X25

ETH_ZLEN

ETXTBSY

EV_CNT

EV_MAX

EWOULDBLOCK

EXDEV

FALLOC_FL_COLLAPSE_RANGE

FALLOC_FL_INSERT_RANGE

FALLOC_FL_KEEP_SIZE

FALLOC_FL_PUNCH_HOLE

FALLOC_FL_UNSHARE_RANGE

FALLOC_FL_ZERO_RANGE

FANOTIFY_METADATA_VERSION

FAN_ACCESS

FAN_ACCESS_PERM

FAN_ALLOW

FAN_CLASS_CONTENT

FAN_CLASS_NOTIF

FAN_CLASS_PRE_CONTENT

FAN_CLOEXEC

FAN_CLOSE

FAN_CLOSE_NOWRITE

FAN_CLOSE_WRITE

FAN_DENY

FAN_EVENT_ON_CHILD

FAN_MARK_ADD

FAN_MARK_DONT_FOLLOW

FAN_MARK_FLUSH

FAN_MARK_IGNORED_MASK

FAN_MARK_IGNORED_SURV_MODIFY

FAN_MARK_ONLYDIR

FAN_MARK_REMOVE

FAN_MODIFY

FAN_NOFD

FAN_NONBLOCK

FAN_ONDIR

FAN_OPEN

FAN_OPEN_PERM

FAN_Q_OVERFLOW

FAN_UNLIMITED_MARKS

FAN_UNLIMITED_QUEUE

FF_CNT

FF_MAX

FUTEX_CLOCK_REALTIME

FUTEX_CMD_MASK

FUTEX_CMP_REQUEUE

FUTEX_CMP_REQUEUE_PI

FUTEX_FD

FUTEX_LOCK_PI

FUTEX_PRIVATE_FLAG

FUTEX_REQUEUE

FUTEX_TRYLOCK_PI

FUTEX_UNLOCK_PI

FUTEX_WAIT

FUTEX_WAIT_BITSET

FUTEX_WAIT_REQUEUE_PI

FUTEX_WAKE

FUTEX_WAKE_BITSET

FUTEX_WAKE_OP

F_LOCK

F_SEAL_FUTURE_WRITE

F_TEST

F_TLOCK

F_ULOCK

GENL_ADMIN_PERM

GENL_CMD_CAP_DO

GENL_CMD_CAP_DUMP

GENL_CMD_CAP_HASPOL

GENL_ID_CTRL

GENL_MAX_ID

GENL_MIN_ID

GENL_NAMSIZ

GLOB_ABORTED

GLOB_APPEND

GLOB_DOOFFS

GLOB_ERR

GLOB_MARK

GLOB_NOCHECK

GLOB_NOESCAPE

GLOB_NOMATCH

GLOB_NOSORT

GLOB_NOSPACE

GRND_NONBLOCK

GRND_RANDOM

IFA_ADDRESS

IFA_ANYCAST

IFA_BROADCAST

IFA_CACHEINFO

IFA_F_DADFAILED

IFA_F_DEPRECATED

IFA_F_HOMEADDRESS

IFA_F_NODAD

IFA_F_OPTIMISTIC

IFA_F_PERMANENT

IFA_F_SECONDARY

IFA_F_TEMPORARY

IFA_F_TENTATIVE

IFA_LABEL

IFA_LOCAL

IFA_MULTICAST

IFA_UNSPEC

IFF_ATTACH_QUEUE

IFF_DETACH_QUEUE

IFF_DORMANT

IFF_ECHO

IFF_LOWER_UP

IFF_MULTI_QUEUE

IFF_NOFILTER

IFF_NO_PI

IFF_ONE_QUEUE

IFF_PERSIST

IFF_TAP

IFF_TUN

IFF_TUN_EXCL

IFF_VNET_HDR

IFLA_ADDRESS

IFLA_AF_SPEC

IFLA_ALT_IFNAME

IFLA_BROADCAST

IFLA_CARRIER

IFLA_CARRIER_CHANGES

IFLA_CARRIER_DOWN_COUNT

IFLA_CARRIER_UP_COUNT

IFLA_COST

IFLA_EVENT

IFLA_EXT_MASK

IFLA_GROUP

IFLA_GSO_MAX_SEGS

IFLA_GSO_MAX_SIZE

IFLA_IFALIAS

IFLA_IFNAME

IFLA_IF_NETNSID

IFLA_INFO_DATA

IFLA_INFO_KIND

IFLA_INFO_SLAVE_DATA

IFLA_INFO_SLAVE_KIND

IFLA_INFO_UNSPEC

IFLA_INFO_XSTATS

IFLA_LINK

IFLA_LINKINFO

IFLA_LINKMODE

IFLA_LINK_NETNSID

IFLA_MAP

IFLA_MASTER

IFLA_MAX_MTU

IFLA_MIN_MTU

IFLA_MTU

IFLA_NET_NS_FD

IFLA_NET_NS_PID

IFLA_NEW_IFINDEX

IFLA_NEW_NETNSID

IFLA_NUM_RX_QUEUES

IFLA_NUM_TX_QUEUES

IFLA_NUM_VF

IFLA_OPERSTATE

IFLA_PAD

IFLA_PERM_ADDRESS

IFLA_PHYS_PORT_ID

IFLA_PHYS_PORT_NAME

IFLA_PHYS_SWITCH_ID

IFLA_PORT_SELF

IFLA_PRIORITY

IFLA_PROMISCUITY

IFLA_PROP_LIST

IFLA_PROTINFO

IFLA_PROTO_DOWN

IFLA_PROTO_DOWN_REASON

IFLA_QDISC

IFLA_STATS

IFLA_STATS64

IFLA_TARGET_NETNSID

IFLA_TXQLEN

IFLA_UNSPEC

IFLA_VFINFO_LIST

IFLA_VF_PORTS

IFLA_WEIGHT

IFLA_WIRELESS

IFLA_XDP

INPUT_PROP_CNT

INPUT_PROP_MAX

IN_ACCESS

IN_ALL_EVENTS

IN_ATTRIB

IN_CLOEXEC

IN_CLOSE

IN_CLOSE_NOWRITE

IN_CLOSE_WRITE

IN_CREATE

IN_DELETE

IN_DELETE_SELF

IN_DONT_FOLLOW

IN_IGNORED

IN_ISDIR

IN_MODIFY

IN_MOVE

IN_MOVED_FROM

IN_MOVED_TO

IN_MOVE_SELF

IN_NONBLOCK

IN_ONESHOT

IN_ONLYDIR

IN_OPEN

IN_Q_OVERFLOW

IN_UNMOUNT

IP6T_SO_ORIGINAL_DST

IPC_CREAT

IPC_EXCL

IPC_INFO

IPC_NOWAIT

IPC_PRIVATE

IPC_RMID

IPC_SET

IPC_STAT

IPPROTO_MAXDeprecated

IPPROTO_MPTCP

Multipath TCP

IPTOS_ECN_NOT_ECT

IPTOS_PREC_MASK

IPTOS_TOS_MASK

IPV6_FLOWINFO

IPV6_FLOWINFO_FLOWLABEL

IPV6_FLOWINFO_PRIORITY

IPV6_FLOWINFO_SEND

IPV6_FLOWLABEL_MGR

IPV6_FREEBIND

IPV6_MULTICAST_ALL

IPV6_RECVFRAGSIZE

IPV6_ROUTER_ALERT_ISOLATE

IPV6_RTHDR_LOOSE

IPV6_RTHDR_STRICT

IP_RECVFRAGSIZE

ITIMER_PROF

ITIMER_REAL

ITIMER_VIRTUAL

IUTF8

KEYCTL_ASSUME_AUTHORITY

KEYCTL_CHOWN

KEYCTL_CLEAR

KEYCTL_DESCRIBE

KEYCTL_GET_KEYRING_ID

KEYCTL_GET_PERSISTENT

KEYCTL_GET_SECURITY

KEYCTL_INSTANTIATE

KEYCTL_INSTANTIATE_IOV

KEYCTL_INVALIDATE

KEYCTL_JOIN_SESSION_KEYRING

KEYCTL_LINK

KEYCTL_NEGATE

KEYCTL_READ

KEYCTL_REJECT

KEYCTL_REVOKE

KEYCTL_SEARCH

KEYCTL_SESSION_TO_PARENT

KEYCTL_SETPERM

KEYCTL_SET_REQKEY_KEYRING

KEYCTL_SET_TIMEOUT

KEYCTL_UNLINK

KEYCTL_UPDATE

KEY_CNT

KEY_MAX

KEY_REQKEY_DEFL_DEFAULT

KEY_REQKEY_DEFL_GROUP_KEYRING

KEY_REQKEY_DEFL_NO_CHANGE

KEY_REQKEY_DEFL_PROCESS_KEYRING

KEY_REQKEY_DEFL_REQUESTOR_KEYRING

KEY_REQKEY_DEFL_SESSION_KEYRING

KEY_REQKEY_DEFL_THREAD_KEYRING

KEY_REQKEY_DEFL_USER_KEYRING

KEY_REQKEY_DEFL_USER_SESSION_KEYRING

KEY_SPEC_GROUP_KEYRING

KEY_SPEC_PROCESS_KEYRING

KEY_SPEC_REQKEY_AUTH_KEY

KEY_SPEC_REQUESTOR_KEYRING

KEY_SPEC_SESSION_KEYRING

KEY_SPEC_THREAD_KEYRING

KEY_SPEC_USER_KEYRING

KEY_SPEC_USER_SESSION_KEYRING

LED_CNT

LED_MAX

LINUX_REBOOT_CMD_CAD_OFF

LINUX_REBOOT_CMD_CAD_ON

LINUX_REBOOT_CMD_HALT

LINUX_REBOOT_CMD_KEXEC

LINUX_REBOOT_CMD_POWER_OFF

LINUX_REBOOT_CMD_RESTART

LINUX_REBOOT_CMD_RESTART2

LINUX_REBOOT_CMD_SW_SUSPEND

LINUX_REBOOT_MAGIC1

LINUX_REBOOT_MAGIC2

LINUX_REBOOT_MAGIC2A

LINUX_REBOOT_MAGIC2B

LINUX_REBOOT_MAGIC2C

LIO_NOP

LIO_NOWAIT

LIO_READ

LIO_WAIT

LIO_WRITE

LOG_NFACILITIES

L_tmpnam

MAP_FIXED_NOREPLACE

MAP_SHARED_VALIDATE

MAX_ADDR_LEN

MFD_ALLOW_SEALING

MFD_CLOEXEC

MFD_HUGETLB

MLOCK_ONFAULT

MODULE_INIT_IGNORE_MODVERSIONS

MODULE_INIT_IGNORE_VERMAGIC

MON_1

MON_2

MON_3

MON_4

MON_5

MON_6

MON_7

MON_8

MON_9

MON_10

MON_11

MON_12

MREMAP_DONTUNMAP

MREMAP_FIXED

MREMAP_MAYMOVE

MSC_CNT

MSC_MAX

MSG_COPY

MSG_EXCEPT

MSG_INFO

MSG_NOERROR

MSG_STAT

MS_NOUSER

NDA_CACHEINFO

NDA_DST

NDA_IFINDEX

NDA_LLADDR

NDA_PORT

NDA_PROBES

NDA_UNSPEC

NDA_VLAN

NDA_VNI

NETLINK_ADD_MEMBERSHIP

NETLINK_AUDIT

NETLINK_BROADCAST_ERROR

NETLINK_CAP_ACK

NETLINK_CONNECTOR

NETLINK_CRYPTO

NETLINK_DNRTMSG

NETLINK_DROP_MEMBERSHIP

NETLINK_ECRYPTFS

NETLINK_EXT_ACK

NETLINK_FIB_LOOKUP

NETLINK_FIREWALL

NETLINK_GENERIC

NETLINK_GET_STRICT_CHK

NETLINK_INET_DIAG

NETLINK_IP6_FW

NETLINK_ISCSI

NETLINK_KOBJECT_UEVENT

NETLINK_LISTEN_ALL_NSID

NETLINK_LIST_MEMBERSHIPS

NETLINK_NETFILTER

NETLINK_NFLOG

NETLINK_NO_ENOBUFS

NETLINK_PKTINFO

NETLINK_RDMA

NETLINK_ROUTE

NETLINK_RX_RING

NETLINK_SCSITRANSPORT

NETLINK_SELINUX

NETLINK_SOCK_DIAG

NETLINK_TX_RING

NETLINK_UNUSED

NETLINK_USERSOCK

NETLINK_XFRM

NFNETLINK_V0

NFNLGRP_ACCT_QUOTA

NFNLGRP_CONNTRACK_DESTROY

NFNLGRP_CONNTRACK_EXP_DESTROY

NFNLGRP_CONNTRACK_EXP_NEW

NFNLGRP_CONNTRACK_EXP_UPDATE

NFNLGRP_CONNTRACK_NEW

NFNLGRP_CONNTRACK_UPDATE

NFNLGRP_NFTABLES

NFNLGRP_NONE

NFNL_MSG_BATCH_BEGIN

NFNL_MSG_BATCH_END

NFNL_SUBSYS_ACCT

NFNL_SUBSYS_COUNT

NFNL_SUBSYS_CTHELPER

NFNL_SUBSYS_CTNETLINK

NFNL_SUBSYS_CTNETLINK_EXP

NFNL_SUBSYS_CTNETLINK_TIMEOUT

NFNL_SUBSYS_IPSET

NFNL_SUBSYS_NFTABLES

NFNL_SUBSYS_NFT_COMPAT

NFNL_SUBSYS_NONE

NFNL_SUBSYS_OSF

NFNL_SUBSYS_QUEUE

NFNL_SUBSYS_ULOG

NFPROTO_ARP

NFPROTO_BRIDGE

NFPROTO_DECNET

NFPROTO_INET

NFPROTO_IPV4

NFPROTO_IPV6

NFPROTO_NETDEV

NFPROTO_NUMPROTO

NFPROTO_UNSPEC

NFQA_CAP_LEN

NFQA_CFG_CMD

NFQA_CFG_FLAGS

NFQA_CFG_F_CONNTRACK

NFQA_CFG_F_FAIL_OPEN

NFQA_CFG_F_GSO

NFQA_CFG_F_MAX

NFQA_CFG_F_SECCTX

NFQA_CFG_F_UID_GID

NFQA_CFG_MASK

NFQA_CFG_PARAMS

NFQA_CFG_QUEUE_MAXLEN

NFQA_CFG_UNSPEC

NFQA_CT

NFQA_CT_INFO

NFQA_EXP

NFQA_GID

NFQA_HWADDR

NFQA_IFINDEX_INDEV

NFQA_IFINDEX_OUTDEV

NFQA_IFINDEX_PHYSINDEV

NFQA_IFINDEX_PHYSOUTDEV

NFQA_MARK

NFQA_PACKET_HDR

NFQA_PAYLOAD

NFQA_SECCTX

NFQA_SKB_CSUMNOTREADY

NFQA_SKB_CSUM_NOTVERIFIED

NFQA_SKB_GSO

NFQA_SKB_INFO

NFQA_TIMESTAMP

NFQA_UID

NFQA_UNSPEC

NFQA_VERDICT_HDR

NFQNL_CFG_CMD_BIND

NFQNL_CFG_CMD_NONE

NFQNL_CFG_CMD_PF_BIND

NFQNL_CFG_CMD_PF_UNBIND

NFQNL_CFG_CMD_UNBIND

NFQNL_COPY_META

NFQNL_COPY_NONE

NFQNL_COPY_PACKET

NFQNL_MSG_CONFIG

NFQNL_MSG_PACKET

NFQNL_MSG_VERDICT

NFQNL_MSG_VERDICT_BATCH

NFT_BREAK

NFT_BYTEORDER_HTON

NFT_BYTEORDER_NTOH

NFT_CHAIN_MAXNAMELEN

NFT_CMP_EQ

NFT_CMP_GT

NFT_CMP_GTE

NFT_CMP_LT

NFT_CMP_LTE

NFT_CMP_NEQ

NFT_CONTINUE

NFT_CT_BYTES

NFT_CT_DIRECTION

NFT_CT_DST

NFT_CT_EXPIRATION

NFT_CT_HELPER

NFT_CT_L3PROTOCOL

NFT_CT_LABELS

NFT_CT_MARK

NFT_CT_PKTS

NFT_CT_PROTOCOL

NFT_CT_PROTO_DST

NFT_CT_PROTO_SRC

NFT_CT_SECMARK

NFT_CT_SRC

NFT_CT_STATE

NFT_CT_STATUS

NFT_DATA_RESERVED_MASK

NFT_DATA_VALUE

NFT_DATA_VALUE_MAXLEN

NFT_DATA_VERDICT

NFT_DYNSET_F_INV

NFT_DYNSET_OP_ADD

NFT_DYNSET_OP_UPDATE

NFT_GOTO

NFT_JUMP

NFT_LIMIT_F_INV

NFT_LIMIT_PKTS

NFT_LIMIT_PKT_BYTES

NFT_LOOKUP_F_INV

NFT_META_BRI_IIFNAME

NFT_META_BRI_OIFNAME

NFT_META_CGROUP

NFT_META_CPU

NFT_META_IIF

NFT_META_IIFGROUP

NFT_META_IIFNAME

NFT_META_IIFTYPE

NFT_META_L4PROTO

NFT_META_LEN

NFT_META_MARK

NFT_META_NFPROTO

NFT_META_NFTRACE

NFT_META_OIF

NFT_META_OIFGROUP

NFT_META_OIFNAME

NFT_META_OIFTYPE

NFT_META_PKTTYPE

NFT_META_PRANDOM

NFT_META_PRIORITY

NFT_META_PROTOCOL

NFT_META_RTCLASSID

NFT_META_SECMARK

NFT_META_SKGID

NFT_META_SKUID

NFT_MSG_DELCHAIN

NFT_MSG_DELOBJ

NFT_MSG_DELRULE

NFT_MSG_DELSET

NFT_MSG_DELSETELEM

NFT_MSG_DELTABLE

NFT_MSG_GETCHAIN

NFT_MSG_GETGEN

NFT_MSG_GETOBJ

NFT_MSG_GETOBJ_RESET

NFT_MSG_GETRULE

NFT_MSG_GETSET

NFT_MSG_GETSETELEM

NFT_MSG_GETTABLE

NFT_MSG_MAX

NFT_MSG_NEWCHAIN

NFT_MSG_NEWGEN

NFT_MSG_NEWOBJ

NFT_MSG_NEWRULE

NFT_MSG_NEWSET

NFT_MSG_NEWSETELEM

NFT_MSG_NEWTABLE

NFT_MSG_TRACE

NFT_NAT_DNAT

NFT_NAT_SNAT

NFT_NG_INCREMENTAL

NFT_NG_RANDOM

NFT_OBJ_MAXNAMELEN

NFT_PAYLOAD_CSUM_INET

NFT_PAYLOAD_CSUM_NONE

NFT_PAYLOAD_LL_HEADER

NFT_PAYLOAD_NETWORK_HEADER

NFT_PAYLOAD_TRANSPORT_HEADER

NFT_QUEUE_FLAG_BYPASS

NFT_QUEUE_FLAG_CPU_FANOUT

NFT_QUEUE_FLAG_MASK

NFT_QUOTA_F_INV

NFT_RANGE_EQ

NFT_RANGE_NEQ

NFT_REG32_00

NFT_REG32_01

NFT_REG32_02

NFT_REG32_03

NFT_REG32_04

NFT_REG32_05

NFT_REG32_06

NFT_REG32_07

NFT_REG32_08

NFT_REG32_09

NFT_REG32_10

NFT_REG32_11

NFT_REG32_12

NFT_REG32_13

NFT_REG32_14

NFT_REG32_15

NFT_REG32_SIZE

NFT_REG_1

NFT_REG_2

NFT_REG_3

NFT_REG_4

NFT_REG_SIZE

NFT_REG_VERDICT

NFT_REJECT_ICMPX_ADMIN_PROHIBITED

NFT_REJECT_ICMPX_HOST_UNREACH

NFT_REJECT_ICMPX_NO_ROUTE

NFT_REJECT_ICMPX_PORT_UNREACH

NFT_REJECT_ICMPX_UNREACH

NFT_REJECT_ICMP_UNREACH

NFT_REJECT_TCP_RST

NFT_RETURN

NFT_SET_ANONYMOUS

NFT_SET_CONSTANT

NFT_SET_ELEM_INTERVAL_END

NFT_SET_EVAL

NFT_SET_INTERVAL

NFT_SET_MAP

NFT_SET_MAXNAMELEN

NFT_SET_POL_MEMORY

NFT_SET_POL_PERFORMANCE

NFT_SET_TIMEOUT

NFT_TABLE_MAXNAMELEN

NFT_TRACETYPE_POLICY

NFT_TRACETYPE_RETURN

NFT_TRACETYPE_RULE

NFT_TRACETYPE_UNSPEC

NFT_USERDATA_MAXLEN

NFULA_CFG_CMD

NFULA_CFG_FLAGS

NFULA_CFG_MODE

NFULA_CFG_NLBUFSIZ

NFULA_CFG_QTHRESH

NFULA_CFG_TIMEOUT

NFULA_CFG_UNSPEC

NFULA_CT

NFULA_CT_INFO

NFULA_GID

NFULA_HWADDR

NFULA_HWHEADER

NFULA_HWLEN

NFULA_HWTYPE

NFULA_IFINDEX_INDEV

NFULA_IFINDEX_OUTDEV

NFULA_IFINDEX_PHYSINDEV

NFULA_IFINDEX_PHYSOUTDEV

NFULA_MARK

NFULA_PACKET_HDR

NFULA_PAYLOAD

NFULA_PREFIX

NFULA_SEQ

NFULA_SEQ_GLOBAL

NFULA_TIMESTAMP

NFULA_UID

NFULA_UNSPEC

NFULNL_CFG_CMD_BIND

NFULNL_CFG_CMD_NONE

NFULNL_CFG_CMD_PF_BIND

NFULNL_CFG_CMD_PF_UNBIND

NFULNL_CFG_CMD_UNBIND

NFULNL_CFG_F_CONNTRACK

NFULNL_CFG_F_SEQ

NFULNL_CFG_F_SEQ_GLOBAL

NFULNL_COPY_META

NFULNL_COPY_NONE

NFULNL_COPY_PACKET

NFULNL_MSG_CONFIG

NFULNL_MSG_PACKET

NF_ACCEPT

NF_DROP

NF_INET_FORWARD

NF_INET_LOCAL_IN

NF_INET_LOCAL_OUT

NF_INET_NUMHOOKS

NF_INET_POST_ROUTING

NF_INET_PRE_ROUTING

NF_IP6_FORWARD

NF_IP6_LOCAL_IN

NF_IP6_LOCAL_OUT

NF_IP6_NUMHOOKS

NF_IP6_POST_ROUTING

NF_IP6_PRE_ROUTING

NF_IP6_PRI_CONNTRACK

NF_IP6_PRI_CONNTRACK_DEFRAG

NF_IP6_PRI_CONNTRACK_HELPER

NF_IP6_PRI_FILTER

NF_IP6_PRI_FIRST

NF_IP6_PRI_LAST

NF_IP6_PRI_MANGLE

NF_IP6_PRI_NAT_DST

NF_IP6_PRI_NAT_SRC

NF_IP6_PRI_RAW

NF_IP6_PRI_SECURITY

NF_IP6_PRI_SELINUX_FIRST

NF_IP6_PRI_SELINUX_LAST

NF_IP_FORWARD

NF_IP_LOCAL_IN

NF_IP_LOCAL_OUT

NF_IP_NUMHOOKS

NF_IP_POST_ROUTING

NF_IP_PRE_ROUTING

NF_IP_PRI_CONNTRACK

NF_IP_PRI_CONNTRACK_CONFIRM

NF_IP_PRI_CONNTRACK_DEFRAG

NF_IP_PRI_CONNTRACK_HELPER

NF_IP_PRI_FILTER

NF_IP_PRI_FIRST

NF_IP_PRI_LAST

NF_IP_PRI_MANGLE

NF_IP_PRI_NAT_DST

NF_IP_PRI_NAT_SRC

NF_IP_PRI_RAW

NF_IP_PRI_SECURITY

NF_IP_PRI_SELINUX_FIRST

NF_IP_PRI_SELINUX_LAST

NF_MAX_VERDICT

NF_NETDEV_INGRESS

NF_NETDEV_NUMHOOKS

NF_QUEUE

NF_REPEAT

NF_STOLEN

NF_STOP

NF_VERDICT_BITS

NF_VERDICT_FLAG_QUEUE_BYPASS

NF_VERDICT_MASK

NF_VERDICT_QBITS

NF_VERDICT_QMASK

NI_DGRAM

NI_NAMEREQD

NI_NOFQDN

NI_NUMERICHOST

NI_NUMERICSERV

NLA_ALIGNTO

NLA_F_NESTED

NLA_F_NET_BYTEORDER

NLA_TYPE_MASK

NLMSG_DONE

NLMSG_ERROR

NLMSG_MIN_TYPE

NLMSG_NOOP

NLMSG_OVERRUN

NLM_F_ACK

NLM_F_APPEND

NLM_F_ATOMIC

NLM_F_CREATE

NLM_F_DUMP

NLM_F_DUMP_FILTERED

NLM_F_DUMP_INTR

NLM_F_ECHO

NLM_F_EXCL

NLM_F_MATCH

NLM_F_MULTI

NLM_F_REPLACE

NLM_F_REQUEST

NLM_F_ROOT

NOEXPR

NOSTR

NTF_MASTER

NTF_PROXY

NTF_ROUTER

NTF_SELF

NTF_USE

NUD_DELAY

NUD_FAILED

NUD_INCOMPLETE

NUD_NOARP

NUD_NONE

NUD_PERMANENT

NUD_PROBE

NUD_REACHABLE

NUD_STALE

PACKET_ADD_MEMBERSHIP

PACKET_DROP_MEMBERSHIP

PACKET_MR_ALLMULTI

PACKET_MR_MULTICAST

PACKET_MR_PROMISC

PACKET_MR_UNICAST

PM_STR

POSIX_MADV_NORMAL

POSIX_MADV_RANDOM

POSIX_MADV_SEQUENTIAL

POSIX_MADV_WILLNEED

POSIX_SPAWN_RESETIDS

POSIX_SPAWN_SETPGROUP

POSIX_SPAWN_SETSCHEDPARAM

POSIX_SPAWN_SETSCHEDULER

POSIX_SPAWN_SETSIGDEF

POSIX_SPAWN_SETSIGMASK

POSIX_SPAWN_USEVFORK

PR_CAPBSET_DROP

PR_CAPBSET_READ

PR_CAP_AMBIENT

PR_CAP_AMBIENT_CLEAR_ALL

PR_CAP_AMBIENT_IS_SET

PR_CAP_AMBIENT_LOWER

PR_CAP_AMBIENT_RAISE

PR_ENDIAN_BIG

PR_ENDIAN_LITTLE

PR_ENDIAN_PPC_LITTLE

PR_FPEMU_NOPRINT

PR_FPEMU_SIGFPE

PR_FP_EXC_ASYNC

PR_FP_EXC_DISABLED

PR_FP_EXC_DIV

PR_FP_EXC_INV

PR_FP_EXC_NONRECOV

PR_FP_EXC_OVF

PR_FP_EXC_PRECISE

PR_FP_EXC_RES

PR_FP_EXC_SW_ENABLE

PR_FP_EXC_UND

PR_FP_MODE_FR

PR_FP_MODE_FRE

PR_GET_CHILD_SUBREAPER

PR_GET_DUMPABLE

PR_GET_ENDIAN

PR_GET_FPEMU

PR_GET_FPEXC

PR_GET_FP_MODE

PR_GET_KEEPCAPS

PR_GET_NAME

PR_GET_NO_NEW_PRIVS

PR_GET_PDEATHSIG

PR_GET_SECCOMP

PR_GET_SECUREBITS

PR_GET_THP_DISABLE

PR_GET_TID_ADDRESS

PR_GET_TIMERSLACK

PR_GET_TIMING

PR_GET_TSC

PR_GET_UNALIGN

PR_MCE_KILL

PR_MCE_KILL_CLEAR

PR_MCE_KILL_DEFAULT

PR_MCE_KILL_EARLY

PR_MCE_KILL_GET

PR_MCE_KILL_LATE

PR_MCE_KILL_SET

PR_MPX_DISABLE_MANAGEMENT

PR_MPX_ENABLE_MANAGEMENT

PR_SET_CHILD_SUBREAPER

PR_SET_DUMPABLE

PR_SET_ENDIAN

PR_SET_FPEMU

PR_SET_FPEXC

PR_SET_FP_MODE

PR_SET_KEEPCAPS

PR_SET_MM

PR_SET_MM_ARG_END

PR_SET_MM_ARG_START

PR_SET_MM_AUXV

PR_SET_MM_BRK

PR_SET_MM_END_CODE

PR_SET_MM_END_DATA

PR_SET_MM_ENV_END

PR_SET_MM_ENV_START

PR_SET_MM_EXE_FILE

PR_SET_MM_MAP

PR_SET_MM_MAP_SIZE

PR_SET_MM_START_BRK

PR_SET_MM_START_CODE

PR_SET_MM_START_DATA

PR_SET_MM_START_STACK

PR_SET_NAME

PR_SET_NO_NEW_PRIVS

PR_SET_PDEATHSIG

PR_SET_PTRACER

PR_SET_SECCOMP

PR_SET_SECUREBITS

PR_SET_THP_DISABLE

PR_SET_TIMERSLACK

PR_SET_TIMING

PR_SET_TSC

PR_SET_UNALIGN

PR_TASK_PERF_EVENTS_DISABLE

PR_TASK_PERF_EVENTS_ENABLE

PR_TIMING_STATISTICAL

PR_TIMING_TIMESTAMP

PR_TSC_ENABLE

PR_TSC_SIGSEGV

PR_UNALIGN_NOPRINT

PR_UNALIGN_SIGBUS

PTHREAD_COND_INITIALIZER

PTHREAD_MUTEX_DEFAULT

PTHREAD_MUTEX_ERRORCHECK

PTHREAD_MUTEX_INITIALIZER

PTHREAD_MUTEX_NORMAL

PTHREAD_MUTEX_RECURSIVE

PTHREAD_MUTEX_ROBUST

PTHREAD_MUTEX_STALLED

PTHREAD_PROCESS_PRIVATE

PTHREAD_PROCESS_SHARED

PTHREAD_RWLOCK_INITIALIZER

PTRACE_EVENT_STOP

PT_DYNAMIC

PT_GNU_EH_FRAME

PT_GNU_RELRO

PT_GNU_STACK

PT_HIOS

PT_HIPROC

PT_INTERP

PT_LOAD

PT_LOOS

PT_LOPROC

PT_NOTE

PT_NULL

PT_NUM

PT_PHDR

PT_SHLIB

PT_TLS

QFMT_VFS_OLD

QFMT_VFS_V0

QFMT_VFS_V1

RADIXCHAR

RB_AUTOBOOT

RB_DISABLE_CAD

RB_ENABLE_CAD

RB_HALT_SYSTEM

RB_KEXEC

RB_POWER_OFF

RB_SW_SUSPEND

REG_BADBR

REG_BADPAT

REG_BADRPT

REG_EBRACE

REG_EBRACK

REG_ECOLLATE

REG_ECTYPE

REG_EESCAPE

REG_ENOSYS

REG_EPAREN

REG_ERANGE

REG_ESPACE

REG_ESUBREG

REG_EXTENDED

REG_ICASE

REG_NEWLINE

REG_NOMATCH

REG_NOSUB

REG_NOTBOL

REG_NOTEOL

REL_CNT

REL_MAX

RENAME_EXCHANGE

RENAME_NOREPLACE

RENAME_WHITEOUT

REP_CNT

REP_MAX

RESOLVE_BENEATH

RESOLVE_CACHED

RESOLVE_IN_ROOT

RESOLVE_NO_MAGICLINKS

RESOLVE_NO_SYMLINKS

RESOLVE_NO_XDEV

RLIM_SAVED_CUR

RLIM_SAVED_MAX

RTA_CACHEINFO

RTA_DST

RTA_FLOW

RTA_GATEWAY

RTA_IIF

RTA_MARK

RTA_METRICS

RTA_MFC_STATS

RTA_MP_ALGO

RTA_MULTIPATH

RTA_OIF

RTA_PREFSRC

RTA_PRIORITY

RTA_PROTOINFO

RTA_SESSION

RTA_SRC

RTA_TABLE

RTA_UNSPEC

RTCF_DIRECTSRC

RTCF_DOREDIRECT

RTCF_LOG

RTCF_MASQ

RTCF_NAT

RTCF_VALVE

RTF_ADDRCLASSMASK

RTF_ADDRCONF

RTF_ALLONLINK

RTF_BROADCAST

RTF_CACHE

RTF_DEFAULT

RTF_DYNAMIC

RTF_FLOW

RTF_GATEWAY

RTF_HOST

RTF_INTERFACE

RTF_IRTT

RTF_LINKRT

RTF_LOCAL

RTF_MODIFIED

RTF_MSS

RTF_MTU

RTF_MULTICAST

RTF_NAT

RTF_NOFORWARD

RTF_NONEXTHOP

RTF_NOPMTUDISC

RTF_POLICY

RTF_REINSTATE

RTF_REJECT

RTF_STATIC

RTF_THROW

RTF_UP

RTF_WINDOW

RTF_XRESOLVE

RTLD_DEFAULT

RTLD_NEXT

RTLD_NODELETE

RTLD_NOW

RTMSG_AR_FAILED

RTMSG_CONTROL

RTMSG_DELDEVICE

RTMSG_DELROUTE

RTMSG_DELRULE

RTMSG_NEWDEVICE

RTMSG_NEWROUTE

RTMSG_NEWRULE

RTMSG_OVERRUN

RTM_DELACTION

RTM_DELADDR

RTM_DELADDRLABEL

RTM_DELLINK

RTM_DELMDB

RTM_DELNEIGH

RTM_DELNSID

RTM_DELQDISC

RTM_DELROUTE

RTM_DELRULE

RTM_DELTCLASS

RTM_DELTFILTER

RTM_F_CLONED

RTM_F_EQUALIZE

RTM_F_NOTIFY

RTM_F_PREFIX

RTM_GETACTION

RTM_GETADDR

RTM_GETADDRLABEL

RTM_GETANYCAST

RTM_GETDCB

RTM_GETLINK

RTM_GETMDB

RTM_GETMULTICAST

RTM_GETNEIGH

RTM_GETNEIGHTBL

RTM_GETNETCONF

RTM_GETNSID

RTM_GETQDISC

RTM_GETROUTE

RTM_GETRULE

RTM_GETTCLASS

RTM_GETTFILTER

RTM_NEWACTION

RTM_NEWADDR

RTM_NEWADDRLABEL

RTM_NEWLINK

RTM_NEWMDB

RTM_NEWNDUSEROPT

RTM_NEWNEIGH

RTM_NEWNEIGHTBL

RTM_NEWNETCONF

RTM_NEWNSID

RTM_NEWPREFIX

RTM_NEWQDISC

RTM_NEWROUTE

RTM_NEWRULE

RTM_NEWTCLASS

RTM_NEWTFILTER

RTM_SETDCB

RTM_SETLINK

RTM_SETNEIGHTBL

RTN_ANYCAST

RTN_BLACKHOLE

RTN_BROADCAST

RTN_LOCAL

RTN_MULTICAST

RTN_NAT

RTN_PROHIBIT

RTN_THROW

RTN_UNICAST

RTN_UNREACHABLE

RTN_UNSPEC

RTN_XRESOLVE

RTPROT_BOOT

RTPROT_KERNEL

RTPROT_REDIRECT

RTPROT_STATIC

RTPROT_UNSPEC

RT_CLASS_DEFAULT

RT_CLASS_LOCAL

RT_CLASS_MAIN

RT_CLASS_MAX

RT_CLASS_UNSPEC

RT_SCOPE_HOST

RT_SCOPE_LINK

RT_SCOPE_NOWHERE

RT_SCOPE_SITE

RT_SCOPE_UNIVERSE

RT_TABLE_COMPAT

RT_TABLE_DEFAULT

RT_TABLE_LOCAL

RT_TABLE_MAIN

RT_TABLE_UNSPEC

RUSAGE_CHILDREN

RUSAGE_THREAD

SCHED_BATCH

SCHED_FIFO

SCHED_IDLE

SCHED_OTHER

SCHED_RESET_ON_FORK

SCHED_RR

SECCOMP_FILTER_FLAG_LOG

SECCOMP_FILTER_FLAG_SPEC_ALLOW

SECCOMP_FILTER_FLAG_TSYNC

SECCOMP_MODE_DISABLED

SECCOMP_MODE_FILTER

SECCOMP_MODE_STRICT

SECCOMP_RET_ACTION

SECCOMP_RET_ACTION_FULL

SECCOMP_RET_ALLOW

SECCOMP_RET_DATA

SECCOMP_RET_ERRNO

SECCOMP_RET_KILL

SECCOMP_RET_KILL_PROCESS

SECCOMP_RET_KILL_THREAD

SECCOMP_RET_LOG

SECCOMP_RET_TRACE

SECCOMP_RET_TRAP

SEEK_DATA

SEEK_HOLE

SEM_FAILED

SHM_EXEC

SHM_HUGETLB

SHM_LOCK

SHM_NORESERVE

SHM_R

SHM_RDONLY

SHM_REMAP

SHM_RND

SHM_UNLOCK

SHM_W

SIOCADDMULTI

SIOCADDRT

SIOCDARP

SIOCDELMULTI

SIOCDELRT

SIOCDRARP

SIOCGARP

SIOCGIFADDR

SIOCGIFBRDADDR

SIOCGIFCONF

SIOCGIFDSTADDR

SIOCGIFENCAP

SIOCGIFFLAGS

SIOCGIFHWADDR

SIOCGIFMAP

SIOCGIFMEM

SIOCGIFMETRIC

SIOCGIFMTU

SIOCGIFNAME

SIOCGIFNETMASK

SIOCGIFSLAVE

SIOCGRARP

SIOCSARP

SIOCSIFADDR

SIOCSIFBRDADDR

SIOCSIFDSTADDR

SIOCSIFENCAP

SIOCSIFFLAGS

SIOCSIFHWADDR

SIOCSIFLINK

SIOCSIFMAP

SIOCSIFMEM

SIOCSIFMETRIC

SIOCSIFMTU

SIOCSIFNETMASK

SIOCSIFSLAVE

SIOCSRARP

SND_CNT

SND_MAX

SOF_TIMESTAMPING_RAW_HARDWARE

SOF_TIMESTAMPING_RX_HARDWARE

SOF_TIMESTAMPING_RX_SOFTWARE

SOF_TIMESTAMPING_SOFTWARE

SOF_TIMESTAMPING_SYS_HARDWARE

SOF_TIMESTAMPING_TX_HARDWARE

SOF_TIMESTAMPING_TX_SOFTWARE

SOF_TXTIME_DEADLINE_MODE

SOF_TXTIME_REPORT_ERRORS

SOL_CAN_BASE

SOL_CAN_RAW

SO_EE_ORIGIN_ICMP

SO_EE_ORIGIN_ICMP6

SO_EE_ORIGIN_LOCAL

SO_EE_ORIGIN_NONE

SO_EE_ORIGIN_TIMESTAMPING

SO_EE_ORIGIN_TXSTATUS

SO_ORIGINAL_DST

ST_APPEND

ST_IMMUTABLE

ST_MANDLOCK

ST_NOATIME

ST_NODEV

ST_NODIRATIME

ST_NOEXEC

ST_NOSUID

ST_RDONLY

ST_SYNCHRONOUS

ST_WRITE

SW_CNT

SW_MAX

SYNC_FILE_RANGE_WAIT_AFTER

SYNC_FILE_RANGE_WAIT_BEFORE

SYNC_FILE_RANGE_WRITE

SYN_CNT

SYN_MAX

S_IEXEC

S_IREAD

S_IWRITE

TCA_FCNT

TCA_KIND

TCA_OPTIONS

TCA_RATE

TCA_STAB

TCA_STATS

TCA_STATS2

TCA_UNSPEC

TCA_XSTATS

TFD_CLOEXEC

TFD_NONBLOCK

TFD_TIMER_ABSTIME

THOUSEP

TUN_READQ_SIZE

TUN_TAP_DEV

TUN_TUN_DEV

TUN_TYPE_MASK

T_FMT

T_FMT_AMPM

UDP_CORK

UDP_ENCAP

UDP_GRO

UDP_NO_CHECK6_RX

UDP_NO_CHECK6_TX

UDP_SEGMENT

UINPUT_MAX_NAME_SIZE

UINPUT_VERSION

VMADDR_CID_ANY

VMADDR_CID_HOST

VMADDR_CID_HYPERVISOR

VMADDR_CID_LOCAL

VMADDR_CID_RESERVEDDeprecated

VMADDR_PORT_ANY

YESEXPR

YESSTR

_PC_2_SYMLINKS

_PC_ALLOC_SIZE_MIN

_PC_ASYNC_IO

_PC_CHOWN_RESTRICTED

_PC_FILESIZEBITS

_PC_LINK_MAX

_PC_MAX_CANON

_PC_MAX_INPUT

_PC_NAME_MAX

_PC_NO_TRUNC

_PC_PATH_MAX

_PC_PIPE_BUF

_PC_PRIO_IO

_PC_REC_INCR_XFER_SIZE

_PC_REC_MAX_XFER_SIZE

_PC_REC_MIN_XFER_SIZE

_PC_REC_XFER_ALIGN

_PC_SOCK_MAXBUF

_PC_SYMLINK_MAX

_PC_SYNC_IO

_PC_VDISABLE

_POSIX_VDISABLE

_SC_2_CHAR_TERM

_SC_2_C_BIND

_SC_2_C_DEV

_SC_2_FORT_DEV

_SC_2_FORT_RUN

_SC_2_LOCALEDEF

_SC_2_PBS

_SC_2_PBS_ACCOUNTING

_SC_2_PBS_CHECKPOINT

_SC_2_PBS_LOCATE

_SC_2_PBS_MESSAGE

_SC_2_PBS_TRACK

_SC_2_SW_DEV

_SC_2_UPE

_SC_2_VERSION

_SC_ADVISORY_INFO

_SC_AIO_LISTIO_MAX

_SC_AIO_MAX

_SC_AIO_PRIO_DELTA_MAX

_SC_ARG_MAX

_SC_ASYNCHRONOUS_IO

_SC_ATEXIT_MAX

_SC_AVPHYS_PAGES

_SC_BARRIERS

_SC_BC_BASE_MAX

_SC_BC_DIM_MAX

_SC_BC_SCALE_MAX

_SC_BC_STRING_MAX

_SC_CHILD_MAX

_SC_CLK_TCK

_SC_CLOCK_SELECTION

_SC_COLL_WEIGHTS_MAX

_SC_CPUTIME

_SC_DELAYTIMER_MAX

_SC_EXPR_NEST_MAX

_SC_FSYNC

_SC_GETGR_R_SIZE_MAX

_SC_GETPW_R_SIZE_MAX

_SC_HOST_NAME_MAX

_SC_IOV_MAX

_SC_IPV6

_SC_JOB_CONTROL

_SC_LINE_MAX

_SC_LOGIN_NAME_MAX

_SC_MAPPED_FILES

_SC_MEMLOCK

_SC_MEMLOCK_RANGE

_SC_MEMORY_PROTECTION

_SC_MESSAGE_PASSING

_SC_MONOTONIC_CLOCK

_SC_MQ_OPEN_MAX

_SC_MQ_PRIO_MAX

_SC_NGROUPS_MAX

_SC_NPROCESSORS_CONF

_SC_NPROCESSORS_ONLN

_SC_NZERO

_SC_OPEN_MAX

_SC_PAGESIZE

_SC_PAGE_SIZE

_SC_PASS_MAX

_SC_PHYS_PAGES

_SC_PRIORITIZED_IO

_SC_PRIORITY_SCHEDULING

_SC_RAW_SOCKETS

_SC_READER_WRITER_LOCKS

_SC_REALTIME_SIGNALS

_SC_REGEXP

_SC_RE_DUP_MAX

_SC_RTSIG_MAX

_SC_SAVED_IDS

_SC_SEMAPHORES

_SC_SEM_NSEMS_MAX

_SC_SEM_VALUE_MAX

_SC_SHARED_MEMORY_OBJECTS

_SC_SHELL

_SC_SIGQUEUE_MAX

_SC_SPAWN

_SC_SPIN_LOCKS

_SC_SPORADIC_SERVER

_SC_SS_REPL_MAX

_SC_STREAMS

_SC_STREAM_MAX

_SC_SYMLOOP_MAX

_SC_SYNCHRONIZED_IO

_SC_THREADS

_SC_THREAD_ATTR_STACKADDR

_SC_THREAD_ATTR_STACKSIZE

_SC_THREAD_CPUTIME

_SC_THREAD_DESTRUCTOR_ITERATIONS

_SC_THREAD_KEYS_MAX

_SC_THREAD_PRIORITY_SCHEDULING

_SC_THREAD_PRIO_INHERIT

_SC_THREAD_PRIO_PROTECT

_SC_THREAD_PROCESS_SHARED

_SC_THREAD_ROBUST_PRIO_INHERIT

_SC_THREAD_ROBUST_PRIO_PROTECT

_SC_THREAD_SAFE_FUNCTIONS

_SC_THREAD_SPORADIC_SERVER

_SC_THREAD_STACK_MIN

_SC_THREAD_THREADS_MAX

_SC_TIMEOUTS

_SC_TIMERS

_SC_TIMER_MAX

_SC_TRACE

_SC_TRACE_EVENT_FILTER

_SC_TRACE_EVENT_NAME_MAX

_SC_TRACE_INHERIT

_SC_TRACE_LOG

_SC_TRACE_NAME_MAX

_SC_TRACE_SYS_MAX

_SC_TRACE_USER_EVENT_MAX

_SC_TTY_NAME_MAX

_SC_TYPED_MEMORY_OBJECTS

_SC_TZNAME_MAX

_SC_UIO_MAXIOV

_SC_V6_ILP32_OFF32

_SC_V6_ILP32_OFFBIG

_SC_V6_LP64_OFF64

_SC_V6_LPBIG_OFFBIG

_SC_V7_ILP32_OFF32

_SC_V7_ILP32_OFFBIG

_SC_V7_LP64_OFF64

_SC_V7_LPBIG_OFFBIG

_SC_VERSION

_SC_XBS5_ILP32_OFF32

_SC_XBS5_ILP32_OFFBIG

_SC_XBS5_LP64_OFF64

_SC_XBS5_LPBIG_OFFBIG

_SC_XOPEN_CRYPT

_SC_XOPEN_ENH_I18N

_SC_XOPEN_LEGACY

_SC_XOPEN_REALTIME

_SC_XOPEN_REALTIME_THREADS

_SC_XOPEN_SHM

_SC_XOPEN_STREAMS

_SC_XOPEN_UNIX

_SC_XOPEN_VERSION

_SC_XOPEN_XCU_VERSION

_SC_XOPEN_XPG2

_SC_XOPEN_XPG3

_SC_XOPEN_XPG4

__NFT_REG_MAX

__SIZEOF_PTHREAD_COND_T

Functions

CMSG_NXTHDR⚠

CPU_ALLOC_SIZE⚠

CPU_CLR⚠

CPU_COUNT⚠

CPU_COUNT_S⚠

CPU_EQUAL⚠

CPU_ISSET⚠

CPU_SET⚠

CPU_ZERO⚠

IPTOS_PREC⚠

IPTOS_TOS⚠

NLA_ALIGN⚠

RT_ADDRCLASS⚠

RT_LOCALADDR⚠

RT_TOS⚠

SO_EE_OFFENDER⚠

__errno_location⚠

abs⚠

accept4⚠

addmntent⚠

aio_cancel⚠

aio_error⚠

aio_fsync⚠

aio_read⚠

aio_return⚠

aio_suspend⚠

aio_write⚠

atof⚠

clock_nanosleep⚠

clone⚠

daemon⚠

dl_iterate_phdr⚠

dup3⚠

endgrent⚠

endmntent⚠

endpwent⚠

endspent⚠

epoll_create⚠

epoll_create1⚠

epoll_ctl⚠

epoll_pwait⚠

epoll_wait⚠

eventfd⚠

faccessat⚠

fallocate⚠

fallocate64⚠

fanotify_init⚠

fgetpos64⚠

fgetxattr⚠

flistxattr⚠

fopen64⚠

fread_unlocked⚠

fremovexattr⚠

freopen64⚠

fseeko64⚠

fsetpos64⚠

fsetxattr⚠

ftello64⚠

ftok⚠

futimes⚠

getdomainname⚠

getdtablesize⚠

getgrent⚠

getgrgid⚠

getgrgid_r⚠

getgrnam⚠

getgrnam_r⚠

getgrouplist⚠

gethostid⚠

getloadavg⚠

getmntent⚠

getnameinfo⚠

getpwent⚠

getpwnam_r⚠

getpwuid_r⚠

getspent⚠

getspnam⚠

getspnam_r⚠

gettid⚠

getxattr⚠

glob⚠

globfree⚠

hasmntopt⚠

iconv⚠

iconv_close⚠

iconv_open⚠

if_freenameindex⚠

if_nameindex⚠

initgroups⚠

inotify_add_watch⚠

inotify_init⚠

inotify_init1⚠

inotify_rm_watch⚠

labs⚠

lgetxattr⚠

lio_listio⚠

listxattr⚠

llistxattr⚠

lremovexattr⚠

lsetxattr⚠

lutimes⚠

madvise⚠

major⚠

makedev⚠

memmem⚠

minor⚠

mkfifoat⚠

mkostemp⚠

mkostemps⚠

mkstemps⚠

mlock2⚠

mount⚠

mprotect⚠

mq_close⚠

mq_getattr⚠

mq_open⚠

mq_receive⚠

mq_send⚠

mq_setattr⚠

mq_timedreceive⚠

mq_timedsend⚠

mq_unlink⚠

mremap⚠

msgctl⚠

msgget⚠

msgrcv⚠

msgsnd⚠

msync⚠

nl_langinfo⚠

nl_langinfo_l⚠

personality⚠

popen⚠

posix_fallocate⚠

posix_fallocate64⚠

posix_madvise⚠

posix_spawn⚠

posix_spawn_file_actions_addclose⚠

posix_spawn_file_actions_adddup2⚠

posix_spawn_file_actions_addopen⚠

posix_spawn_file_actions_destroy⚠

posix_spawn_file_actions_init⚠

posix_spawnattr_destroy⚠

posix_spawnattr_getflags⚠

posix_spawnattr_getpgroup⚠

posix_spawnattr_getschedparam⚠

posix_spawnattr_getschedpolicy⚠

posix_spawnattr_getsigdefault⚠

posix_spawnattr_getsigmask⚠

posix_spawnattr_init⚠

posix_spawnattr_setflags⚠

posix_spawnattr_setpgroup⚠

posix_spawnattr_setschedparam⚠

posix_spawnattr_setschedpolicy⚠

posix_spawnattr_setsigdefault⚠

posix_spawnattr_setsigmask⚠

posix_spawnp⚠

ppoll⚠

prctl⚠

preadv⚠

process_vm_readv⚠

process_vm_writev⚠

pthread_atfork⚠

pthread_attr_getguardsize⚠

pthread_cancel⚠

pthread_condattr_getpshared⚠

pthread_create⚠

pthread_getaffinity_np⚠

pthread_getcpuclockid⚠

pthread_getschedparam⚠

pthread_kill⚠

pthread_mutex_consistent⚠

pthread_mutex_timedlock⚠

pthread_mutexattr_getpshared⚠

pthread_mutexattr_getrobust⚠

pthread_mutexattr_setrobust⚠

pthread_setaffinity_np⚠

pthread_setschedparam⚠

pthread_setschedprio⚠

pthread_sigmask⚠

pthread_spin_destroy⚠

pthread_spin_init⚠

pthread_spin_lock⚠

pthread_spin_trylock⚠

pthread_spin_unlock⚠

pwritev⚠

quotactl⚠

rand⚠

readahead⚠

reboot⚠

recvfrom⚠

regcomp⚠

regerror⚠

regexec⚠

regfree⚠

remap_file_pages⚠

removexattr⚠

sched_get_priority_max⚠

sched_get_priority_min⚠

sched_getaffinity⚠

sched_getcpu⚠

sched_getparam⚠

sched_getscheduler⚠

sched_rr_get_interval⚠

sched_setaffinity⚠

sched_setparam⚠

sched_setscheduler⚠

seekdir⚠

sem_close⚠

sem_getvalue⚠

sem_open⚠

sem_timedwait⚠

sem_unlink⚠

semctl⚠

semget⚠

semop⚠

sendfile⚠

sendfile64⚠

setdomainname⚠

setfsgid⚠

setfsuid⚠

setgrent⚠

sethostname⚠

setmntent⚠

setns⚠

setpwent⚠

setspent⚠

settimeofday⚠

setxattr⚠

shm_open⚠

shm_unlink⚠

shmat⚠

shmctl⚠

shmdt⚠

shmget⚠

sigaltstack⚠

signalfd⚠

sigsuspend⚠

sigtimedwait⚠

sigwait⚠

sigwaitinfo⚠

splice⚠

srand⚠

strerror_r⚠

swapoff⚠

swapon⚠

sync⚠

sync_file_range⚠

syncfs⚠

syscall⚠

sysinfo⚠

tee⚠

telldir⚠

timer_create⚠

timer_delete⚠

timer_getoverrun⚠

timer_gettime⚠

timer_settime⚠

timerfd_create⚠

timerfd_gettime⚠

timerfd_settime⚠

tmpfile64⚠

umount⚠

umount2⚠

unshare⚠

vhangup⚠

vmsplice⚠

Type Definitions

Elf32_Addr

Elf32_Half

Elf32_Off

Elf32_Section

Elf32_Word

Elf64_Addr

Elf64_Half

Elf64_Off

Elf64_Section

Elf64_Sxword

Elf64_Word

Elf64_Xword

__s16

__s32

__u8

__u16

__u32

blkcnt64_t

can_err_mask_t

canid_t

dev_t

iconv_t

idtype_t

ino64_t

loff_t

mode_t

mqd_t

nfds_t

nl_item

off64_t

pthread_key_t

pthread_spinlock_t

rlim64_t

socklen_t

useconds_t

Unions

__c_anonymous_sockaddr_can_can_addr